Network.AWS.RDS

Description

Amazon Relational Database Service (Amazon RDS) is a web service that makes it easy to set up, operate, and scale a relational database in the cloud. It provides cost-efficient and resizable capacity while managing time-consuming database administration tasks, freeing you up to focus on your applications and business.

Documentation

module Network.AWS.RDS.AddSourceIdentifierToSubscription

module Network.AWS.RDS.AddTagsToResource

module Network.AWS.RDS.ApplyPendingMaintenanceAction

module Network.AWS.RDS.AuthorizeDBSecurityGroupIngress

module Network.AWS.RDS.CopyDBParameterGroup

module Network.AWS.RDS.CopyDBSnapshot

module Network.AWS.RDS.CopyOptionGroup

module Network.AWS.RDS.CreateDBInstance

module Network.AWS.RDS.CreateDBInstanceReadReplica

module Network.AWS.RDS.CreateDBParameterGroup

module Network.AWS.RDS.CreateDBSecurityGroup

module Network.AWS.RDS.CreateDBSnapshot

module Network.AWS.RDS.CreateDBSubnetGroup

module Network.AWS.RDS.CreateEventSubscription

module Network.AWS.RDS.CreateOptionGroup

module Network.AWS.RDS.DeleteDBInstance

module Network.AWS.RDS.DeleteDBParameterGroup

module Network.AWS.RDS.DeleteDBSecurityGroup

module Network.AWS.RDS.DeleteDBSnapshot

module Network.AWS.RDS.DeleteDBSubnetGroup

module Network.AWS.RDS.DeleteEventSubscription

module Network.AWS.RDS.DeleteOptionGroup

module Network.AWS.RDS.DescribeAccountAttributes

module Network.AWS.RDS.DescribeCertificates

module Network.AWS.RDS.DescribeDBEngineVersions

module Network.AWS.RDS.DescribeDBInstances

module Network.AWS.RDS.DescribeDBLogFiles

module Network.AWS.RDS.DescribeDBParameterGroups

module Network.AWS.RDS.DescribeDBParameters

module Network.AWS.RDS.DescribeDBSecurityGroups

module Network.AWS.RDS.DescribeDBSnapshots

module Network.AWS.RDS.DescribeDBSubnetGroups

module Network.AWS.RDS.DescribeEngineDefaultParameters

module Network.AWS.RDS.DescribeEventCategories

module Network.AWS.RDS.DescribeEventSubscriptions

module Network.AWS.RDS.DescribeEvents

module Network.AWS.RDS.DescribeOptionGroupOptions

module Network.AWS.RDS.DescribeOptionGroups

module Network.AWS.RDS.DescribeOrderableDBInstanceOptions

module Network.AWS.RDS.DescribePendingMaintenanceActions

module Network.AWS.RDS.DescribeReservedDBInstances

module Network.AWS.RDS.DescribeReservedDBInstancesOfferings

module Network.AWS.RDS.DownloadDBLogFilePortion

module Network.AWS.RDS.ListTagsForResource

module Network.AWS.RDS.ModifyDBInstance

module Network.AWS.RDS.ModifyDBParameterGroup

module Network.AWS.RDS.ModifyDBSubnetGroup

module Network.AWS.RDS.ModifyEventSubscription

module Network.AWS.RDS.ModifyOptionGroup

module Network.AWS.RDS.PromoteReadReplica

module Network.AWS.RDS.PurchaseReservedDBInstancesOffering

module Network.AWS.RDS.RebootDBInstance

module Network.AWS.RDS.RemoveSourceIdentifierFromSubscription

module Network.AWS.RDS.RemoveTagsFromResource

module Network.AWS.RDS.ResetDBParameterGroup

module Network.AWS.RDS.RestoreDBInstanceFromDBSnapshot

module Network.AWS.RDS.RestoreDBInstanceToPointInTime

module Network.AWS.RDS.RevokeDBSecurityGroupIngress

module Network.AWS.RDS.Types

module Network.AWS.RDS.Waiters