Network.AWS.RDS.AddTagsToResource

Description

Adds metadata tags to an Amazon RDS resource. These tags can also be used with cost allocation reporting to track cost associated with Amazon RDS resources, or used in Condition statement in IAM policy for Amazon RDS.

For an overview on tagging Amazon RDS resources, see Tagging Amazon RDSResources.

http://docs.aws.amazon.com/AmazonRDS/latest/APIReference/API_AddTagsToResource.html

Request

data AddTagsToResource

Request constructor

addTagsToResource

Arguments

:: Text
attrResourceName

-> AddTagsToResource

AddTagsToResource constructor.

The fields accessible through corresponding lenses are:

attrResourceName :: Text

attrTags :: [Tag]

Request lenses

attrResourceName :: Lens' AddTagsToResource Text

The Amazon RDS resource the tags will be added to. This value is an Amazon Resource Name (ARN). For information about creating an ARN, see Constructingan RDS Amazon Resource Name (ARN).

attrTags :: Lens' AddTagsToResource [Tag]

The tags to be assigned to the Amazon RDS resource.

Response

data AddTagsToResourceResponse

Response constructor

addTagsToResourceResponse :: AddTagsToResourceResponse

AddTagsToResourceResponse constructor.