Network.AWS.RDS.ApplyPendingMaintenanceAction

Description

Applies a pending maintenance action to a resource (for example, a DB instance).

http://docs.aws.amazon.com/AmazonRDS/latest/APIReference/API_ApplyPendingMaintenanceAction.html

Request

data ApplyPendingMaintenanceAction

Request constructor

applyPendingMaintenanceAction

Arguments

:: Text
apmaResourceIdentifier

-> Text
apmaApplyAction

-> Text
apmaOptInType

-> ApplyPendingMaintenanceAction

ApplyPendingMaintenanceAction constructor.

The fields accessible through corresponding lenses are:

apmaApplyAction :: Text

apmaOptInType :: Text

apmaResourceIdentifier :: Text

Request lenses

apmaApplyAction :: Lens' ApplyPendingMaintenanceAction Text

The pending maintenance action to apply to this resource.

apmaOptInType :: Lens' ApplyPendingMaintenanceAction Text

A value that specifies the type of opt-in request, or undoes an opt-in request. An opt-in request of type immediate cannot be undone.

Valid values:

immediate - Apply the maintenance action immediately. 'next-maintenance' - Apply the maintenance action during the next maintenance window for the resource. 'undo-opt-in' - Cancel any existing 'next-maintenance' opt-in requests.

apmaResourceIdentifier :: Lens' ApplyPendingMaintenanceAction Text

The ARN of the resource that the pending maintenance action applies to.

Response

data ApplyPendingMaintenanceActionResponse

Response constructor

applyPendingMaintenanceActionResponse :: ApplyPendingMaintenanceActionResponse

ApplyPendingMaintenanceActionResponse constructor.

The fields accessible through corresponding lenses are:

apmarResourcePendingMaintenanceActions :: Maybe ResourcePendingMaintenanceActions

Response lenses

apmarResourcePendingMaintenanceActions :: Lens' ApplyPendingMaintenanceActionResponse (Maybe ResourcePendingMaintenanceActions)