Network.AWS.RDS.AuthorizeDBSecurityGroupIngress

Description

Enables ingress to a DBSecurityGroup using one of two forms of authorization. First, EC2 or VPC security groups can be added to the DBSecurityGroup if the application using the database is running on EC2 or VPC instances. Second, IP ranges are available if the application accessing your database is running on the Internet. Required parameters for this API are one of CIDR range, EC2SecurityGroupId for VPC, or (EC2SecurityGroupOwnerId and either EC2SecurityGroupName or EC2SecurityGroupId for non-VPC).

For an overview of CIDR ranges, go to the Wikipedia Tutorial .

Creating a Request

authorizeDBSecurityGroupIngress

Arguments

:: Text
adsgiDBSecurityGroupName

-> AuthorizeDBSecurityGroupIngress

Creates a value of AuthorizeDBSecurityGroupIngress with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

adsgiEC2SecurityGroupOwnerId - AWS account number of the owner of the EC2 security group specified in the EC2SecurityGroupName parameter. The AWS Access Key ID is not an acceptable value. For VPC DB security groups, EC2SecurityGroupId must be provided. Otherwise, EC2SecurityGroupOwnerId and either EC2SecurityGroupName or EC2SecurityGroupId must be provided.

adsgiEC2SecurityGroupName - Name of the EC2 security group to authorize. For VPC DB security groups, EC2SecurityGroupId must be provided. Otherwise, EC2SecurityGroupOwnerId and either EC2SecurityGroupName or EC2SecurityGroupId must be provided.

adsgiCIdRIP - The IP range to authorize.

adsgiEC2SecurityGroupId - Id of the EC2 security group to authorize. For VPC DB security groups, EC2SecurityGroupId must be provided. Otherwise, EC2SecurityGroupOwnerId and either EC2SecurityGroupName or EC2SecurityGroupId must be provided.

adsgiDBSecurityGroupName - The name of the DB security group to add authorization to.

data AuthorizeDBSecurityGroupIngress

See: authorizeDBSecurityGroupIngress smart constructor.

Request Lenses

adsgiEC2SecurityGroupOwnerId :: Lens' AuthorizeDBSecurityGroupIngress (Maybe Text)

AWS account number of the owner of the EC2 security group specified in the EC2SecurityGroupName parameter. The AWS Access Key ID is not an acceptable value. For VPC DB security groups, EC2SecurityGroupId must be provided. Otherwise, EC2SecurityGroupOwnerId and either EC2SecurityGroupName or EC2SecurityGroupId must be provided.

adsgiEC2SecurityGroupName :: Lens' AuthorizeDBSecurityGroupIngress (Maybe Text)

Name of the EC2 security group to authorize. For VPC DB security groups, EC2SecurityGroupId must be provided. Otherwise, EC2SecurityGroupOwnerId and either EC2SecurityGroupName or EC2SecurityGroupId must be provided.

adsgiCIdRIP :: Lens' AuthorizeDBSecurityGroupIngress (Maybe Text)

The IP range to authorize.

adsgiEC2SecurityGroupId :: Lens' AuthorizeDBSecurityGroupIngress (Maybe Text)

Id of the EC2 security group to authorize. For VPC DB security groups, EC2SecurityGroupId must be provided. Otherwise, EC2SecurityGroupOwnerId and either EC2SecurityGroupName or EC2SecurityGroupId must be provided.

adsgiDBSecurityGroupName :: Lens' AuthorizeDBSecurityGroupIngress Text

The name of the DB security group to add authorization to.

Destructuring the Response

authorizeDBSecurityGroupIngressResponse

Arguments

:: Int
adsgirsResponseStatus

-> AuthorizeDBSecurityGroupIngressResponse

Creates a value of AuthorizeDBSecurityGroupIngressResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

adsgirsDBSecurityGroup - Undocumented member.

adsgirsResponseStatus - -- | The response status code.

data AuthorizeDBSecurityGroupIngressResponse

See: authorizeDBSecurityGroupIngressResponse smart constructor.

Response Lenses

adsgirsDBSecurityGroup :: Lens' AuthorizeDBSecurityGroupIngressResponse (Maybe DBSecurityGroup)

Undocumented member.

adsgirsResponseStatus :: Lens' AuthorizeDBSecurityGroupIngressResponse Int

- | The response status code.