Network.AWS.RDS.CopyDBClusterParameterGroup

Description

Copies the specified DB cluster parameter group.

Creating a Request

copyDBClusterParameterGroup

Arguments

:: Text
cdbcpgSourceDBClusterParameterGroupIdentifier

-> Text
cdbcpgTargetDBClusterParameterGroupIdentifier

-> Text
cdbcpgTargetDBClusterParameterGroupDescription

-> CopyDBClusterParameterGroup

Creates a value of CopyDBClusterParameterGroup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdbcpgTags - Undocumented member.

cdbcpgSourceDBClusterParameterGroupIdentifier - The identifier or Amazon Resource Name (ARN) for the source DB cluster parameter group. For information about creating an ARN, see Constructing an RDS Amazon Resource Name (ARN) . Constraints: * Must specify a valid DB cluster parameter group. * If the source DB cluster parameter group is in the same AWS Region as the copy, specify a valid DB parameter group identifier, for example my-db-cluster-param-group , or a valid ARN. * If the source DB parameter group is in a different AWS Region than the copy, specify a valid DB cluster parameter group ARN, for example arn:aws:rds:us-east-1:123456789012:cluster-pg:custom-cluster-group1 .

cdbcpgTargetDBClusterParameterGroupIdentifier - The identifier for the copied DB cluster parameter group. Constraints: * Cannot be null, empty, or blank * Must contain from 1 to 255 alphanumeric characters or hyphens * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens Example: my-cluster-param-group1

cdbcpgTargetDBClusterParameterGroupDescription - A description for the copied DB cluster parameter group.

data CopyDBClusterParameterGroup

See: copyDBClusterParameterGroup smart constructor.

Request Lenses

cdbcpgTags :: Lens' CopyDBClusterParameterGroup [Tag]

Undocumented member.

cdbcpgSourceDBClusterParameterGroupIdentifier :: Lens' CopyDBClusterParameterGroup Text

The identifier or Amazon Resource Name (ARN) for the source DB cluster parameter group. For information about creating an ARN, see Constructing an RDS Amazon Resource Name (ARN) . Constraints: * Must specify a valid DB cluster parameter group. * If the source DB cluster parameter group is in the same AWS Region as the copy, specify a valid DB parameter group identifier, for example my-db-cluster-param-group , or a valid ARN. * If the source DB parameter group is in a different AWS Region than the copy, specify a valid DB cluster parameter group ARN, for example arn:aws:rds:us-east-1:123456789012:cluster-pg:custom-cluster-group1 .

cdbcpgTargetDBClusterParameterGroupIdentifier :: Lens' CopyDBClusterParameterGroup Text

The identifier for the copied DB cluster parameter group. Constraints: * Cannot be null, empty, or blank * Must contain from 1 to 255 alphanumeric characters or hyphens * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens Example: my-cluster-param-group1

cdbcpgTargetDBClusterParameterGroupDescription :: Lens' CopyDBClusterParameterGroup Text

A description for the copied DB cluster parameter group.

Destructuring the Response

copyDBClusterParameterGroupResponse

Arguments

:: Int
cdcpgrsResponseStatus

-> CopyDBClusterParameterGroupResponse

Creates a value of CopyDBClusterParameterGroupResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdcpgrsDBClusterParameterGroup - Undocumented member.

cdcpgrsResponseStatus - -- | The response status code.

data CopyDBClusterParameterGroupResponse

See: copyDBClusterParameterGroupResponse smart constructor.

Response Lenses

cdcpgrsDBClusterParameterGroup :: Lens' CopyDBClusterParameterGroupResponse (Maybe DBClusterParameterGroup)

Undocumented member.

cdcpgrsResponseStatus :: Lens' CopyDBClusterParameterGroupResponse Int

- | The response status code.