Network.AWS.RDS.CopyDBParameterGroup

Description

Copies the specified DB parameter group.

http://docs.aws.amazon.com/AmazonRDS/latest/APIReference/API_CopyDBParameterGroup.html

Request

data CopyDBParameterGroup

Request constructor

copyDBParameterGroup

Arguments

:: Text
cdbpgSourceDBParameterGroupIdentifier

-> Text
cdbpgTargetDBParameterGroupIdentifier

-> Text
cdbpgTargetDBParameterGroupDescription

-> CopyDBParameterGroup

CopyDBParameterGroup constructor.

The fields accessible through corresponding lenses are:

cdbpgSourceDBParameterGroupIdentifier :: Text

cdbpgTags :: [Tag]

cdbpgTargetDBParameterGroupDescription :: Text

cdbpgTargetDBParameterGroupIdentifier :: Text

Request lenses

cdbpgSourceDBParameterGroupIdentifier :: Lens' CopyDBParameterGroup Text

The identifier or ARN for the source DB parameter group.

Constraints:

Must specify a valid DB parameter group. If the source DB parameter group is in the same region as the copy, specify a valid DB parameter group identifier, for example 'my-db-param-group', or a valid ARN. If the source DB parameter group is in a different region than the copy, specify a valid DB parameter group ARN, for example 'arn:aws:rds:us-west-2:123456789012:pg:special-parameters'.

cdbpgTags :: Lens' CopyDBParameterGroup [Tag]

cdbpgTargetDBParameterGroupDescription :: Lens' CopyDBParameterGroup Text

A description for the copied DB parameter group.

cdbpgTargetDBParameterGroupIdentifier :: Lens' CopyDBParameterGroup Text

The identifier for the copied DB parameter group.

Constraints:

Cannot be null, empty, or blank Must contain from 1 to 255 alphanumeric characters or hyphens First character must be a letter Cannot end with a hyphen or contain two consecutive hyphens Example: 'my-db-parameter-group'

Response

data CopyDBParameterGroupResponse

Response constructor

copyDBParameterGroupResponse :: CopyDBParameterGroupResponse

CopyDBParameterGroupResponse constructor.

The fields accessible through corresponding lenses are:

cdbpgr1DBParameterGroup :: Maybe DBParameterGroup

Response lenses

cdbpgr1DBParameterGroup :: Lens' CopyDBParameterGroupResponse (Maybe DBParameterGroup)