Network.AWS.RDS.CopyDBSnapshot

Description

Copies the specified DBSnapshot. The source DBSnapshot must be in the "available" state. https:/rds.amazonaws.com ?Action=CopyDBSnapshot &SourceDBSnapshotIdentifier=rds:simcoprod01-2012-04-02-00-01 &TargetDBSnapshotIdentifier=mydbsnapshot &Version=2013-05-15 &SignatureVersion=2&SignatureMethod=HmacSHA256 &Timestamp=2011-12-12T06%3A27%3A42.551Z &AWSAccessKeyId= &Signature= 3306 mysql available us-east-1a general-public-license 2011-05-23T06:06:43.110Z 10 simcoprod01 5.1.50 mydbsnapshot manual master c4181d1d-8505-11e0-90aa-eb648410240d.

Request

data CopyDBSnapshot

Request constructor

copyDBSnapshot

Arguments

:: Text
cdbsSourceDBSnapshotIdentifier

-> Text
cdbsTargetDBSnapshotIdentifier

-> CopyDBSnapshot

Smart constructor for the minimum required parameters to construct a valid CopyDBSnapshot request.

The fields accessible through corresponding lenses are:

SourceDBSnapshotIdentifier :: Text

TargetDBSnapshotIdentifier :: Text

Tags :: [Tag]

Request lenses

cdbsSourceDBSnapshotIdentifier :: Lens' CopyDBSnapshot Text

The identifier for the source DB snapshot. Constraints: Must specify a valid system snapshot in the "available" state. If the source snapshot is in the same region as the copy, specify a valid DB snapshot identifier. If the source snapshot is in a different region than the copy, specify valid DB snapshot ARN. For more information, go to Copying a DB Snapshot. Example: rds:mydb-2012-04-02-00-01 Example: arn:aws:rds:rr-regn-1:123456789012:snapshot:mysql-instance1-snapshot-20130805.

cdbsTargetDBSnapshotIdentifier :: Lens' CopyDBSnapshot Text

The identifier for the copied snapshot. Constraints: Cannot be null, empty, or blank Must contain from 1 to 255 alphanumeric characters or hyphens First character must be a letter Cannot end with a hyphen or contain two consecutive hyphens Example: my-db-snapshot.

cdbsTags :: Lens' CopyDBSnapshot [Tag]

A list of tags.

Response

data CopyDBSnapshotResponse

Response constructor

copyDBSnapshotResponse :: CopyDBSnapshotResponse

Smart constructor for the minimum required parameters to construct a valid CopyDBSnapshotResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

DBSnapshot :: Maybe DBSnapshot

Response lenses

cdbsrDBSnapshot :: Lens' CopyDBSnapshotResponse (Maybe DBSnapshot)

Contains the result of a successful invocation of the following actions: CreateDBSnapshot DeleteDBSnapshot This data type is used as a response element in the DescribeDBSnapshots action.