Network.AWS.RDS.CopyOptionGroup

Description

Copies the specified option group.

http://docs.aws.amazon.com/AmazonRDS/latest/APIReference/API_CopyOptionGroup.html

Request

data CopyOptionGroup

Request constructor

copyOptionGroup

Arguments

:: Text
cog1SourceOptionGroupIdentifier

-> Text
cog1TargetOptionGroupIdentifier

-> Text
cog1TargetOptionGroupDescription

-> CopyOptionGroup

CopyOptionGroup constructor.

The fields accessible through corresponding lenses are:

cog1SourceOptionGroupIdentifier :: Text

cog1Tags :: [Tag]

cog1TargetOptionGroupDescription :: Text

cog1TargetOptionGroupIdentifier :: Text

Request lenses

cog1SourceOptionGroupIdentifier :: Lens' CopyOptionGroup Text

The identifier or ARN for the source option group.

Constraints:

Must specify a valid option group. If the source option group is in the same region as the copy, specify a valid option group identifier, for example 'my-option-group', or a valid ARN. If the source option group is in a different region than the copy, specify a valid option group ARN, for example 'arn:aws:rds:us-west-2:123456789012:og:special-options'.

cog1Tags :: Lens' CopyOptionGroup [Tag]

cog1TargetOptionGroupDescription :: Lens' CopyOptionGroup Text

The description for the copied option group.

cog1TargetOptionGroupIdentifier :: Lens' CopyOptionGroup Text

The identifier for the copied option group.

Constraints:

Cannot be null, empty, or blank Must contain from 1 to 255 alphanumeric characters or hyphens First character must be a letter Cannot end with a hyphen or contain two consecutive hyphens Example: 'my-option-group'

Response

data CopyOptionGroupResponse

Response constructor

copyOptionGroupResponse :: CopyOptionGroupResponse

CopyOptionGroupResponse constructor.

The fields accessible through corresponding lenses are:

cogrOptionGroup :: Maybe OptionGroup

Response lenses

cogrOptionGroup :: Lens' CopyOptionGroupResponse (Maybe OptionGroup)