Network.AWS.RDS.CreateDBCluster

Description

Creates a new Amazon Aurora DB cluster.

You can use the ReplicationSourceIdentifier parameter to create the DB cluster as a Read Replica of another DB cluster or Amazon RDS MySQL DB instance. For cross-region replication where the DB cluster identified by ReplicationSourceIdentifier is encrypted, you must also specify the PreSignedUrl parameter.

For more information on Amazon Aurora, see Aurora on Amazon RDS in the Amazon RDS User Guide.

Creating a Request

createDBCluster

Arguments

:: Text
cdcDBClusterIdentifier

-> Text
cdcEngine

-> CreateDBCluster

Creates a value of CreateDBCluster with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdcEngineVersion - The version number of the database engine to use. Aurora Example: 5.6.10a

cdcStorageEncrypted - Specifies whether the DB cluster is encrypted.

cdcMasterUserPassword - The password for the master database user. This password can contain any printable ASCII character except "/", """, or "@". Constraints: Must contain from 8 to 41 characters.

cdcReplicationSourceIdentifier - The Amazon Resource Name (ARN) of the source DB instance or DB cluster if this DB cluster is created as a Read Replica.

cdcMasterUsername - The name of the master user for the DB cluster. Constraints: * Must be 1 to 16 alphanumeric characters. * First character must be a letter. * Cannot be a reserved word for the chosen database engine.

cdcDBSubnetGroupName - A DB subnet group to associate with this DB cluster. Constraints: Must contain no more than 255 alphanumeric characters, periods, underscores, spaces, or hyphens. Must not be default. Example: mySubnetgroup

cdcPreSignedURL - A URL that contains a Signature Version 4 signed request for the CreateDBCluster action to be called in the source AWS Region where the DB cluster will be replicated from. You only need to specify PreSignedUrl when you are performing cross-region replication from an encrypted DB cluster. The pre-signed URL must be a valid request for the CreateDBCluster API action that can be executed in the source AWS Region that contains the encrypted DB cluster to be copied. The pre-signed URL request must contain the following parameter values: * KmsKeyId - The KMS key identifier for the key to use to encrypt the copy of the DB cluster in the destination AWS Region. This should refer to the same KMS key for both the CreateDBCluster action that is called in the destination AWS Region, and the action contained in the pre-signed URL. * DestinationRegion - The name of the AWS Region that Aurora Read Replica will be created in. * ReplicationSourceIdentifier - The DB cluster identifier for the encrypted DB cluster to be copied. This identifier must be in the Amazon Resource Name (ARN) format for the source AWS Region. For example, if you are copying an encrypted DB cluster from the us-west-2 region, then your ReplicationSourceIdentifier would look like Example: arn:aws:rds:us-west-2:123456789012:cluster:aurora-cluster1 . To learn how to generate a Signature Version 4 signed request, see Authenticating Requests: Using Query Parameters (AWS Signature Version 4) and Signature Version 4 Signing Process .

cdcPreferredMaintenanceWindow - The weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC). Format: ddd:hh24:mi-ddd:hh24:mi Default: A 30-minute window selected at random from an 8-hour block of time per AWS Region, occurring on a random day of the week. To see the time blocks available, see Adjusting the Preferred Maintenance Window in the Amazon RDS User Guide. Valid Days: Mon, Tue, Wed, Thu, Fri, Sat, Sun Constraints: Minimum 30-minute window.

cdcAvailabilityZones - A list of EC2 Availability Zones that instances in the DB cluster can be created in. For information on regions and Availability Zones, see Regions and Availability Zones .

cdcCharacterSetName - A value that indicates that the DB cluster should be associated with the specified CharacterSet.

cdcKMSKeyId - The KMS key identifier for an encrypted DB cluster. The KMS key identifier is the Amazon Resource Name (ARN) for the KMS encryption key. If you are creating a DB cluster with the same AWS account that owns the KMS encryption key used to encrypt the new DB cluster, then you can use the KMS key alias instead of the ARN for the KMS encryption key. If the StorageEncrypted parameter is true, and you do not specify a value for the KmsKeyId parameter, then Amazon RDS will use your default encryption key. AWS KMS creates the default encryption key for your AWS account. Your AWS account has a different default encryption key for each AWS Region. If you create a Read Replica of an encrypted DB cluster in another AWS Region, you must set KmsKeyId to a KMS key ID that is valid in the destination AWS Region. This key is used to encrypt the Read Replica in that AWS Region.

cdcPreferredBackupWindow - The daily time range during which automated backups are created if automated backups are enabled using the BackupRetentionPeriod parameter. Default: A 30-minute window selected at random from an 8-hour block of time per AWS Region. To see the time blocks available, see Adjusting the Preferred Maintenance Window in the Amazon RDS User Guide. Constraints: * Must be in the format hh24:mi-hh24:mi . * Times should be in Universal Coordinated Time (UTC). * Must not conflict with the preferred maintenance window. * Must be at least 30 minutes.

cdcBackupRetentionPeriod - The number of days for which automated backups are retained. You must specify a minimum value of 1. Default: 1 Constraints: * Must be a value from 1 to 35

cdcVPCSecurityGroupIds - A list of EC2 VPC security groups to associate with this DB cluster.

cdcDatabaseName - The name for your database of up to 64 alpha-numeric characters. If you do not provide a name, Amazon RDS will not create a database in the DB cluster you are creating.

cdcDBClusterParameterGroupName - The name of the DB cluster parameter group to associate with this DB cluster. If this argument is omitted, default.aurora5.6 will be used. Constraints: * Must be 1 to 255 alphanumeric characters * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens

cdcOptionGroupName - A value that indicates that the DB cluster should be associated with the specified option group. Permanent options cannot be removed from an option group. The option group cannot be removed from a DB cluster once it is associated with a DB cluster.

cdcTags - Undocumented member.

cdcPort - The port number on which the instances in the DB cluster accept connections. Default: 3306

cdcEnableIAMDatabaseAuthentication - A Boolean value that is true to enable mapping of AWS Identity and Access Management (IAM) accounts to database accounts, and otherwise false. Default: false

cdcDBClusterIdentifier - The DB cluster identifier. This parameter is stored as a lowercase string. Constraints: * Must contain from 1 to 63 alphanumeric characters or hyphens. * First character must be a letter. * Cannot end with a hyphen or contain two consecutive hyphens. Example: my-cluster1

cdcEngine - The name of the database engine to be used for this DB cluster. Valid Values: aurora

data CreateDBCluster

See: createDBCluster smart constructor.

Request Lenses

cdcEngineVersion :: Lens' CreateDBCluster (Maybe Text)

The version number of the database engine to use. Aurora Example: 5.6.10a

cdcStorageEncrypted :: Lens' CreateDBCluster (Maybe Bool)

Specifies whether the DB cluster is encrypted.

cdcMasterUserPassword :: Lens' CreateDBCluster (Maybe Text)

The password for the master database user. This password can contain any printable ASCII character except "/", """, or "@". Constraints: Must contain from 8 to 41 characters.

cdcReplicationSourceIdentifier :: Lens' CreateDBCluster (Maybe Text)

The Amazon Resource Name (ARN) of the source DB instance or DB cluster if this DB cluster is created as a Read Replica.

cdcMasterUsername :: Lens' CreateDBCluster (Maybe Text)

The name of the master user for the DB cluster. Constraints: * Must be 1 to 16 alphanumeric characters. * First character must be a letter. * Cannot be a reserved word for the chosen database engine.

cdcDBSubnetGroupName :: Lens' CreateDBCluster (Maybe Text)

A DB subnet group to associate with this DB cluster. Constraints: Must contain no more than 255 alphanumeric characters, periods, underscores, spaces, or hyphens. Must not be default. Example: mySubnetgroup

cdcPreSignedURL :: Lens' CreateDBCluster (Maybe Text)

A URL that contains a Signature Version 4 signed request for the CreateDBCluster action to be called in the source AWS Region where the DB cluster will be replicated from. You only need to specify PreSignedUrl when you are performing cross-region replication from an encrypted DB cluster. The pre-signed URL must be a valid request for the CreateDBCluster API action that can be executed in the source AWS Region that contains the encrypted DB cluster to be copied. The pre-signed URL request must contain the following parameter values: * KmsKeyId - The KMS key identifier for the key to use to encrypt the copy of the DB cluster in the destination AWS Region. This should refer to the same KMS key for both the CreateDBCluster action that is called in the destination AWS Region, and the action contained in the pre-signed URL. * DestinationRegion - The name of the AWS Region that Aurora Read Replica will be created in. * ReplicationSourceIdentifier - The DB cluster identifier for the encrypted DB cluster to be copied. This identifier must be in the Amazon Resource Name (ARN) format for the source AWS Region. For example, if you are copying an encrypted DB cluster from the us-west-2 region, then your ReplicationSourceIdentifier would look like Example: arn:aws:rds:us-west-2:123456789012:cluster:aurora-cluster1 . To learn how to generate a Signature Version 4 signed request, see Authenticating Requests: Using Query Parameters (AWS Signature Version 4) and Signature Version 4 Signing Process .

cdcPreferredMaintenanceWindow :: Lens' CreateDBCluster (Maybe Text)

The weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC). Format: ddd:hh24:mi-ddd:hh24:mi Default: A 30-minute window selected at random from an 8-hour block of time per AWS Region, occurring on a random day of the week. To see the time blocks available, see Adjusting the Preferred Maintenance Window in the Amazon RDS User Guide. Valid Days: Mon, Tue, Wed, Thu, Fri, Sat, Sun Constraints: Minimum 30-minute window.

cdcAvailabilityZones :: Lens' CreateDBCluster [Text]

A list of EC2 Availability Zones that instances in the DB cluster can be created in. For information on regions and Availability Zones, see Regions and Availability Zones .

cdcCharacterSetName :: Lens' CreateDBCluster (Maybe Text)

A value that indicates that the DB cluster should be associated with the specified CharacterSet.

cdcKMSKeyId :: Lens' CreateDBCluster (Maybe Text)

The KMS key identifier for an encrypted DB cluster. The KMS key identifier is the Amazon Resource Name (ARN) for the KMS encryption key. If you are creating a DB cluster with the same AWS account that owns the KMS encryption key used to encrypt the new DB cluster, then you can use the KMS key alias instead of the ARN for the KMS encryption key. If the StorageEncrypted parameter is true, and you do not specify a value for the KmsKeyId parameter, then Amazon RDS will use your default encryption key. AWS KMS creates the default encryption key for your AWS account. Your AWS account has a different default encryption key for each AWS Region. If you create a Read Replica of an encrypted DB cluster in another AWS Region, you must set KmsKeyId to a KMS key ID that is valid in the destination AWS Region. This key is used to encrypt the Read Replica in that AWS Region.

cdcPreferredBackupWindow :: Lens' CreateDBCluster (Maybe Text)

The daily time range during which automated backups are created if automated backups are enabled using the BackupRetentionPeriod parameter. Default: A 30-minute window selected at random from an 8-hour block of time per AWS Region. To see the time blocks available, see Adjusting the Preferred Maintenance Window in the Amazon RDS User Guide. Constraints: * Must be in the format hh24:mi-hh24:mi . * Times should be in Universal Coordinated Time (UTC). * Must not conflict with the preferred maintenance window. * Must be at least 30 minutes.

cdcBackupRetentionPeriod :: Lens' CreateDBCluster (Maybe Int)

The number of days for which automated backups are retained. You must specify a minimum value of 1. Default: 1 Constraints: * Must be a value from 1 to 35

cdcVPCSecurityGroupIds :: Lens' CreateDBCluster [Text]

A list of EC2 VPC security groups to associate with this DB cluster.

cdcDatabaseName :: Lens' CreateDBCluster (Maybe Text)

The name for your database of up to 64 alpha-numeric characters. If you do not provide a name, Amazon RDS will not create a database in the DB cluster you are creating.

cdcDBClusterParameterGroupName :: Lens' CreateDBCluster (Maybe Text)

The name of the DB cluster parameter group to associate with this DB cluster. If this argument is omitted, default.aurora5.6 will be used. Constraints: * Must be 1 to 255 alphanumeric characters * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens

cdcOptionGroupName :: Lens' CreateDBCluster (Maybe Text)

A value that indicates that the DB cluster should be associated with the specified option group. Permanent options cannot be removed from an option group. The option group cannot be removed from a DB cluster once it is associated with a DB cluster.

cdcTags :: Lens' CreateDBCluster [Tag]

Undocumented member.

cdcPort :: Lens' CreateDBCluster (Maybe Int)

The port number on which the instances in the DB cluster accept connections. Default: 3306

cdcEnableIAMDatabaseAuthentication :: Lens' CreateDBCluster (Maybe Bool)

A Boolean value that is true to enable mapping of AWS Identity and Access Management (IAM) accounts to database accounts, and otherwise false. Default: false

cdcDBClusterIdentifier :: Lens' CreateDBCluster Text

The DB cluster identifier. This parameter is stored as a lowercase string. Constraints: * Must contain from 1 to 63 alphanumeric characters or hyphens. * First character must be a letter. * Cannot end with a hyphen or contain two consecutive hyphens. Example: my-cluster1

cdcEngine :: Lens' CreateDBCluster Text

The name of the database engine to be used for this DB cluster. Valid Values: aurora

Destructuring the Response

createDBClusterResponse

Arguments

:: Int
cdcrsResponseStatus

-> CreateDBClusterResponse

Creates a value of CreateDBClusterResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdcrsDBCluster - Undocumented member.

cdcrsResponseStatus - -- | The response status code.

data CreateDBClusterResponse

See: createDBClusterResponse smart constructor.

Response Lenses

cdcrsDBCluster :: Lens' CreateDBClusterResponse (Maybe DBCluster)

Undocumented member.

cdcrsResponseStatus :: Lens' CreateDBClusterResponse Int

- | The response status code.