Network.AWS.RDS.CreateDBInstanceReadReplica

Description

Creates a DB instance that acts as a Read Replica of a source DB instance.

All Read Replica DB instances are created as Single-AZ deployments with backups disabled. All other DB instance attributes (including DB security groups and DB parameter groups) are inherited from the source DB instance, except as specified below.

The source DB instance must have backup retention enabled.

http://docs.aws.amazon.com/AmazonRDS/latest/APIReference/API_CreateDBInstanceReadReplica.html

Request

data CreateDBInstanceReadReplica

Request constructor

createDBInstanceReadReplica

Arguments

:: Text
cdbirrDBInstanceIdentifier

-> Text
cdbirrSourceDBInstanceIdentifier

-> CreateDBInstanceReadReplica

CreateDBInstanceReadReplica constructor.

The fields accessible through corresponding lenses are:

cdbirrAutoMinorVersionUpgrade :: Maybe Bool

cdbirrAvailabilityZone :: Maybe Text

cdbirrDBInstanceClass :: Maybe Text

cdbirrDBInstanceIdentifier :: Text

cdbirrDBSubnetGroupName :: Maybe Text

cdbirrIops :: Maybe Int

cdbirrOptionGroupName :: Maybe Text

cdbirrPort :: Maybe Int

cdbirrPubliclyAccessible :: Maybe Bool

cdbirrSourceDBInstanceIdentifier :: Text

cdbirrStorageType :: Maybe Text

cdbirrTags :: [Tag]

Request lenses

cdbirrAutoMinorVersionUpgrade :: Lens' CreateDBInstanceReadReplica (Maybe Bool)

Indicates that minor engine upgrades will be applied automatically to the Read Replica during the maintenance window.

Default: Inherits from the source DB instance

cdbirrAvailabilityZone :: Lens' CreateDBInstanceReadReplica (Maybe Text)

The Amazon EC2 Availability Zone that the Read Replica will be created in.

Default: A random, system-chosen Availability Zone in the endpoint's region.

Example: 'us-east-1d'

cdbirrDBInstanceClass :: Lens' CreateDBInstanceReadReplica (Maybe Text)

The compute and memory capacity of the Read Replica.

Valid Values: 'db.m1.small | db.m1.medium | db.m1.large | db.m1.xlarge |db.m2.xlarge |db.m2.2xlarge | db.m2.4xlarge | db.m3.medium | db.m3.large |db.m3.xlarge | db.m3.2xlarge | db.r3.large | db.r3.xlarge | db.r3.2xlarge |db.r3.4xlarge | db.r3.8xlarge | db.t2.micro | db.t2.small | db.t2.medium'

Default: Inherits from the source DB instance.

cdbirrDBInstanceIdentifier :: Lens' CreateDBInstanceReadReplica Text

The DB instance identifier of the Read Replica. This is the unique key that identifies a DB instance. This parameter is stored as a lowercase string.

cdbirrDBSubnetGroupName :: Lens' CreateDBInstanceReadReplica (Maybe Text)

Specifies a DB subnet group for the DB instance. The new DB instance will be created in the VPC associated with the DB subnet group. If no DB subnet group is specified, then the new DB instance is not created in a VPC.

Constraints:

Can only be specified if the source DB instance identifier specifies a DB instance in another region. The specified DB subnet group must be in the same region in which the operation is running. All Read Replicas in one region that are created from the same source DB instance must either: Specify DB subnet groups from the same VPC. All these Read Replicas will be created in the same VPC. Not specify a DB subnet group. All these Read Replicas will be created outside of any VPC.

cdbirrIops :: Lens' CreateDBInstanceReadReplica (Maybe Int)

The amount of Provisioned IOPS (input/output operations per second) to be initially allocated for the DB instance.

cdbirrOptionGroupName :: Lens' CreateDBInstanceReadReplica (Maybe Text)

The option group the DB instance will be associated with. If omitted, the default option group for the engine specified will be used.

cdbirrPort :: Lens' CreateDBInstanceReadReplica (Maybe Int)

The port number that the DB instance uses for connections.

Default: Inherits from the source DB instance

Valid Values: '1150-65535'

cdbirrPubliclyAccessible :: Lens' CreateDBInstanceReadReplica (Maybe Bool)

Specifies the accessibility options for the DB instance. A value of true specifies an Internet-facing instance with a publicly resolvable DNS name, which resolves to a public IP address. A value of false specifies an internal instance with a DNS name that resolves to a private IP address.

Default: The default behavior varies depending on whether a VPC has been requested or not. The following list shows the default behavior in each case.

Default VPC:true VPC:false If no DB subnet group has been specified as part of the request and the PubliclyAccessible value has not been set, the DB instance will be publicly accessible. If a specific DB subnet group has been specified as part of the request and the PubliclyAccessible value has not been set, the DB instance will be private.

cdbirrSourceDBInstanceIdentifier :: Lens' CreateDBInstanceReadReplica Text

The identifier of the DB instance that will act as the source for the Read Replica. Each DB instance can have up to five Read Replicas.

Constraints:

Must be the identifier of an existing DB instance. Can specify a DB instance that is a MySQL Read Replica only if the source is running MySQL 5.6. Can specify a DB instance that is a PostgreSQL Read Replica only if the source is running PostgreSQL 9.3.5. The specified DB instance must have automatic backups enabled, its backup retention period must be greater than 0. If the source DB instance is in the same region as the Read Replica, specify a valid DB instance identifier. If the source DB instance is in a different region than the Read Replica, specify a valid DB instance ARN. For more information, go to Constructing a Amazon RDS Amazon Resource Name (ARN).

cdbirrStorageType :: Lens' CreateDBInstanceReadReplica (Maybe Text)

Specifies the storage type to be associated with the Read Replica.

Valid values: 'standard | gp2 | io1'

If you specify io1, you must also include a value for the Iops parameter.

Default: io1 if the Iops parameter is specified; otherwise standard

cdbirrTags :: Lens' CreateDBInstanceReadReplica [Tag]

Response

data CreateDBInstanceReadReplicaResponse

Response constructor

createDBInstanceReadReplicaResponse :: CreateDBInstanceReadReplicaResponse

CreateDBInstanceReadReplicaResponse constructor.

The fields accessible through corresponding lenses are:

cdbirrrDBInstance :: Maybe DBInstance

Response lenses

cdbirrrDBInstance :: Lens' CreateDBInstanceReadReplicaResponse (Maybe DBInstance)