Network.AWS.RDS.CreateDBSecurityGroup

Description

Creates a new DB security group. DB security groups control access to a DB instance.

http://docs.aws.amazon.com/AmazonRDS/latest/APIReference/API_CreateDBSecurityGroup.html

Request

data CreateDBSecurityGroup

Request constructor

createDBSecurityGroup

Arguments

:: Text
cdbsgDBSecurityGroupName

-> Text
cdbsgDBSecurityGroupDescription

-> CreateDBSecurityGroup

CreateDBSecurityGroup constructor.

The fields accessible through corresponding lenses are:

cdbsgDBSecurityGroupDescription :: Text

cdbsgDBSecurityGroupName :: Text

cdbsgTags :: [Tag]

Request lenses

cdbsgDBSecurityGroupDescription :: Lens' CreateDBSecurityGroup Text

The description for the DB security group.

cdbsgDBSecurityGroupName :: Lens' CreateDBSecurityGroup Text

The name for the DB security group. This value is stored as a lowercase string.

Constraints:

Must be 1 to 255 alphanumeric characters First character must be a letter Cannot end with a hyphen or contain two consecutive hyphens Must not be Default May not contain spaces Example: mysecuritygroup

cdbsgTags :: Lens' CreateDBSecurityGroup [Tag]

Response

data CreateDBSecurityGroupResponse

Response constructor

createDBSecurityGroupResponse :: CreateDBSecurityGroupResponse

CreateDBSecurityGroupResponse constructor.

The fields accessible through corresponding lenses are:

cdbsgrDBSecurityGroup :: Maybe DBSecurityGroup

Response lenses

cdbsgrDBSecurityGroup :: Lens' CreateDBSecurityGroupResponse (Maybe DBSecurityGroup)