Network.AWS.RDS.CreateDBSnapshot

Description

Creates a DBSnapshot. The source DBInstance must be in "available" state.

http://docs.aws.amazon.com/AmazonRDS/latest/APIReference/API_CreateDBSnapshot.html

Request

data CreateDBSnapshot

Request constructor

createDBSnapshot

Arguments

:: Text
cdbs1DBSnapshotIdentifier

-> Text
cdbs1DBInstanceIdentifier

-> CreateDBSnapshot

CreateDBSnapshot constructor.

The fields accessible through corresponding lenses are:

cdbs1DBInstanceIdentifier :: Text

cdbs1DBSnapshotIdentifier :: Text

cdbs1Tags :: [Tag]

Request lenses

cdbs1DBInstanceIdentifier :: Lens' CreateDBSnapshot Text

The DB instance identifier. This is the unique key that identifies a DB instance.

Constraints:

Must contain from 1 to 63 alphanumeric characters or hyphens First character must be a letter Cannot end with a hyphen or contain two consecutive hyphens

cdbs1DBSnapshotIdentifier :: Lens' CreateDBSnapshot Text

The identifier for the DB snapshot.

Constraints:

Cannot be null, empty, or blank Must contain from 1 to 255 alphanumeric characters or hyphens First character must be a letter Cannot end with a hyphen or contain two consecutive hyphens Example: 'my-snapshot-id'

cdbs1Tags :: Lens' CreateDBSnapshot [Tag]

Response

data CreateDBSnapshotResponse

Response constructor

createDBSnapshotResponse :: CreateDBSnapshotResponse

CreateDBSnapshotResponse constructor.

The fields accessible through corresponding lenses are:

cdbsr1DBSnapshot :: Maybe DBSnapshot

Response lenses

cdbsr1DBSnapshot :: Lens' CreateDBSnapshotResponse (Maybe DBSnapshot)