Network.AWS.RDS.CreateDBSubnetGroup

Description

Creates a new DB subnet group. DB subnet groups must contain at least one subnet in at least two AZs in the AWS Region.

Creating a Request

createDBSubnetGroup

Arguments

:: Text
cdbsgDBSubnetGroupName

-> Text
cdbsgDBSubnetGroupDescription

-> CreateDBSubnetGroup

Creates a value of CreateDBSubnetGroup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdbsgTags - Undocumented member.

cdbsgDBSubnetGroupName - The name for the DB subnet group. This value is stored as a lowercase string. Constraints: Must contain no more than 255 alphanumeric characters, periods, underscores, spaces, or hyphens. Must not be default. Example: mySubnetgroup

cdbsgDBSubnetGroupDescription - The description for the DB subnet group.

cdbsgSubnetIds - The EC2 Subnet IDs for the DB subnet group.

data CreateDBSubnetGroup

See: createDBSubnetGroup smart constructor.

Request Lenses

cdbsgTags :: Lens' CreateDBSubnetGroup [Tag]

Undocumented member.

cdbsgDBSubnetGroupName :: Lens' CreateDBSubnetGroup Text

The name for the DB subnet group. This value is stored as a lowercase string. Constraints: Must contain no more than 255 alphanumeric characters, periods, underscores, spaces, or hyphens. Must not be default. Example: mySubnetgroup

cdbsgDBSubnetGroupDescription :: Lens' CreateDBSubnetGroup Text

The description for the DB subnet group.

cdbsgSubnetIds :: Lens' CreateDBSubnetGroup [Text]

The EC2 Subnet IDs for the DB subnet group.

Destructuring the Response

createDBSubnetGroupResponse

Arguments

:: Int
cdsgrsResponseStatus

-> CreateDBSubnetGroupResponse

Creates a value of CreateDBSubnetGroupResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdsgrsDBSubnetGroup - Undocumented member.

cdsgrsResponseStatus - -- | The response status code.

data CreateDBSubnetGroupResponse

See: createDBSubnetGroupResponse smart constructor.

Response Lenses

cdsgrsDBSubnetGroup :: Lens' CreateDBSubnetGroupResponse (Maybe DBSubnetGroup)

Undocumented member.

cdsgrsResponseStatus :: Lens' CreateDBSubnetGroupResponse Int

- | The response status code.