Network.AWS.RDS.CreateEventSubscription

Description

Creates an RDS event notification subscription. This action requires a topic ARN (Amazon Resource Name) created by either the RDS console, the SNS console, or the SNS API. To obtain an ARN with SNS, you must create a topic in Amazon SNS and subscribe to the topic. The ARN is displayed in the SNS console.

You can specify the type of source (SourceType) you want to be notified of, provide a list of RDS sources (SourceIds) that triggers the events, and provide a list of event categories (EventCategories) for events you want to be notified of. For example, you can specify SourceType = db-instance, SourceIds = mydbinstance1, mydbinstance2 and EventCategories = Availability, Backup.

If you specify both the SourceType and SourceIds, such as SourceType = db-instance and SourceIdentifier = myDBInstance1, you will be notified of all the db-instance events for the specified source. If you specify a SourceType but do not specify a SourceIdentifier, you will receive notice of the events for that source type for all your RDS sources. If you do not specify either the SourceType nor the SourceIdentifier, you will be notified of events generated from all RDS sources belonging to your customer account.

http://docs.aws.amazon.com/AmazonRDS/latest/APIReference/API_CreateEventSubscription.html

Request

data CreateEventSubscription

Request constructor

createEventSubscription

Arguments

:: Text
cesSubscriptionName

-> Text
cesSnsTopicArn

-> CreateEventSubscription

CreateEventSubscription constructor.

The fields accessible through corresponding lenses are:

cesEnabled :: Maybe Bool

cesEventCategories :: [Text]

cesSnsTopicArn :: Text

cesSourceIds :: [Text]

cesSourceType :: Maybe Text

cesSubscriptionName :: Text

cesTags :: [Tag]

Request lenses

cesEnabled :: Lens' CreateEventSubscription (Maybe Bool)

A Boolean value; set to true to activate the subscription, set to false to create the subscription but not active it.

cesEventCategories :: Lens' CreateEventSubscription [Text]

A list of event categories for a SourceType that you want to subscribe to. You can see a list of the categories for a given SourceType in the Events topic in the Amazon RDS User Guide or by using the DescribeEventCategories action.

cesSnsTopicArn :: Lens' CreateEventSubscription Text

The Amazon Resource Name (ARN) of the SNS topic created for event notification. The ARN is created by Amazon SNS when you create a topic and subscribe to it.

cesSourceIds :: Lens' CreateEventSubscription [Text]

The list of identifiers of the event sources for which events will be returned. If not specified, then all sources are included in the response. An identifier must begin with a letter and must contain only ASCII letters, digits, and hyphens; it cannot end with a hyphen or contain two consecutive hyphens.

Constraints:

If SourceIds are supplied, SourceType must also be provided. If the source type is a DB instance, then a DBInstanceIdentifier must be supplied. If the source type is a DB security group, a DBSecurityGroupName must be supplied. If the source type is a DB parameter group, a DBParameterGroupName must be supplied. If the source type is a DB snapshot, a DBSnapshotIdentifier must be supplied.

cesSourceType :: Lens' CreateEventSubscription (Maybe Text)

The type of source that will be generating the events. For example, if you want to be notified of events generated by a DB instance, you would set this parameter to db-instance. if this value is not specified, all events are returned.

Valid values: db-instance | db-parameter-group | db-security-group | db-snapshot

cesSubscriptionName :: Lens' CreateEventSubscription Text

The name of the subscription.

Constraints: The name must be less than 255 characters.

cesTags :: Lens' CreateEventSubscription [Tag]

Response

data CreateEventSubscriptionResponse

Response constructor

createEventSubscriptionResponse :: CreateEventSubscriptionResponse

CreateEventSubscriptionResponse constructor.

The fields accessible through corresponding lenses are:

cesrEventSubscription :: Maybe EventSubscription

Response lenses

cesrEventSubscription :: Lens' CreateEventSubscriptionResponse (Maybe EventSubscription)