Network.AWS.RDS.CreateOptionGroup

Description

Creates a new option group. You can create up to 20 option groups. https:/rds.amazonaws.com ?Action=CreateOptionGroup &OptionGroupName=myoptiongroup &EngineName=oracle-se1 &MajorEngineVersion=11.2 &OptionGroupDescription=Test option group 11.2 myoptiongroup oracle-se1 Test option group b2416a8a-84c9-11e1-a264-0b23c28bc344.

Request

data CreateOptionGroup

Request constructor

mkCreateOptionGroup

Arguments

:: Text
cogOptionGroupName

-> Text
cogEngineName

-> Text
cogMajorEngineVersion

-> Text
cogOptionGroupDescription

-> CreateOptionGroup

Smart constructor for the minimum required parameters to construct a valid CreateOptionGroup request.

The fields accessible through corresponding lenses are:

OptionGroupName :: Text

EngineName :: Text

MajorEngineVersion :: Text

OptionGroupDescription :: Text

Tags :: [Tag]

Request lenses

cogOptionGroupName :: Lens' CreateOptionGroup Text

Specifies the name of the option group to be created. Constraints: Must be 1 to 255 alphanumeric characters or hyphens First character must be a letter Cannot end with a hyphen or contain two consecutive hyphens Example: myoptiongroup.

cogEngineName :: Lens' CreateOptionGroup Text

Specifies the name of the engine that this option group should be associated with.

cogMajorEngineVersion :: Lens' CreateOptionGroup Text

Specifies the major version of the engine that this option group should be associated with.

cogOptionGroupDescription :: Lens' CreateOptionGroup Text

The description of the option group.

cogTags :: Lens' CreateOptionGroup [Tag]

A list of tags.

Response

data CreateOptionGroupResponse

Response constructor

mkCreateOptionGroupResponse :: CreateOptionGroupResponse

Smart constructor for the minimum required parameters to construct a valid CreateOptionGroupResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

OptionGroup :: Maybe OptionGroup

Response lenses

cogrOptionGroup :: Lens' CreateOptionGroupResponse (Maybe OptionGroup)