Network.AWS.RDS.DeleteDBCluster

Description

The DeleteDBCluster action deletes a previously provisioned DB cluster. When you delete a DB cluster, all automated backups for that DB cluster are deleted and cannot be recovered. Manual DB cluster snapshots of the specified DB cluster are not deleted.

For more information on Amazon Aurora, see Aurora on Amazon RDS in the Amazon RDS User Guide.

Creating a Request

deleteDBCluster

Arguments

:: Text
ddbcDBClusterIdentifier

-> DeleteDBCluster

Creates a value of DeleteDBCluster with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddbcFinalDBSnapshotIdentifier - The DB cluster snapshot identifier of the new DB cluster snapshot created when SkipFinalSnapshot is set to false . Constraints: * Must be 1 to 255 alphanumeric characters * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens

ddbcSkipFinalSnapshot - Determines whether a final DB cluster snapshot is created before the DB cluster is deleted. If true is specified, no DB cluster snapshot is created. If false is specified, a DB cluster snapshot is created before the DB cluster is deleted. Default: false

ddbcDBClusterIdentifier - The DB cluster identifier for the DB cluster to be deleted. This parameter isn't case-sensitive. Constraints: * Must contain from 1 to 63 alphanumeric characters or hyphens * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens

data DeleteDBCluster

See: deleteDBCluster smart constructor.

Request Lenses

ddbcFinalDBSnapshotIdentifier :: Lens' DeleteDBCluster (Maybe Text)

The DB cluster snapshot identifier of the new DB cluster snapshot created when SkipFinalSnapshot is set to false . Constraints: * Must be 1 to 255 alphanumeric characters * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens

ddbcSkipFinalSnapshot :: Lens' DeleteDBCluster (Maybe Bool)

Determines whether a final DB cluster snapshot is created before the DB cluster is deleted. If true is specified, no DB cluster snapshot is created. If false is specified, a DB cluster snapshot is created before the DB cluster is deleted. Default: false

ddbcDBClusterIdentifier :: Lens' DeleteDBCluster Text

The DB cluster identifier for the DB cluster to be deleted. This parameter isn't case-sensitive. Constraints: * Must contain from 1 to 63 alphanumeric characters or hyphens * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens

Destructuring the Response

deleteDBClusterResponse

Arguments

:: Int
ddbcrsResponseStatus

-> DeleteDBClusterResponse

Creates a value of DeleteDBClusterResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddbcrsDBCluster - Undocumented member.

ddbcrsResponseStatus - -- | The response status code.

data DeleteDBClusterResponse

See: deleteDBClusterResponse smart constructor.

Response Lenses

ddbcrsDBCluster :: Lens' DeleteDBClusterResponse (Maybe DBCluster)

Undocumented member.

ddbcrsResponseStatus :: Lens' DeleteDBClusterResponse Int

- | The response status code.