Network.AWS.RDS.DeleteDBClusterSnapshot

Description

Deletes a DB cluster snapshot. If the snapshot is being copied, the copy operation is terminated.

For more information on Amazon Aurora, see Aurora on Amazon RDS in the Amazon RDS User Guide.

Creating a Request

deleteDBClusterSnapshot

Arguments

:: Text
ddcsDBClusterSnapshotIdentifier

-> DeleteDBClusterSnapshot

Creates a value of DeleteDBClusterSnapshot with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddcsDBClusterSnapshotIdentifier - The identifier of the DB cluster snapshot to delete. Constraints: Must be the name of an existing DB cluster snapshot in the available state.

data DeleteDBClusterSnapshot

See: deleteDBClusterSnapshot smart constructor.

Request Lenses

ddcsDBClusterSnapshotIdentifier :: Lens' DeleteDBClusterSnapshot Text

The identifier of the DB cluster snapshot to delete. Constraints: Must be the name of an existing DB cluster snapshot in the available state.

Destructuring the Response

deleteDBClusterSnapshotResponse

Arguments

:: Int
ddcsrsResponseStatus

-> DeleteDBClusterSnapshotResponse

Creates a value of DeleteDBClusterSnapshotResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddcsrsDBClusterSnapshot - Undocumented member.

ddcsrsResponseStatus - -- | The response status code.

data DeleteDBClusterSnapshotResponse

See: deleteDBClusterSnapshotResponse smart constructor.

Response Lenses

ddcsrsDBClusterSnapshot :: Lens' DeleteDBClusterSnapshotResponse (Maybe DBClusterSnapshot)

Undocumented member.

ddcsrsResponseStatus :: Lens' DeleteDBClusterSnapshotResponse Int

- | The response status code.