Network.AWS.RDS.DeleteDBInstance

Description

The DeleteDBInstance action deletes a previously provisioned DB instance. A successful response from the web service indicates the request was received correctly. When you delete a DB instance, all automated backups for that instance are deleted and cannot be recovered. Manual DB snapshots of the DB instance to be deleted are not deleted.

If a final DB snapshot is requested the status of the RDS instance will be "deleting" until the DB snapshot is created. The API action DescribeDBInstance is used to monitor the status of this operation. The action cannot be canceled or reverted once submitted.

http://docs.aws.amazon.com/AmazonRDS/latest/APIReference/API_DeleteDBInstance.html

Request

data DeleteDBInstance

Request constructor

deleteDBInstance

Arguments

:: Text
ddbiDBInstanceIdentifier

-> DeleteDBInstance

DeleteDBInstance constructor.

The fields accessible through corresponding lenses are:

ddbiDBInstanceIdentifier :: Text

ddbiFinalDBSnapshotIdentifier :: Maybe Text

ddbiSkipFinalSnapshot :: Maybe Bool

Request lenses

ddbiDBInstanceIdentifier :: Lens' DeleteDBInstance Text

The DB instance identifier for the DB instance to be deleted. This parameter isn't case sensitive.

Constraints:

Must contain from 1 to 63 alphanumeric characters or hyphens First character must be a letter Cannot end with a hyphen or contain two consecutive hyphens

ddbiFinalDBSnapshotIdentifier :: Lens' DeleteDBInstance (Maybe Text)

The DBSnapshotIdentifier of the new DBSnapshot created when SkipFinalSnapshot is set to false.

Specifying this parameter and also setting the SkipFinalShapshot parameter to true results in an error. Constraints:

Must be 1 to 255 alphanumeric characters First character must be a letter Cannot end with a hyphen or contain two consecutive hyphens Cannot be specified when deleting a Read Replica.

ddbiSkipFinalSnapshot :: Lens' DeleteDBInstance (Maybe Bool)

Determines whether a final DB snapshot is created before the DB instance is deleted. If true is specified, no DBSnapshot is created. If false is specified, a DB snapshot is created before the DB instance is deleted.

Specify true when deleting a Read Replica.

The FinalDBSnapshotIdentifier parameter must be specified if SkipFinalSnapshot is false. Default: false

Response

data DeleteDBInstanceResponse

Response constructor

deleteDBInstanceResponse :: DeleteDBInstanceResponse

DeleteDBInstanceResponse constructor.

The fields accessible through corresponding lenses are:

ddbirDBInstance :: Maybe DBInstance

Response lenses

ddbirDBInstance :: Lens' DeleteDBInstanceResponse (Maybe DBInstance)