Network.AWS.RDS.DeleteDBParameterGroup

Description

Deletes a specified DBParameterGroup. The DBParameterGroup to be deleted cannot be associated with any DB instances.

http://docs.aws.amazon.com/AmazonRDS/latest/APIReference/API_DeleteDBParameterGroup.html

Request

data DeleteDBParameterGroup

Request constructor

deleteDBParameterGroup

Arguments

:: Text
ddbpg1DBParameterGroupName

-> DeleteDBParameterGroup

DeleteDBParameterGroup constructor.

The fields accessible through corresponding lenses are:

ddbpg1DBParameterGroupName :: Text

Request lenses

ddbpg1DBParameterGroupName :: Lens' DeleteDBParameterGroup Text

The name of the DB parameter group.

Constraints:

Must be the name of an existing DB parameter group You cannot delete a default DB parameter group Cannot be associated with any DB instances

Response

data DeleteDBParameterGroupResponse

Response constructor

deleteDBParameterGroupResponse :: DeleteDBParameterGroupResponse

DeleteDBParameterGroupResponse constructor.