Network.AWS.RDS.DeleteDBSecurityGroup

Description

Deletes a DB security group.

The specified DB security group must not be associated with any DB instances.

http://docs.aws.amazon.com/AmazonRDS/latest/APIReference/API_DeleteDBSecurityGroup.html

Request

data DeleteDBSecurityGroup

Request constructor

deleteDBSecurityGroup

Arguments

:: Text
ddbsgDBSecurityGroupName

-> DeleteDBSecurityGroup

DeleteDBSecurityGroup constructor.

The fields accessible through corresponding lenses are:

ddbsgDBSecurityGroupName :: Text

Request lenses

ddbsgDBSecurityGroupName :: Lens' DeleteDBSecurityGroup Text

The name of the DB security group to delete.

You cannot delete the default DB security group. Constraints:

Must be 1 to 255 alphanumeric characters First character must be a letter Cannot end with a hyphen or contain two consecutive hyphens Must not be Default May not contain spaces

Response

data DeleteDBSecurityGroupResponse

Response constructor

deleteDBSecurityGroupResponse :: DeleteDBSecurityGroupResponse

DeleteDBSecurityGroupResponse constructor.