Network.AWS.RDS.DeleteDBSnapshot

Description

Deletes a DBSnapshot. If the snapshot is being copied, the copy operation is terminated.

The DBSnapshot must be in the available state to be deleted.

http://docs.aws.amazon.com/AmazonRDS/latest/APIReference/API_DeleteDBSnapshot.html

Request

data DeleteDBSnapshot

Request constructor

deleteDBSnapshot

Arguments

:: Text
ddbs1DBSnapshotIdentifier

-> DeleteDBSnapshot

DeleteDBSnapshot constructor.

The fields accessible through corresponding lenses are:

ddbs1DBSnapshotIdentifier :: Text

Request lenses

ddbs1DBSnapshotIdentifier :: Lens' DeleteDBSnapshot Text

The DBSnapshot identifier.

Constraints: Must be the name of an existing DB snapshot in the available state.

Response

data DeleteDBSnapshotResponse

Response constructor

deleteDBSnapshotResponse :: DeleteDBSnapshotResponse

DeleteDBSnapshotResponse constructor.

The fields accessible through corresponding lenses are:

ddbsrDBSnapshot :: Maybe DBSnapshot

Response lenses

ddbsrDBSnapshot :: Lens' DeleteDBSnapshotResponse (Maybe DBSnapshot)