Network.AWS.RDS.DescribeAccountAttributes

Description

Lists all of the attributes for a customer account. The attributes include Amazon RDS quotas for the account, such as the number of DB instances allowed. The description for a quota includes the quota name, current usage toward that quota, and the quota's maximum value.

This command does not take any parameters.

http://docs.aws.amazon.com/AmazonRDS/latest/APIReference/API_DescribeAccountAttributes.html

Request

data DescribeAccountAttributes

Request constructor

describeAccountAttributes :: DescribeAccountAttributes

DescribeAccountAttributes constructor.

Response

data DescribeAccountAttributesResponse

Response constructor

describeAccountAttributesResponse :: DescribeAccountAttributesResponse

DescribeAccountAttributesResponse constructor.

The fields accessible through corresponding lenses are:

daarAccountQuotas :: [AccountQuota]

Response lenses

daarAccountQuotas :: Lens' DescribeAccountAttributesResponse [AccountQuota]

A list of AccountQuota objects. Within this list, each quota has a name, a count of usage toward the quota maximum, and a maximum value for the quota.