Network.AWS.RDS.DescribeCertificates

Description

Lists the set of CA certificates provided by Amazon RDS for this AWS account.

http://docs.aws.amazon.com/AmazonRDS/latest/APIReference/API_DescribeCertificates.html

Request

data DescribeCertificates

Request constructor

describeCertificates :: DescribeCertificates

DescribeCertificates constructor.

The fields accessible through corresponding lenses are:

dcCertificateIdentifier :: Maybe Text

dcFilters :: [Filter]

dcMarker :: Maybe Text

dcMaxRecords :: Maybe Int

Request lenses

dcCertificateIdentifier :: Lens' DescribeCertificates (Maybe Text)

The user-supplied certificate identifier. If this parameter is specified, information for only the identified certificate is returned. This parameter isn't case-sensitive.

Constraints:

Must contain from 1 to 63 alphanumeric characters or hyphens First character must be a letter Cannot end with a hyphen or contain two consecutive hyphens

dcFilters :: Lens' DescribeCertificates [Filter]

This parameter is not currently supported.

dcMarker :: Lens' DescribeCertificates (Maybe Text)

An optional pagination token provided by a previous DescribeCertificates request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

dcMaxRecords :: Lens' DescribeCertificates (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100

Response

data DescribeCertificatesResponse

Response constructor

describeCertificatesResponse :: DescribeCertificatesResponse

DescribeCertificatesResponse constructor.

The fields accessible through corresponding lenses are:

dcrCertificates :: [Certificate]

dcrMarker :: Maybe Text

Response lenses

dcrCertificates :: Lens' DescribeCertificatesResponse [Certificate]

The list of Certificate objects for the AWS account.

dcrMarker :: Lens' DescribeCertificatesResponse (Maybe Text)

An optional pagination token provided by a previous DescribeCertificates request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .