Network.AWS.RDS.DescribeDBClusterParameterGroups

Description

Returns a list of DBClusterParameterGroup descriptions. If a DBClusterParameterGroupName parameter is specified, the list will contain only the description of the specified DB cluster parameter group.

For more information on Amazon Aurora, see Aurora on Amazon RDS in the Amazon RDS User Guide.

Creating a Request

describeDBClusterParameterGroups :: DescribeDBClusterParameterGroups

Creates a value of DescribeDBClusterParameterGroups with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddcpgFilters - This parameter is not currently supported.

ddcpgMarker - An optional pagination token provided by a previous DescribeDBClusterParameterGroups request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

ddcpgMaxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: Minimum 20, maximum 100.

ddcpgDBClusterParameterGroupName - The name of a specific DB cluster parameter group to return details for. Constraints: * Must be 1 to 255 alphanumeric characters * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens

data DescribeDBClusterParameterGroups

See: describeDBClusterParameterGroups smart constructor.

Request Lenses

ddcpgFilters :: Lens' DescribeDBClusterParameterGroups [Filter]

This parameter is not currently supported.

ddcpgMarker :: Lens' DescribeDBClusterParameterGroups (Maybe Text)

An optional pagination token provided by a previous DescribeDBClusterParameterGroups request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

ddcpgMaxRecords :: Lens' DescribeDBClusterParameterGroups (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: Minimum 20, maximum 100.

ddcpgDBClusterParameterGroupName :: Lens' DescribeDBClusterParameterGroups (Maybe Text)

The name of a specific DB cluster parameter group to return details for. Constraints: * Must be 1 to 255 alphanumeric characters * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens

Destructuring the Response

describeDBClusterParameterGroupsResponse

Arguments

:: Int
ddcpgrsResponseStatus

-> DescribeDBClusterParameterGroupsResponse

Creates a value of DescribeDBClusterParameterGroupsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddcpgrsMarker - An optional pagination token provided by a previous DescribeDBClusterParameterGroups request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

ddcpgrsDBClusterParameterGroups - A list of DB cluster parameter groups.

ddcpgrsResponseStatus - -- | The response status code.

data DescribeDBClusterParameterGroupsResponse

See: describeDBClusterParameterGroupsResponse smart constructor.

Response Lenses

ddcpgrsMarker :: Lens' DescribeDBClusterParameterGroupsResponse (Maybe Text)

An optional pagination token provided by a previous DescribeDBClusterParameterGroups request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

ddcpgrsDBClusterParameterGroups :: Lens' DescribeDBClusterParameterGroupsResponse [DBClusterParameterGroup]

A list of DB cluster parameter groups.

ddcpgrsResponseStatus :: Lens' DescribeDBClusterParameterGroupsResponse Int

- | The response status code.