Network.AWS.RDS.DescribeDBClusterParameters

Description

Returns the detailed parameter list for a particular DB cluster parameter group.

For more information on Amazon Aurora, see Aurora on Amazon RDS in the Amazon RDS User Guide.

Creating a Request

describeDBClusterParameters

Arguments

:: Text
ddcpDBClusterParameterGroupName

-> DescribeDBClusterParameters

Creates a value of DescribeDBClusterParameters with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddcpFilters - This parameter is not currently supported.

ddcpMarker - An optional pagination token provided by a previous DescribeDBClusterParameters request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

ddcpMaxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: Minimum 20, maximum 100.

ddcpSource - A value that indicates to return only parameters for a specific source. Parameter sources can be engine , service , or customer .

ddcpDBClusterParameterGroupName - The name of a specific DB cluster parameter group to return parameter details for. Constraints: * Must be 1 to 255 alphanumeric characters * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens

data DescribeDBClusterParameters

See: describeDBClusterParameters smart constructor.

Request Lenses

ddcpFilters :: Lens' DescribeDBClusterParameters [Filter]

This parameter is not currently supported.

ddcpMarker :: Lens' DescribeDBClusterParameters (Maybe Text)

An optional pagination token provided by a previous DescribeDBClusterParameters request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

ddcpMaxRecords :: Lens' DescribeDBClusterParameters (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: Minimum 20, maximum 100.

ddcpSource :: Lens' DescribeDBClusterParameters (Maybe Text)

A value that indicates to return only parameters for a specific source. Parameter sources can be engine , service , or customer .

ddcpDBClusterParameterGroupName :: Lens' DescribeDBClusterParameters Text

The name of a specific DB cluster parameter group to return parameter details for. Constraints: * Must be 1 to 255 alphanumeric characters * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens

Destructuring the Response

describeDBClusterParametersResponse

Arguments

:: Int
ddcprsResponseStatus

-> DescribeDBClusterParametersResponse

Creates a value of DescribeDBClusterParametersResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddcprsMarker - An optional pagination token provided by a previous DescribeDBClusterParameters request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

ddcprsParameters - Provides a list of parameters for the DB cluster parameter group.

ddcprsResponseStatus - -- | The response status code.

data DescribeDBClusterParametersResponse

Provides details about a DB cluster parameter group including the parameters in the DB cluster parameter group.

See: describeDBClusterParametersResponse smart constructor.

Response Lenses

ddcprsMarker :: Lens' DescribeDBClusterParametersResponse (Maybe Text)

An optional pagination token provided by a previous DescribeDBClusterParameters request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

ddcprsParameters :: Lens' DescribeDBClusterParametersResponse [Parameter]

Provides a list of parameters for the DB cluster parameter group.

ddcprsResponseStatus :: Lens' DescribeDBClusterParametersResponse Int

- | The response status code.