Network.AWS.RDS.DescribeDBClusterSnapshotAttributes

Description

Returns a list of DB cluster snapshot attribute names and values for a manual DB cluster snapshot.

When sharing snapshots with other AWS accounts, DescribeDBClusterSnapshotAttributes returns the restore attribute and a list of IDs for the AWS accounts that are authorized to copy or restore the manual DB cluster snapshot. If all is included in the list of values for the restore attribute, then the manual DB cluster snapshot is public and can be copied or restored by all AWS accounts.

To add or remove access for an AWS account to copy or restore a manual DB cluster snapshot, or to make the manual DB cluster snapshot public or private, use the ModifyDBClusterSnapshotAttribute API action.

Creating a Request

describeDBClusterSnapshotAttributes

Arguments

:: Text
ddcsaDBClusterSnapshotIdentifier

-> DescribeDBClusterSnapshotAttributes

Creates a value of DescribeDBClusterSnapshotAttributes with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddcsaDBClusterSnapshotIdentifier - The identifier for the DB cluster snapshot to describe the attributes for.

data DescribeDBClusterSnapshotAttributes

See: describeDBClusterSnapshotAttributes smart constructor.

Request Lenses

ddcsaDBClusterSnapshotIdentifier :: Lens' DescribeDBClusterSnapshotAttributes Text

The identifier for the DB cluster snapshot to describe the attributes for.

Destructuring the Response

describeDBClusterSnapshotAttributesResponse

Arguments

:: Int
ddcsarsResponseStatus

-> DescribeDBClusterSnapshotAttributesResponse

Creates a value of DescribeDBClusterSnapshotAttributesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddcsarsDBClusterSnapshotAttributesResult - Undocumented member.

ddcsarsResponseStatus - -- | The response status code.

data DescribeDBClusterSnapshotAttributesResponse

See: describeDBClusterSnapshotAttributesResponse smart constructor.

Response Lenses

ddcsarsDBClusterSnapshotAttributesResult :: Lens' DescribeDBClusterSnapshotAttributesResponse (Maybe DBClusterSnapshotAttributesResult)

Undocumented member.

ddcsarsResponseStatus :: Lens' DescribeDBClusterSnapshotAttributesResponse Int

- | The response status code.