Network.AWS.RDS.DescribeDBClusterSnapshots

Description

Returns information about DB cluster snapshots. This API action supports pagination.

For more information on Amazon Aurora, see Aurora on Amazon RDS in the Amazon RDS User Guide.

This operation returns paginated results.

Creating a Request

describeDBClusterSnapshots :: DescribeDBClusterSnapshots

Creates a value of DescribeDBClusterSnapshots with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddbcsDBClusterIdentifier - The ID of the DB cluster to retrieve the list of DB cluster snapshots for. This parameter cannot be used in conjunction with the DBClusterSnapshotIdentifier parameter. This parameter is not case-sensitive. Constraints: * Must contain from 1 to 63 alphanumeric characters or hyphens * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens

ddbcsIncludeShared - Set this value to true to include shared manual DB cluster snapshots from other AWS accounts that this AWS account has been given permission to copy or restore, otherwise set this value to false . The default is false . You can give an AWS account permission to restore a manual DB cluster snapshot from another AWS account by the ModifyDBClusterSnapshotAttribute API action.

ddbcsDBClusterSnapshotIdentifier - A specific DB cluster snapshot identifier to describe. This parameter cannot be used in conjunction with the DBClusterIdentifier parameter. This value is stored as a lowercase string. Constraints: * Must be 1 to 255 alphanumeric characters * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens * If this identifier is for an automated snapshot, the SnapshotType parameter must also be specified.

ddbcsFilters - This parameter is not currently supported.

ddbcsSnapshotType - The type of DB cluster snapshots to be returned. You can specify one of the following values: * automated - Return all DB cluster snapshots that have been automatically taken by Amazon RDS for my AWS account. * manual - Return all DB cluster snapshots that have been taken by my AWS account. * shared - Return all manual DB cluster snapshots that have been shared to my AWS account. * public - Return all DB cluster snapshots that have been marked as public. If you don't specify a SnapshotType value, then both automated and manual DB cluster snapshots are returned. You can include shared DB cluster snapshots with these results by setting the IncludeShared parameter to true . You can include public DB cluster snapshots with these results by setting the IncludePublic parameter to true . The IncludeShared and IncludePublic parameters don't apply for SnapshotType values of manual or automated . The IncludePublic parameter doesn't apply when SnapshotType is set to shared . The IncludeShared parameter doesn't apply when SnapshotType is set to public .

ddbcsMarker - An optional pagination token provided by a previous DescribeDBClusterSnapshots request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

ddbcsMaxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: Minimum 20, maximum 100.

ddbcsIncludePublic - Set this value to true to include manual DB cluster snapshots that are public and can be copied or restored by any AWS account, otherwise set this value to false . The default is false . The default is false. You can share a manual DB cluster snapshot as public by using the ModifyDBClusterSnapshotAttribute API action.

data DescribeDBClusterSnapshots

See: describeDBClusterSnapshots smart constructor.

Request Lenses

ddbcsDBClusterIdentifier :: Lens' DescribeDBClusterSnapshots (Maybe Text)

The ID of the DB cluster to retrieve the list of DB cluster snapshots for. This parameter cannot be used in conjunction with the DBClusterSnapshotIdentifier parameter. This parameter is not case-sensitive. Constraints: * Must contain from 1 to 63 alphanumeric characters or hyphens * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens

ddbcsIncludeShared :: Lens' DescribeDBClusterSnapshots (Maybe Bool)

Set this value to true to include shared manual DB cluster snapshots from other AWS accounts that this AWS account has been given permission to copy or restore, otherwise set this value to false . The default is false . You can give an AWS account permission to restore a manual DB cluster snapshot from another AWS account by the ModifyDBClusterSnapshotAttribute API action.

ddbcsDBClusterSnapshotIdentifier :: Lens' DescribeDBClusterSnapshots (Maybe Text)

A specific DB cluster snapshot identifier to describe. This parameter cannot be used in conjunction with the DBClusterIdentifier parameter. This value is stored as a lowercase string. Constraints: * Must be 1 to 255 alphanumeric characters * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens * If this identifier is for an automated snapshot, the SnapshotType parameter must also be specified.

ddbcsFilters :: Lens' DescribeDBClusterSnapshots [Filter]

This parameter is not currently supported.

ddbcsSnapshotType :: Lens' DescribeDBClusterSnapshots (Maybe Text)

The type of DB cluster snapshots to be returned. You can specify one of the following values: * automated - Return all DB cluster snapshots that have been automatically taken by Amazon RDS for my AWS account. * manual - Return all DB cluster snapshots that have been taken by my AWS account. * shared - Return all manual DB cluster snapshots that have been shared to my AWS account. * public - Return all DB cluster snapshots that have been marked as public. If you don't specify a SnapshotType value, then both automated and manual DB cluster snapshots are returned. You can include shared DB cluster snapshots with these results by setting the IncludeShared parameter to true . You can include public DB cluster snapshots with these results by setting the IncludePublic parameter to true . The IncludeShared and IncludePublic parameters don't apply for SnapshotType values of manual or automated . The IncludePublic parameter doesn't apply when SnapshotType is set to shared . The IncludeShared parameter doesn't apply when SnapshotType is set to public .

ddbcsMarker :: Lens' DescribeDBClusterSnapshots (Maybe Text)

An optional pagination token provided by a previous DescribeDBClusterSnapshots request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

ddbcsMaxRecords :: Lens' DescribeDBClusterSnapshots (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: Minimum 20, maximum 100.

ddbcsIncludePublic :: Lens' DescribeDBClusterSnapshots (Maybe Bool)

Set this value to true to include manual DB cluster snapshots that are public and can be copied or restored by any AWS account, otherwise set this value to false . The default is false . The default is false. You can share a manual DB cluster snapshot as public by using the ModifyDBClusterSnapshotAttribute API action.

Destructuring the Response

describeDBClusterSnapshotsResponse

Arguments

:: Int
ddbcsrsResponseStatus

-> DescribeDBClusterSnapshotsResponse

Creates a value of DescribeDBClusterSnapshotsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddbcsrsMarker - An optional pagination token provided by a previous DescribeDBClusterSnapshots request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

ddbcsrsDBClusterSnapshots - Provides a list of DB cluster snapshots for the user.

ddbcsrsResponseStatus - -- | The response status code.

data DescribeDBClusterSnapshotsResponse

Provides a list of DB cluster snapshots for the user as the result of a call to the DescribeDBClusterSnapshots action.

See: describeDBClusterSnapshotsResponse smart constructor.

Response Lenses

ddbcsrsMarker :: Lens' DescribeDBClusterSnapshotsResponse (Maybe Text)

An optional pagination token provided by a previous DescribeDBClusterSnapshots request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

ddbcsrsDBClusterSnapshots :: Lens' DescribeDBClusterSnapshotsResponse [DBClusterSnapshot]

Provides a list of DB cluster snapshots for the user.

ddbcsrsResponseStatus :: Lens' DescribeDBClusterSnapshotsResponse Int

- | The response status code.