Network.AWS.RDS.DescribeDBClusters

Description

Returns information about provisioned Aurora DB clusters. This API supports pagination.

For more information on Amazon Aurora, see Aurora on Amazon RDS in the Amazon RDS User Guide.

Creating a Request

describeDBClusters :: DescribeDBClusters

Creates a value of DescribeDBClusters with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddcDBClusterIdentifier - The user-supplied DB cluster identifier. If this parameter is specified, information from only the specific DB cluster is returned. This parameter isn't case-sensitive. Constraints: * Must contain from 1 to 63 alphanumeric characters or hyphens * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens

ddcFilters - A filter that specifies one or more DB clusters to describe. Supported filters: * db-cluster-id - Accepts DB cluster identifiers and DB cluster Amazon Resource Names (ARNs). The results list will only include information about the DB clusters identified by these ARNs.

ddcMarker - An optional pagination token provided by a previous DescribeDBClusters request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

ddcMaxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: Minimum 20, maximum 100.

data DescribeDBClusters

See: describeDBClusters smart constructor.

Request Lenses

ddcDBClusterIdentifier :: Lens' DescribeDBClusters (Maybe Text)

The user-supplied DB cluster identifier. If this parameter is specified, information from only the specific DB cluster is returned. This parameter isn't case-sensitive. Constraints: * Must contain from 1 to 63 alphanumeric characters or hyphens * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens

ddcFilters :: Lens' DescribeDBClusters [Filter]

A filter that specifies one or more DB clusters to describe. Supported filters: * db-cluster-id - Accepts DB cluster identifiers and DB cluster Amazon Resource Names (ARNs). The results list will only include information about the DB clusters identified by these ARNs.

ddcMarker :: Lens' DescribeDBClusters (Maybe Text)

An optional pagination token provided by a previous DescribeDBClusters request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

ddcMaxRecords :: Lens' DescribeDBClusters (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: Minimum 20, maximum 100.

Destructuring the Response

describeDBClustersResponse

Arguments

:: Int
ddcrsResponseStatus

-> DescribeDBClustersResponse

Creates a value of DescribeDBClustersResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddcrsDBClusters - Contains a list of DB clusters for the user.

ddcrsMarker - A pagination token that can be used in a subsequent DescribeDBClusters request.

ddcrsResponseStatus - -- | The response status code.

data DescribeDBClustersResponse

Contains the result of a successful invocation of the DescribeDBClusters action.

See: describeDBClustersResponse smart constructor.

Response Lenses

ddcrsDBClusters :: Lens' DescribeDBClustersResponse [DBCluster]

Contains a list of DB clusters for the user.

ddcrsMarker :: Lens' DescribeDBClustersResponse (Maybe Text)

A pagination token that can be used in a subsequent DescribeDBClusters request.

ddcrsResponseStatus :: Lens' DescribeDBClustersResponse Int

- | The response status code.