Network.AWS.RDS.DescribeDBEngineVersions

Description

Returns a list of the available DB engines.

http://docs.aws.amazon.com/AmazonRDS/latest/APIReference/API_DescribeDBEngineVersions.html

Request

data DescribeDBEngineVersions

Request constructor

describeDBEngineVersions :: DescribeDBEngineVersions

DescribeDBEngineVersions constructor.

The fields accessible through corresponding lenses are:

ddbevDBParameterGroupFamily :: Maybe Text

ddbevDefaultOnly :: Maybe Bool

ddbevEngine :: Maybe Text

ddbevEngineVersion :: Maybe Text

ddbevFilters :: [Filter]

ddbevListSupportedCharacterSets :: Maybe Bool

ddbevMarker :: Maybe Text

ddbevMaxRecords :: Maybe Int

Request lenses

ddbevDBParameterGroupFamily :: Lens' DescribeDBEngineVersions (Maybe Text)

The name of a specific DB parameter group family to return details for.

Constraints:

Must be 1 to 255 alphanumeric characters First character must be a letter Cannot end with a hyphen or contain two consecutive hyphens

ddbevDefaultOnly :: Lens' DescribeDBEngineVersions (Maybe Bool)

Indicates that only the default version of the specified engine or engine and major version combination is returned.

ddbevEngine :: Lens' DescribeDBEngineVersions (Maybe Text)

The database engine to return.

ddbevEngineVersion :: Lens' DescribeDBEngineVersions (Maybe Text)

The database engine version to return.

Example: '5.1.49'

ddbevFilters :: Lens' DescribeDBEngineVersions [Filter]

Not currently supported.

ddbevListSupportedCharacterSets :: Lens' DescribeDBEngineVersions (Maybe Bool)

If this parameter is specified, and if the requested engine supports the CharacterSetName parameter for CreateDBInstance, the response includes a list of supported character sets for each engine version.

ddbevMarker :: Lens' DescribeDBEngineVersions (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

ddbevMaxRecords :: Lens' DescribeDBEngineVersions (Maybe Int)

The maximum number of records to include in the response. If more than the MaxRecords value is available, a pagination token called a marker is included in the response so that the following results can be retrieved.

Default: 100

Constraints: minimum 20, maximum 100

Response

data DescribeDBEngineVersionsResponse

Response constructor

describeDBEngineVersionsResponse :: DescribeDBEngineVersionsResponse

DescribeDBEngineVersionsResponse constructor.

The fields accessible through corresponding lenses are:

ddbevrDBEngineVersions :: [DBEngineVersion]

ddbevrMarker :: Maybe Text

Response lenses

ddbevrDBEngineVersions :: Lens' DescribeDBEngineVersionsResponse [DBEngineVersion]

A list of DBEngineVersion elements.

ddbevrMarker :: Lens' DescribeDBEngineVersionsResponse (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.