Network.AWS.RDS.DescribeDBInstances

Description

Returns information about provisioned RDS instances. This API supports pagination.

http://docs.aws.amazon.com/AmazonRDS/latest/APIReference/API_DescribeDBInstances.html

Request

data DescribeDBInstances

Request constructor

describeDBInstances :: DescribeDBInstances

DescribeDBInstances constructor.

The fields accessible through corresponding lenses are:

ddbi1DBInstanceIdentifier :: Maybe Text

ddbi1Filters :: [Filter]

ddbi1Marker :: Maybe Text

ddbi1MaxRecords :: Maybe Int

Request lenses

ddbi1DBInstanceIdentifier :: Lens' DescribeDBInstances (Maybe Text)

The user-supplied instance identifier. If this parameter is specified, information from only the specific DB instance is returned. This parameter isn't case sensitive.

Constraints:

Must contain from 1 to 63 alphanumeric characters or hyphens First character must be a letter Cannot end with a hyphen or contain two consecutive hyphens

ddbi1Filters :: Lens' DescribeDBInstances [Filter]

This parameter is not currently supported.

ddbi1Marker :: Lens' DescribeDBInstances (Maybe Text)

An optional pagination token provided by a previous DescribeDBInstances request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

ddbi1MaxRecords :: Lens' DescribeDBInstances (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results may be retrieved.

Default: 100

Constraints: minimum 20, maximum 100

Response

data DescribeDBInstancesResponse

Response constructor

describeDBInstancesResponse :: DescribeDBInstancesResponse

DescribeDBInstancesResponse constructor.

The fields accessible through corresponding lenses are:

ddbirDBInstances :: [DBInstance]

ddbirMarker :: Maybe Text

Response lenses

ddbirDBInstances :: Lens' DescribeDBInstancesResponse [DBInstance]

A list of DBInstance instances.

ddbirMarker :: Lens' DescribeDBInstancesResponse (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .