Network.AWS.RDS.DescribeDBLogFiles

Description

Returns a list of DB log files for the DB instance.

http://docs.aws.amazon.com/AmazonRDS/latest/APIReference/API_DescribeDBLogFiles.html

Request

data DescribeDBLogFiles

Request constructor

describeDBLogFiles

Arguments

:: Text
ddblfDBInstanceIdentifier

-> DescribeDBLogFiles

DescribeDBLogFiles constructor.

The fields accessible through corresponding lenses are:

ddblfDBInstanceIdentifier :: Text

ddblfFileLastWritten :: Maybe Integer

ddblfFileSize :: Maybe Integer

ddblfFilenameContains :: Maybe Text

ddblfFilters :: [Filter]

ddblfMarker :: Maybe Text

ddblfMaxRecords :: Maybe Int

Request lenses

ddblfDBInstanceIdentifier :: Lens' DescribeDBLogFiles Text

The customer-assigned name of the DB instance that contains the log files you want to list.

Constraints:

Must contain from 1 to 63 alphanumeric characters or hyphens First character must be a letter Cannot end with a hyphen or contain two consecutive hyphens

ddblfFileLastWritten :: Lens' DescribeDBLogFiles (Maybe Integer)

Filters the available log files for files written since the specified date, in POSIX timestamp format.

ddblfFileSize :: Lens' DescribeDBLogFiles (Maybe Integer)

Filters the available log files for files larger than the specified size.

ddblfFilenameContains :: Lens' DescribeDBLogFiles (Maybe Text)

Filters the available log files for log file names that contain the specified string.

ddblfFilters :: Lens' DescribeDBLogFiles [Filter]

This parameter is not currently supported.

ddblfMarker :: Lens' DescribeDBLogFiles (Maybe Text)

The pagination token provided in the previous request. If this parameter is specified the response includes only records beyond the marker, up to MaxRecords.

ddblfMaxRecords :: Lens' DescribeDBLogFiles (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

Response

data DescribeDBLogFilesResponse

Response constructor

describeDBLogFilesResponse :: DescribeDBLogFilesResponse

DescribeDBLogFilesResponse constructor.

The fields accessible through corresponding lenses are:

ddblfrDescribeDBLogFiles :: [DescribeDBLogFilesDetails]

ddblfrMarker :: Maybe Text

Response lenses

ddblfrDescribeDBLogFiles :: Lens' DescribeDBLogFilesResponse [DescribeDBLogFilesDetails]

The DB log files returned.

ddblfrMarker :: Lens' DescribeDBLogFilesResponse (Maybe Text)

A pagination token that can be used in a subsequent DescribeDBLogFiles request.