Network.AWS.RDS.DescribeDBParameterGroups

Description

Returns a list of DBParameterGroup descriptions. If a DBParameterGroupName is specified, the list will contain only the description of the specified DB parameter group.

http://docs.aws.amazon.com/AmazonRDS/latest/APIReference/API_DescribeDBParameterGroups.html

Request

data DescribeDBParameterGroups

Request constructor

describeDBParameterGroups :: DescribeDBParameterGroups

DescribeDBParameterGroups constructor.

The fields accessible through corresponding lenses are:

ddbpgDBParameterGroupName :: Maybe Text

ddbpgFilters :: [Filter]

ddbpgMarker :: Maybe Text

ddbpgMaxRecords :: Maybe Int

Request lenses

ddbpgDBParameterGroupName :: Lens' DescribeDBParameterGroups (Maybe Text)

The name of a specific DB parameter group to return details for.

Constraints:

Must be 1 to 255 alphanumeric characters First character must be a letter Cannot end with a hyphen or contain two consecutive hyphens

ddbpgFilters :: Lens' DescribeDBParameterGroups [Filter]

This parameter is not currently supported.

ddbpgMarker :: Lens' DescribeDBParameterGroups (Maybe Text)

An optional pagination token provided by a previous DescribeDBParameterGroups request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

ddbpgMaxRecords :: Lens' DescribeDBParameterGroups (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results may be retrieved.

Default: 100

Constraints: minimum 20, maximum 100

Response

data DescribeDBParameterGroupsResponse

Response constructor

describeDBParameterGroupsResponse :: DescribeDBParameterGroupsResponse

DescribeDBParameterGroupsResponse constructor.

The fields accessible through corresponding lenses are:

ddbpgrDBParameterGroups :: [DBParameterGroup]

ddbpgrMarker :: Maybe Text

Response lenses

ddbpgrDBParameterGroups :: Lens' DescribeDBParameterGroupsResponse [DBParameterGroup]

A list of DBParameterGroup instances.

ddbpgrMarker :: Lens' DescribeDBParameterGroupsResponse (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.