Network.AWS.RDS.DescribeDBParameters

Description

Returns the detailed parameter list for a particular DB parameter group.

http://docs.aws.amazon.com/AmazonRDS/latest/APIReference/API_DescribeDBParameters.html

Request

data DescribeDBParameters

Request constructor

describeDBParameters

Arguments

:: Text
ddbpDBParameterGroupName

-> DescribeDBParameters

DescribeDBParameters constructor.

The fields accessible through corresponding lenses are:

ddbpDBParameterGroupName :: Text

ddbpFilters :: [Filter]

ddbpMarker :: Maybe Text

ddbpMaxRecords :: Maybe Int

ddbpSource :: Maybe Text

Request lenses

ddbpDBParameterGroupName :: Lens' DescribeDBParameters Text

The name of a specific DB parameter group to return details for.

Constraints:

Must be 1 to 255 alphanumeric characters First character must be a letter Cannot end with a hyphen or contain two consecutive hyphens

ddbpFilters :: Lens' DescribeDBParameters [Filter]

This parameter is not currently supported.

ddbpMarker :: Lens' DescribeDBParameters (Maybe Text)

An optional pagination token provided by a previous DescribeDBParameters request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

ddbpMaxRecords :: Lens' DescribeDBParameters (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results may be retrieved.

Default: 100

Constraints: minimum 20, maximum 100

ddbpSource :: Lens' DescribeDBParameters (Maybe Text)

The parameter types to return.

Default: All parameter types returned

Valid Values: 'user | system | engine-default'

Response

data DescribeDBParametersResponse

Response constructor

describeDBParametersResponse :: DescribeDBParametersResponse

DescribeDBParametersResponse constructor.

The fields accessible through corresponding lenses are:

ddbprMarker :: Maybe Text

ddbprParameters :: [Parameter]

Response lenses

ddbprMarker :: Lens' DescribeDBParametersResponse (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

ddbprParameters :: Lens' DescribeDBParametersResponse [Parameter]

A list of Parameter values.