Network.AWS.RDS.DescribeDBSecurityGroups

Description

Returns a list of DBSecurityGroup descriptions. If a DBSecurityGroupName is specified, the list will contain only the descriptions of the specified DB security group. https:/rds.amazonaws.com ?Action=DescribeDBSecurityGroups &Version=2013-05-15 &MaxRecords=100 &SignatureVersion=2 &SignatureMethod=HmacSHA256 &Timestamp=2011-02-15T19%3A40%3A19.926Z &AWSAccessKeyId= &Signature= authorized myec2securitygroup 054794666394 default 127.0.0.1/30 authorized 621567473609 default vpc-1ab2c3d4 My new DBSecurityGroup 192.168.1.1/24 authorized 621567473609 mydbsecuritygroup vpc-1ab2c3d5 My new DBSecurityGroup 621567473609 mydbsecuritygroup4 vpc-1ab2c3d6 bbdad154-bf42-11de-86a4-97241dfaadff.

Request

data DescribeDBSecurityGroups

Request constructor

describeDBSecurityGroups :: DescribeDBSecurityGroups

Smart constructor for the minimum required parameters to construct a valid DescribeDBSecurityGroups request.

The fields accessible through corresponding lenses are:

DBSecurityGroupName :: Maybe Text

MaxRecords :: Maybe Integer

Marker :: Maybe Text

Request lenses

ddbsg2DBSecurityGroupName :: Lens' DescribeDBSecurityGroups (Maybe Text)

The name of the DB security group to return details for.

ddbsg2MaxRecords :: Lens' DescribeDBSecurityGroups (Maybe Integer)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results may be retrieved. Default: 100 Constraints: minimum 20, maximum 100.

ddbsg2Marker :: Lens' DescribeDBSecurityGroups (Maybe Text)

An optional pagination token provided by a previous DescribeDBSecurityGroups request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

Response

data DescribeDBSecurityGroupsResponse

Contains the result of a successful invocation of the DescribeDBSecurityGroups action.

Response constructor

describeDBSecurityGroupsResponse :: DescribeDBSecurityGroupsResponse

Smart constructor for the minimum required parameters to construct a valid DescribeDBSecurityGroupsResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Marker :: Maybe Text

DBSecurityGroups :: [DBSecurityGroup]

Response lenses

ddbsgrMarker :: Lens' DescribeDBSecurityGroupsResponse (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

ddbsgrDBSecurityGroups :: Lens' DescribeDBSecurityGroupsResponse [DBSecurityGroup]

A list of DBSecurityGroup instances.