Network.AWS.RDS.DescribeDBSnapshotAttributes

Description

Returns a list of DB snapshot attribute names and values for a manual DB snapshot.

When sharing snapshots with other AWS accounts, DescribeDBSnapshotAttributes returns the restore attribute and a list of IDs for the AWS accounts that are authorized to copy or restore the manual DB snapshot. If all is included in the list of values for the restore attribute, then the manual DB snapshot is public and can be copied or restored by all AWS accounts.

To add or remove access for an AWS account to copy or restore a manual DB snapshot, or to make the manual DB snapshot public or private, use the ModifyDBSnapshotAttribute API action.

Creating a Request

describeDBSnapshotAttributes

Arguments

:: Text
ddsaDBSnapshotIdentifier

-> DescribeDBSnapshotAttributes

Creates a value of DescribeDBSnapshotAttributes with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddsaDBSnapshotIdentifier - The identifier for the DB snapshot to describe the attributes for.

data DescribeDBSnapshotAttributes

See: describeDBSnapshotAttributes smart constructor.

Request Lenses

ddsaDBSnapshotIdentifier :: Lens' DescribeDBSnapshotAttributes Text

The identifier for the DB snapshot to describe the attributes for.

Destructuring the Response

describeDBSnapshotAttributesResponse

Arguments

:: Int
ddsarsResponseStatus

-> DescribeDBSnapshotAttributesResponse

Creates a value of DescribeDBSnapshotAttributesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddsarsDBSnapshotAttributesResult - Undocumented member.

ddsarsResponseStatus - -- | The response status code.

data DescribeDBSnapshotAttributesResponse

See: describeDBSnapshotAttributesResponse smart constructor.

Response Lenses

ddsarsDBSnapshotAttributesResult :: Lens' DescribeDBSnapshotAttributesResponse (Maybe DBSnapshotAttributesResult)

Undocumented member.

ddsarsResponseStatus :: Lens' DescribeDBSnapshotAttributesResponse Int

- | The response status code.