Network.AWS.RDS.DescribeDBSnapshots

Description

Returns information about DB snapshots. This API supports pagination.

http://docs.aws.amazon.com/AmazonRDS/latest/APIReference/API_DescribeDBSnapshots.html

Request

data DescribeDBSnapshots

Request constructor

describeDBSnapshots :: DescribeDBSnapshots

DescribeDBSnapshots constructor.

The fields accessible through corresponding lenses are:

ddbsDBInstanceIdentifier :: Maybe Text

ddbsDBSnapshotIdentifier :: Maybe Text

ddbsFilters :: [Filter]

ddbsMarker :: Maybe Text

ddbsMaxRecords :: Maybe Int

ddbsSnapshotType :: Maybe Text

Request lenses

ddbsDBInstanceIdentifier :: Lens' DescribeDBSnapshots (Maybe Text)

A DB instance identifier to retrieve the list of DB snapshots for. Cannot be used in conjunction with DBSnapshotIdentifier. This parameter is not case sensitive.

Constraints:

Must contain from 1 to 63 alphanumeric characters or hyphens First character must be a letter Cannot end with a hyphen or contain two consecutive hyphens

ddbsDBSnapshotIdentifier :: Lens' DescribeDBSnapshots (Maybe Text)

A specific DB snapshot identifier to describe. Cannot be used in conjunction with DBInstanceIdentifier. This value is stored as a lowercase string.

Constraints:

Must be 1 to 255 alphanumeric characters First character must be a letter Cannot end with a hyphen or contain two consecutive hyphens If this is the identifier of an automated snapshot, the SnapshotType parameter must also be specified.

ddbsFilters :: Lens' DescribeDBSnapshots [Filter]

This parameter is not currently supported.

ddbsMarker :: Lens' DescribeDBSnapshots (Maybe Text)

An optional pagination token provided by a previous DescribeDBSnapshots request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

ddbsMaxRecords :: Lens' DescribeDBSnapshots (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results may be retrieved.

Default: 100

Constraints: minimum 20, maximum 100

ddbsSnapshotType :: Lens' DescribeDBSnapshots (Maybe Text)

The type of snapshots that will be returned. Values can be "automated" or "manual." If not specified, the returned results will include all snapshots types.

Response

data DescribeDBSnapshotsResponse

Response constructor

describeDBSnapshotsResponse :: DescribeDBSnapshotsResponse

DescribeDBSnapshotsResponse constructor.

The fields accessible through corresponding lenses are:

ddbsrDBSnapshots :: [DBSnapshot]

ddbsrMarker :: Maybe Text

Response lenses

ddbsrDBSnapshots :: Lens' DescribeDBSnapshotsResponse [DBSnapshot]

A list of DBSnapshot instances.

ddbsrMarker :: Lens' DescribeDBSnapshotsResponse (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.