Network.AWS.RDS.DescribeDBSubnetGroups

Description

Returns a list of DBSubnetGroup descriptions. If a DBSubnetGroupName is specified, the list will contain only the descriptions of the specified DBSubnetGroup. For an overview of CIDR ranges, go to the Wikipedia Tutorial. https:/rds.amazonaws.com ?Action=DescribeDBSubnetGroups &Version=2013-05-15 &MaxRecords=100 &SignatureVersion=2 &SignatureMethod=HmacSHA256 &Timestamp=2011-02-15T19%3A40%3A19.926Z &AWSAccessKeyId= &Signature= 990524496922 Complete description subnet_grp1 Active subnet-7c5b4115 us-east-1c Active subnet-7b5b4112 us-east-1b Active subnet-3ea6bd57 us-east-1d 990524496922 Complete description subnet_grp2 Active subnet-7c5b4115 us-east-1c Active subnet-7b5b4112 us-east-1b Active subnet-3ea6bd57 us-east-1d 31d0faee-229b-11e1-81f1-df3a2a803dad.

Request

data DescribeDBSubnetGroups

Request constructor

describeDBSubnetGroups :: DescribeDBSubnetGroups

Smart constructor for the minimum required parameters to construct a valid DescribeDBSubnetGroups request.

The fields accessible through corresponding lenses are:

DBSubnetGroupName :: Maybe Text

MaxRecords :: Maybe Integer

Marker :: Maybe Text

Request lenses

ddbsg3DBSubnetGroupName :: Lens' DescribeDBSubnetGroups (Maybe Text)

The name of the DB subnet group to return details for.

ddbsg3MaxRecords :: Lens' DescribeDBSubnetGroups (Maybe Integer)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results may be retrieved. Default: 100 Constraints: minimum 20, maximum 100.

ddbsg3Marker :: Lens' DescribeDBSubnetGroups (Maybe Text)

An optional pagination token provided by a previous DescribeDBSubnetGroups request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

Response

data DescribeDBSubnetGroupsResponse

Contains the result of a successful invocation of the DescribeDBSubnetGroups action.

Response constructor

describeDBSubnetGroupsResponse :: DescribeDBSubnetGroupsResponse

Smart constructor for the minimum required parameters to construct a valid DescribeDBSubnetGroupsResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Marker :: Maybe Text

DBSubnetGroups :: [DBSubnetGroup]

Response lenses

ddbsgrrMarker :: Lens' DescribeDBSubnetGroupsResponse (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

ddbsgrrDBSubnetGroups :: Lens' DescribeDBSubnetGroupsResponse [DBSubnetGroup]

A list of DBSubnetGroup instances.