Network.AWS.RDS.DescribeEngineDefaultClusterParameters

Description

Returns the default engine and system parameter information for the cluster database engine.

For more information on Amazon Aurora, see Aurora on Amazon RDS in the Amazon RDS User Guide.

Creating a Request

describeEngineDefaultClusterParameters

Arguments

:: Text
dedcpDBParameterGroupFamily

-> DescribeEngineDefaultClusterParameters

Creates a value of DescribeEngineDefaultClusterParameters with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dedcpFilters - This parameter is not currently supported.

dedcpMarker - An optional pagination token provided by a previous DescribeEngineDefaultClusterParameters request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

dedcpMaxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: Minimum 20, maximum 100.

dedcpDBParameterGroupFamily - The name of the DB cluster parameter group family to return engine parameter information for.

data DescribeEngineDefaultClusterParameters

See: describeEngineDefaultClusterParameters smart constructor.

Request Lenses

dedcpFilters :: Lens' DescribeEngineDefaultClusterParameters [Filter]

This parameter is not currently supported.

dedcpMarker :: Lens' DescribeEngineDefaultClusterParameters (Maybe Text)

An optional pagination token provided by a previous DescribeEngineDefaultClusterParameters request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

dedcpMaxRecords :: Lens' DescribeEngineDefaultClusterParameters (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: Minimum 20, maximum 100.

dedcpDBParameterGroupFamily :: Lens' DescribeEngineDefaultClusterParameters Text

The name of the DB cluster parameter group family to return engine parameter information for.

Destructuring the Response

describeEngineDefaultClusterParametersResponse

Arguments

:: Int
dedcprsResponseStatus

-> DescribeEngineDefaultClusterParametersResponse

Creates a value of DescribeEngineDefaultClusterParametersResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dedcprsEngineDefaults - Undocumented member.

dedcprsResponseStatus - -- | The response status code.

data DescribeEngineDefaultClusterParametersResponse

See: describeEngineDefaultClusterParametersResponse smart constructor.

Response Lenses

dedcprsEngineDefaults :: Lens' DescribeEngineDefaultClusterParametersResponse (Maybe EngineDefaults)

Undocumented member.

dedcprsResponseStatus :: Lens' DescribeEngineDefaultClusterParametersResponse Int

- | The response status code.