Network.AWS.RDS.DescribeEngineDefaultParameters

Description

Returns the default engine and system parameter information for the specified database engine.

http://docs.aws.amazon.com/AmazonRDS/latest/APIReference/API_DescribeEngineDefaultParameters.html

Request

data DescribeEngineDefaultParameters

Request constructor

describeEngineDefaultParameters

Arguments

:: Text
dedpDBParameterGroupFamily

-> DescribeEngineDefaultParameters

DescribeEngineDefaultParameters constructor.

The fields accessible through corresponding lenses are:

dedpDBParameterGroupFamily :: Text

dedpFilters :: [Filter]

dedpMarker :: Maybe Text

dedpMaxRecords :: Maybe Int

Request lenses

dedpDBParameterGroupFamily :: Lens' DescribeEngineDefaultParameters Text

The name of the DB parameter group family.

dedpFilters :: Lens' DescribeEngineDefaultParameters [Filter]

Not currently supported.

dedpMarker :: Lens' DescribeEngineDefaultParameters (Maybe Text)

An optional pagination token provided by a previous DescribeEngineDefaultParameters request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

dedpMaxRecords :: Lens' DescribeEngineDefaultParameters (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results may be retrieved.

Default: 100

Constraints: minimum 20, maximum 100

Response

data DescribeEngineDefaultParametersResponse

Response constructor

describeEngineDefaultParametersResponse

Arguments

:: EngineDefaults
dedprEngineDefaults

-> DescribeEngineDefaultParametersResponse

DescribeEngineDefaultParametersResponse constructor.

The fields accessible through corresponding lenses are:

dedprEngineDefaults :: EngineDefaults

Response lenses

dedprEngineDefaults :: Lens' DescribeEngineDefaultParametersResponse EngineDefaults