Network.AWS.RDS.DescribeEventCategories

Description

Displays a list of categories for all event source types, or, if specified, for a specified source type. You can see a list of the event categories and source types in the Events topic in the Amazon RDS User Guide. https:/rds.us-east-1.amazonaws.com ?Action=DescribeEventCategories &SourceType=db-instance &Version=2013-01-10 &SignatureVersion=4 &SignatureMethod=HmacSHA256 &Timestamp=20130128T013452Z &AWSAccessKeyId= &Signature= db-instance failover low storage maintenance recovery restoration deletion configuration change failover availability creation backup notification ea3bf54b-68ea-11e2-bd13-a92da73b3119.

Request

data DescribeEventCategories

Request constructor

describeEventCategories :: DescribeEventCategories

Smart constructor for the minimum required parameters to construct a valid DescribeEventCategories request.

The fields accessible through corresponding lenses are:

SourceType :: Maybe Text

Request lenses

decSourceType :: Lens' DescribeEventCategories (Maybe Text)

The type of source that will be generating the events. Valid values: db-instance | db-parameter-group | db-security-group | db-snapshot.

Response

data DescribeEventCategoriesResponse

Data returned from the DescribeEventCategories action.

Response constructor

describeEventCategoriesResponse :: DescribeEventCategoriesResponse

Smart constructor for the minimum required parameters to construct a valid DescribeEventCategoriesResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

EventCategoriesMapList :: [EventCategoriesMap]

Response lenses

decrEventCategoriesMapList :: Lens' DescribeEventCategoriesResponse [EventCategoriesMap]

A list of EventCategoriesMap data types.