Network.AWS.RDS.DescribeEventSubscriptions

Description

Lists all the subscription descriptions for a customer account. The description for a subscription includes SubscriptionName, SNSTopicARN, CustomerID, SourceType, SourceID, CreationTime, and Status. If you specify a SubscriptionName, lists the description for that subscription. https:/rds.us-east-1.amazonaws.com ?Action=DescribeEventSubscriptions &MaxRecords=100 &Version=2013-01-10 &SignatureVersion=4 &SignatureMethod=HmacSHA256 &Timestamp=20130128T004543Z &AWSAccessKeyId= &Signature= true 012345678901 active 2013-01-28 00:29:23.736 EventSubscription01 arn:aws:sns:us-east-1:012345678901:EventSubscription01 true 012345678901 active 2013-01-28 00:29:42.851 EventSubscription02 arn:aws:sns:us-east-1:012345678901:EventSubscription01 0ce48079-68e4-11e2-91fe-5daa8e68c7d4.

Request

data DescribeEventSubscriptions

Request constructor

describeEventSubscriptions :: DescribeEventSubscriptions

Smart constructor for the minimum required parameters to construct a valid DescribeEventSubscriptions request.

The fields accessible through corresponding lenses are:

SubscriptionName :: Maybe Text

MaxRecords :: Maybe Integer

Marker :: Maybe Text

Request lenses

des1SubscriptionName :: Lens' DescribeEventSubscriptions (Maybe Text)

The name of the RDS event notification subscription you want to describe.

des1MaxRecords :: Lens' DescribeEventSubscriptions (Maybe Integer)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: minimum 20, maximum 100.

des1Marker :: Lens' DescribeEventSubscriptions (Maybe Text)

An optional pagination token provided by a previous DescribeOrderableDBInstanceOptions request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

Response

data DescribeEventSubscriptionsResponse

Data returned by the DescribeEventSubscriptions action.

Response constructor

describeEventSubscriptionsResponse :: DescribeEventSubscriptionsResponse

Smart constructor for the minimum required parameters to construct a valid DescribeEventSubscriptionsResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Marker :: Maybe Text

EventSubscriptionsList :: [EventSubscription]

Response lenses

desrrMarker :: Lens' DescribeEventSubscriptionsResponse (Maybe Text)

An optional pagination token provided by a previous DescribeOrderableDBInstanceOptions request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

desrrEventSubscriptionsList :: Lens' DescribeEventSubscriptionsResponse [EventSubscription]

A list of EventSubscriptions data types.