Network.AWS.RDS.DescribeOptionGroupOptions

Description

Describes all available options.

http://docs.aws.amazon.com/AmazonRDS/latest/APIReference/API_DescribeOptionGroupOptions.html

Request

data DescribeOptionGroupOptions

Request constructor

describeOptionGroupOptions

Arguments

:: Text
dogoEngineName

-> DescribeOptionGroupOptions

DescribeOptionGroupOptions constructor.

The fields accessible through corresponding lenses are:

dogoEngineName :: Text

dogoFilters :: [Filter]

dogoMajorEngineVersion :: Maybe Text

dogoMarker :: Maybe Text

dogoMaxRecords :: Maybe Int

Request lenses

dogoEngineName :: Lens' DescribeOptionGroupOptions Text

A required parameter. Options available for the given Engine name will be described.

dogoFilters :: Lens' DescribeOptionGroupOptions [Filter]

This parameter is not currently supported.

dogoMajorEngineVersion :: Lens' DescribeOptionGroupOptions (Maybe Text)

If specified, filters the results to include only options for the specified major engine version.

dogoMarker :: Lens' DescribeOptionGroupOptions (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

dogoMaxRecords :: Lens' DescribeOptionGroupOptions (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: minimum 20, maximum 100

Response

data DescribeOptionGroupOptionsResponse

Response constructor

describeOptionGroupOptionsResponse :: DescribeOptionGroupOptionsResponse

DescribeOptionGroupOptionsResponse constructor.

The fields accessible through corresponding lenses are:

dogorMarker :: Maybe Text

dogorOptionGroupOptions :: [OptionGroupOption]

Response lenses

dogorMarker :: Lens' DescribeOptionGroupOptionsResponse (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

dogorOptionGroupOptions :: Lens' DescribeOptionGroupOptionsResponse [OptionGroupOption]