Network.AWS.RDS.DescribeOptionGroups

Description

Describes the available option groups.

http://docs.aws.amazon.com/AmazonRDS/latest/APIReference/API_DescribeOptionGroups.html

Request

data DescribeOptionGroups

Request constructor

describeOptionGroups :: DescribeOptionGroups

DescribeOptionGroups constructor.

The fields accessible through corresponding lenses are:

dogEngineName :: Maybe Text

dogFilters :: [Filter]

dogMajorEngineVersion :: Maybe Text

dogMarker :: Maybe Text

dogMaxRecords :: Maybe Int

dogOptionGroupName :: Maybe Text

Request lenses

dogEngineName :: Lens' DescribeOptionGroups (Maybe Text)

Filters the list of option groups to only include groups associated with a specific database engine.

dogFilters :: Lens' DescribeOptionGroups [Filter]

This parameter is not currently supported.

dogMajorEngineVersion :: Lens' DescribeOptionGroups (Maybe Text)

Filters the list of option groups to only include groups associated with a specific database engine version. If specified, then EngineName must also be specified.

dogMarker :: Lens' DescribeOptionGroups (Maybe Text)

An optional pagination token provided by a previous DescribeOptionGroups request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

dogMaxRecords :: Lens' DescribeOptionGroups (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: minimum 20, maximum 100

dogOptionGroupName :: Lens' DescribeOptionGroups (Maybe Text)

The name of the option group to describe. Cannot be supplied together with EngineName or MajorEngineVersion.

Response

data DescribeOptionGroupsResponse

Response constructor

describeOptionGroupsResponse :: DescribeOptionGroupsResponse

DescribeOptionGroupsResponse constructor.

The fields accessible through corresponding lenses are:

dogrMarker :: Maybe Text

dogrOptionGroupsList :: [OptionGroup]

Response lenses

dogrMarker :: Lens' DescribeOptionGroupsResponse (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

dogrOptionGroupsList :: Lens' DescribeOptionGroupsResponse [OptionGroup]

List of option groups.