Network.AWS.RDS.DescribeOrderableDBInstanceOptions

Description

Returns a list of orderable DB instance options for the specified engine. https:/rds.amazonaws.com ?Action=DescribeOrderableDBInstanceOptions &Engine=mysql &MaxRecords=100 &Version=2013-05-15 &Timestamp=2011-05-23T07%3A49%3A17.749Z &SignatureVersion=2 &SignatureMethod=HmacSHA256 &AWSAccessKeyId= &Signature= true mysql general-public-license true 5.1.45 db.m1.large us-east-1a yes us-east-1b no us-east-1d yes true mysql general-public-license true 5.1.45 db.m1.small us-east-1a yes us-east-1b yes us-east-1d yes true mysql general-public-license true 5.1.45 db.m1.xlarge us-east-1a yes us-east-1b yes us-east-1d yes true mysql general-public-license true 5.1.45 db.m2.2xlarge us-east-1a yes us-east-1b yes us-east-1d yes true mysql general-public-license true 5.1.45 db.m2.4xlarge us-east-1a yes us-east-1b no us-east-1d no 2a0406d7-8511-11e0-90aa-eb648410240d.

Request

data DescribeOrderableDBInstanceOptions

Request constructor

describeOrderableDBInstanceOptions

Arguments

:: Text
dodbioEngine

-> DescribeOrderableDBInstanceOptions

Smart constructor for the minimum required parameters to construct a valid DescribeOrderableDBInstanceOptions request.

The fields accessible through corresponding lenses are:

Engine :: Text

EngineVersion :: Maybe Text

DBInstanceClass :: Maybe Text

LicenseModel :: Maybe Text

Vpc :: Maybe Bool

MaxRecords :: Maybe Integer

Marker :: Maybe Text

Request lenses

dodbioEngine :: Lens' DescribeOrderableDBInstanceOptions Text

The name of the engine to retrieve DB instance options for.

dodbioEngineVersion :: Lens' DescribeOrderableDBInstanceOptions (Maybe Text)

The engine version filter value. Specify this parameter to show only the available offerings matching the specified engine version.

dodbioDBInstanceClass :: Lens' DescribeOrderableDBInstanceOptions (Maybe Text)

The DB instance class filter value. Specify this parameter to show only the available offerings matching the specified DB instance class.

dodbioLicenseModel :: Lens' DescribeOrderableDBInstanceOptions (Maybe Text)

The license model filter value. Specify this parameter to show only the available offerings matching the specified license model.

dodbioVpc :: Lens' DescribeOrderableDBInstanceOptions (Maybe Bool)

The VPC filter value. Specify this parameter to show only the available VPC or non-VPC offerings.

dodbioMaxRecords :: Lens' DescribeOrderableDBInstanceOptions (Maybe Integer)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: minimum 20, maximum 100.

dodbioMarker :: Lens' DescribeOrderableDBInstanceOptions (Maybe Text)

An optional pagination token provided by a previous DescribeOrderableDBInstanceOptions request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

Response

data DescribeOrderableDBInstanceOptionsResponse

Contains the result of a successful invocation of the DescribeOrderableDBInstanceOptions action.

Response constructor

describeOrderableDBInstanceOptionsResponse :: DescribeOrderableDBInstanceOptionsResponse

Smart constructor for the minimum required parameters to construct a valid DescribeOrderableDBInstanceOptionsResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

OrderableDBInstanceOptions :: [OrderableDBInstanceOption]

Marker :: Maybe Text

Response lenses

dodbiorOrderableDBInstanceOptions :: Lens' DescribeOrderableDBInstanceOptionsResponse [OrderableDBInstanceOption]

An OrderableDBInstanceOption structure containing information about orderable options for the DB instance.

dodbiorMarker :: Lens' DescribeOrderableDBInstanceOptionsResponse (Maybe Text)

An optional pagination token provided by a previous OrderableDBInstanceOptions request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .