Network.AWS.RDS.DescribePendingMaintenanceActions

Description

Returns a list of resources (for example, DB instances) that have at least one pending maintenance action.

http://docs.aws.amazon.com/AmazonRDS/latest/APIReference/API_DescribePendingMaintenanceActions.html

Request

data DescribePendingMaintenanceActions

Request constructor

describePendingMaintenanceActions :: DescribePendingMaintenanceActions

DescribePendingMaintenanceActions constructor.

The fields accessible through corresponding lenses are:

dpmaFilters :: [Filter]

dpmaMarker :: Maybe Text

dpmaMaxRecords :: Maybe Int

dpmaResourceIdentifier :: Maybe Text

Request lenses

dpmaFilters :: Lens' DescribePendingMaintenanceActions [Filter]

A filter that specifies one or more resources to return pending maintenance actions for.

Supported filters:

'db-instance-id' - Accepts DB instance identifiers and DB instance Amazon Resource Names (ARNs). The results list will only include pending maintenance actions for the DB instances identified by these ARNs.

dpmaMarker :: Lens' DescribePendingMaintenanceActions (Maybe Text)

An optional pagination token provided by a previous DescribePendingMaintenanceActions request. If this parameter is specified, the response includes only records beyond the marker, up to a number of records specified by MaxRecords.

dpmaMaxRecords :: Lens' DescribePendingMaintenanceActions (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: minimum 20, maximum 100

dpmaResourceIdentifier :: Lens' DescribePendingMaintenanceActions (Maybe Text)

The ARN of a resource to return pending maintenance actions for.

Response

data DescribePendingMaintenanceActionsResponse

Response constructor

describePendingMaintenanceActionsResponse :: DescribePendingMaintenanceActionsResponse

DescribePendingMaintenanceActionsResponse constructor.

The fields accessible through corresponding lenses are:

dpmarMarker :: Maybe Text

dpmarPendingMaintenanceActions :: [ResourcePendingMaintenanceActions]

Response lenses

dpmarMarker :: Lens' DescribePendingMaintenanceActionsResponse (Maybe Text)

An optional pagination token provided by a previous DescribePendingMaintenanceActions request. If this parameter is specified, the response includes only records beyond the marker, up to a number of records specified by MaxRecords.

dpmarPendingMaintenanceActions :: Lens' DescribePendingMaintenanceActionsResponse [ResourcePendingMaintenanceActions]

A list of the pending maintenance actions for the resource.