Network.AWS.RDS.DescribeReservedDBInstancesOfferings

Description

Lists available reserved DB instance offerings.

http://docs.aws.amazon.com/AmazonRDS/latest/APIReference/API_DescribeReservedDBInstancesOfferings.html

Request

data DescribeReservedDBInstancesOfferings

Request constructor

describeReservedDBInstancesOfferings :: DescribeReservedDBInstancesOfferings

DescribeReservedDBInstancesOfferings constructor.

The fields accessible through corresponding lenses are:

drdbioDBInstanceClass :: Maybe Text

drdbioDuration :: Maybe Text

drdbioFilters :: [Filter]

drdbioMarker :: Maybe Text

drdbioMaxRecords :: Maybe Int

drdbioMultiAZ :: Maybe Bool

drdbioOfferingType :: Maybe Text

drdbioProductDescription :: Maybe Text

drdbioReservedDBInstancesOfferingId :: Maybe Text

Request lenses

drdbioDBInstanceClass :: Lens' DescribeReservedDBInstancesOfferings (Maybe Text)

The DB instance class filter value. Specify this parameter to show only the available offerings matching the specified DB instance class.

drdbioDuration :: Lens' DescribeReservedDBInstancesOfferings (Maybe Text)

Duration filter value, specified in years or seconds. Specify this parameter to show only reservations for this duration.

Valid Values: '1 | 3 | 31536000 | 94608000'

drdbioFilters :: Lens' DescribeReservedDBInstancesOfferings [Filter]

This parameter is not currently supported.

drdbioMarker :: Lens' DescribeReservedDBInstancesOfferings (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

drdbioMaxRecords :: Lens' DescribeReservedDBInstancesOfferings (Maybe Int)

The maximum number of records to include in the response. If more than the MaxRecords value is available, a pagination token called a marker is included in the response so that the following results can be retrieved.

Default: 100

Constraints: minimum 20, maximum 100

drdbioMultiAZ :: Lens' DescribeReservedDBInstancesOfferings (Maybe Bool)

The Multi-AZ filter value. Specify this parameter to show only the available offerings matching the specified Multi-AZ parameter.

drdbioOfferingType :: Lens' DescribeReservedDBInstancesOfferings (Maybe Text)

The offering type filter value. Specify this parameter to show only the available offerings matching the specified offering type.

Valid Values: '"Light Utilization" | "Medium Utilization" | HeavyUtilization '

drdbioProductDescription :: Lens' DescribeReservedDBInstancesOfferings (Maybe Text)

Product description filter value. Specify this parameter to show only the available offerings matching the specified product description.

drdbioReservedDBInstancesOfferingId :: Lens' DescribeReservedDBInstancesOfferings (Maybe Text)

The offering identifier filter value. Specify this parameter to show only the available offering that matches the specified reservation identifier.

Example: '438012d3-4052-4cc7-b2e3-8d3372e0e706'

Response

data DescribeReservedDBInstancesOfferingsResponse

Response constructor

describeReservedDBInstancesOfferingsResponse :: DescribeReservedDBInstancesOfferingsResponse

DescribeReservedDBInstancesOfferingsResponse constructor.

The fields accessible through corresponding lenses are:

drdbiorMarker :: Maybe Text

drdbiorReservedDBInstancesOfferings :: [ReservedDBInstancesOffering]

Response lenses

drdbiorMarker :: Lens' DescribeReservedDBInstancesOfferingsResponse (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

drdbiorReservedDBInstancesOfferings :: Lens' DescribeReservedDBInstancesOfferingsResponse [ReservedDBInstancesOffering]

A list of reserved DB instance offerings.