Network.AWS.RDS.DescribeSourceRegions

Description

Returns a list of the source AWS regions where the current AWS Region can create a Read Replica or copy a DB snapshot from. This API action supports pagination.

Creating a Request

describeSourceRegions :: DescribeSourceRegions

Creates a value of DescribeSourceRegions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsrRegionName - The source AWS Region name. For example, us-east-1 . Constraints: * Must specify a valid AWS Region name.

dsrFilters - This parameter is not currently supported.

dsrMarker - An optional pagination token provided by a previous DescribeSourceRegions request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

dsrMaxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: Minimum 20, maximum 100.

data DescribeSourceRegions

See: describeSourceRegions smart constructor.

Request Lenses

dsrRegionName :: Lens' DescribeSourceRegions (Maybe Text)

The source AWS Region name. For example, us-east-1 . Constraints: * Must specify a valid AWS Region name.

dsrFilters :: Lens' DescribeSourceRegions [Filter]

This parameter is not currently supported.

dsrMarker :: Lens' DescribeSourceRegions (Maybe Text)

An optional pagination token provided by a previous DescribeSourceRegions request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

dsrMaxRecords :: Lens' DescribeSourceRegions (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: Minimum 20, maximum 100.

Destructuring the Response

describeSourceRegionsResponse

Arguments

:: Int
dsrrsResponseStatus

-> DescribeSourceRegionsResponse

Creates a value of DescribeSourceRegionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsrrsMarker - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

dsrrsSourceRegions - A list of SourceRegion instances that contains each source AWS Region that the current AWS Region can get a Read Replica or a DB snapshot from.

dsrrsResponseStatus - -- | The response status code.

data DescribeSourceRegionsResponse

Contains the result of a successful invocation of the DescribeSourceRegions action.

See: describeSourceRegionsResponse smart constructor.

Response Lenses

dsrrsMarker :: Lens' DescribeSourceRegionsResponse (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

dsrrsSourceRegions :: Lens' DescribeSourceRegionsResponse [SourceRegion]

A list of SourceRegion instances that contains each source AWS Region that the current AWS Region can get a Read Replica or a DB snapshot from.

dsrrsResponseStatus :: Lens' DescribeSourceRegionsResponse Int

- | The response status code.