Network.AWS.RDS.DownloadDBLogFilePortion

Description

Downloads all or a portion of the specified log file.

http://docs.aws.amazon.com/AmazonRDS/latest/APIReference/API_DownloadDBLogFilePortion.html

Request

data DownloadDBLogFilePortion

Request constructor

downloadDBLogFilePortion

Arguments

:: Text
ddblfpDBInstanceIdentifier

-> Text
ddblfpLogFileName

-> DownloadDBLogFilePortion

DownloadDBLogFilePortion constructor.

The fields accessible through corresponding lenses are:

ddblfpDBInstanceIdentifier :: Text

ddblfpLogFileName :: Text

ddblfpMarker :: Maybe Text

ddblfpNumberOfLines :: Maybe Int

Request lenses

ddblfpDBInstanceIdentifier :: Lens' DownloadDBLogFilePortion Text

The customer-assigned name of the DB instance that contains the log files you want to list.

Constraints:

Must contain from 1 to 63 alphanumeric characters or hyphens First character must be a letter Cannot end with a hyphen or contain two consecutive hyphens

ddblfpLogFileName :: Lens' DownloadDBLogFilePortion Text

The name of the log file to be downloaded.

ddblfpMarker :: Lens' DownloadDBLogFilePortion (Maybe Text)

The pagination token provided in the previous request or "0". If the Marker parameter is specified the response includes only records beyond the marker until the end of the file or up to NumberOfLines.

ddblfpNumberOfLines :: Lens' DownloadDBLogFilePortion (Maybe Int)

The number of lines to download.

If the NumberOfLines parameter is specified, then the block of lines returned can be from the beginning or the end of the log file, depending on the value of the Marker parameter. If neither Marker or NumberOfLines are specified, the entire log file is returned.

If NumberOfLines is specified and Marker is not specified, then the most recent lines from the end of the log file are returned.

If Marker is specified as "0", then the specified number of lines from the beginning of the log file are returned.

You can download the log file in blocks of lines by specifying the size of the block using the NumberOfLines parameter, and by specifying a value of "0" for the Marker parameter in your first request. Include the Marker value returned in the response as the Marker value for the next request, continuing until the AdditionalDataPending response element returns false.

Response

data DownloadDBLogFilePortionResponse

Response constructor

downloadDBLogFilePortionResponse :: DownloadDBLogFilePortionResponse

DownloadDBLogFilePortionResponse constructor.

The fields accessible through corresponding lenses are:

ddblfprAdditionalDataPending :: Maybe Bool

ddblfprLogFileData :: Maybe Text

ddblfprMarker :: Maybe Text

Response lenses

ddblfprAdditionalDataPending :: Lens' DownloadDBLogFilePortionResponse (Maybe Bool)

Boolean value that if true, indicates there is more data to be downloaded.

ddblfprLogFileData :: Lens' DownloadDBLogFilePortionResponse (Maybe Text)

Entries from the specified log file.

ddblfprMarker :: Lens' DownloadDBLogFilePortionResponse (Maybe Text)

A pagination token that can be used in a subsequent DownloadDBLogFilePortion request.