Network.AWS.RDS.FailoverDBCluster

Description

Forces a failover for a DB cluster.

A failover for a DB cluster promotes one of the Aurora Replicas (read-only instances) in the DB cluster to be the primary instance (the cluster writer).

Amazon Aurora will automatically fail over to an Aurora Replica, if one exists, when the primary instance fails. You can force a failover when you want to simulate a failure of a primary instance for testing. Because each instance in a DB cluster has its own endpoint address, you will need to clean up and re-establish any existing connections that use those endpoint addresses when the failover is complete.

For more information on Amazon Aurora, see Aurora on Amazon RDS in the Amazon RDS User Guide.

Creating a Request

failoverDBCluster :: FailoverDBCluster

Creates a value of FailoverDBCluster with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

fdcDBClusterIdentifier - A DB cluster identifier to force a failover for. This parameter is not case-sensitive. Constraints: * Must contain from 1 to 63 alphanumeric characters or hyphens * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens

fdcTargetDBInstanceIdentifier - The name of the instance to promote to the primary instance. You must specify the instance identifier for an Aurora Replica in the DB cluster. For example, mydbcluster-replica1 .

data FailoverDBCluster

See: failoverDBCluster smart constructor.

Request Lenses

fdcDBClusterIdentifier :: Lens' FailoverDBCluster (Maybe Text)

A DB cluster identifier to force a failover for. This parameter is not case-sensitive. Constraints: * Must contain from 1 to 63 alphanumeric characters or hyphens * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens

fdcTargetDBInstanceIdentifier :: Lens' FailoverDBCluster (Maybe Text)

The name of the instance to promote to the primary instance. You must specify the instance identifier for an Aurora Replica in the DB cluster. For example, mydbcluster-replica1 .

Destructuring the Response

failoverDBClusterResponse

Arguments

:: Int
fdcrsResponseStatus

-> FailoverDBClusterResponse

Creates a value of FailoverDBClusterResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

fdcrsDBCluster - Undocumented member.

fdcrsResponseStatus - -- | The response status code.

data FailoverDBClusterResponse

See: failoverDBClusterResponse smart constructor.

Response Lenses

fdcrsDBCluster :: Lens' FailoverDBClusterResponse (Maybe DBCluster)

Undocumented member.

fdcrsResponseStatus :: Lens' FailoverDBClusterResponse Int

- | The response status code.