Network.AWS.RDS.ModifyDBCluster

Description

Modify a setting for an Amazon Aurora DB cluster. You can change one or more database configuration parameters by specifying these parameters and the new values in the request. For more information on Amazon Aurora, see Aurora on Amazon RDS in the Amazon RDS User Guide.

Creating a Request

modifyDBCluster

Arguments

:: Text
mdcDBClusterIdentifier

-> ModifyDBCluster

Creates a value of ModifyDBCluster with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mdcMasterUserPassword - The new password for the master database user. This password can contain any printable ASCII character except "/", """, or "@". Constraints: Must contain from 8 to 41 characters.

mdcPreferredMaintenanceWindow - The weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC). Format: ddd:hh24:mi-ddd:hh24:mi Default: A 30-minute window selected at random from an 8-hour block of time per AWS Region, occurring on a random day of the week. To see the time blocks available, see Adjusting the Preferred Maintenance Window in the Amazon RDS User Guide. Valid Days: Mon, Tue, Wed, Thu, Fri, Sat, Sun Constraints: Minimum 30-minute window.

mdcPreferredBackupWindow - The daily time range during which automated backups are created if automated backups are enabled, using the BackupRetentionPeriod parameter. Default: A 30-minute window selected at random from an 8-hour block of time per AWS Region. To see the time blocks available, see Adjusting the Preferred Maintenance Window in the Amazon RDS User Guide. Constraints: * Must be in the format hh24:mi-hh24:mi . * Times should be in Universal Coordinated Time (UTC). * Must not conflict with the preferred maintenance window. * Must be at least 30 minutes.

mdcBackupRetentionPeriod - The number of days for which automated backups are retained. You must specify a minimum value of 1. Default: 1 Constraints: * Must be a value from 1 to 35

mdcVPCSecurityGroupIds - A list of VPC security groups that the DB cluster will belong to.

mdcDBClusterParameterGroupName - The name of the DB cluster parameter group to use for the DB cluster.

mdcApplyImmediately - A value that specifies whether the modifications in this request and any pending modifications are asynchronously applied as soon as possible, regardless of the PreferredMaintenanceWindow setting for the DB cluster. If this parameter is set to false , changes to the DB cluster are applied during the next maintenance window. The ApplyImmediately parameter only affects the NewDBClusterIdentifier and MasterUserPassword values. If you set the ApplyImmediately parameter value to false, then changes to the NewDBClusterIdentifier and MasterUserPassword values are applied during the next maintenance window. All other changes are applied immediately, regardless of the value of the ApplyImmediately parameter. Default: false

mdcOptionGroupName - A value that indicates that the DB cluster should be associated with the specified option group. Changing this parameter does not result in an outage except in the following case, and the change is applied during the next maintenance window unless the ApplyImmediately parameter is set to true for this request. If the parameter change results in an option group that enables OEM, this change can cause a brief (sub-second) period during which new connections are rejected but existing connections are not interrupted. Permanent options cannot be removed from an option group. The option group cannot be removed from a DB cluster once it is associated with a DB cluster.

mdcNewDBClusterIdentifier - The new DB cluster identifier for the DB cluster when renaming a DB cluster. This value is stored as a lowercase string. Constraints: * Must contain from 1 to 63 alphanumeric characters or hyphens * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens Example: my-cluster2

mdcPort - The port number on which the DB cluster accepts connections. Constraints: Value must be 1150-65535 Default: The same port as the original DB cluster.

mdcEnableIAMDatabaseAuthentication - A Boolean value that is true to enable mapping of AWS Identity and Access Management (IAM) accounts to database accounts, and otherwise false. Default: false

mdcDBClusterIdentifier - The DB cluster identifier for the cluster being modified. This parameter is not case-sensitive. Constraints: * Must be the identifier for an existing DB cluster. * Must contain from 1 to 63 alphanumeric characters or hyphens. * First character must be a letter. * Cannot end with a hyphen or contain two consecutive hyphens.

data ModifyDBCluster

See: modifyDBCluster smart constructor.

Request Lenses

mdcMasterUserPassword :: Lens' ModifyDBCluster (Maybe Text)

The new password for the master database user. This password can contain any printable ASCII character except "/", """, or "@". Constraints: Must contain from 8 to 41 characters.

mdcPreferredMaintenanceWindow :: Lens' ModifyDBCluster (Maybe Text)

The weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC). Format: ddd:hh24:mi-ddd:hh24:mi Default: A 30-minute window selected at random from an 8-hour block of time per AWS Region, occurring on a random day of the week. To see the time blocks available, see Adjusting the Preferred Maintenance Window in the Amazon RDS User Guide. Valid Days: Mon, Tue, Wed, Thu, Fri, Sat, Sun Constraints: Minimum 30-minute window.

mdcPreferredBackupWindow :: Lens' ModifyDBCluster (Maybe Text)

The daily time range during which automated backups are created if automated backups are enabled, using the BackupRetentionPeriod parameter. Default: A 30-minute window selected at random from an 8-hour block of time per AWS Region. To see the time blocks available, see Adjusting the Preferred Maintenance Window in the Amazon RDS User Guide. Constraints: * Must be in the format hh24:mi-hh24:mi . * Times should be in Universal Coordinated Time (UTC). * Must not conflict with the preferred maintenance window. * Must be at least 30 minutes.

mdcBackupRetentionPeriod :: Lens' ModifyDBCluster (Maybe Int)

The number of days for which automated backups are retained. You must specify a minimum value of 1. Default: 1 Constraints: * Must be a value from 1 to 35

mdcVPCSecurityGroupIds :: Lens' ModifyDBCluster [Text]

A list of VPC security groups that the DB cluster will belong to.

mdcDBClusterParameterGroupName :: Lens' ModifyDBCluster (Maybe Text)

The name of the DB cluster parameter group to use for the DB cluster.

mdcApplyImmediately :: Lens' ModifyDBCluster (Maybe Bool)

A value that specifies whether the modifications in this request and any pending modifications are asynchronously applied as soon as possible, regardless of the PreferredMaintenanceWindow setting for the DB cluster. If this parameter is set to false , changes to the DB cluster are applied during the next maintenance window. The ApplyImmediately parameter only affects the NewDBClusterIdentifier and MasterUserPassword values. If you set the ApplyImmediately parameter value to false, then changes to the NewDBClusterIdentifier and MasterUserPassword values are applied during the next maintenance window. All other changes are applied immediately, regardless of the value of the ApplyImmediately parameter. Default: false

mdcOptionGroupName :: Lens' ModifyDBCluster (Maybe Text)

A value that indicates that the DB cluster should be associated with the specified option group. Changing this parameter does not result in an outage except in the following case, and the change is applied during the next maintenance window unless the ApplyImmediately parameter is set to true for this request. If the parameter change results in an option group that enables OEM, this change can cause a brief (sub-second) period during which new connections are rejected but existing connections are not interrupted. Permanent options cannot be removed from an option group. The option group cannot be removed from a DB cluster once it is associated with a DB cluster.

mdcNewDBClusterIdentifier :: Lens' ModifyDBCluster (Maybe Text)

The new DB cluster identifier for the DB cluster when renaming a DB cluster. This value is stored as a lowercase string. Constraints: * Must contain from 1 to 63 alphanumeric characters or hyphens * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens Example: my-cluster2

mdcPort :: Lens' ModifyDBCluster (Maybe Int)

The port number on which the DB cluster accepts connections. Constraints: Value must be 1150-65535 Default: The same port as the original DB cluster.

mdcEnableIAMDatabaseAuthentication :: Lens' ModifyDBCluster (Maybe Bool)

A Boolean value that is true to enable mapping of AWS Identity and Access Management (IAM) accounts to database accounts, and otherwise false. Default: false

mdcDBClusterIdentifier :: Lens' ModifyDBCluster Text

The DB cluster identifier for the cluster being modified. This parameter is not case-sensitive. Constraints: * Must be the identifier for an existing DB cluster. * Must contain from 1 to 63 alphanumeric characters or hyphens. * First character must be a letter. * Cannot end with a hyphen or contain two consecutive hyphens.

Destructuring the Response

modifyDBClusterResponse

Arguments

:: Int
mdcrsResponseStatus

-> ModifyDBClusterResponse

Creates a value of ModifyDBClusterResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mdcrsDBCluster - Undocumented member.

mdcrsResponseStatus - -- | The response status code.

data ModifyDBClusterResponse

See: modifyDBClusterResponse smart constructor.

Response Lenses

mdcrsDBCluster :: Lens' ModifyDBClusterResponse (Maybe DBCluster)

Undocumented member.

mdcrsResponseStatus :: Lens' ModifyDBClusterResponse Int

- | The response status code.