Network.AWS.RDS.ModifyDBClusterParameterGroup

Description

Modifies the parameters of a DB cluster parameter group. To modify more than one parameter, submit a list of the following: ParameterName , ParameterValue , and ApplyMethod . A maximum of 20 parameters can be modified in a single request.

For more information on Amazon Aurora, see Aurora on Amazon RDS in the Amazon RDS User Guide.

Important: After you create a DB cluster parameter group, you should wait at least 5 minutes before creating your first DB cluster that uses that DB cluster parameter group as the default parameter group. This allows Amazon RDS to fully complete the create action before the parameter group is used as the default for a new DB cluster. This is especially important for parameters that are critical when creating the default database for a DB cluster, such as the character set for the default database defined by the character_set_database parameter. You can use the Parameter Groups option of the Amazon RDS console or the DescribeDBClusterParameters command to verify that your DB cluster parameter group has been created or modified.

Creating a Request

modifyDBClusterParameterGroup

Arguments

:: Text
mdcpgDBClusterParameterGroupName

-> ModifyDBClusterParameterGroup

Creates a value of ModifyDBClusterParameterGroup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mdcpgDBClusterParameterGroupName - The name of the DB cluster parameter group to modify.

mdcpgParameters - A list of parameters in the DB cluster parameter group to modify.

data ModifyDBClusterParameterGroup

See: modifyDBClusterParameterGroup smart constructor.

Request Lenses

mdcpgDBClusterParameterGroupName :: Lens' ModifyDBClusterParameterGroup Text

The name of the DB cluster parameter group to modify.

mdcpgParameters :: Lens' ModifyDBClusterParameterGroup [Parameter]

A list of parameters in the DB cluster parameter group to modify.

Destructuring the Response

dbClusterParameterGroupNameMessage :: DBClusterParameterGroupNameMessage

Creates a value of DBClusterParameterGroupNameMessage with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcpgnmDBClusterParameterGroupName - The name of the DB cluster parameter group. Constraints: * Must be 1 to 255 alphanumeric characters * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens

data DBClusterParameterGroupNameMessage

See: dbClusterParameterGroupNameMessage smart constructor.

Response Lenses

dcpgnmDBClusterParameterGroupName :: Lens' DBClusterParameterGroupNameMessage (Maybe Text)

The name of the DB cluster parameter group. Constraints: * Must be 1 to 255 alphanumeric characters * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens