Network.AWS.RDS.ModifyDBParameterGroup

Description

Modifies the parameters of a DB parameter group. To modify more than one parameter, submit a list of the following: ParameterName, ParameterValue, and ApplyMethod. A maximum of 20 parameters can be modified in a single request.

Changes to dynamic parameters are applied immediately. Changes to static parameters require a reboot without failover to the DB instance associated with the parameter group before the change can take effect.

After you modify a DB parameter group, you should wait at least 5 minutes before creating your first DB instance that uses that DB parameter group as the default parameter group. This allows Amazon RDS to fully complete the modify action before the parameter group is used as the default for a new DB instance. This is especially important for parameters that are critical when creating the default database for a DB instance, such as the character set for the default database defined by the character_set_database parameter. You can use the Parameter Groups option of the Amazon RDS console or the DescribeDBParameters command to verify that your DB parameter group has been created or modified.

http://docs.aws.amazon.com/AmazonRDS/latest/APIReference/API_ModifyDBParameterGroup.html

Request

data ModifyDBParameterGroup

Request constructor

modifyDBParameterGroup

Arguments

:: Text
mdbpgDBParameterGroupName

-> ModifyDBParameterGroup

ModifyDBParameterGroup constructor.

The fields accessible through corresponding lenses are:

mdbpgDBParameterGroupName :: Text

mdbpgParameters :: [Parameter]

Request lenses

mdbpgDBParameterGroupName :: Lens' ModifyDBParameterGroup Text

The name of the DB parameter group.

Constraints:

Must be the name of an existing DB parameter group Must be 1 to 255 alphanumeric characters First character must be a letter Cannot end with a hyphen or contain two consecutive hyphens

mdbpgParameters :: Lens' ModifyDBParameterGroup [Parameter]

An array of parameter names, values, and the apply method for the parameter update. At least one parameter name, value, and apply method must be supplied; subsequent arguments are optional. A maximum of 20 parameters may be modified in a single request.

Valid Values (for the application method): 'immediate | pending-reboot'

You can use the immediate value with dynamic parameters only. You can use the pending-reboot value for both dynamic and static parameters, and changes are applied when you reboot the DB instance without failover.

Response

data ModifyDBParameterGroupResponse

Response constructor

modifyDBParameterGroupResponse :: ModifyDBParameterGroupResponse

ModifyDBParameterGroupResponse constructor.

The fields accessible through corresponding lenses are:

mdbpgrDBParameterGroupName :: Maybe Text

Response lenses

mdbpgrDBParameterGroupName :: Lens' ModifyDBParameterGroupResponse (Maybe Text)

The name of the DB parameter group.