Network.AWS.RDS.ModifyEventSubscription

Description

Modifies an existing RDS event notification subscription. Note that you cannot modify the source identifiers using this call; to change source identifiers for a subscription, use the AddSourceIdentifierToSubscription and RemoveSourceIdentifierFromSubscription calls. You can see a list of the event categories for a given SourceType in the <http://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/USER_Events.html Events> topic in the Amazon RDS User Guide or by using the DescribeEventCategories action.

http://docs.aws.amazon.com/AmazonRDS/latest/APIReference/API_ModifyEventSubscription.html

Request

data ModifyEventSubscription

Request constructor

modifyEventSubscription

Arguments

:: Text
mesSubscriptionName

-> ModifyEventSubscription

ModifyEventSubscription constructor.

The fields accessible through corresponding lenses are:

mesEnabled :: Maybe Bool

mesEventCategories :: [Text]

mesSnsTopicArn :: Maybe Text

mesSourceType :: Maybe Text

mesSubscriptionName :: Text

Request lenses

mesEnabled :: Lens' ModifyEventSubscription (Maybe Bool)

A Boolean value; set to true to activate the subscription.

mesEventCategories :: Lens' ModifyEventSubscription [Text]

A list of event categories for a SourceType that you want to subscribe to. You can see a list of the categories for a given SourceType in the <http://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/USER_Events.html Events> topic in the Amazon RDS User Guide or by using the DescribeEventCategories action.

mesSnsTopicArn :: Lens' ModifyEventSubscription (Maybe Text)

The Amazon Resource Name (ARN) of the SNS topic created for event notification. The ARN is created by Amazon SNS when you create a topic and subscribe to it.

mesSourceType :: Lens' ModifyEventSubscription (Maybe Text)

The type of source that will be generating the events. For example, if you want to be notified of events generated by a DB instance, you would set this parameter to db-instance. if this value is not specified, all events are returned. Valid values: db-instance | db-parameter-group | db-security-group | db-snapshot.

mesSubscriptionName :: Lens' ModifyEventSubscription Text

The name of the RDS event notification subscription.

Response

data ModifyEventSubscriptionResponse

Response constructor

modifyEventSubscriptionResponse :: ModifyEventSubscriptionResponse

ModifyEventSubscriptionResponse constructor.

The fields accessible through corresponding lenses are:

mesrEventSubscription :: Maybe EventSubscription

Response lenses

mesrEventSubscription :: Lens' ModifyEventSubscriptionResponse (Maybe EventSubscription)