Network.AWS.RDS.RemoveRoleFromDBCluster

Description

Disassociates an Identity and Access Management (IAM) role from an Aurora DB cluster. For more information, see Authorizing Amazon Aurora to Access Other AWS Services On Your Behalf .

Creating a Request

removeRoleFromDBCluster

Arguments

:: Text
rrfdcDBClusterIdentifier

-> Text
rrfdcRoleARN

-> RemoveRoleFromDBCluster

Creates a value of RemoveRoleFromDBCluster with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rrfdcDBClusterIdentifier - The name of the DB cluster to disassociate the IAM role from.

rrfdcRoleARN - The Amazon Resource Name (ARN) of the IAM role to disassociate from the Aurora DB cluster, for example arn:aws:iam::123456789012:role/AuroraAccessRole .

data RemoveRoleFromDBCluster

See: removeRoleFromDBCluster smart constructor.

Request Lenses

rrfdcDBClusterIdentifier :: Lens' RemoveRoleFromDBCluster Text

The name of the DB cluster to disassociate the IAM role from.

rrfdcRoleARN :: Lens' RemoveRoleFromDBCluster Text

The Amazon Resource Name (ARN) of the IAM role to disassociate from the Aurora DB cluster, for example arn:aws:iam::123456789012:role/AuroraAccessRole .

Destructuring the Response

removeRoleFromDBClusterResponse :: RemoveRoleFromDBClusterResponse

Creates a value of RemoveRoleFromDBClusterResponse with the minimum fields required to make a request.

data RemoveRoleFromDBClusterResponse

See: removeRoleFromDBClusterResponse smart constructor.