Network.AWS.RDS.RemoveSourceIdentifierFromSubscription

Description

Removes a source identifier from an existing RDS event notification subscription.

http://docs.aws.amazon.com/AmazonRDS/latest/APIReference/API_RemoveSourceIdentifierFromSubscription.html

Request

data RemoveSourceIdentifierFromSubscription

Request constructor

removeSourceIdentifierFromSubscription

Arguments

:: Text
rsifsSubscriptionName

-> Text
rsifsSourceIdentifier

-> RemoveSourceIdentifierFromSubscription

RemoveSourceIdentifierFromSubscription constructor.

The fields accessible through corresponding lenses are:

rsifsSourceIdentifier :: Text

rsifsSubscriptionName :: Text

Request lenses

rsifsSourceIdentifier :: Lens' RemoveSourceIdentifierFromSubscription Text

The source identifier to be removed from the subscription, such as the DB instance identifier for a DB instance or the name of a security group.

rsifsSubscriptionName :: Lens' RemoveSourceIdentifierFromSubscription Text

The name of the RDS event notification subscription you want to remove a source identifier from.

Response

data RemoveSourceIdentifierFromSubscriptionResponse

Response constructor

removeSourceIdentifierFromSubscriptionResponse :: RemoveSourceIdentifierFromSubscriptionResponse

RemoveSourceIdentifierFromSubscriptionResponse constructor.

The fields accessible through corresponding lenses are:

rsifsrEventSubscription :: Maybe EventSubscription

Response lenses

rsifsrEventSubscription :: Lens' RemoveSourceIdentifierFromSubscriptionResponse (Maybe EventSubscription)