Network.AWS.RDS.ResetDBParameterGroup

Description

Modifies the parameters of a DB parameter group to the engine/system default value. To reset specific parameters submit a list of the following: ParameterName and ApplyMethod. To reset the entire DB parameter group, specify the DBParameterGroup name and ResetAllParameters parameters. When resetting the entire group, dynamic parameters are updated immediately and static parameters are set to pending-reboot to take effect on the next DB instance restart or RebootDBInstance request.

http://docs.aws.amazon.com/AmazonRDS/latest/APIReference/API_ResetDBParameterGroup.html

Request

data ResetDBParameterGroup

Request constructor

resetDBParameterGroup

Arguments

:: Text
rdbpgDBParameterGroupName

-> ResetDBParameterGroup

ResetDBParameterGroup constructor.

The fields accessible through corresponding lenses are:

rdbpgDBParameterGroupName :: Text

rdbpgParameters :: [Parameter]

rdbpgResetAllParameters :: Maybe Bool

Request lenses

rdbpgDBParameterGroupName :: Lens' ResetDBParameterGroup Text

The name of the DB parameter group. Constraints: Must be 1 to 255 alphanumeric characters First character must be a letter Cannot end with a hyphen or contain two consecutive hyphens.

rdbpgParameters :: Lens' ResetDBParameterGroup [Parameter]

An array of parameter names, values, and the apply method for the parameter update. At least one parameter name, value, and apply method must be supplied; subsequent arguments are optional. A maximum of 20 parameters may be modified in a single request. MySQL Valid Values (for Apply method): immediate | pending-reboot You can use the immediate value with dynamic parameters only. You can use the pending-reboot value for both dynamic and static parameters, and changes are applied when DB instance reboots. Oracle Valid Values (for Apply method): pending-reboot.

rdbpgResetAllParameters :: Lens' ResetDBParameterGroup (Maybe Bool)

Specifies whether (true) or not (false) to reset all parameters in the DB parameter group to default values. Default: true.

Response

data ResetDBParameterGroupResponse

Response constructor

resetDBParameterGroupResponse :: ResetDBParameterGroupResponse

ResetDBParameterGroupResponse constructor.

The fields accessible through corresponding lenses are:

rdbpgrDBParameterGroupName :: Maybe Text

Response lenses

rdbpgrDBParameterGroupName :: Lens' ResetDBParameterGroupResponse (Maybe Text)

The name of the DB parameter group.