Network.AWS.RDS.RestoreDBClusterFromS3

Description

Creates an Amazon Aurora DB cluster from data stored in an Amazon S3 bucket. Amazon RDS must be authorized to access the Amazon S3 bucket and the data must be created using the Percona XtraBackup utility as described in Migrating Data from MySQL by Using an Amazon S3 Bucket .

Creating a Request

restoreDBClusterFromS3

Arguments

:: Text
rdcfsDBClusterIdentifier

-> Text
rdcfsEngine

-> Text
rdcfsMasterUsername

-> Text
rdcfsMasterUserPassword

-> Text
rdcfsSourceEngine

-> Text
rdcfsSourceEngineVersion

-> Text
rdcfsS3BucketName

-> Text
rdcfsS3IngestionRoleARN

-> RestoreDBClusterFromS3

Creates a value of RestoreDBClusterFromS3 with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rdcfsEngineVersion - The version number of the database engine to use. Aurora Example: 5.6.10a

rdcfsStorageEncrypted - Specifies whether the restored DB cluster is encrypted.

rdcfsDBSubnetGroupName - A DB subnet group to associate with the restored DB cluster. Constraints: Must contain no more than 255 alphanumeric characters, periods, underscores, spaces, or hyphens. Must not be default. Example: mySubnetgroup

rdcfsPreferredMaintenanceWindow - The weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC). Format: ddd:hh24:mi-ddd:hh24:mi Default: A 30-minute window selected at random from an 8-hour block of time per AWS Region, occurring on a random day of the week. To see the time blocks available, see Adjusting the Preferred Maintenance Window in the Amazon RDS User Guide. Valid Days: Mon, Tue, Wed, Thu, Fri, Sat, Sun Constraints: Minimum 30-minute window.

rdcfsAvailabilityZones - A list of EC2 Availability Zones that instances in the restored DB cluster can be created in.

rdcfsCharacterSetName - A value that indicates that the restored DB cluster should be associated with the specified CharacterSet.

rdcfsKMSKeyId - The KMS key identifier for an encrypted DB cluster. The KMS key identifier is the Amazon Resource Name (ARN) for the KMS encryption key. If you are creating a DB cluster with the same AWS account that owns the KMS encryption key used to encrypt the new DB cluster, then you can use the KMS key alias instead of the ARN for the KM encryption key. If the StorageEncrypted parameter is true, and you do not specify a value for the KmsKeyId parameter, then Amazon RDS will use your default encryption key. AWS KMS creates the default encryption key for your AWS account. Your AWS account has a different default encryption key for each AWS Region.

rdcfsPreferredBackupWindow - The daily time range during which automated backups are created if automated backups are enabled using the BackupRetentionPeriod parameter. Default: A 30-minute window selected at random from an 8-hour block of time per AWS Region. To see the time blocks available, see Adjusting the Preferred Maintenance Window in the Amazon RDS User Guide. Constraints: * Must be in the format hh24:mi-hh24:mi . * Times should be in Universal Coordinated Time (UTC). * Must not conflict with the preferred maintenance window. * Must be at least 30 minutes.

rdcfsBackupRetentionPeriod - The number of days for which automated backups of the restored DB cluster are retained. You must specify a minimum value of 1. Default: 1 Constraints: * Must be a value from 1 to 35

rdcfsVPCSecurityGroupIds - A list of EC2 VPC security groups to associate with the restored DB cluster.

rdcfsDatabaseName - The database name for the restored DB cluster.

rdcfsDBClusterParameterGroupName - The name of the DB cluster parameter group to associate with the restored DB cluster. If this argument is omitted, default.aurora5.6 will be used. Constraints: * Must be 1 to 255 alphanumeric characters * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens

rdcfsS3Prefix - The prefix for all of the file names that contain the data used to create the Amazon Aurora DB cluster. If you do not specify a SourceS3Prefix value, then the Amazon Aurora DB cluster is created by using all of the files in the Amazon S3 bucket.

rdcfsOptionGroupName - A value that indicates that the restored DB cluster should be associated with the specified option group. Permanent options cannot be removed from an option group. An option group cannot be removed from a DB cluster once it is associated with a DB cluster.

rdcfsTags - Undocumented member.

rdcfsPort - The port number on which the instances in the restored DB cluster accept connections. Default: 3306

rdcfsEnableIAMDatabaseAuthentication - A Boolean value that is true to enable mapping of AWS Identity and Access Management (IAM) accounts to database accounts, and otherwise false. Default: false

rdcfsDBClusterIdentifier - The name of the DB cluster to create from the source data in the S3 bucket. This parameter is isn't case-sensitive. Constraints: * Must contain from 1 to 63 alphanumeric characters or hyphens. * First character must be a letter. * Cannot end with a hyphen or contain two consecutive hyphens. Example: my-cluster1

rdcfsEngine - The name of the database engine to be used for the restored DB cluster. Valid Values: aurora

rdcfsMasterUsername - The name of the master user for the restored DB cluster. Constraints: * Must be 1 to 16 alphanumeric characters. * First character must be a letter. * Cannot be a reserved word for the chosen database engine.

rdcfsMasterUserPassword - The password for the master database user. This password can contain any printable ASCII character except "/", """, or "@". Constraints: Must contain from 8 to 41 characters.

rdcfsSourceEngine - The identifier for the database engine that was backed up to create the files stored in the Amazon S3 bucket. Valid values: mysql

rdcfsSourceEngineVersion - The version of the database that the backup files were created from. MySQL version 5.5 and 5.6 are supported. Example: 5.6.22

rdcfsS3BucketName - The name of the Amazon S3 bucket that contains the data used to create the Amazon Aurora DB cluster.

rdcfsS3IngestionRoleARN - The Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role that authorizes Amazon RDS to access the Amazon S3 bucket on your behalf.

data RestoreDBClusterFromS3

See: restoreDBClusterFromS3 smart constructor.

Request Lenses

rdcfsEngineVersion :: Lens' RestoreDBClusterFromS3 (Maybe Text)

The version number of the database engine to use. Aurora Example: 5.6.10a

rdcfsStorageEncrypted :: Lens' RestoreDBClusterFromS3 (Maybe Bool)

Specifies whether the restored DB cluster is encrypted.

rdcfsDBSubnetGroupName :: Lens' RestoreDBClusterFromS3 (Maybe Text)

A DB subnet group to associate with the restored DB cluster. Constraints: Must contain no more than 255 alphanumeric characters, periods, underscores, spaces, or hyphens. Must not be default. Example: mySubnetgroup

rdcfsPreferredMaintenanceWindow :: Lens' RestoreDBClusterFromS3 (Maybe Text)

The weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC). Format: ddd:hh24:mi-ddd:hh24:mi Default: A 30-minute window selected at random from an 8-hour block of time per AWS Region, occurring on a random day of the week. To see the time blocks available, see Adjusting the Preferred Maintenance Window in the Amazon RDS User Guide. Valid Days: Mon, Tue, Wed, Thu, Fri, Sat, Sun Constraints: Minimum 30-minute window.

rdcfsAvailabilityZones :: Lens' RestoreDBClusterFromS3 [Text]

A list of EC2 Availability Zones that instances in the restored DB cluster can be created in.

rdcfsCharacterSetName :: Lens' RestoreDBClusterFromS3 (Maybe Text)

A value that indicates that the restored DB cluster should be associated with the specified CharacterSet.

rdcfsKMSKeyId :: Lens' RestoreDBClusterFromS3 (Maybe Text)

The KMS key identifier for an encrypted DB cluster. The KMS key identifier is the Amazon Resource Name (ARN) for the KMS encryption key. If you are creating a DB cluster with the same AWS account that owns the KMS encryption key used to encrypt the new DB cluster, then you can use the KMS key alias instead of the ARN for the KM encryption key. If the StorageEncrypted parameter is true, and you do not specify a value for the KmsKeyId parameter, then Amazon RDS will use your default encryption key. AWS KMS creates the default encryption key for your AWS account. Your AWS account has a different default encryption key for each AWS Region.

rdcfsPreferredBackupWindow :: Lens' RestoreDBClusterFromS3 (Maybe Text)

The daily time range during which automated backups are created if automated backups are enabled using the BackupRetentionPeriod parameter. Default: A 30-minute window selected at random from an 8-hour block of time per AWS Region. To see the time blocks available, see Adjusting the Preferred Maintenance Window in the Amazon RDS User Guide. Constraints: * Must be in the format hh24:mi-hh24:mi . * Times should be in Universal Coordinated Time (UTC). * Must not conflict with the preferred maintenance window. * Must be at least 30 minutes.

rdcfsBackupRetentionPeriod :: Lens' RestoreDBClusterFromS3 (Maybe Int)

The number of days for which automated backups of the restored DB cluster are retained. You must specify a minimum value of 1. Default: 1 Constraints: * Must be a value from 1 to 35

rdcfsVPCSecurityGroupIds :: Lens' RestoreDBClusterFromS3 [Text]

A list of EC2 VPC security groups to associate with the restored DB cluster.

rdcfsDatabaseName :: Lens' RestoreDBClusterFromS3 (Maybe Text)

The database name for the restored DB cluster.

rdcfsDBClusterParameterGroupName :: Lens' RestoreDBClusterFromS3 (Maybe Text)

The name of the DB cluster parameter group to associate with the restored DB cluster. If this argument is omitted, default.aurora5.6 will be used. Constraints: * Must be 1 to 255 alphanumeric characters * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens

rdcfsS3Prefix :: Lens' RestoreDBClusterFromS3 (Maybe Text)

The prefix for all of the file names that contain the data used to create the Amazon Aurora DB cluster. If you do not specify a SourceS3Prefix value, then the Amazon Aurora DB cluster is created by using all of the files in the Amazon S3 bucket.

rdcfsOptionGroupName :: Lens' RestoreDBClusterFromS3 (Maybe Text)

A value that indicates that the restored DB cluster should be associated with the specified option group. Permanent options cannot be removed from an option group. An option group cannot be removed from a DB cluster once it is associated with a DB cluster.

rdcfsTags :: Lens' RestoreDBClusterFromS3 [Tag]

Undocumented member.

rdcfsPort :: Lens' RestoreDBClusterFromS3 (Maybe Int)

The port number on which the instances in the restored DB cluster accept connections. Default: 3306

rdcfsEnableIAMDatabaseAuthentication :: Lens' RestoreDBClusterFromS3 (Maybe Bool)

A Boolean value that is true to enable mapping of AWS Identity and Access Management (IAM) accounts to database accounts, and otherwise false. Default: false

rdcfsDBClusterIdentifier :: Lens' RestoreDBClusterFromS3 Text

The name of the DB cluster to create from the source data in the S3 bucket. This parameter is isn't case-sensitive. Constraints: * Must contain from 1 to 63 alphanumeric characters or hyphens. * First character must be a letter. * Cannot end with a hyphen or contain two consecutive hyphens. Example: my-cluster1

rdcfsEngine :: Lens' RestoreDBClusterFromS3 Text

The name of the database engine to be used for the restored DB cluster. Valid Values: aurora

rdcfsMasterUsername :: Lens' RestoreDBClusterFromS3 Text

The name of the master user for the restored DB cluster. Constraints: * Must be 1 to 16 alphanumeric characters. * First character must be a letter. * Cannot be a reserved word for the chosen database engine.

rdcfsMasterUserPassword :: Lens' RestoreDBClusterFromS3 Text

The password for the master database user. This password can contain any printable ASCII character except "/", """, or "@". Constraints: Must contain from 8 to 41 characters.

rdcfsSourceEngine :: Lens' RestoreDBClusterFromS3 Text

The identifier for the database engine that was backed up to create the files stored in the Amazon S3 bucket. Valid values: mysql

rdcfsSourceEngineVersion :: Lens' RestoreDBClusterFromS3 Text

The version of the database that the backup files were created from. MySQL version 5.5 and 5.6 are supported. Example: 5.6.22

rdcfsS3BucketName :: Lens' RestoreDBClusterFromS3 Text

The name of the Amazon S3 bucket that contains the data used to create the Amazon Aurora DB cluster.

rdcfsS3IngestionRoleARN :: Lens' RestoreDBClusterFromS3 Text

The Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role that authorizes Amazon RDS to access the Amazon S3 bucket on your behalf.

Destructuring the Response

restoreDBClusterFromS3Response

Arguments

:: Int
rdcfsrsResponseStatus

-> RestoreDBClusterFromS3Response

Creates a value of RestoreDBClusterFromS3Response with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rdcfsrsDBCluster - Undocumented member.

rdcfsrsResponseStatus - -- | The response status code.

data RestoreDBClusterFromS3Response

See: restoreDBClusterFromS3Response smart constructor.

Response Lenses

rdcfsrsDBCluster :: Lens' RestoreDBClusterFromS3Response (Maybe DBCluster)

Undocumented member.

rdcfsrsResponseStatus :: Lens' RestoreDBClusterFromS3Response Int

- | The response status code.