Network.AWS.RDS.RestoreDBClusterFromSnapshot

Description

Creates a new DB cluster from a DB cluster snapshot. The target DB cluster is created from the source DB cluster restore point with the same configuration as the original source DB cluster, except that the new DB cluster is created with the default security group.

For more information on Amazon Aurora, see Aurora on Amazon RDS in the Amazon RDS User Guide.

Creating a Request

restoreDBClusterFromSnapshot

Arguments

:: Text
rdbcfsDBClusterIdentifier

-> Text
rdbcfsSnapshotIdentifier

-> Text
rdbcfsEngine

-> RestoreDBClusterFromSnapshot

Creates a value of RestoreDBClusterFromSnapshot with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rdbcfsEngineVersion - The version of the database engine to use for the new DB cluster.

rdbcfsDBSubnetGroupName - The name of the DB subnet group to use for the new DB cluster. Constraints: Must contain no more than 255 alphanumeric characters, periods, underscores, spaces, or hyphens. Must not be default. Example: mySubnetgroup

rdbcfsAvailabilityZones - Provides the list of EC2 Availability Zones that instances in the restored DB cluster can be created in.

rdbcfsKMSKeyId - The KMS key identifier to use when restoring an encrypted DB cluster from a DB cluster snapshot. The KMS key identifier is the Amazon Resource Name (ARN) for the KMS encryption key. If you are restoring a DB cluster with the same AWS account that owns the KMS encryption key used to encrypt the new DB cluster, then you can use the KMS key alias instead of the ARN for the KMS encryption key. If you do not specify a value for the KmsKeyId parameter, then the following will occur: * If the DB cluster snapshot is encrypted, then the restored DB cluster is encrypted using the KMS key that was used to encrypt the DB cluster snapshot. * If the DB cluster snapshot is not encrypted, then the restored DB cluster is encrypted using the specified encryption key.

rdbcfsVPCSecurityGroupIds - A list of VPC security groups that the new DB cluster will belong to.

rdbcfsDatabaseName - The database name for the restored DB cluster.

rdbcfsOptionGroupName - The name of the option group to use for the restored DB cluster.

rdbcfsTags - The tags to be assigned to the restored DB cluster.

rdbcfsPort - The port number on which the new DB cluster accepts connections. Constraints: Value must be 1150-65535 Default: The same port as the original DB cluster.

rdbcfsEnableIAMDatabaseAuthentication - A Boolean value that is true to enable mapping of AWS Identity and Access Management (IAM) accounts to database accounts, and otherwise false. Default: false

rdbcfsDBClusterIdentifier - The name of the DB cluster to create from the DB cluster snapshot. This parameter isn't case-sensitive. Constraints: * Must contain from 1 to 255 alphanumeric characters or hyphens * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens Example: my-snapshot-id

rdbcfsSnapshotIdentifier - The identifier for the DB cluster snapshot to restore from. Constraints: * Must contain from 1 to 63 alphanumeric characters or hyphens * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens

rdbcfsEngine - The database engine to use for the new DB cluster. Default: The same as source Constraint: Must be compatible with the engine of the source

data RestoreDBClusterFromSnapshot

See: restoreDBClusterFromSnapshot smart constructor.

Request Lenses

rdbcfsEngineVersion :: Lens' RestoreDBClusterFromSnapshot (Maybe Text)

The version of the database engine to use for the new DB cluster.

rdbcfsDBSubnetGroupName :: Lens' RestoreDBClusterFromSnapshot (Maybe Text)

The name of the DB subnet group to use for the new DB cluster. Constraints: Must contain no more than 255 alphanumeric characters, periods, underscores, spaces, or hyphens. Must not be default. Example: mySubnetgroup

rdbcfsAvailabilityZones :: Lens' RestoreDBClusterFromSnapshot [Text]

Provides the list of EC2 Availability Zones that instances in the restored DB cluster can be created in.

rdbcfsKMSKeyId :: Lens' RestoreDBClusterFromSnapshot (Maybe Text)

The KMS key identifier to use when restoring an encrypted DB cluster from a DB cluster snapshot. The KMS key identifier is the Amazon Resource Name (ARN) for the KMS encryption key. If you are restoring a DB cluster with the same AWS account that owns the KMS encryption key used to encrypt the new DB cluster, then you can use the KMS key alias instead of the ARN for the KMS encryption key. If you do not specify a value for the KmsKeyId parameter, then the following will occur: * If the DB cluster snapshot is encrypted, then the restored DB cluster is encrypted using the KMS key that was used to encrypt the DB cluster snapshot. * If the DB cluster snapshot is not encrypted, then the restored DB cluster is encrypted using the specified encryption key.

rdbcfsVPCSecurityGroupIds :: Lens' RestoreDBClusterFromSnapshot [Text]

A list of VPC security groups that the new DB cluster will belong to.

rdbcfsDatabaseName :: Lens' RestoreDBClusterFromSnapshot (Maybe Text)

The database name for the restored DB cluster.

rdbcfsOptionGroupName :: Lens' RestoreDBClusterFromSnapshot (Maybe Text)

The name of the option group to use for the restored DB cluster.

rdbcfsTags :: Lens' RestoreDBClusterFromSnapshot [Tag]

The tags to be assigned to the restored DB cluster.

rdbcfsPort :: Lens' RestoreDBClusterFromSnapshot (Maybe Int)

The port number on which the new DB cluster accepts connections. Constraints: Value must be 1150-65535 Default: The same port as the original DB cluster.

rdbcfsEnableIAMDatabaseAuthentication :: Lens' RestoreDBClusterFromSnapshot (Maybe Bool)

A Boolean value that is true to enable mapping of AWS Identity and Access Management (IAM) accounts to database accounts, and otherwise false. Default: false

rdbcfsDBClusterIdentifier :: Lens' RestoreDBClusterFromSnapshot Text

The name of the DB cluster to create from the DB cluster snapshot. This parameter isn't case-sensitive. Constraints: * Must contain from 1 to 255 alphanumeric characters or hyphens * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens Example: my-snapshot-id

rdbcfsSnapshotIdentifier :: Lens' RestoreDBClusterFromSnapshot Text

The identifier for the DB cluster snapshot to restore from. Constraints: * Must contain from 1 to 63 alphanumeric characters or hyphens * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens

rdbcfsEngine :: Lens' RestoreDBClusterFromSnapshot Text

The database engine to use for the new DB cluster. Default: The same as source Constraint: Must be compatible with the engine of the source

Destructuring the Response

restoreDBClusterFromSnapshotResponse

Arguments

:: Int
rdbcfsrsResponseStatus

-> RestoreDBClusterFromSnapshotResponse

Creates a value of RestoreDBClusterFromSnapshotResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rdbcfsrsDBCluster - Undocumented member.

rdbcfsrsResponseStatus - -- | The response status code.

data RestoreDBClusterFromSnapshotResponse

See: restoreDBClusterFromSnapshotResponse smart constructor.

Response Lenses

rdbcfsrsDBCluster :: Lens' RestoreDBClusterFromSnapshotResponse (Maybe DBCluster)

Undocumented member.

rdbcfsrsResponseStatus :: Lens' RestoreDBClusterFromSnapshotResponse Int

- | The response status code.