Network.AWS.RDS.RestoreDBClusterToPointInTime

Description

Restores a DB cluster to an arbitrary point in time. Users can restore to any point in time before LatestRestorableTime for up to BackupRetentionPeriod days. The target DB cluster is created from the source DB cluster with the same configuration as the original DB cluster, except that the new DB cluster is created with the default DB security group.

For more information on Amazon Aurora, see Aurora on Amazon RDS in the Amazon RDS User Guide.

Creating a Request

restoreDBClusterToPointInTime

Arguments

:: Text
rdctpitDBClusterIdentifier

-> Text
rdctpitSourceDBClusterIdentifier

-> RestoreDBClusterToPointInTime

Creates a value of RestoreDBClusterToPointInTime with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rdctpitUseLatestRestorableTime - A value that is set to true to restore the DB cluster to the latest restorable backup time, and false otherwise. Default: false Constraints: Cannot be specified if RestoreToTime parameter is provided.

rdctpitDBSubnetGroupName - The DB subnet group name to use for the new DB cluster. Constraints: Must contain no more than 255 alphanumeric characters, periods, underscores, spaces, or hyphens. Must not be default. Example: mySubnetgroup

rdctpitKMSKeyId - The KMS key identifier to use when restoring an encrypted DB cluster from an encrypted DB cluster. The KMS key identifier is the Amazon Resource Name (ARN) for the KMS encryption key. If you are restoring a DB cluster with the same AWS account that owns the KMS encryption key used to encrypt the new DB cluster, then you can use the KMS key alias instead of the ARN for the KMS encryption key. You can restore to a new DB cluster and encrypt the new DB cluster with a KMS key that is different than the KMS key used to encrypt the source DB cluster. The new DB cluster will be encrypted with the KMS key identified by the KmsKeyId parameter. If you do not specify a value for the KmsKeyId parameter, then the following will occur: * If the DB cluster is encrypted, then the restored DB cluster is encrypted using the KMS key that was used to encrypt the source DB cluster. * If the DB cluster is not encrypted, then the restored DB cluster is not encrypted. If DBClusterIdentifier refers to a DB cluster that is not encrypted, then the restore request is rejected.

rdctpitVPCSecurityGroupIds - A list of VPC security groups that the new DB cluster belongs to.

rdctpitRestoreType - The type of restore to be performed. You can specify one of the following values: * full-copy - The new DB cluster is restored as a full copy of the source DB cluster. * copy-on-write - The new DB cluster is restored as a clone of the source DB cluster. Constraints: You cannot specify copy-on-write if the engine version of the source DB cluster is earlier than 1.11. If you don't specify a RestoreType value, then the new DB cluster is restored as a full copy of the source DB cluster.

rdctpitOptionGroupName - The name of the option group for the new DB cluster.

rdctpitRestoreToTime - The date and time to restore the DB cluster to. Valid Values: Value must be a time in Universal Coordinated Time (UTC) format Constraints: * Must be before the latest restorable time for the DB instance * Must be specified if UseLatestRestorableTime parameter is not provided * Cannot be specified if UseLatestRestorableTime parameter is true * Cannot be specified if RestoreType parameter is copy-on-write Example: 2015-03-07T23:45:00Z

rdctpitTags - Undocumented member.

rdctpitPort - The port number on which the new DB cluster accepts connections. Constraints: Value must be 1150-65535 Default: The same port as the original DB cluster.

rdctpitEnableIAMDatabaseAuthentication - A Boolean value that is true to enable mapping of AWS Identity and Access Management (IAM) accounts to database accounts, and otherwise false. Default: false

rdctpitDBClusterIdentifier - The name of the new DB cluster to be created. Constraints: * Must contain from 1 to 63 alphanumeric characters or hyphens * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens

rdctpitSourceDBClusterIdentifier - The identifier of the source DB cluster from which to restore. Constraints: * Must be the identifier of an existing database instance * Must contain from 1 to 63 alphanumeric characters or hyphens * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens

data RestoreDBClusterToPointInTime

See: restoreDBClusterToPointInTime smart constructor.

Request Lenses

rdctpitUseLatestRestorableTime :: Lens' RestoreDBClusterToPointInTime (Maybe Bool)

A value that is set to true to restore the DB cluster to the latest restorable backup time, and false otherwise. Default: false Constraints: Cannot be specified if RestoreToTime parameter is provided.

rdctpitDBSubnetGroupName :: Lens' RestoreDBClusterToPointInTime (Maybe Text)

The DB subnet group name to use for the new DB cluster. Constraints: Must contain no more than 255 alphanumeric characters, periods, underscores, spaces, or hyphens. Must not be default. Example: mySubnetgroup

rdctpitKMSKeyId :: Lens' RestoreDBClusterToPointInTime (Maybe Text)

The KMS key identifier to use when restoring an encrypted DB cluster from an encrypted DB cluster. The KMS key identifier is the Amazon Resource Name (ARN) for the KMS encryption key. If you are restoring a DB cluster with the same AWS account that owns the KMS encryption key used to encrypt the new DB cluster, then you can use the KMS key alias instead of the ARN for the KMS encryption key. You can restore to a new DB cluster and encrypt the new DB cluster with a KMS key that is different than the KMS key used to encrypt the source DB cluster. The new DB cluster will be encrypted with the KMS key identified by the KmsKeyId parameter. If you do not specify a value for the KmsKeyId parameter, then the following will occur: * If the DB cluster is encrypted, then the restored DB cluster is encrypted using the KMS key that was used to encrypt the source DB cluster. * If the DB cluster is not encrypted, then the restored DB cluster is not encrypted. If DBClusterIdentifier refers to a DB cluster that is not encrypted, then the restore request is rejected.

rdctpitVPCSecurityGroupIds :: Lens' RestoreDBClusterToPointInTime [Text]

A list of VPC security groups that the new DB cluster belongs to.

rdctpitRestoreType :: Lens' RestoreDBClusterToPointInTime (Maybe Text)

The type of restore to be performed. You can specify one of the following values: * full-copy - The new DB cluster is restored as a full copy of the source DB cluster. * copy-on-write - The new DB cluster is restored as a clone of the source DB cluster. Constraints: You cannot specify copy-on-write if the engine version of the source DB cluster is earlier than 1.11. If you don't specify a RestoreType value, then the new DB cluster is restored as a full copy of the source DB cluster.

rdctpitOptionGroupName :: Lens' RestoreDBClusterToPointInTime (Maybe Text)

The name of the option group for the new DB cluster.

rdctpitRestoreToTime :: Lens' RestoreDBClusterToPointInTime (Maybe UTCTime)

The date and time to restore the DB cluster to. Valid Values: Value must be a time in Universal Coordinated Time (UTC) format Constraints: * Must be before the latest restorable time for the DB instance * Must be specified if UseLatestRestorableTime parameter is not provided * Cannot be specified if UseLatestRestorableTime parameter is true * Cannot be specified if RestoreType parameter is copy-on-write Example: 2015-03-07T23:45:00Z

rdctpitTags :: Lens' RestoreDBClusterToPointInTime [Tag]

Undocumented member.

rdctpitPort :: Lens' RestoreDBClusterToPointInTime (Maybe Int)

The port number on which the new DB cluster accepts connections. Constraints: Value must be 1150-65535 Default: The same port as the original DB cluster.

rdctpitEnableIAMDatabaseAuthentication :: Lens' RestoreDBClusterToPointInTime (Maybe Bool)

A Boolean value that is true to enable mapping of AWS Identity and Access Management (IAM) accounts to database accounts, and otherwise false. Default: false

rdctpitDBClusterIdentifier :: Lens' RestoreDBClusterToPointInTime Text

The name of the new DB cluster to be created. Constraints: * Must contain from 1 to 63 alphanumeric characters or hyphens * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens

rdctpitSourceDBClusterIdentifier :: Lens' RestoreDBClusterToPointInTime Text

The identifier of the source DB cluster from which to restore. Constraints: * Must be the identifier of an existing database instance * Must contain from 1 to 63 alphanumeric characters or hyphens * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens

Destructuring the Response

restoreDBClusterToPointInTimeResponse

Arguments

:: Int
rdctpitrsResponseStatus

-> RestoreDBClusterToPointInTimeResponse

Creates a value of RestoreDBClusterToPointInTimeResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rdctpitrsDBCluster - Undocumented member.

rdctpitrsResponseStatus - -- | The response status code.

data RestoreDBClusterToPointInTimeResponse

See: restoreDBClusterToPointInTimeResponse smart constructor.

Response Lenses

rdctpitrsDBCluster :: Lens' RestoreDBClusterToPointInTimeResponse (Maybe DBCluster)

Undocumented member.

rdctpitrsResponseStatus :: Lens' RestoreDBClusterToPointInTimeResponse Int

- | The response status code.