Network.AWS.RDS.RestoreDBInstanceFromDBSnapshot

Description

Creates a new DB instance from a DB snapshot. The target database is created from the source database restore point with the same configuration as the original source database, except that the new RDS instance is created with the default security group. https:/rds.amazon.com ?Action=RestoreDBInstanceFromDBSnapshot &DBSnapshotIdentifier=mydbsnapshot &DBInstanceIdentifier=myrestoreddbinstance &Version=2013-05-15 &Timestamp=2011-05-23T06%3A47%3A11.071Z &SignatureVersion=2 &SignatureMethod=HmacSHA256 &AWSAccessKeyId= &Signature= mysql 1 false general-public-license creating 5.1.50 myrestoreddbinstance in-sync default.mysql5.1 active default 00:00-00:30 true sat:07:30-sat:08:00 10 db.m1.large master 7ca622e8-8508-11e0-bd9b-a7b1ece36d51.

Request

data RestoreDBInstanceFromDBSnapshot

Request constructor

restoreDBInstanceFromDBSnapshot

Arguments

:: Text
rdbifdbsDBInstanceIdentifier

-> Text
rdbifdbsDBSnapshotIdentifier

-> RestoreDBInstanceFromDBSnapshot

Smart constructor for the minimum required parameters to construct a valid RestoreDBInstanceFromDBSnapshot request.

The fields accessible through corresponding lenses are:

DBInstanceIdentifier :: Text

DBSnapshotIdentifier :: Text

DBInstanceClass :: Maybe Text

Port :: Maybe Integer

AvailabilityZone :: Maybe Text

DBSubnetGroupName :: Maybe Text

MultiAZ :: Maybe Bool

PubliclyAccessible :: Maybe Bool

AutoMinorVersionUpgrade :: Maybe Bool

LicenseModel :: Maybe Text

DBName :: Maybe Text

Engine :: Maybe Text

Iops :: Maybe Integer

OptionGroupName :: Maybe Text

Tags :: [Tag]

Request lenses

rdbifdbsDBInstanceIdentifier :: Lens' RestoreDBInstanceFromDBSnapshot Text

The identifier for the DB snapshot to restore from. Constraints: Must contain from 1 to 63 alphanumeric characters or hyphens First character must be a letter Cannot end with a hyphen or contain two consecutive hyphens.

rdbifdbsDBSnapshotIdentifier :: Lens' RestoreDBInstanceFromDBSnapshot Text

Name of the DB instance to create from the DB snapshot. This parameter isn't case sensitive. Constraints: Must contain from 1 to 255 alphanumeric characters or hyphens First character must be a letter Cannot end with a hyphen or contain two consecutive hyphens Example: my-snapshot-id.

rdbifdbsDBInstanceClass :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Text)

The compute and memory capacity of the Amazon RDS DB instance. Valid Values: db.t1.micro | db.m1.small | db.m1.medium | db.m1.large | db.m1.xlarge | db.m2.2xlarge | db.m2.4xlarge.

rdbifdbsPort :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Integer)

The port number on which the database accepts connections. Default: The same port as the original DB instance Constraints: Value must be 1150-65535.

rdbifdbsAvailabilityZone :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Text)

The EC2 Availability Zone that the database instance will be created in. Default: A random, system-chosen Availability Zone. Constraint: You cannot specify the AvailabilityZone parameter if the MultiAZ parameter is set to true. Example: us-east-1a.

rdbifdbsDBSubnetGroupName :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Text)

The DB subnet group name to use for the new instance.

rdbifdbsMultiAZ :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Bool)

Specifies if the DB instance is a Multi-AZ deployment. Constraint: You cannot specify the AvailabilityZone parameter if the MultiAZ parameter is set to true.

rdbifdbsPubliclyAccessible :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Bool)

Specifies the accessibility options for the DB instance. A value of true specifies an Internet-facing instance with a publicly resolvable DNS name, which resolves to a public IP address. A value of false specifies an internal instance with a DNS name that resolves to a private IP address. Default: The default behavior varies depending on whether a VPC has been requested or not. The following list shows the default behavior in each case. Default VPC:true VPC:false If no DB subnet group has been specified as part of the request and the PubliclyAccessible value has not been set, the DB instance will be publicly accessible. If a specific DB subnet group has been specified as part of the request and the PubliclyAccessible value has not been set, the DB instance will be private.

rdbifdbsAutoMinorVersionUpgrade :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Bool)

Indicates that minor version upgrades will be applied automatically to the DB instance during the maintenance window.

rdbifdbsLicenseModel :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Text)

License model information for the restored DB instance. Default: Same as source. Valid values: license-included | bring-your-own-license | general-public-license.

rdbifdbsDBName :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Text)

The database name for the restored DB instance. This parameter doesn't apply to the MySQL engine.

rdbifdbsEngine :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Text)

The database engine to use for the new instance. Default: The same as source Constraint: Must be compatible with the engine of the source Example: oracle-ee.

rdbifdbsIops :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Integer)

Specifies the amount of provisioned IOPS for the DB instance, expressed in I/O operations per second. If this parameter is not specified, the IOPS value will be taken from the backup. If this parameter is set to 0, the new instance will be converted to a non-PIOPS instance, which will take additional time, though your DB instance will be available for connections before the conversion starts. Constraints: Must be an integer greater than 1000.

rdbifdbsOptionGroupName :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Text)

The name of the option group to be used for the restored DB instance. cannot be removed from an option group while DB instances are associated with the option group. --> Permanent options, such as the TDE option for Oracle Advanced Security TDE, cannot be removed from an option group, and that option group cannot be removed from a DB instance once it is associated with a DB instance.

rdbifdbsTags :: Lens' RestoreDBInstanceFromDBSnapshot [Tag]

A list of tags.

Response

data RestoreDBInstanceFromDBSnapshotResponse

Response constructor

restoreDBInstanceFromDBSnapshotResponse :: RestoreDBInstanceFromDBSnapshotResponse

Smart constructor for the minimum required parameters to construct a valid RestoreDBInstanceFromDBSnapshotResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

DBInstance :: Maybe DBInstance

Response lenses

rdbifdbsrDBInstance :: Lens' RestoreDBInstanceFromDBSnapshotResponse (Maybe DBInstance)

Contains the result of a successful invocation of the following actions: CreateDBInstance DeleteDBInstance ModifyDBInstance This data type is used as a response element in the DescribeDBInstances action.