Network.AWS.RDS.RevokeDBSecurityGroupIngress

Description

Revokes ingress from a DBSecurityGroup for previously authorized IP ranges or EC2 or VPC Security Groups. Required parameters for this API are one of CIDRIP, EC2SecurityGroupId for VPC, or (EC2SecurityGroupOwnerId and either EC2SecurityGroupName or EC2SecurityGroupId).

Creating a Request

revokeDBSecurityGroupIngress

Arguments

:: Text
rdsgiDBSecurityGroupName

-> RevokeDBSecurityGroupIngress

Creates a value of RevokeDBSecurityGroupIngress with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rdsgiEC2SecurityGroupOwnerId - The AWS Account Number of the owner of the EC2 security group specified in the EC2SecurityGroupName parameter. The AWS Access Key ID is not an acceptable value. For VPC DB security groups, EC2SecurityGroupId must be provided. Otherwise, EC2SecurityGroupOwnerId and either EC2SecurityGroupName or EC2SecurityGroupId must be provided.

rdsgiEC2SecurityGroupName - The name of the EC2 security group to revoke access from. For VPC DB security groups, EC2SecurityGroupId must be provided. Otherwise, EC2SecurityGroupOwnerId and either EC2SecurityGroupName or EC2SecurityGroupId must be provided.

rdsgiCIdRIP - The IP range to revoke access from. Must be a valid CIDR range. If CIDRIP is specified, EC2SecurityGroupName , EC2SecurityGroupId and EC2SecurityGroupOwnerId cannot be provided.

rdsgiEC2SecurityGroupId - The id of the EC2 security group to revoke access from. For VPC DB security groups, EC2SecurityGroupId must be provided. Otherwise, EC2SecurityGroupOwnerId and either EC2SecurityGroupName or EC2SecurityGroupId must be provided.

rdsgiDBSecurityGroupName - The name of the DB security group to revoke ingress from.

data RevokeDBSecurityGroupIngress

See: revokeDBSecurityGroupIngress smart constructor.

Request Lenses

rdsgiEC2SecurityGroupOwnerId :: Lens' RevokeDBSecurityGroupIngress (Maybe Text)

The AWS Account Number of the owner of the EC2 security group specified in the EC2SecurityGroupName parameter. The AWS Access Key ID is not an acceptable value. For VPC DB security groups, EC2SecurityGroupId must be provided. Otherwise, EC2SecurityGroupOwnerId and either EC2SecurityGroupName or EC2SecurityGroupId must be provided.

rdsgiEC2SecurityGroupName :: Lens' RevokeDBSecurityGroupIngress (Maybe Text)

The name of the EC2 security group to revoke access from. For VPC DB security groups, EC2SecurityGroupId must be provided. Otherwise, EC2SecurityGroupOwnerId and either EC2SecurityGroupName or EC2SecurityGroupId must be provided.

rdsgiCIdRIP :: Lens' RevokeDBSecurityGroupIngress (Maybe Text)

The IP range to revoke access from. Must be a valid CIDR range. If CIDRIP is specified, EC2SecurityGroupName , EC2SecurityGroupId and EC2SecurityGroupOwnerId cannot be provided.

rdsgiEC2SecurityGroupId :: Lens' RevokeDBSecurityGroupIngress (Maybe Text)

The id of the EC2 security group to revoke access from. For VPC DB security groups, EC2SecurityGroupId must be provided. Otherwise, EC2SecurityGroupOwnerId and either EC2SecurityGroupName or EC2SecurityGroupId must be provided.

rdsgiDBSecurityGroupName :: Lens' RevokeDBSecurityGroupIngress Text

The name of the DB security group to revoke ingress from.

Destructuring the Response

revokeDBSecurityGroupIngressResponse

Arguments

:: Int
rdsgirsResponseStatus

-> RevokeDBSecurityGroupIngressResponse

Creates a value of RevokeDBSecurityGroupIngressResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rdsgirsDBSecurityGroup - Undocumented member.

rdsgirsResponseStatus - -- | The response status code.

data RevokeDBSecurityGroupIngressResponse

See: revokeDBSecurityGroupIngressResponse smart constructor.

Response Lenses

rdsgirsDBSecurityGroup :: Lens' RevokeDBSecurityGroupIngressResponse (Maybe DBSecurityGroup)

Undocumented member.

rdsgirsResponseStatus :: Lens' RevokeDBSecurityGroupIngressResponse Int

- | The response status code.