Network.AWS.RDS.Types

Service

data RDS

Version 2014-10-31 of the Amazon Relational Database Service service.

Error

data RESTError :: *

XML

ns :: Text

PendingMaintenanceAction

data PendingMaintenanceAction

pendingMaintenanceAction :: PendingMaintenanceAction

PendingMaintenanceAction constructor.

The fields accessible through corresponding lenses are:

pmaAction :: Maybe Text

pmaAutoAppliedAfterDate :: Maybe UTCTime

pmaCurrentApplyDate :: Maybe UTCTime

pmaDescription :: Maybe Text

pmaForcedApplyDate :: Maybe UTCTime

pmaOptInStatus :: Maybe Text

pmaAction :: Lens' PendingMaintenanceAction (Maybe Text)

The type of pending maintenance action that is available for the resource.

pmaAutoAppliedAfterDate :: Lens' PendingMaintenanceAction (Maybe UTCTime)

The date of the maintenance window when the action will be applied. The maintenance action will be applied to the resource during its first maintenance window after this date. If this date is specified, any 'next-maintenance' opt-in requests are ignored.

pmaCurrentApplyDate :: Lens' PendingMaintenanceAction (Maybe UTCTime)

The effective date when the pending maintenance action will be applied to the resource. This date takes into account opt-in requests received from the ApplyPendingMaintenanceAction API, the AutoAppliedAfterDate, and the ForcedApplyDate. This value is blank if an opt-in request has not been received and nothing has been specified as AutoAppliedAfterDate or ForcedApplyDate.

pmaDescription :: Lens' PendingMaintenanceAction (Maybe Text)

A description providing more detail about the maintenance action.

pmaForcedApplyDate :: Lens' PendingMaintenanceAction (Maybe UTCTime)

The date when the maintenance action will be automatically applied. The maintenance action will be applied to the resource on this date regardless of the maintenance window for the resource. If this date is specified, any immediate opt-in requests are ignored.

pmaOptInStatus :: Lens' PendingMaintenanceAction (Maybe Text)

Indicates the type of opt-in request that has been received for the resource.

OptionGroup

data OptionGroup

optionGroup :: OptionGroup

OptionGroup constructor.

The fields accessible through corresponding lenses are:

ogAllowsVpcAndNonVpcInstanceMemberships :: Maybe Bool

ogEngineName :: Maybe Text

ogMajorEngineVersion :: Maybe Text

ogOptionGroupDescription :: Maybe Text

ogOptionGroupName :: Maybe Text

ogOptions :: [Option]

ogVpcId :: Maybe Text

ogAllowsVpcAndNonVpcInstanceMemberships :: Lens' OptionGroup (Maybe Bool)

Indicates whether this option group can be applied to both VPC and non-VPC instances. The value true indicates the option group can be applied to both VPC and non-VPC instances.

ogEngineName :: Lens' OptionGroup (Maybe Text)

Engine name that this option group can be applied to.

ogMajorEngineVersion :: Lens' OptionGroup (Maybe Text)

Indicates the major engine version associated with this option group.

ogOptionGroupDescription :: Lens' OptionGroup (Maybe Text)

Provides a description of the option group.

ogOptionGroupName :: Lens' OptionGroup (Maybe Text)

Specifies the name of the option group.

ogOptions :: Lens' OptionGroup [Option]

Indicates what options are available in the option group.

ogVpcId :: Lens' OptionGroup (Maybe Text)

If AllowsVpcAndNonVpcInstanceMemberships is false, this field is blank. If AllowsVpcAndNonVpcInstanceMemberships is true and this field is blank, then this option group can be applied to both VPC and non-VPC instances. If this field contains a value, then this option group can only be applied to instances that are in the VPC indicated by this field.

DBParameterGroupStatus

data DBParameterGroupStatus

dbparameterGroupStatus :: DBParameterGroupStatus

DBParameterGroupStatus constructor.

The fields accessible through corresponding lenses are:

dbpgsDBParameterGroupName :: Maybe Text

dbpgsParameterApplyStatus :: Maybe Text

dbpgsDBParameterGroupName :: Lens' DBParameterGroupStatus (Maybe Text)

The name of the DP parameter group.

dbpgsParameterApplyStatus :: Lens' DBParameterGroupStatus (Maybe Text)

The status of parameter updates.

Event

data Event

event :: Event

Event constructor.

The fields accessible through corresponding lenses are:

eDate :: Maybe UTCTime

eEventCategories :: [Text]

eMessage :: Maybe Text

eSourceIdentifier :: Maybe Text

eSourceType :: Maybe SourceType

eDate :: Lens' Event (Maybe UTCTime)

Specifies the date and time of the event.

eEventCategories :: Lens' Event [Text]

Specifies the category for the event.

eMessage :: Lens' Event (Maybe Text)

Provides the text of this event.

eSourceIdentifier :: Lens' Event (Maybe Text)

Provides the identifier for the source of the event.

eSourceType :: Lens' Event (Maybe SourceType)

Specifies the source type for this event.

DBSecurityGroup

data DBSecurityGroup

dbsecurityGroup :: DBSecurityGroup

DBSecurityGroup constructor.

The fields accessible through corresponding lenses are:

dbsgDBSecurityGroupDescription :: Maybe Text

dbsgDBSecurityGroupName :: Maybe Text

dbsgEC2SecurityGroups :: [EC2SecurityGroup]

dbsgIPRanges :: [IPRange]

dbsgOwnerId :: Maybe Text

dbsgVpcId :: Maybe Text

dbsgDBSecurityGroupDescription :: Lens' DBSecurityGroup (Maybe Text)

Provides the description of the DB security group.

dbsgDBSecurityGroupName :: Lens' DBSecurityGroup (Maybe Text)

Specifies the name of the DB security group.

dbsgEC2SecurityGroups :: Lens' DBSecurityGroup [EC2SecurityGroup]

Contains a list of EC2SecurityGroup elements.

dbsgIPRanges :: Lens' DBSecurityGroup [IPRange]

Contains a list of IPRange elements.

dbsgOwnerId :: Lens' DBSecurityGroup (Maybe Text)

Provides the AWS ID of the owner of a specific DB security group.

dbsgVpcId :: Lens' DBSecurityGroup (Maybe Text)

Provides the VpcId of the DB security group.

Tag

data Tag

tag :: Tag

Tag constructor.

The fields accessible through corresponding lenses are:

tagKey :: Maybe Text

tagValue :: Maybe Text

tagKey :: Lens' Tag (Maybe Text)

A key is the required name of the tag. The string value can be from 1 to 128 Unicode characters in length and cannot be prefixed with "aws:" or "rds:". The string may only contain only the set of Unicode letters, digits, white-space, '_', ., /, '=', +, - (Java regex: "^([\p{L}\p{Z}\p{N}_.:/=+\-]*)$").

tagValue :: Lens' Tag (Maybe Text)

A value is the optional value of the tag. The string value can be from 1 to 256 Unicode characters in length and cannot be prefixed with "aws:" or "rds:". The string may only contain only the set of Unicode letters, digits, white-space, '_', ., /, '=', +, - (Java regex: "^([\p{L}\p{Z}\p{N}_.:/=+\-]*)$").

DBEngineVersion

data DBEngineVersion

dbengineVersion :: DBEngineVersion

DBEngineVersion constructor.

The fields accessible through corresponding lenses are:

dbevDBEngineDescription :: Maybe Text

dbevDBEngineVersionDescription :: Maybe Text

dbevDBParameterGroupFamily :: Maybe Text

dbevDefaultCharacterSet :: Maybe CharacterSet

dbevEngine :: Maybe Text

dbevEngineVersion :: Maybe Text

dbevSupportedCharacterSets :: [CharacterSet]

dbevDBEngineDescription :: Lens' DBEngineVersion (Maybe Text)

The description of the database engine.

dbevDBEngineVersionDescription :: Lens' DBEngineVersion (Maybe Text)

The description of the database engine version.

dbevDBParameterGroupFamily :: Lens' DBEngineVersion (Maybe Text)

The name of the DB parameter group family for the database engine.

dbevDefaultCharacterSet :: Lens' DBEngineVersion (Maybe CharacterSet)

The default character set for new instances of this engine version, if the CharacterSetName parameter of the CreateDBInstance API is not specified.

dbevEngine :: Lens' DBEngineVersion (Maybe Text)

The name of the database engine.

dbevEngineVersion :: Lens' DBEngineVersion (Maybe Text)

The version number of the database engine.

dbevSupportedCharacterSets :: Lens' DBEngineVersion [CharacterSet]

A list of the character sets supported by this engine for the CharacterSetName parameter of the CreateDBInstance API.

DBSnapshot

data DBSnapshot

dbsnapshot :: DBSnapshot

DBSnapshot constructor.

The fields accessible through corresponding lenses are:

dbsAllocatedStorage :: Maybe Int

dbsAvailabilityZone :: Maybe Text

dbsDBInstanceIdentifier :: Maybe Text

dbsDBSnapshotIdentifier :: Maybe Text

dbsEncrypted :: Maybe Bool

dbsEngine :: Maybe Text

dbsEngineVersion :: Maybe Text

dbsInstanceCreateTime :: Maybe UTCTime

dbsIops :: Maybe Int

dbsKmsKeyId :: Maybe Text

dbsLicenseModel :: Maybe Text

dbsMasterUsername :: Maybe Text

dbsOptionGroupName :: Maybe Text

dbsPercentProgress :: Maybe Int

dbsPort :: Maybe Int

dbsSnapshotCreateTime :: Maybe UTCTime

dbsSnapshotType :: Maybe Text

dbsSourceRegion :: Maybe Text

dbsStatus :: Maybe Text

dbsStorageType :: Maybe Text

dbsTdeCredentialArn :: Maybe Text

dbsVpcId :: Maybe Text

dbsAllocatedStorage :: Lens' DBSnapshot (Maybe Int)

Specifies the allocated storage size in gigabytes (GB).

dbsAvailabilityZone :: Lens' DBSnapshot (Maybe Text)

Specifies the name of the Availability Zone the DB instance was located in at the time of the DB snapshot.

dbsDBInstanceIdentifier :: Lens' DBSnapshot (Maybe Text)

Specifies the DB instance identifier of the DB instance this DB snapshot was created from.

dbsDBSnapshotIdentifier :: Lens' DBSnapshot (Maybe Text)

Specifies the identifier for the DB snapshot.

dbsEncrypted :: Lens' DBSnapshot (Maybe Bool)

Specifies whether the DB snapshot is encrypted.

dbsEngine :: Lens' DBSnapshot (Maybe Text)

Specifies the name of the database engine.

dbsEngineVersion :: Lens' DBSnapshot (Maybe Text)

Specifies the version of the database engine.

dbsInstanceCreateTime :: Lens' DBSnapshot (Maybe UTCTime)

Specifies the time (UTC) when the snapshot was taken.

dbsIops :: Lens' DBSnapshot (Maybe Int)

Specifies the Provisioned IOPS (I/O operations per second) value of the DB instance at the time of the snapshot.

dbsKmsKeyId :: Lens' DBSnapshot (Maybe Text)

If Encrypted is true, the KMS key identifier for the encrypted DB snapshot.

dbsLicenseModel :: Lens' DBSnapshot (Maybe Text)

License model information for the restored DB instance.

dbsMasterUsername :: Lens' DBSnapshot (Maybe Text)

Provides the master username for the DB snapshot.

dbsOptionGroupName :: Lens' DBSnapshot (Maybe Text)

Provides the option group name for the DB snapshot.

dbsPercentProgress :: Lens' DBSnapshot (Maybe Int)

The percentage of the estimated data that has been transferred.

dbsPort :: Lens' DBSnapshot (Maybe Int)

Specifies the port that the database engine was listening on at the time of the snapshot.

dbsSnapshotCreateTime :: Lens' DBSnapshot (Maybe UTCTime)

Provides the time (UTC) when the snapshot was taken.

dbsSnapshotType :: Lens' DBSnapshot (Maybe Text)

Provides the type of the DB snapshot.

dbsSourceRegion :: Lens' DBSnapshot (Maybe Text)

The region that the DB snapshot was created in or copied from.

dbsStatus :: Lens' DBSnapshot (Maybe Text)

Specifies the status of this DB snapshot.

dbsStorageType :: Lens' DBSnapshot (Maybe Text)

Specifies the storage type associated with DB Snapshot.

dbsTdeCredentialArn :: Lens' DBSnapshot (Maybe Text)

The ARN from the Key Store with which to associate the instance for TDE encryption.

dbsVpcId :: Lens' DBSnapshot (Maybe Text)

Provides the Vpc Id associated with the DB snapshot.

DBSecurityGroupMembership

data DBSecurityGroupMembership

dbsecurityGroupMembership :: DBSecurityGroupMembership

DBSecurityGroupMembership constructor.

The fields accessible through corresponding lenses are:

dbsgmDBSecurityGroupName :: Maybe Text

dbsgmStatus :: Maybe Text

dbsgmDBSecurityGroupName :: Lens' DBSecurityGroupMembership (Maybe Text)

The name of the DB security group.

dbsgmStatus :: Lens' DBSecurityGroupMembership (Maybe Text)

The status of the DB security group.

EC2SecurityGroup

data EC2SecurityGroup

ec2SecurityGroup :: EC2SecurityGroup

EC2SecurityGroup constructor.

The fields accessible through corresponding lenses are:

ecsgEC2SecurityGroupId :: Maybe Text

ecsgEC2SecurityGroupName :: Maybe Text

ecsgEC2SecurityGroupOwnerId :: Maybe Text

ecsgStatus :: Maybe Text

ecsgEC2SecurityGroupId :: Lens' EC2SecurityGroup (Maybe Text)

Specifies the id of the EC2 security group.

ecsgEC2SecurityGroupName :: Lens' EC2SecurityGroup (Maybe Text)

Specifies the name of the EC2 security group.

ecsgEC2SecurityGroupOwnerId :: Lens' EC2SecurityGroup (Maybe Text)

Specifies the AWS ID of the owner of the EC2 security group specified in the EC2SecurityGroupName field.

ecsgStatus :: Lens' EC2SecurityGroup (Maybe Text)

Provides the status of the EC2 security group. Status can be "authorizing", "authorized", "revoking", and "revoked".

SourceType

data SourceType

Constructors

DbInstance
db-instance

DbParameterGroup
db-parameter-group

DbSecurityGroup
db-security-group

DbSnapshot
db-snapshot

ResourcePendingMaintenanceActions

data ResourcePendingMaintenanceActions

resourcePendingMaintenanceActions :: ResourcePendingMaintenanceActions

ResourcePendingMaintenanceActions constructor.

The fields accessible through corresponding lenses are:

rpmaPendingMaintenanceActionDetails :: [PendingMaintenanceAction]

rpmaResourceIdentifier :: Maybe Text

rpmaPendingMaintenanceActionDetails :: Lens' ResourcePendingMaintenanceActions [PendingMaintenanceAction]

A list that provides details about the pending maintenance actions for the resource.

rpmaResourceIdentifier :: Lens' ResourcePendingMaintenanceActions (Maybe Text)

The ARN of the resource that has pending maintenance actions.

DBParameterGroup

data DBParameterGroup

dbparameterGroup :: DBParameterGroup

DBParameterGroup constructor.

The fields accessible through corresponding lenses are:

dbpgDBParameterGroupFamily :: Maybe Text

dbpgDBParameterGroupName :: Maybe Text

dbpgDescription :: Maybe Text

dbpgDBParameterGroupFamily :: Lens' DBParameterGroup (Maybe Text)

Provides the name of the DB parameter group family that this DB parameter group is compatible with.

dbpgDBParameterGroupName :: Lens' DBParameterGroup (Maybe Text)

Provides the name of the DB parameter group.

dbpgDescription :: Lens' DBParameterGroup (Maybe Text)

Provides the customer-specified description for this DB parameter group.

ReservedDBInstancesOffering

data ReservedDBInstancesOffering

reservedDBInstancesOffering :: ReservedDBInstancesOffering

ReservedDBInstancesOffering constructor.

The fields accessible through corresponding lenses are:

rdbioCurrencyCode :: Maybe Text

rdbioDBInstanceClass :: Maybe Text

rdbioDuration :: Maybe Int

rdbioFixedPrice :: Maybe Double

rdbioMultiAZ :: Maybe Bool

rdbioOfferingType :: Maybe Text

rdbioProductDescription :: Maybe Text

rdbioRecurringCharges :: [RecurringCharge]

rdbioReservedDBInstancesOfferingId :: Maybe Text

rdbioUsagePrice :: Maybe Double

rdbioCurrencyCode :: Lens' ReservedDBInstancesOffering (Maybe Text)

The currency code for the reserved DB instance offering.

rdbioDBInstanceClass :: Lens' ReservedDBInstancesOffering (Maybe Text)

The DB instance class for the reserved DB instance.

rdbioDuration :: Lens' ReservedDBInstancesOffering (Maybe Int)

The duration of the offering in seconds.

rdbioFixedPrice :: Lens' ReservedDBInstancesOffering (Maybe Double)

The fixed price charged for this offering.

rdbioMultiAZ :: Lens' ReservedDBInstancesOffering (Maybe Bool)

Indicates if the offering applies to Multi-AZ deployments.

rdbioOfferingType :: Lens' ReservedDBInstancesOffering (Maybe Text)

The offering type.

rdbioProductDescription :: Lens' ReservedDBInstancesOffering (Maybe Text)

The database engine used by the offering.

rdbioRecurringCharges :: Lens' ReservedDBInstancesOffering [RecurringCharge]

The recurring price charged to run this reserved DB instance.

rdbioReservedDBInstancesOfferingId :: Lens' ReservedDBInstancesOffering (Maybe Text)

The offering identifier.

rdbioUsagePrice :: Lens' ReservedDBInstancesOffering (Maybe Double)

The hourly price charged for this offering.

ApplyMethod

data ApplyMethod

Constructors

Immediate
immediate

PendingReboot
pending-reboot

CharacterSet

data CharacterSet

characterSet :: CharacterSet

CharacterSet constructor.

The fields accessible through corresponding lenses are:

csCharacterSetDescription :: Maybe Text

csCharacterSetName :: Maybe Text

csCharacterSetDescription :: Lens' CharacterSet (Maybe Text)

The description of the character set.

csCharacterSetName :: Lens' CharacterSet (Maybe Text)

The name of the character set.

Subnet

data Subnet

subnet :: Subnet

Subnet constructor.

The fields accessible through corresponding lenses are:

sSubnetAvailabilityZone :: Maybe AvailabilityZone

sSubnetIdentifier :: Maybe Text

sSubnetStatus :: Maybe Text

sSubnetAvailabilityZone :: Lens' Subnet (Maybe AvailabilityZone)

sSubnetIdentifier :: Lens' Subnet (Maybe Text)

Specifies the identifier of the subnet.

sSubnetStatus :: Lens' Subnet (Maybe Text)

Specifies the status of the subnet.

ReservedDBInstance

data ReservedDBInstance

reservedDBInstance :: ReservedDBInstance

ReservedDBInstance constructor.

The fields accessible through corresponding lenses are:

rdbiCurrencyCode :: Maybe Text

rdbiDBInstanceClass :: Maybe Text

rdbiDBInstanceCount :: Maybe Int

rdbiDuration :: Maybe Int

rdbiFixedPrice :: Maybe Double

rdbiMultiAZ :: Maybe Bool

rdbiOfferingType :: Maybe Text

rdbiProductDescription :: Maybe Text

rdbiRecurringCharges :: [RecurringCharge]

rdbiReservedDBInstanceId :: Maybe Text

rdbiReservedDBInstancesOfferingId :: Maybe Text

rdbiStartTime :: Maybe UTCTime

rdbiState :: Maybe Text

rdbiUsagePrice :: Maybe Double

rdbiCurrencyCode :: Lens' ReservedDBInstance (Maybe Text)

The currency code for the reserved DB instance.

rdbiDBInstanceClass :: Lens' ReservedDBInstance (Maybe Text)

The DB instance class for the reserved DB instance.

rdbiDBInstanceCount :: Lens' ReservedDBInstance (Maybe Int)

The number of reserved DB instances.

rdbiDuration :: Lens' ReservedDBInstance (Maybe Int)

The duration of the reservation in seconds.

rdbiFixedPrice :: Lens' ReservedDBInstance (Maybe Double)

The fixed price charged for this reserved DB instance.

rdbiMultiAZ :: Lens' ReservedDBInstance (Maybe Bool)

Indicates if the reservation applies to Multi-AZ deployments.

rdbiOfferingType :: Lens' ReservedDBInstance (Maybe Text)

The offering type of this reserved DB instance.

rdbiProductDescription :: Lens' ReservedDBInstance (Maybe Text)

The description of the reserved DB instance.

rdbiRecurringCharges :: Lens' ReservedDBInstance [RecurringCharge]

The recurring price charged to run this reserved DB instance.

rdbiReservedDBInstanceId :: Lens' ReservedDBInstance (Maybe Text)

The unique identifier for the reservation.

rdbiReservedDBInstancesOfferingId :: Lens' ReservedDBInstance (Maybe Text)

The offering identifier.

rdbiStartTime :: Lens' ReservedDBInstance (Maybe UTCTime)

The time the reservation started.

rdbiState :: Lens' ReservedDBInstance (Maybe Text)

The state of the reserved DB instance.

rdbiUsagePrice :: Lens' ReservedDBInstance (Maybe Double)

The hourly price charged for this reserved DB instance.

EngineDefaults

data EngineDefaults

engineDefaults :: EngineDefaults

EngineDefaults constructor.

The fields accessible through corresponding lenses are:

edDBParameterGroupFamily :: Maybe Text

edMarker :: Maybe Text

edParameters :: [Parameter]

edDBParameterGroupFamily :: Lens' EngineDefaults (Maybe Text)

Specifies the name of the DB parameter group family which the engine default parameters apply to.

edMarker :: Lens' EngineDefaults (Maybe Text)

An optional pagination token provided by a previous EngineDefaults request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

edParameters :: Lens' EngineDefaults [Parameter]

Contains a list of engine default parameters.

DBParameterGroupNameMessage

data DBParameterGroupNameMessage

dbparameterGroupNameMessage :: DBParameterGroupNameMessage

DBParameterGroupNameMessage constructor.

The fields accessible through corresponding lenses are:

dbpgnmDBParameterGroupName :: Maybe Text

dbpgnmDBParameterGroupName :: Lens' DBParameterGroupNameMessage (Maybe Text)

The name of the DB parameter group.

OptionGroupOption

data OptionGroupOption

optionGroupOption :: OptionGroupOption

OptionGroupOption constructor.

The fields accessible through corresponding lenses are:

ogoDefaultPort :: Maybe Int

ogoDescription :: Maybe Text

ogoEngineName :: Maybe Text

ogoMajorEngineVersion :: Maybe Text

ogoMinimumRequiredMinorEngineVersion :: Maybe Text

ogoName :: Maybe Text

ogoOptionGroupOptionSettings :: [OptionGroupOptionSetting]

ogoOptionsDependedOn :: [Text]

ogoPermanent :: Maybe Bool

ogoPersistent :: Maybe Bool

ogoPortRequired :: Maybe Bool

ogoDefaultPort :: Lens' OptionGroupOption (Maybe Int)

If the option requires a port, specifies the default port for the option.

ogoDescription :: Lens' OptionGroupOption (Maybe Text)

The description of the option.

ogoEngineName :: Lens' OptionGroupOption (Maybe Text)

The name of the engine that this option can be applied to.

ogoMajorEngineVersion :: Lens' OptionGroupOption (Maybe Text)

Indicates the major engine version that the option is available for.

ogoMinimumRequiredMinorEngineVersion :: Lens' OptionGroupOption (Maybe Text)

The minimum required engine version for the option to be applied.

ogoName :: Lens' OptionGroupOption (Maybe Text)

The name of the option.

ogoOptionGroupOptionSettings :: Lens' OptionGroupOption [OptionGroupOptionSetting]

Specifies the option settings that are available (and the default value) for each option in an option group.

ogoOptionsDependedOn :: Lens' OptionGroupOption [Text]

List of all options that are prerequisites for this option.

ogoPermanent :: Lens' OptionGroupOption (Maybe Bool)

A permanent option cannot be removed from the option group once the option group is used, and it cannot be removed from the db instance after assigning an option group with this permanent option.

ogoPersistent :: Lens' OptionGroupOption (Maybe Bool)

A persistent option cannot be removed from the option group once the option group is used, but this option can be removed from the db instance while modifying the related data and assigning another option group without this option.

ogoPortRequired :: Lens' OptionGroupOption (Maybe Bool)

Specifies whether the option requires a port.

DBInstance

data DBInstance

dbinstance :: DBInstance

DBInstance constructor.

The fields accessible through corresponding lenses are:

dbiAllocatedStorage :: Maybe Int

dbiAutoMinorVersionUpgrade :: Maybe Bool

dbiAvailabilityZone :: Maybe Text

dbiBackupRetentionPeriod :: Maybe Int

dbiCACertificateIdentifier :: Maybe Text

dbiCharacterSetName :: Maybe Text

dbiDBInstanceClass :: Maybe Text

dbiDBInstanceIdentifier :: Maybe Text

dbiDBInstanceStatus :: Maybe Text

dbiDBName :: Maybe Text

dbiDBParameterGroups :: [DBParameterGroupStatus]

dbiDBSecurityGroups :: [DBSecurityGroupMembership]

dbiDBSubnetGroup :: Maybe DBSubnetGroup

dbiDbiResourceId :: Maybe Text

dbiEndpoint :: Maybe Endpoint

dbiEngine :: Maybe Text

dbiEngineVersion :: Maybe Text

dbiInstanceCreateTime :: Maybe UTCTime

dbiIops :: Maybe Int

dbiKmsKeyId :: Maybe Text

dbiLatestRestorableTime :: Maybe UTCTime

dbiLicenseModel :: Maybe Text

dbiMasterUsername :: Maybe Text

dbiMultiAZ :: Maybe Bool

dbiOptionGroupMemberships :: [OptionGroupMembership]

dbiPendingModifiedValues :: Maybe PendingModifiedValues

dbiPreferredBackupWindow :: Maybe Text

dbiPreferredMaintenanceWindow :: Maybe Text

dbiPubliclyAccessible :: Maybe Bool

dbiReadReplicaDBInstanceIdentifiers :: [Text]

dbiReadReplicaSourceDBInstanceIdentifier :: Maybe Text

dbiSecondaryAvailabilityZone :: Maybe Text

dbiStatusInfos :: [DBInstanceStatusInfo]

dbiStorageEncrypted :: Maybe Bool

dbiStorageType :: Maybe Text

dbiTdeCredentialArn :: Maybe Text

dbiVpcSecurityGroups :: [VpcSecurityGroupMembership]

dbiAllocatedStorage :: Lens' DBInstance (Maybe Int)

Specifies the allocated storage size specified in gigabytes.

dbiAutoMinorVersionUpgrade :: Lens' DBInstance (Maybe Bool)

Indicates that minor version patches are applied automatically.

dbiAvailabilityZone :: Lens' DBInstance (Maybe Text)

Specifies the name of the Availability Zone the DB instance is located in.

dbiBackupRetentionPeriod :: Lens' DBInstance (Maybe Int)

Specifies the number of days for which automatic DB snapshots are retained.

dbiCACertificateIdentifier :: Lens' DBInstance (Maybe Text)

The identifier of the CA certificate for this DB instance.

dbiCharacterSetName :: Lens' DBInstance (Maybe Text)

If present, specifies the name of the character set that this instance is associated with.

dbiDBInstanceClass :: Lens' DBInstance (Maybe Text)

Contains the name of the compute and memory capacity class of the DB instance.

dbiDBInstanceIdentifier :: Lens' DBInstance (Maybe Text)

Contains a user-supplied database identifier. This is the unique key that identifies a DB instance.

dbiDBInstanceStatus :: Lens' DBInstance (Maybe Text)

Specifies the current state of this database.

dbiDBName :: Lens' DBInstance (Maybe Text)

The meaning of this parameter differs according to the database engine you use. For example, this value returns either MySQL or PostgreSQL information when returning values from CreateDBInstanceReadReplica since Read Replicas are only supported for MySQL and PostgreSQL.

MySQL, SQL Server, PostgreSQL

Contains the name of the initial database of this instance that was provided at create time, if one was specified when the DB instance was created. This same name is returned for the life of the DB instance.

Type: String

Oracle

Contains the Oracle System ID (SID) of the created DB instance. Not shown when the returned parameters do not apply to an Oracle DB instance.

dbiDBParameterGroups :: Lens' DBInstance [DBParameterGroupStatus]

Provides the list of DB parameter groups applied to this DB instance.

dbiDBSecurityGroups :: Lens' DBInstance [DBSecurityGroupMembership]

Provides List of DB security group elements containing only Name and Status subelements.

dbiDBSubnetGroup :: Lens' DBInstance (Maybe DBSubnetGroup)

Specifies information on the subnet group associated with the DB instance, including the name, description, and subnets in the subnet group.

dbiDbiResourceId :: Lens' DBInstance (Maybe Text)

If StorageEncrypted is true, the region-unique, immutable identifier for the encrypted DB instance. This identifier is found in AWS CloudTrail log entries whenever the KMS key for the DB instance is accessed.

dbiEndpoint :: Lens' DBInstance (Maybe Endpoint)

Specifies the connection endpoint.

dbiEngine :: Lens' DBInstance (Maybe Text)

Provides the name of the database engine to be used for this DB instance.

dbiEngineVersion :: Lens' DBInstance (Maybe Text)

Indicates the database engine version.

dbiInstanceCreateTime :: Lens' DBInstance (Maybe UTCTime)

Provides the date and time the DB instance was created.

dbiIops :: Lens' DBInstance (Maybe Int)

Specifies the Provisioned IOPS (I/O operations per second) value.

dbiKmsKeyId :: Lens' DBInstance (Maybe Text)

If StorageEncrypted is true, the KMS key identifier for the encrypted DB instance.

dbiLatestRestorableTime :: Lens' DBInstance (Maybe UTCTime)

Specifies the latest time to which a database can be restored with point-in-time restore.

dbiLicenseModel :: Lens' DBInstance (Maybe Text)

License model information for this DB instance.

dbiMasterUsername :: Lens' DBInstance (Maybe Text)

Contains the master username for the DB instance.

dbiMultiAZ :: Lens' DBInstance (Maybe Bool)

Specifies if the DB instance is a Multi-AZ deployment.

dbiOptionGroupMemberships :: Lens' DBInstance [OptionGroupMembership]

Provides the list of option group memberships for this DB instance.

dbiPendingModifiedValues :: Lens' DBInstance (Maybe PendingModifiedValues)

Specifies that changes to the DB instance are pending. This element is only included when changes are pending. Specific changes are identified by subelements.

dbiPreferredBackupWindow :: Lens' DBInstance (Maybe Text)

Specifies the daily time range during which automated backups are created if automated backups are enabled, as determined by the BackupRetentionPeriod.

dbiPreferredMaintenanceWindow :: Lens' DBInstance (Maybe Text)

Specifies the weekly time range (in UTC) during which system maintenance can occur.

dbiPubliclyAccessible :: Lens' DBInstance (Maybe Bool)

Specifies the accessibility options for the DB instance. A value of true specifies an Internet-facing instance with a publicly resolvable DNS name, which resolves to a public IP address. A value of false specifies an internal instance with a DNS name that resolves to a private IP address.

Default: The default behavior varies depending on whether a VPC has been requested or not. The following list shows the default behavior in each case.

Default VPC:true VPC:false If no DB subnet group has been specified as part of the request and the PubliclyAccessible value has not been set, the DB instance will be publicly accessible. If a specific DB subnet group has been specified as part of the request and the PubliclyAccessible value has not been set, the DB instance will be private.

dbiReadReplicaDBInstanceIdentifiers :: Lens' DBInstance [Text]

Contains one or more identifiers of the Read Replicas associated with this DB instance.

dbiReadReplicaSourceDBInstanceIdentifier :: Lens' DBInstance (Maybe Text)

Contains the identifier of the source DB instance if this DB instance is a Read Replica.

dbiSecondaryAvailabilityZone :: Lens' DBInstance (Maybe Text)

If present, specifies the name of the secondary Availability Zone for a DB instance with multi-AZ support.

dbiStatusInfos :: Lens' DBInstance [DBInstanceStatusInfo]

The status of a Read Replica. If the instance is not a Read Replica, this will be blank.

dbiStorageEncrypted :: Lens' DBInstance (Maybe Bool)

Specifies whether the DB instance is encrypted.

dbiStorageType :: Lens' DBInstance (Maybe Text)

Specifies the storage type associated with DB instance.

dbiTdeCredentialArn :: Lens' DBInstance (Maybe Text)

The ARN from the Key Store with which the instance is associated for TDE encryption.

dbiVpcSecurityGroups :: Lens' DBInstance [VpcSecurityGroupMembership]

Provides List of VPC security group elements that the DB instance belongs to.

AccountQuota

data AccountQuota

accountQuota :: AccountQuota

AccountQuota constructor.

The fields accessible through corresponding lenses are:

aqAccountQuotaName :: Maybe Text

aqMax :: Maybe Integer

aqUsed :: Maybe Integer

aqAccountQuotaName :: Lens' AccountQuota (Maybe Text)

The name of the Amazon RDS quota for this AWS account.

aqMax :: Lens' AccountQuota (Maybe Integer)

The maximum allowed value for the quota.

aqUsed :: Lens' AccountQuota (Maybe Integer)

The amount currently used toward the quota maximum.

AvailabilityZone

data AvailabilityZone

availabilityZone :: AvailabilityZone

AvailabilityZone constructor.

The fields accessible through corresponding lenses are:

azName :: Maybe Text

azName :: Lens' AvailabilityZone (Maybe Text)

The name of the availability zone.

EventSubscription

data EventSubscription

eventSubscription :: EventSubscription

EventSubscription constructor.

The fields accessible through corresponding lenses are:

esCustSubscriptionId :: Maybe Text

esCustomerAwsId :: Maybe Text

esEnabled :: Maybe Bool

esEventCategoriesList :: [Text]

esSnsTopicArn :: Maybe Text

esSourceIdsList :: [Text]

esSourceType :: Maybe Text

esStatus :: Maybe Text

esSubscriptionCreationTime :: Maybe Text

esCustSubscriptionId :: Lens' EventSubscription (Maybe Text)

The RDS event notification subscription Id.

esCustomerAwsId :: Lens' EventSubscription (Maybe Text)

The AWS customer account associated with the RDS event notification subscription.

esEnabled :: Lens' EventSubscription (Maybe Bool)

A Boolean value indicating if the subscription is enabled. True indicates the subscription is enabled.

esEventCategoriesList :: Lens' EventSubscription [Text]

A list of event categories for the RDS event notification subscription.

esSnsTopicArn :: Lens' EventSubscription (Maybe Text)

The topic ARN of the RDS event notification subscription.

esSourceIdsList :: Lens' EventSubscription [Text]

A list of source IDs for the RDS event notification subscription.

esSourceType :: Lens' EventSubscription (Maybe Text)

The source type for the RDS event notification subscription.

esStatus :: Lens' EventSubscription (Maybe Text)

The status of the RDS event notification subscription.

Constraints:

Can be one of the following: creating | modifying | deleting | active | no-permission | topic-not-exist

The status "no-permission" indicates that RDS no longer has permission to post to the SNS topic. The status "topic-not-exist" indicates that the topic was deleted after the subscription was created.

esSubscriptionCreationTime :: Lens' EventSubscription (Maybe Text)

The time the RDS event notification subscription was created.

DBSubnetGroup

data DBSubnetGroup

dbsubnetGroup :: DBSubnetGroup

DBSubnetGroup constructor.

The fields accessible through corresponding lenses are:

dbsg1DBSubnetGroupDescription :: Maybe Text

dbsg1DBSubnetGroupName :: Maybe Text

dbsg1SubnetGroupStatus :: Maybe Text

dbsg1Subnets :: [Subnet]

dbsg1VpcId :: Maybe Text

dbsg1DBSubnetGroupDescription :: Lens' DBSubnetGroup (Maybe Text)

Provides the description of the DB subnet group.

dbsg1DBSubnetGroupName :: Lens' DBSubnetGroup (Maybe Text)

Specifies the name of the DB subnet group.

dbsg1SubnetGroupStatus :: Lens' DBSubnetGroup (Maybe Text)

Provides the status of the DB subnet group.

dbsg1Subnets :: Lens' DBSubnetGroup [Subnet]

Contains a list of Subnet elements.

dbsg1VpcId :: Lens' DBSubnetGroup (Maybe Text)

Provides the VpcId of the DB subnet group.

Certificate

data Certificate

certificate :: Certificate

Certificate constructor.

The fields accessible through corresponding lenses are:

cCertificateIdentifier :: Maybe Text

cCertificateType :: Maybe Text

cThumbprint :: Maybe Text

cValidFrom :: Maybe UTCTime

cValidTill :: Maybe UTCTime

cCertificateIdentifier :: Lens' Certificate (Maybe Text)

The unique key that identifies a certificate.

cCertificateType :: Lens' Certificate (Maybe Text)

The type of the certificate.

cThumbprint :: Lens' Certificate (Maybe Text)

The thumbprint of the certificate.

cValidFrom :: Lens' Certificate (Maybe UTCTime)

The starting date from which the certificate is valid.

cValidTill :: Lens' Certificate (Maybe UTCTime)

The final date that the certificate continues to be valid.

DBInstanceStatusInfo

data DBInstanceStatusInfo

dbinstanceStatusInfo :: DBInstanceStatusInfo

DBInstanceStatusInfo constructor.

The fields accessible through corresponding lenses are:

dbisiMessage :: Maybe Text

dbisiNormal :: Maybe Bool

dbisiStatus :: Maybe Text

dbisiStatusType :: Maybe Text

dbisiMessage :: Lens' DBInstanceStatusInfo (Maybe Text)

Details of the error if there is an error for the instance. If the instance is not in an error state, this value is blank.

dbisiNormal :: Lens' DBInstanceStatusInfo (Maybe Bool)

Boolean value that is true if the instance is operating normally, or false if the instance is in an error state.

dbisiStatus :: Lens' DBInstanceStatusInfo (Maybe Text)

Status of the DB instance. For a StatusType of read replica, the values can be replicating, error, stopped, or terminated.

dbisiStatusType :: Lens' DBInstanceStatusInfo (Maybe Text)

This value is currently "read replication."

OptionSetting

data OptionSetting

optionSetting :: OptionSetting

OptionSetting constructor.

The fields accessible through corresponding lenses are:

osAllowedValues :: Maybe Text

osApplyType :: Maybe Text

osDataType :: Maybe Text

osDefaultValue :: Maybe Text

osDescription :: Maybe Text

osIsCollection :: Maybe Bool

osIsModifiable :: Maybe Bool

osName :: Maybe Text

osValue :: Maybe Text

osAllowedValues :: Lens' OptionSetting (Maybe Text)

The allowed values of the option setting.

osApplyType :: Lens' OptionSetting (Maybe Text)

The DB engine specific parameter type.

osDataType :: Lens' OptionSetting (Maybe Text)

The data type of the option setting.

osDefaultValue :: Lens' OptionSetting (Maybe Text)

The default value of the option setting.

osDescription :: Lens' OptionSetting (Maybe Text)

The description of the option setting.

osIsCollection :: Lens' OptionSetting (Maybe Bool)

Indicates if the option setting is part of a collection.

osIsModifiable :: Lens' OptionSetting (Maybe Bool)

A Boolean value that, when true, indicates the option setting can be modified from the default.

osName :: Lens' OptionSetting (Maybe Text)

The name of the option that has settings that you can set.

osValue :: Lens' OptionSetting (Maybe Text)

The current value of the option setting.

DescribeDBLogFilesDetails

data DescribeDBLogFilesDetails

describeDBLogFilesDetails :: DescribeDBLogFilesDetails

DescribeDBLogFilesDetails constructor.

The fields accessible through corresponding lenses are:

ddblfdLastWritten :: Maybe Integer

ddblfdLogFileName :: Maybe Text

ddblfdSize :: Maybe Integer

ddblfdLastWritten :: Lens' DescribeDBLogFilesDetails (Maybe Integer)

A POSIX timestamp when the last log entry was written.

ddblfdLogFileName :: Lens' DescribeDBLogFilesDetails (Maybe Text)

The name of the log file for the specified DB instance.

ddblfdSize :: Lens' DescribeDBLogFilesDetails (Maybe Integer)

The size, in bytes, of the log file for the specified DB instance.

OrderableDBInstanceOption

data OrderableDBInstanceOption

orderableDBInstanceOption :: OrderableDBInstanceOption

OrderableDBInstanceOption constructor.

The fields accessible through corresponding lenses are:

odbioAvailabilityZones :: [AvailabilityZone]

odbioDBInstanceClass :: Maybe Text

odbioEngine :: Maybe Text

odbioEngineVersion :: Maybe Text

odbioLicenseModel :: Maybe Text

odbioMultiAZCapable :: Maybe Bool

odbioReadReplicaCapable :: Maybe Bool

odbioStorageType :: Maybe Text

odbioSupportsIops :: Maybe Bool

odbioSupportsStorageEncryption :: Maybe Bool

odbioVpc :: Maybe Bool

odbioAvailabilityZones :: Lens' OrderableDBInstanceOption [AvailabilityZone]

A list of availability zones for the orderable DB instance.

odbioDBInstanceClass :: Lens' OrderableDBInstanceOption (Maybe Text)

The DB instance Class for the orderable DB instance

odbioEngine :: Lens' OrderableDBInstanceOption (Maybe Text)

The engine type of the orderable DB instance.

odbioEngineVersion :: Lens' OrderableDBInstanceOption (Maybe Text)

The engine version of the orderable DB instance.

odbioLicenseModel :: Lens' OrderableDBInstanceOption (Maybe Text)

The license model for the orderable DB instance.

odbioMultiAZCapable :: Lens' OrderableDBInstanceOption (Maybe Bool)

Indicates whether this orderable DB instance is multi-AZ capable.

odbioReadReplicaCapable :: Lens' OrderableDBInstanceOption (Maybe Bool)

Indicates whether this orderable DB instance can have a Read Replica.

odbioStorageType :: Lens' OrderableDBInstanceOption (Maybe Text)

The storage type for this orderable DB instance.

odbioSupportsIops :: Lens' OrderableDBInstanceOption (Maybe Bool)

Indicates whether this orderable DB instance supports provisioned IOPS.

odbioSupportsStorageEncryption :: Lens' OrderableDBInstanceOption (Maybe Bool)

Indicates whether this orderable DB instance supports encrypted storage.

odbioVpc :: Lens' OrderableDBInstanceOption (Maybe Bool)

Indicates whether this is a VPC orderable DB instance.

Filter

data Filter

filter'

Arguments

:: Text
fName

-> Filter

Filter constructor.

The fields accessible through corresponding lenses are:

fName :: Text

fValues :: [Text]

fName :: Lens' Filter Text

This parameter is not currently supported.

fValues :: Lens' Filter [Text]

This parameter is not currently supported.

RecurringCharge

data RecurringCharge

recurringCharge :: RecurringCharge

RecurringCharge constructor.

The fields accessible through corresponding lenses are:

rcRecurringChargeAmount :: Maybe Double

rcRecurringChargeFrequency :: Maybe Text

rcRecurringChargeAmount :: Lens' RecurringCharge (Maybe Double)

The amount of the recurring charge.

rcRecurringChargeFrequency :: Lens' RecurringCharge (Maybe Text)

The frequency of the recurring charge.

Endpoint

data Endpoint

endpoint :: Endpoint

Endpoint constructor.

The fields accessible through corresponding lenses are:

eAddress :: Maybe Text

ePort :: Maybe Int

eAddress :: Lens' Endpoint (Maybe Text)

Specifies the DNS address of the DB instance.

ePort :: Lens' Endpoint (Maybe Int)

Specifies the port that the database engine is listening on.

OptionConfiguration

data OptionConfiguration

optionConfiguration

Arguments

:: Text
ocOptionName

-> OptionConfiguration

OptionConfiguration constructor.

The fields accessible through corresponding lenses are:

ocDBSecurityGroupMemberships :: [Text]

ocOptionName :: Text

ocOptionSettings :: [OptionSetting]

ocPort :: Maybe Int

ocVpcSecurityGroupMemberships :: [Text]

ocDBSecurityGroupMemberships :: Lens' OptionConfiguration [Text]

A list of DBSecurityGroupMemebrship name strings used for this option.

ocOptionName :: Lens' OptionConfiguration Text

The configuration of options to include in a group.

ocOptionSettings :: Lens' OptionConfiguration [OptionSetting]

The option settings to include in an option group.

ocPort :: Lens' OptionConfiguration (Maybe Int)

The optional port for the option.

ocVpcSecurityGroupMemberships :: Lens' OptionConfiguration [Text]

A list of VpcSecurityGroupMemebrship name strings used for this option.

Option

data Option

option :: Option

Option constructor.

The fields accessible through corresponding lenses are:

oDBSecurityGroupMemberships :: [DBSecurityGroupMembership]

oOptionDescription :: Maybe Text

oOptionName :: Maybe Text

oOptionSettings :: [OptionSetting]

oPermanent :: Maybe Bool

oPersistent :: Maybe Bool

oPort :: Maybe Int

oVpcSecurityGroupMemberships :: [VpcSecurityGroupMembership]

oDBSecurityGroupMemberships :: Lens' Option [DBSecurityGroupMembership]

If the option requires access to a port, then this DB security group allows access to the port.

oOptionDescription :: Lens' Option (Maybe Text)

The description of the option.

oOptionName :: Lens' Option (Maybe Text)

The name of the option.

oOptionSettings :: Lens' Option [OptionSetting]

The option settings for this option.

oPermanent :: Lens' Option (Maybe Bool)

Indicate if this option is permanent.

oPersistent :: Lens' Option (Maybe Bool)

Indicate if this option is persistent.

oPort :: Lens' Option (Maybe Int)

If required, the port configured for this option to use.

oVpcSecurityGroupMemberships :: Lens' Option [VpcSecurityGroupMembership]

If the option requires access to a port, then this VPC security group allows access to the port.

IPRange

data IPRange

iprange :: IPRange

IPRange constructor.

The fields accessible through corresponding lenses are:

iprCIDRIP :: Maybe Text

iprStatus :: Maybe Text

iprCIDRIP :: Lens' IPRange (Maybe Text)

Specifies the IP range.

iprStatus :: Lens' IPRange (Maybe Text)

Specifies the status of the IP range. Status can be "authorizing", "authorized", "revoking", and "revoked".

OptionGroupMembership

data OptionGroupMembership

optionGroupMembership :: OptionGroupMembership

OptionGroupMembership constructor.

The fields accessible through corresponding lenses are:

ogmOptionGroupName :: Maybe Text

ogmStatus :: Maybe Text

ogmOptionGroupName :: Lens' OptionGroupMembership (Maybe Text)

The name of the option group that the instance belongs to.

ogmStatus :: Lens' OptionGroupMembership (Maybe Text)

The status of the DB instance's option group membership (e.g. in-sync, pending, pending-maintenance, applying).

EventCategoriesMap

data EventCategoriesMap

eventCategoriesMap :: EventCategoriesMap

EventCategoriesMap constructor.

The fields accessible through corresponding lenses are:

ecmEventCategories :: [Text]

ecmSourceType :: Maybe Text

ecmEventCategories :: Lens' EventCategoriesMap [Text]

The event categories for the specified source type

ecmSourceType :: Lens' EventCategoriesMap (Maybe Text)

The source type that the returned categories belong to

PendingModifiedValues

data PendingModifiedValues

pendingModifiedValues :: PendingModifiedValues

PendingModifiedValues constructor.

The fields accessible through corresponding lenses are:

pmvAllocatedStorage :: Maybe Int

pmvBackupRetentionPeriod :: Maybe Int

pmvCACertificateIdentifier :: Maybe Text

pmvDBInstanceClass :: Maybe Text

pmvDBInstanceIdentifier :: Maybe Text

pmvEngineVersion :: Maybe Text

pmvIops :: Maybe Int

pmvMasterUserPassword :: Maybe Text

pmvMultiAZ :: Maybe Bool

pmvPort :: Maybe Int

pmvStorageType :: Maybe Text

pmvAllocatedStorage :: Lens' PendingModifiedValues (Maybe Int)

Contains the new AllocatedStorage size for the DB instance that will be applied or is in progress.

pmvBackupRetentionPeriod :: Lens' PendingModifiedValues (Maybe Int)

Specifies the pending number of days for which automated backups are retained.

pmvCACertificateIdentifier :: Lens' PendingModifiedValues (Maybe Text)

Specifies the identifier of the CA certificate for the DB instance.

pmvDBInstanceClass :: Lens' PendingModifiedValues (Maybe Text)

Contains the new DBInstanceClass for the DB instance that will be applied or is in progress.

pmvDBInstanceIdentifier :: Lens' PendingModifiedValues (Maybe Text)

Contains the new DBInstanceIdentifier for the DB instance that will be applied or is in progress.

pmvEngineVersion :: Lens' PendingModifiedValues (Maybe Text)

Indicates the database engine version.

pmvIops :: Lens' PendingModifiedValues (Maybe Int)

Specifies the new Provisioned IOPS value for the DB instance that will be applied or is being applied.

pmvMasterUserPassword :: Lens' PendingModifiedValues (Maybe Text)

Contains the pending or in-progress change of the master credentials for the DB instance.

pmvMultiAZ :: Lens' PendingModifiedValues (Maybe Bool)

Indicates that the Single-AZ DB instance is to change to a Multi-AZ deployment.

pmvPort :: Lens' PendingModifiedValues (Maybe Int)

Specifies the pending port for the DB instance.

pmvStorageType :: Lens' PendingModifiedValues (Maybe Text)

Specifies the storage type to be associated with the DB instance.

VpcSecurityGroupMembership

data VpcSecurityGroupMembership

vpcSecurityGroupMembership :: VpcSecurityGroupMembership

VpcSecurityGroupMembership constructor.

The fields accessible through corresponding lenses are:

vsgmStatus :: Maybe Text

vsgmVpcSecurityGroupId :: Maybe Text

vsgmStatus :: Lens' VpcSecurityGroupMembership (Maybe Text)

The status of the VPC security group.

vsgmVpcSecurityGroupId :: Lens' VpcSecurityGroupMembership (Maybe Text)

The name of the VPC security group.

Parameter

data Parameter

parameter :: Parameter

Parameter constructor.

The fields accessible through corresponding lenses are:

pAllowedValues :: Maybe Text

pApplyMethod :: Maybe ApplyMethod

pApplyType :: Maybe Text

pDataType :: Maybe Text

pDescription :: Maybe Text

pIsModifiable :: Maybe Bool

pMinimumEngineVersion :: Maybe Text

pParameterName :: Maybe Text

pParameterValue :: Maybe Text

pSource :: Maybe Text

pAllowedValues :: Lens' Parameter (Maybe Text)

Specifies the valid range of values for the parameter.

pApplyMethod :: Lens' Parameter (Maybe ApplyMethod)

Indicates when to apply parameter updates.

pApplyType :: Lens' Parameter (Maybe Text)

Specifies the engine specific parameters type.

pDataType :: Lens' Parameter (Maybe Text)

Specifies the valid data type for the parameter.

pDescription :: Lens' Parameter (Maybe Text)

Provides a description of the parameter.

pIsModifiable :: Lens' Parameter (Maybe Bool)

Indicates whether (true) or not (false) the parameter can be modified. Some parameters have security or operational implications that prevent them from being changed.

pMinimumEngineVersion :: Lens' Parameter (Maybe Text)

The earliest engine version to which the parameter can apply.

pParameterName :: Lens' Parameter (Maybe Text)

Specifies the name of the parameter.

pParameterValue :: Lens' Parameter (Maybe Text)

Specifies the value of the parameter.

pSource :: Lens' Parameter (Maybe Text)

Indicates the source of the parameter value.

OptionGroupOptionSetting

data OptionGroupOptionSetting

optionGroupOptionSetting :: OptionGroupOptionSetting

OptionGroupOptionSetting constructor.

The fields accessible through corresponding lenses are:

ogosAllowedValues :: Maybe Text

ogosApplyType :: Maybe Text

ogosDefaultValue :: Maybe Text

ogosIsModifiable :: Maybe Bool

ogosSettingDescription :: Maybe Text

ogosSettingName :: Maybe Text

ogosAllowedValues :: Lens' OptionGroupOptionSetting (Maybe Text)

Indicates the acceptable values for the option group option.

ogosApplyType :: Lens' OptionGroupOptionSetting (Maybe Text)

The DB engine specific parameter type for the option group option.

ogosDefaultValue :: Lens' OptionGroupOptionSetting (Maybe Text)

The default value for the option group option.

ogosIsModifiable :: Lens' OptionGroupOptionSetting (Maybe Bool)

Boolean value where true indicates that this option group option can be changed from the default value.

ogosSettingDescription :: Lens' OptionGroupOptionSetting (Maybe Text)

The description of the option group option.

ogosSettingName :: Lens' OptionGroupOptionSetting (Maybe Text)

The name of the option group option.