Network.AWS.RDS.V2013_09_09

Description

Amazon Relational Database Service (Amazon RDS) is a web service that makes it easy to set up, operate, and scale a relational database in the cloud. It provides cost-efficient and resizable capacity while managing time-consuming database administration tasks, freeing you up to focus on your applications and business.

Documentation

module Network.AWS.RDS.V2013_09_09.AddSourceIdentifierToSubscription

module Network.AWS.RDS.V2013_09_09.AddTagsToResource

module Network.AWS.RDS.V2013_09_09.AuthorizeDBSecurityGroupIngress

module Network.AWS.RDS.V2013_09_09.CopyDBSnapshot

module Network.AWS.RDS.V2013_09_09.CreateDBInstance

module Network.AWS.RDS.V2013_09_09.CreateDBInstanceReadReplica

module Network.AWS.RDS.V2013_09_09.CreateDBParameterGroup

module Network.AWS.RDS.V2013_09_09.CreateDBSecurityGroup

module Network.AWS.RDS.V2013_09_09.CreateDBSnapshot

module Network.AWS.RDS.V2013_09_09.CreateDBSubnetGroup

module Network.AWS.RDS.V2013_09_09.CreateEventSubscription

module Network.AWS.RDS.V2013_09_09.CreateOptionGroup

module Network.AWS.RDS.V2013_09_09.DeleteDBInstance

module Network.AWS.RDS.V2013_09_09.DeleteDBParameterGroup

module Network.AWS.RDS.V2013_09_09.DeleteDBSecurityGroup

module Network.AWS.RDS.V2013_09_09.DeleteDBSnapshot

module Network.AWS.RDS.V2013_09_09.DeleteDBSubnetGroup

module Network.AWS.RDS.V2013_09_09.DeleteEventSubscription

module Network.AWS.RDS.V2013_09_09.DeleteOptionGroup

module Network.AWS.RDS.V2013_09_09.DescribeDBEngineVersions

module Network.AWS.RDS.V2013_09_09.DescribeDBInstances

module Network.AWS.RDS.V2013_09_09.DescribeDBLogFiles

module Network.AWS.RDS.V2013_09_09.DescribeDBParameterGroups

module Network.AWS.RDS.V2013_09_09.DescribeDBParameters

module Network.AWS.RDS.V2013_09_09.DescribeDBSecurityGroups

module Network.AWS.RDS.V2013_09_09.DescribeDBSnapshots

module Network.AWS.RDS.V2013_09_09.DescribeDBSubnetGroups

module Network.AWS.RDS.V2013_09_09.DescribeEngineDefaultParameters

module Network.AWS.RDS.V2013_09_09.DescribeEventCategories

module Network.AWS.RDS.V2013_09_09.DescribeEventSubscriptions

module Network.AWS.RDS.V2013_09_09.DescribeEvents

module Network.AWS.RDS.V2013_09_09.DescribeOptionGroupOptions

module Network.AWS.RDS.V2013_09_09.DescribeOptionGroups

module Network.AWS.RDS.V2013_09_09.DescribeOrderableDBInstanceOptions

module Network.AWS.RDS.V2013_09_09.DescribeReservedDBInstances

module Network.AWS.RDS.V2013_09_09.DescribeReservedDBInstancesOfferings

module Network.AWS.RDS.V2013_09_09.DownloadDBLogFilePortion

module Network.AWS.RDS.V2013_09_09.ListTagsForResource

module Network.AWS.RDS.V2013_09_09.ModifyDBInstance

module Network.AWS.RDS.V2013_09_09.ModifyDBParameterGroup

module Network.AWS.RDS.V2013_09_09.ModifyDBSubnetGroup

module Network.AWS.RDS.V2013_09_09.ModifyEventSubscription

module Network.AWS.RDS.V2013_09_09.ModifyOptionGroup

module Network.AWS.RDS.V2013_09_09.PromoteReadReplica

module Network.AWS.RDS.V2013_09_09.PurchaseReservedDBInstancesOffering

module Network.AWS.RDS.V2013_09_09.RebootDBInstance

module Network.AWS.RDS.V2013_09_09.RemoveSourceIdentifierFromSubscription

module Network.AWS.RDS.V2013_09_09.RemoveTagsFromResource

module Network.AWS.RDS.V2013_09_09.ResetDBParameterGroup

module Network.AWS.RDS.V2013_09_09.RestoreDBInstanceFromDBSnapshot

module Network.AWS.RDS.V2013_09_09.RestoreDBInstanceToPointInTime

module Network.AWS.RDS.V2013_09_09.RevokeDBSecurityGroupIngress

module Network.AWS.RDS.V2013_09_09.Types