Network.AWS.RDS.V2013_09_09.AddTagsToResource

Description

Adds metadata tags to an Amazon RDS resource. These tags can also be used with cost allocation reporting to track cost associated with Amazon RDS resources, or used in Condition statement in IAM policy for Amazon RDS. For an overview on tagging Amazon RDS resources, see Tagging Amazon RDS Resources.

Documentation

data AddTagsToResource

Constructors

AddTagsToResource

Fields

_attrmResourceName :: Text

The Amazon RDS resource the tags will be added to. This value is an Amazon Resource Name (ARN). For information about creating an ARN, see Constructing an RDS Amazon Resource Name (ARN).

_attrmTags :: [Tag]

The tags to be assigned to the Amazon RDS resource.

attrmTags :: Lens' AddTagsToResource [Tag]

attrmResourceName :: Lens' AddTagsToResource Text

data AddTagsToResourceResponse

Constructors

AddTagsToResourceResponse