Network.AWS.RDS.V2013_09_09.AuthorizeDBSecurityGroupIngress

Description

Enables ingress to a DBSecurityGroup using one of two forms of authorization. First, EC2 or VPC security groups can be added to the DBSecurityGroup if the application using the database is running on EC2 or VPC instances. Second, IP ranges are available if the application accessing your database is running on the Internet. Required parameters for this API are one of CIDR range, EC2SecurityGroupId for VPC, or (EC2SecurityGroupOwnerId and either EC2SecurityGroupName or EC2SecurityGroupId for non-VPC). You cannot authorize ingress from an EC2 security group in one Region to an Amazon RDS DB instance in another. You cannot authorize ingress from a VPC security group in one VPC to an Amazon RDS DB instance in another. For an overview of CIDR ranges, go to the Wikipedia Tutorial. https:/rds.amazonaws.com ?CIDRIP=192.168.1.1%2F24 &DBSecurityGroupName=mydbsecuritygroup &Version=2013-05-15 &Action=AuthorizeDBSecurityGroupIngress &SignatureVersion=2 &SignatureMethod=HmacSHA256 &Timestamp=2011-02-15T17%3A10%3A50.274Z &AWSAccessKeyId= &Signature= My new DBSecurityGroup 192.168.1.1/24 authorizing 621567473609 mydbsecuritygroup vpc-1ab2c3d4 d9799197-bf2d-11de-b88d-993294bf1c81.

Documentation

authorizeDBSecurityGroupIngress

Arguments

:: Text
_adbsgimDBSecurityGroupName

-> AuthorizeDBSecurityGroupIngress

Minimum specification for a AuthorizeDBSecurityGroupIngress request.

data AuthorizeDBSecurityGroupIngress

Constructors

AuthorizeDBSecurityGroupIngress

Fields

_adbsgimDBSecurityGroupName :: Text

The name of the DB security group to add authorization to.

_adbsgimEC2SecurityGroupOwnerId :: Maybe Text

AWS Account Number of the owner of the EC2 security group specified in the EC2SecurityGroupName parameter. The AWS Access Key ID is not an acceptable value. For VPC DB security groups, EC2SecurityGroupId must be provided. Otherwise, EC2SecurityGroupOwnerId and either EC2SecurityGroupName or EC2SecurityGroupId must be provided.

_adbsgimEC2SecurityGroupName :: Maybe Text

Name of the EC2 security group to authorize. For VPC DB security groups, EC2SecurityGroupId must be provided. Otherwise, EC2SecurityGroupOwnerId and either EC2SecurityGroupName or EC2SecurityGroupId must be provided.

_adbsgimCIDRIP :: Maybe Text

The IP range to authorize.

_adbsgimEC2SecurityGroupId :: Maybe Text

Id of the EC2 security group to authorize. For VPC DB security groups, EC2SecurityGroupId must be provided. Otherwise, EC2SecurityGroupOwnerId and either EC2SecurityGroupName or EC2SecurityGroupId must be provided.

adbsgimEC2SecurityGroupOwnerId :: Lens' AuthorizeDBSecurityGroupIngress (Maybe Text)

adbsgimEC2SecurityGroupName :: Lens' AuthorizeDBSecurityGroupIngress (Maybe Text)

adbsgimEC2SecurityGroupId :: Lens' AuthorizeDBSecurityGroupIngress (Maybe Text)

adbsgimDBSecurityGroupName :: Lens' AuthorizeDBSecurityGroupIngress Text

adbsgimCIDRIP :: Lens' AuthorizeDBSecurityGroupIngress (Maybe Text)

data AuthorizeDBSecurityGroupIngressResponse

Constructors

AuthorizeDBSecurityGroupIngressResponse

Fields

_dbsgxDBSecurityGroup :: Maybe DBSecurityGroup

Contains the result of a successful invocation of the following actions: DescribeDBSecurityGroups AuthorizeDBSecurityGroupIngress CreateDBSecurityGroup RevokeDBSecurityGroupIngress This data type is used as a response element in the DescribeDBSecurityGroups action.

dbsgxDBSecurityGroup :: Iso' AuthorizeDBSecurityGroupIngressResponse (Maybe DBSecurityGroup)