Network.AWS.RDS.V2013_09_09.CreateOptionGroup

Description

Creates a new option group. You can create up to 20 option groups. https:/rds.amazonaws.com ?Action=CreateOptionGroup &OptionGroupName=myoptiongroup &EngineName=oracle-se1 &MajorEngineVersion=11.2 &OptionGroupDescription=Test option group 11.2 myoptiongroup oracle-se1 Test option group b2416a8a-84c9-11e1-a264-0b23c28bc344.

Documentation

createOptionGroup

Arguments

:: Text
_cogmOptionGroupDescription

-> Text
_cogmEngineName

-> Text
_cogmMajorEngineVersion

-> Text
_cogmOptionGroupName

-> CreateOptionGroup

Minimum specification for a CreateOptionGroup request.

data CreateOptionGroup

Constructors

CreateOptionGroup

Fields

_cogmOptionGroupDescription :: Text

The description of the option group.

_cogmEngineName :: Text

Specifies the name of the engine that this option group should be associated with.

_cogmMajorEngineVersion :: Text

Specifies the major version of the engine that this option group should be associated with.

_cogmOptionGroupName :: Text

Specifies the name of the option group to be created. Constraints: Must be 1 to 255 alphanumeric characters or hyphens First character must be a letter Cannot end with a hyphen or contain two consecutive hyphens Example: myoptiongroup.

_cogmTags :: [Tag]

A list of tags.

cogmTags :: Lens' CreateOptionGroup [Tag]

cogmOptionGroupName :: Lens' CreateOptionGroup Text

cogmOptionGroupDescription :: Lens' CreateOptionGroup Text

cogmMajorEngineVersion :: Lens' CreateOptionGroup Text

cogmEngineName :: Lens' CreateOptionGroup Text

data CreateOptionGroupResponse

Constructors

CreateOptionGroupResponse

Fields

_ogxOptionGroup :: Maybe OptionGroup

ogxOptionGroup :: Iso' CreateOptionGroupResponse (Maybe OptionGroup)