Network.AWS.RDS.V2013_09_09.DeleteDBSnapshot

Description

Deletes a DBSnapshot. If the snapshot is being copied, the copy operation is terminated. The DBSnapshot must be in the available state to be deleted. https:/rds.amazon.com &DBSnapshotIdentifier=mydbsnapshot &Version=2013-05-15 &SignatureVersion=2 &SignatureMethod=HmacSHA256 &Timestamp=2011-05-23T06%3A27%3A42.551Z &AWSAccessKeyId= &Signature= 3306 2011-03-11T07:20:24.082Z mysql deleted us-east-1d general-public-license 2010-07-16T00:06:59.107Z 60 simcoprod01 5.1.47 mysnapshot2 manual master 627a43a1-8507-11e0-bd9b-a7b1ece36d51.

Request

data DeleteDBSnapshot

Request constructor

mkDeleteDBSnapshot

Arguments

:: Text
ddbsDBSnapshotIdentifier

-> DeleteDBSnapshot

Smart constructor for the minimum required parameters to construct a valid DeleteDBSnapshot request.

Request lenses

ddbsDBSnapshotIdentifier :: Lens' DeleteDBSnapshot Text

The DBSnapshot identifier. Constraints: Must be the name of an existing DB snapshot in the available state.

Response

data DeleteDBSnapshotResponse

Response constructor

mkDeleteDBSnapshotResponse :: DeleteDBSnapshotResponse

Smart constructor for the minimum required parameters to construct a valid DeleteDBSnapshotResponse response.

This constructor is provided for convenience and testing purposes.

Response lenses

ddbsrDBSnapshot :: Lens' DeleteDBSnapshotResponse (Maybe DBSnapshot)

Contains the result of a successful invocation of the following actions: CreateDBSnapshot DeleteDBSnapshot This data type is used as a response element in the DescribeDBSnapshots action.