Network.AWS.RDS.V2013_09_09.DeleteEventSubscription

Description

Deletes an RDS event notification subscription. https:/rds.us-east-1.amazonaws.com ?Action=DeleteEventSubscription &SubscriptionName=EventSubscription01 &Version=2013-01-10 &SignatureVersion=4 &SignatureMethod=HmacSHA256 &Timestamp=20130128T012739Z &AWSAccessKeyId= &Signature= true 012345678901 db-instance deleting 2013-01-28 00:29:23.736 creation deletion EventSubscription01 arn:aws:sns:us-east-1:012345678901:EventSubscription01 e7cf30ac-68e9-11e2-bd13-a92da73b3119.

Request

data DeleteEventSubscription

Request constructor

mkDeleteEventSubscription

Arguments

:: Text
desSubscriptionName

-> DeleteEventSubscription

Smart constructor for the minimum required parameters to construct a valid DeleteEventSubscription request.

Request lenses

desSubscriptionName :: Lens' DeleteEventSubscription Text

The name of the RDS event notification subscription you want to delete.

Response

data DeleteEventSubscriptionResponse

Response constructor

mkDeleteEventSubscriptionResponse :: DeleteEventSubscriptionResponse

Smart constructor for the minimum required parameters to construct a valid DeleteEventSubscriptionResponse response.

This constructor is provided for convenience and testing purposes.

Response lenses

desrEventSubscription :: Lens' DeleteEventSubscriptionResponse (Maybe EventSubscription)

Contains the results of a successful invocation of the DescribeEventSubscriptions action.