Network.AWS.RDS.V2013_09_09.DeleteOptionGroup

Description

Deletes an existing option group. https:/rds.amazonaws.com ?Action=DeleteOptionGroup &OptionGroupName=myoptiongroup.

Request

data DeleteOptionGroup

Request constructor

mkDeleteOptionGroup

Arguments

:: Text
dogOptionGroupName

-> DeleteOptionGroup

Smart constructor for the minimum required parameters to construct a valid DeleteOptionGroup request.

Request lenses

dogOptionGroupName :: Lens' DeleteOptionGroup Text

The name of the option group to be deleted. You cannot delete default option groups.

Response

data DeleteOptionGroupResponse