Network.AWS.RDS.V2013_09_09.DescribeDBInstances

Description

Returns information about provisioned RDS instances. This API supports pagination. https:/rds.amazonaws.com ?Action=DescribeDBInstances &Version=2013-05-15 &MaxRecords=100 &SignatureVersion=2 &SignatureMethod=HmacSHA256 &Timestamp=2011-05-23T06%3A54%3A55.116Z &AWSAccessKeyId= &Signature= 2011-05-23T06:50:00Z mysql 1 false general-public-license available 5.1.50 3306 simcoprod01.cu7u2t4uz396.us-east-1.rds.amazonaws.com simcoprod01 in-sync default.mysql5.1 active default 00:00-00:30 true sat:07:30-sat:08:00 us-east-1a 2011-05-23T06:06:43.110Z 10 default.mysql5.1 in-sync db.m1.large master 9135fff3-8509-11e0-bd9b-a7b1ece36d51.

Request

data DescribeDBInstances

Request constructor

mkDescribeDBInstances :: DescribeDBInstances

Smart constructor for the minimum required parameters to construct a valid DescribeDBInstances request.

The fields accessible through corresponding lenses are:

DBInstanceIdentifier :: Maybe Text

MaxRecords :: Maybe Integer

Marker :: Maybe Text

Request lenses

ddbi1DBInstanceIdentifier :: Lens' DescribeDBInstances (Maybe Text)

The user-supplied instance identifier. If this parameter is specified, information from only the specific DB instance is returned. This parameter isn't case sensitive. Constraints: Must contain from 1 to 63 alphanumeric characters or hyphens First character must be a letter Cannot end with a hyphen or contain two consecutive hyphens.

ddbi1MaxRecords :: Lens' DescribeDBInstances (Maybe Integer)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results may be retrieved. Default: 100 Constraints: minimum 20, maximum 100.

ddbi1Marker :: Lens' DescribeDBInstances (Maybe Text)

An optional pagination token provided by a previous DescribeDBInstances request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

Response

data DescribeDBInstancesResponse

Contains the result of a successful invocation of the DescribeDBInstances action.

Response constructor

mkDescribeDBInstancesResponse :: DescribeDBInstancesResponse

Smart constructor for the minimum required parameters to construct a valid DescribeDBInstancesResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Marker :: Maybe Text

DBInstances :: [DBInstance]

Response lenses

ddbirrMarker :: Lens' DescribeDBInstancesResponse (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

ddbirrDBInstances :: Lens' DescribeDBInstancesResponse [DBInstance]

A list of DBInstance instances.