Network.AWS.RDS.V2013_09_09.DescribeDBParameterGroups

Description

Returns a list of DBParameterGroup descriptions. If a DBParameterGroupName is specified, the list will contain only the description of the specified DB parameter group. https:/rds.amazonaws.com ?Action=DescribeDBParameterGroups &DBParameterGroupName=myparamsgroup &MaxRecords=100 &Version=2013-05-15 &SignatureVersion=2 &SignatureMethod=HmacSHA256 &Timestamp=2011-02-15T17%3A54%3A32.899Z &AWSAccessKeyId= &Signature= mysql5.1 Default parameter group for mysql5.1 default.mysql5.1 mysql5.1 My DB Param Group testdbparamgroup cb8d9bb4-a02a-11df-bd60-c955b7d6e8e0.

Request

data DescribeDBParameterGroups

Request constructor

mkDescribeDBParameterGroups :: DescribeDBParameterGroups

Smart constructor for the minimum required parameters to construct a valid DescribeDBParameterGroups request.

The fields accessible through corresponding lenses are:

DBParameterGroupName :: Maybe Text

MaxRecords :: Maybe Integer

Marker :: Maybe Text

Request lenses

ddbpg1DBParameterGroupName :: Lens' DescribeDBParameterGroups (Maybe Text)

The name of a specific DB parameter group to return details for. Constraints: Must be 1 to 255 alphanumeric characters First character must be a letter Cannot end with a hyphen or contain two consecutive hyphens.

ddbpg1MaxRecords :: Lens' DescribeDBParameterGroups (Maybe Integer)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results may be retrieved. Default: 100 Constraints: minimum 20, maximum 100.

ddbpg1Marker :: Lens' DescribeDBParameterGroups (Maybe Text)

An optional pagination token provided by a previous DescribeDBParameterGroups request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

Response

data DescribeDBParameterGroupsResponse

Contains the result of a successful invocation of the DescribeDBParameterGroups action.

Response constructor

mkDescribeDBParameterGroupsResponse :: DescribeDBParameterGroupsResponse

Smart constructor for the minimum required parameters to construct a valid DescribeDBParameterGroupsResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Marker :: Maybe Text

DBParameterGroups :: [DBParameterGroup]

Response lenses

ddbpgrMarker :: Lens' DescribeDBParameterGroupsResponse (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

ddbpgrDBParameterGroups :: Lens' DescribeDBParameterGroupsResponse [DBParameterGroup]

A list of DBParameterGroup instances.