Network.AWS.RDS.V2013_09_09.DescribeDBParameters

Description

Returns the detailed parameter list for a particular DB parameter group. https:/rds.amazonaws.com ?Action=DescribeDBParameters &DBParameterGroupName=mydbparametergroup &Source=system &MaxRecords=100 &Version=2013-05-15 &SignatureVersion=2 &SignatureMethod=HmacSHA256 &Timestamp=2011-05-11T19%3A31%3A42.262Z &AWSAccessKeyId= &Signature= rdsdbbinmysql string system false The MySQL installation base directory. static basedir 32768 integer system true The size of the cache to hold the SQL statements for the binary log during a transaction. dynamic 4096-9223372036854775807 binlog_cache_size 8743f2cf-bf41-11de-8c8e-49155882c409.

Documentation

describeDBParameters

Arguments

:: Text
_ddbpmDBParameterGroupName

-> DescribeDBParameters

Minimum specification for a DescribeDBParameters request.

data DescribeDBParameters

Constructors

DescribeDBParameters

Fields

_ddbpmDBParameterGroupName :: Text

The name of a specific DB parameter group to return details for. Constraints: Must be 1 to 255 alphanumeric characters First character must be a letter Cannot end with a hyphen or contain two consecutive hyphens.

_ddbpmMaxRecords :: Maybe Integer

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results may be retrieved. Default: 100 Constraints: minimum 20, maximum 100.

_ddbpmMarker :: Maybe Text

An optional pagination token provided by a previous DescribeDBParameters request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

_ddbpmSource :: Maybe Text

The parameter types to return. Default: All parameter types returned Valid Values: user | system | engine-default.

ddbpmSource :: Lens' DescribeDBParameters (Maybe Text)

ddbpmMaxRecords :: Lens' DescribeDBParameters (Maybe Integer)

ddbpmMarker :: Lens' DescribeDBParameters (Maybe Text)

ddbpmDBParameterGroupName :: Lens' DescribeDBParameters Text

data DescribeDBParametersResponse

Constructors

DescribeDBParametersResponse

Fields

_dbpgdParameters :: [Parameter]

A list of Parameter values.

_dbpgdMarker :: Maybe Text

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

dbpgdParameters :: Lens' DescribeDBParametersResponse [Parameter]

dbpgdMarker :: Lens' DescribeDBParametersResponse (Maybe Text)