Network.AWS.RDS.V2013_09_09.DescribeDBSecurityGroups

Description

Returns a list of DBSecurityGroup descriptions. If a DBSecurityGroupName is specified, the list will contain only the descriptions of the specified DB security group. https:/rds.amazonaws.com ?Action=DescribeDBSecurityGroups &Version=2013-05-15 &MaxRecords=100 &SignatureVersion=2 &SignatureMethod=HmacSHA256 &Timestamp=2011-02-15T19%3A40%3A19.926Z &AWSAccessKeyId= &Signature= authorized myec2securitygroup 054794666394 default 127.0.0.1/30 authorized 621567473609 default vpc-1ab2c3d4 My new DBSecurityGroup 192.168.1.1/24 authorized 621567473609 mydbsecuritygroup vpc-1ab2c3d5 My new DBSecurityGroup 621567473609 mydbsecuritygroup4 vpc-1ab2c3d6 bbdad154-bf42-11de-86a4-97241dfaadff.

Documentation

describeDBSecurityGroups :: DescribeDBSecurityGroups

Minimum specification for a DescribeDBSecurityGroups request.

data DescribeDBSecurityGroups

Constructors

DescribeDBSecurityGroups

Fields

_ddbsgpMaxRecords :: Maybe Integer

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results may be retrieved. Default: 100 Constraints: minimum 20, maximum 100.

_ddbsgpMarker :: Maybe Text

An optional pagination token provided by a previous DescribeDBSecurityGroups request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

_ddbsgpDBSecurityGroupName :: Maybe Text

The name of the DB security group to return details for.

ddbsgpMaxRecords :: Lens' DescribeDBSecurityGroups (Maybe Integer)

ddbsgpMarker :: Lens' DescribeDBSecurityGroups (Maybe Text)

ddbsgpDBSecurityGroupName :: Lens' DescribeDBSecurityGroups (Maybe Text)

data DescribeDBSecurityGroupsResponse

Constructors

DescribeDBSecurityGroupsResponse

Fields

_dbsgdeDBSecurityGroups :: [DBSecurityGroup]

A list of DBSecurityGroup instances.

_dbsgdeMarker :: Maybe Text

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

dbsgdeMarker :: Lens' DescribeDBSecurityGroupsResponse (Maybe Text)

dbsgdeDBSecurityGroups :: Lens' DescribeDBSecurityGroupsResponse [DBSecurityGroup]