Network.AWS.RDS.V2013_09_09.DescribeDBSubnetGroups

Description

Returns a list of DBSubnetGroup descriptions. If a DBSubnetGroupName is specified, the list will contain only the descriptions of the specified DBSubnetGroup. For an overview of CIDR ranges, go to the Wikipedia Tutorial. https:/rds.amazonaws.com ?Action=DescribeDBSubnetGroups &Version=2013-05-15 &MaxRecords=100 &SignatureVersion=2 &SignatureMethod=HmacSHA256 &Timestamp=2011-02-15T19%3A40%3A19.926Z &AWSAccessKeyId= &Signature= 990524496922 Complete description subnet_grp1 Active subnet-7c5b4115 us-east-1c Active subnet-7b5b4112 us-east-1b Active subnet-3ea6bd57 us-east-1d 990524496922 Complete description subnet_grp2 Active subnet-7c5b4115 us-east-1c Active subnet-7b5b4112 us-east-1b Active subnet-3ea6bd57 us-east-1d 31d0faee-229b-11e1-81f1-df3a2a803dad.

Documentation

describeDBSubnetGroups :: DescribeDBSubnetGroups

Minimum specification for a DescribeDBSubnetGroups request.

data DescribeDBSubnetGroups

Constructors

DescribeDBSubnetGroups

Fields

_ddbsgmMaxRecords :: Maybe Integer

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results may be retrieved. Default: 100 Constraints: minimum 20, maximum 100.

_ddbsgmDBSubnetGroupName :: Maybe Text

The name of the DB subnet group to return details for.

_ddbsgmMarker :: Maybe Text

An optional pagination token provided by a previous DescribeDBSubnetGroups request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

ddbsgmMaxRecords :: Lens' DescribeDBSubnetGroups (Maybe Integer)

ddbsgmMarker :: Lens' DescribeDBSubnetGroups (Maybe Text)

ddbsgmDBSubnetGroupName :: Lens' DescribeDBSubnetGroups (Maybe Text)

data DescribeDBSubnetGroupsResponse

Constructors

DescribeDBSubnetGroupsResponse

Fields

_dbsgrDBSubnetGroups :: [DBSubnetGroup]

A list of DBSubnetGroup instances.

_dbsgrMarker :: Maybe Text

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

dbsgrMarker :: Lens' DescribeDBSubnetGroupsResponse (Maybe Text)

dbsgrDBSubnetGroups :: Lens' DescribeDBSubnetGroupsResponse [DBSubnetGroup]