Network.AWS.RDS.V2013_09_09.DescribeOptionGroups

Description

Describes the available option groups. https:/rds.amazonaws.com ?Action=DescribeOptionGroups &OptionGroupName=myoptiongroup &MaxRecords=100 11.2 myoptiongroup oracle-se1 Test option group 6088823d-84c8-11e1-a264-0b23c28bc344 https:/rds.amazonaws.com ?Action=DescribeOptionGroups &MaxRecords=100 11.2 myoptiongroup oracle-se1 Test option group 11.2 default:oracle-se1-11-2 oracle-se1 Default option group. e4b234d9-84d5-11e1-87a6-71059839a52b.

Request

data DescribeOptionGroups

Request constructor

mkDescribeOptionGroups :: DescribeOptionGroups

Smart constructor for the minimum required parameters to construct a valid DescribeOptionGroups request.

Request lenses

dog1OptionGroupName :: Lens' DescribeOptionGroups (Maybe Text)

The name of the option group to describe. Cannot be supplied together with EngineName or MajorEngineVersion.

dog1Marker :: Lens' DescribeOptionGroups (Maybe Text)

An optional pagination token provided by a previous DescribeOptionGroups request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

dog1MaxRecords :: Lens' DescribeOptionGroups (Maybe Integer)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: minimum 20, maximum 100.

dog1EngineName :: Lens' DescribeOptionGroups (Maybe Text)

Filters the list of option groups to only include groups associated with a specific database engine.

dog1MajorEngineVersion :: Lens' DescribeOptionGroups (Maybe Text)

Filters the list of option groups to only include groups associated with a specific database engine version. If specified, then EngineName must also be specified.

Response

data DescribeOptionGroupsResponse

List of option groups.

Response constructor

mkDescribeOptionGroupsResponse :: DescribeOptionGroupsResponse

Smart constructor for the minimum required parameters to construct a valid DescribeOptionGroupsResponse response.

This constructor is provided for convenience and testing purposes.

Response lenses

dogrOptionGroupsList :: Lens' DescribeOptionGroupsResponse [OptionGroup]

List of option groups.

dogrMarker :: Lens' DescribeOptionGroupsResponse (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.