Network.AWS.RDS.V2013_09_09.DescribeReservedDBInstances

Description

Returns information about reserved DB instances for this account, or about a specified reserved DB instance. https:/rds.amazonaws.com ?Action=DescribeReservedDBInstances &ReservedDBInstanceId=customerSpecifiedID &SignatureVersion=2 &SignatureMethod=HmacSHA256 &Timestamp=2012-12-18T18%3A31%3A36.118Z &AWSAccessKeyId= &Signature= Medium Utilization USD mysql 649fd0c8-cf6d-47a0-bfa6-060f8e75e95f false active myreservationid 1 2010-12-15T00:25:14.131Z 31536000 227.5 0.046 db.m1.small c695119b-2961-11e1-bd06-6fe008f046c3.

Request

data DescribeReservedDBInstances

Request constructor

mkDescribeReservedDBInstances :: DescribeReservedDBInstances

Smart constructor for the minimum required parameters to construct a valid DescribeReservedDBInstances request.

Request lenses

drdbiReservedDBInstanceId :: Lens' DescribeReservedDBInstances (Maybe Text)

The reserved DB instance identifier filter value. Specify this parameter to show only the reservation that matches the specified reservation ID.

drdbiReservedDBInstancesOfferingId :: Lens' DescribeReservedDBInstances (Maybe Text)

The offering identifier filter value. Specify this parameter to show only purchased reservations matching the specified offering identifier.

drdbiDBInstanceClass :: Lens' DescribeReservedDBInstances (Maybe Text)

The DB instance class filter value. Specify this parameter to show only those reservations matching the specified DB instances class.

drdbiDuration :: Lens' DescribeReservedDBInstances (Maybe Text)

The duration filter value, specified in years or seconds. Specify this parameter to show only reservations for this duration. Valid Values: 1 | 3 | 31536000 | 94608000.

drdbiProductDescription :: Lens' DescribeReservedDBInstances (Maybe Text)

The product description filter value. Specify this parameter to show only those reservations matching the specified product description.

drdbiOfferingType :: Lens' DescribeReservedDBInstances (Maybe Text)

The offering type filter value. Specify this parameter to show only the available offerings matching the specified offering type. Valid Values: "Light Utilization" | "Medium Utilization" | "Heavy Utilization".

drdbiMultiAZ :: Lens' DescribeReservedDBInstances (Maybe Bool)

The Multi-AZ filter value. Specify this parameter to show only those reservations matching the specified Multi-AZ parameter.

drdbiMaxRecords :: Lens' DescribeReservedDBInstances (Maybe Integer)

The maximum number of records to include in the response. If more than the MaxRecords value is available, a pagination token called a marker is included in the response so that the following results can be retrieved. Default: 100 Constraints: minimum 20, maximum 100.

drdbiMarker :: Lens' DescribeReservedDBInstances (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

Response

data DescribeReservedDBInstancesResponse

Contains the result of a successful invocation of the DescribeReservedDBInstances action.

Response lenses

drdbirsMarker :: Lens' DescribeReservedDBInstancesResponse (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

drdbirsReservedDBInstances :: Lens' DescribeReservedDBInstancesResponse [ReservedDBInstance]

A list of reserved DB instances.