Network.AWS.RDS.V2013_09_09.DownloadDBLogFilePortion

Description

Downloads the last line of the specified log file. https:/rds.amazonaws.com ?DBInstanceIdentifier=rra-mysql &MaxRecords=100 &Version=2013-05-15 &Action=DescribeDBLogFiles &SignatureVersion=4 &SignatureMethod=HmacSHA256 &Timestamp=20130327T173621Z &X-Amz-Algorithm=AWS4-HMAC-SHA256 &X-Amz-Date=20130327T173621Z &X-Amz-SignedHeaders=Host &X-Amz-Expires=20130327T173621Z &X-Amz-Credential= &X-Amz-Signature= 1364403600000 error/mysql-error-running.log 0 1364338800000 error/mysql-error-running.log.0 0 1364342400000 error/mysql-error-running.log.1 0 1364371200000 errormysql-error-running.log.9 0 1364405700000 errormysql-error.log 0 d70fb3b3-9704-11e2-a0db-871552e0ef19.

Request

data DownloadDBLogFilePortion

Request constructor

mkDownloadDBLogFilePortion

Arguments

:: Text
ddblfpDBInstanceIdentifier

-> Text
ddblfpLogFileName

-> DownloadDBLogFilePortion

Smart constructor for the minimum required parameters to construct a valid DownloadDBLogFilePortion request.

Request lenses

ddblfpDBInstanceIdentifier :: Lens' DownloadDBLogFilePortion Text

The customer-assigned name of the DB instance that contains the log files you want to list. Constraints: Must contain from 1 to 63 alphanumeric characters or hyphens First character must be a letter Cannot end with a hyphen or contain two consecutive hyphens.

ddblfpLogFileName :: Lens' DownloadDBLogFilePortion Text

The name of the log file to be downloaded.

ddblfpMarker :: Lens' DownloadDBLogFilePortion (Maybe Text)

The pagination token provided in the previous request. If this parameter is specified the response includes only records beyond the marker, up to MaxRecords.

ddblfpNumberOfLines :: Lens' DownloadDBLogFilePortion (Maybe Integer)

The number of lines remaining to be downloaded.

Response

data DownloadDBLogFilePortionResponse

This data type is used as a response element to DownloadDBLogFilePortion.

Response lenses

ddblfprAdditionalDataPending :: Lens' DownloadDBLogFilePortionResponse Bool

Boolean value that if true, indicates there is more data to be downloaded.

ddblfprLogFileData :: Lens' DownloadDBLogFilePortionResponse ByteString

Entries from the specified log file.

ddblfprMarker :: Lens' DownloadDBLogFilePortionResponse (Maybe Text)

An optional pagination token provided by a previous DownloadDBLogFilePortion request.