Network.AWS.RDS.V2013_09_09.ModifyDBSubnetGroup

Description

Modifies an existing DB subnet group. DB subnet groups must contain at least one subnet in at least two AZs in the region. https:/rds.amazonaws.com ?Action=ModifyDBSubnetGroup &DBSubnetGroupName=mydbsubnetgroup &DBSubnetGroupDescription=My%20modified%20DBSubnetGroup &Version=2013-05-15 &SignatureVersion=2 &SignatureMethod=HmacSHA256 &Timestamp=2011-02-15T18%3A14%3A49.482Z &AWSAccessKeyId= &Signature= 990524496922 Complete My modified DBSubnetGroup mydbsubnetgroup Active subnet-7c5b4115 us-east-1c Active subnet-7b5b4112 us-east-1b Active subnet-3ea6bd57 us-east-1d ed662948-a57b-11df-9e38-7ffab86c801f.

Request

data ModifyDBSubnetGroup

Request constructor

mkModifyDBSubnetGroup

Arguments

:: Text
mdbsgDBSubnetGroupName

-> [Text]
mdbsgSubnetIds

-> ModifyDBSubnetGroup

Smart constructor for the minimum required parameters to construct a valid ModifyDBSubnetGroup request.

Request lenses

mdbsgDBSubnetGroupName :: Lens' ModifyDBSubnetGroup Text

The name for the DB subnet group. This value is stored as a lowercase string. Constraints: Must contain no more than 255 alphanumeric characters or hyphens. Must not be Default. Example: mySubnetgroup.

mdbsgDBSubnetGroupDescription :: Lens' ModifyDBSubnetGroup (Maybe Text)

The description for the DB subnet group.

mdbsgSubnetIds :: Lens' ModifyDBSubnetGroup [Text]

The EC2 subnet IDs for the DB subnet group.

Response

data ModifyDBSubnetGroupResponse

Response lenses

mdbsgrsDBSubnetGroup :: Lens' ModifyDBSubnetGroupResponse (Maybe DBSubnetGroup)

Contains the result of a successful invocation of the following actions: CreateDBSubnetGroup ModifyDBSubnetGroup DescribeDBSubnetGroups DeleteDBSubnetGroup This data type is used as a response element in the DescribeDBSubnetGroups action.