Network.AWS.RDS.V2013_09_09.ModifyEventSubscription

Description

Modifies an existing RDS event notification subscription. Note that you cannot modify the source identifiers using this call; to change source identifiers for a subscription, use the AddSourceIdentifierToSubscription and RemoveSourceIdentifierFromSubscription calls. You can see a list of the event categories for a given SourceType in the Events topic in the Amazon RDS User Guide or by using the DescribeEventCategories action. https:/rds.us-east-1.amazonaws.com ?Action=ModifyEventSubscription &SubscriptionName=EventSubscription01 &EventCategories.member.1=creation &EventCategories.member.2=deletion &SourceType=db-instance &Enabled=true &Version=2013-01-10 &SignatureVersion=4 &SignatureMethod=HmacSHA256 &Timestamp=20130128T005359Z &AWSAccessKeyId= &Signature= true 012345678901 db-instance modifying 2013-01-28 00:29:23.736 creation deletion EventSubscription01 arn:aws:sns:us-east-1:012345678901:EventSubscription01 34907d48-68e5-11e2-98ef-2b071ac20a57.

Documentation

modifyEventSubscription

Arguments

:: Text
_mesmSubscriptionName

-> ModifyEventSubscription

Minimum specification for a ModifyEventSubscription request.

data ModifyEventSubscription

Constructors

ModifyEventSubscription

Fields

_mesmSubscriptionName :: Text

The name of the RDS event notification subscription.

_mesmEnabled :: Maybe Bool

A Boolean value; set to true to activate the subscription.

_mesmEventCategories :: [Text]

A list of event categories for a SourceType that you want to subscribe to. You can see a list of the categories for a given SourceType in the Events topic in the Amazon RDS User Guide or by using the DescribeEventCategories action.

_mesmSnsTopicArn :: Maybe Text

The Amazon Resource Name (ARN) of the SNS topic created for event notification. The ARN is created by Amazon SNS when you create a topic and subscribe to it.

_mesmSourceType :: Maybe Text

The type of source that will be generating the events. For example, if you want to be notified of events generated by a DB instance, you would set this parameter to db-instance. if this value is not specified, all events are returned. Valid values: db-instance | db-parameter-group | db-security-group | db-snapshot.

mesmSubscriptionName :: Lens' ModifyEventSubscription Text

mesmSourceType :: Lens' ModifyEventSubscription (Maybe Text)

mesmSnsTopicArn :: Lens' ModifyEventSubscription (Maybe Text)

mesmEventCategories :: Lens' ModifyEventSubscription [Text]

mesmEnabled :: Lens' ModifyEventSubscription (Maybe Bool)

data ModifyEventSubscriptionResponse

Constructors

ModifyEventSubscriptionResponse

Fields

_esxEventSubscription :: Maybe EventSubscription

Contains the results of a successful invocation of the DescribeEventSubscriptions action.

esxEventSubscription :: Iso' ModifyEventSubscriptionResponse (Maybe EventSubscription)