Network.AWS.RDS.V2013_09_09.ModifyOptionGroup

Description

Modifies an existing option group. https:/rds.amazonaws.com ?Action=ModifyOptionGroup &OptionGroupName=myoptiongroup &OptionsToInclude=OEM &DBSecurityGroupMemberships=default &ApplyImmediately=true myoptiongroup Test option group oracle-se1 11.2 OEM Oracle Enterprise Manager 1158 default ACTIVE ed662948-a57b-11df-9e38-7ffab86c801f https:/rds.amazonaws.com ?Action=ModifyOptionGroup &OptionGroupName=myoptiongroup &OptionsToRemove=OEM &ApplyImmediately=true myoptiongroup Test option group oracle-se1 11.2 ed662948-a57b-11df-9e38-7ffab86c801f.

Request

data ModifyOptionGroup

Request constructor

mkModifyOptionGroup

Arguments

:: Text
mogOptionGroupName

-> ModifyOptionGroup

Smart constructor for the minimum required parameters to construct a valid ModifyOptionGroup request.

The fields accessible through corresponding lenses are:

OptionGroupName :: Text

OptionsToInclude :: [OptionConfiguration]

OptionsToRemove :: [Text]

ApplyImmediately :: Maybe Bool

Request lenses

mogOptionGroupName :: Lens' ModifyOptionGroup Text

The name of the option group to be modified. cannot be removed from an option group while DB instances are associated with the option group. --> Permanent options, such as the TDE option for Oracle Advanced Security TDE, cannot be removed from an option group, and that option group cannot be removed from a DB instance once it is associated with a DB instance.

mogOptionsToInclude :: Lens' ModifyOptionGroup [OptionConfiguration]

Options in this list are added to the option group or, if already present, the specified configuration is used to update the existing configuration.

mogOptionsToRemove :: Lens' ModifyOptionGroup [Text]

Options in this list are removed from the option group.

mogApplyImmediately :: Lens' ModifyOptionGroup (Maybe Bool)

Indicates whether the changes should be applied immediately, or during the next maintenance window for each instance associated with the option group.

Response

data ModifyOptionGroupResponse

Response constructor

mkModifyOptionGroupResponse :: ModifyOptionGroupResponse

Smart constructor for the minimum required parameters to construct a valid ModifyOptionGroupResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

OptionGroup :: Maybe OptionGroup

Response lenses

mogrOptionGroup :: Lens' ModifyOptionGroupResponse (Maybe OptionGroup)