Network.AWS.RDS.V2013_09_09.RemoveSourceIdentifierFromSubscription

Description

Removes a source identifier from an existing RDS event notification subscription. https:/rds.us-east-1.amazonaws.com ?Action=RemoveSourceIdentifierFromSubscription &SubscriptionName=EventSubscription01 &SourceIdentifier=dbinstance01 &Version=2013-01-10 &SignatureVersion=4 &SignatureMethod=HmacSHA256 &Timestamp=20130128T012415Z &AWSAccessKeyId= &Signature= true 012345678901 db-instance modifying 2013-01-28 00:29:23.736 creation deletion EventSubscription01 arn:aws:sns:us-east-1:012345678901:EventSubscription01 6f0b82bf-68e9-11e2-b97b-43c6362ec60d.

Documentation

data RemoveSourceIdentifierFromSubscription

Constructors

RemoveSourceIdentifierFromSubscription

Fields

_rsifsmSubscriptionName :: Text

The name of the RDS event notification subscription you want to remove a source identifier from.

_rsifsmSourceIdentifier :: Text

The source identifier to be removed from the subscription, such as the DB instance identifier for a DB instance or the name of a security group.

rsifsmSubscriptionName :: Lens' RemoveSourceIdentifierFromSubscription Text

rsifsmSourceIdentifier :: Lens' RemoveSourceIdentifierFromSubscription Text

data RemoveSourceIdentifierFromSubscriptionResponse

Constructors

RemoveSourceIdentifierFromSubscriptionResponse

Fields

_esyEventSubscription :: Maybe EventSubscription

Contains the results of a successful invocation of the DescribeEventSubscriptions action.

esyEventSubscription :: Iso' RemoveSourceIdentifierFromSubscriptionResponse (Maybe EventSubscription)