Network.AWS.RDS.V2013_09_09.RevokeDBSecurityGroupIngress

Description

Revokes ingress from a DBSecurityGroup for previously authorized IP ranges or EC2 or VPC Security Groups. Required parameters for this API are one of CIDRIP, EC2SecurityGroupId for VPC, or (EC2SecurityGroupOwnerId and either EC2SecurityGroupName or EC2SecurityGroupId). https:/rds.amazonaws.com ?Action=RevokeDBSecurityGroupIngress &DBSecurityGroupName=mydbsecuritygroup &CIDRIP=192.168.1.1%2F24 &Version=2013-05-15 &SignatureVersion=2&SignatureMethod=HmacSHA256 &Timestamp=2011-02-15T22%3A32%3A12.515Z &AWSAccessKeyId= &Signature= My new DBSecurityGroup 192.168.1.1/24 revoking 621567473609 mydbsecuritygroup vpc-1ab2c3d4 beecb8ac-bf5a-11de-9f9f-53d6aee22de9.

Documentation

revokeDBSecurityGroupIngress

Arguments

:: Text
_rdbsgimDBSecurityGroupName

-> RevokeDBSecurityGroupIngress

Minimum specification for a RevokeDBSecurityGroupIngress request.

data RevokeDBSecurityGroupIngress

Constructors

RevokeDBSecurityGroupIngress

Fields

_rdbsgimDBSecurityGroupName :: Text

The name of the DB security group to revoke ingress from.

_rdbsgimEC2SecurityGroupOwnerId :: Maybe Text

The AWS Account Number of the owner of the EC2 security group specified in the EC2SecurityGroupName parameter. The AWS Access Key ID is not an acceptable value. For VPC DB security groups, EC2SecurityGroupId must be provided. Otherwise, EC2SecurityGroupOwnerId and either EC2SecurityGroupName or EC2SecurityGroupId must be provided.

_rdbsgimEC2SecurityGroupName :: Maybe Text

The name of the EC2 security group to revoke access from. For VPC DB security groups, EC2SecurityGroupId must be provided. Otherwise, EC2SecurityGroupOwnerId and either EC2SecurityGroupName or EC2SecurityGroupId must be provided.

_rdbsgimCIDRIP :: Maybe Text

The IP range to revoke access from. Must be a valid CIDR range. If CIDRIP is specified, EC2SecurityGroupName, EC2SecurityGroupId and EC2SecurityGroupOwnerId cannot be provided.

_rdbsgimEC2SecurityGroupId :: Maybe Text

The id of the EC2 security group to revoke access from. For VPC DB security groups, EC2SecurityGroupId must be provided. Otherwise, EC2SecurityGroupOwnerId and either EC2SecurityGroupName or EC2SecurityGroupId must be provided.

rdbsgimEC2SecurityGroupOwnerId :: Lens' RevokeDBSecurityGroupIngress (Maybe Text)

rdbsgimEC2SecurityGroupName :: Lens' RevokeDBSecurityGroupIngress (Maybe Text)

rdbsgimEC2SecurityGroupId :: Lens' RevokeDBSecurityGroupIngress (Maybe Text)

rdbsgimDBSecurityGroupName :: Lens' RevokeDBSecurityGroupIngress Text

rdbsgimCIDRIP :: Lens' RevokeDBSecurityGroupIngress (Maybe Text)

data RevokeDBSecurityGroupIngressResponse

Constructors

RevokeDBSecurityGroupIngressResponse

Fields

_dbsgyDBSecurityGroup :: Maybe DBSecurityGroup

Contains the result of a successful invocation of the following actions: DescribeDBSecurityGroups AuthorizeDBSecurityGroupIngress CreateDBSecurityGroup RevokeDBSecurityGroupIngress This data type is used as a response element in the DescribeDBSecurityGroups action.

dbsgyDBSecurityGroup :: Iso' RevokeDBSecurityGroupIngressResponse (Maybe DBSecurityGroup)