Network.AWS.RDS.V2013_09_09.Types

Description

Amazon Relational Database Service (Amazon RDS) is a web service that makes it easy to set up, operate, and scale a relational database in the cloud. It provides cost-efficient and resizable capacity while managing time-consuming database administration tasks, freeing you up to focus on your applications and business.

Service

data RDS

Supported version (2013-09-09) of the Amazon Relational Database Service service.

XML

xmlOptions :: Tagged a XMLOptions

ApplyMethod

data ApplyMethod

Constructors

ApplyMethodImmediate
immediate

ApplyMethodPendingReboot
pending-reboot

SourceType

data SourceType

Constructors

SourceTypeDbInstance
db-instance

SourceTypeDbParameterGroup
db-parameter-group

SourceTypeDbSecurityGroup
db-security-group

SourceTypeDbSnapshot
db-snapshot

AvailabilityZone

data AvailabilityZone

Contains Availability Zone information. This data type is used as an element in the following data type: OrderableDBInstanceOption.

mkAvailabilityZone :: AvailabilityZone

Smart constructor for the minimum required fields to construct a valid AvailabilityZone data type to populate a request.

The fields accessible through corresponding lenses are:

Name :: Maybe Text

ProvisionedIopsCapable :: Maybe Bool

azName :: Lens' AvailabilityZone (Maybe Text)

The name of the availability zone.

azProvisionedIopsCapable :: Lens' AvailabilityZone (Maybe Bool)

True indicates the availability zone is capable of provisioned IOPs.

CharacterSet

data CharacterSet

The default character set for new instances of this engine version, if the CharacterSetName parameter of the CreateDBInstance API is not specified.

mkCharacterSet :: CharacterSet

Smart constructor for the minimum required fields to construct a valid CharacterSet data type to populate a request.

The fields accessible through corresponding lenses are:

CharacterSetName :: Maybe Text

CharacterSetDescription :: Maybe Text

csCharacterSetName :: Lens' CharacterSet (Maybe Text)

The name of the character set.

csCharacterSetDescription :: Lens' CharacterSet (Maybe Text)

The description of the character set.

DBEngineVersion

data DBEngineVersion

This data type is used as a response element in the action DescribeDBEngineVersions.

mkDBEngineVersion :: DBEngineVersion

Smart constructor for the minimum required fields to construct a valid DBEngineVersion data type.

DBEngineVersion is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Engine :: Maybe Text

EngineVersion :: Maybe Text

DBParameterGroupFamily :: Maybe Text

DBEngineDescription :: Maybe Text

DBEngineVersionDescription :: Maybe Text

DefaultCharacterSet :: Maybe CharacterSet

SupportedCharacterSets :: [CharacterSet]

dbevEngine :: Lens' DBEngineVersion (Maybe Text)

The name of the database engine.

dbevEngineVersion :: Lens' DBEngineVersion (Maybe Text)

The version number of the database engine.

dbevDBParameterGroupFamily :: Lens' DBEngineVersion (Maybe Text)

The name of the DB parameter group family for the database engine.

dbevDBEngineDescription :: Lens' DBEngineVersion (Maybe Text)

The description of the database engine.

dbevDBEngineVersionDescription :: Lens' DBEngineVersion (Maybe Text)

The description of the database engine version.

dbevDefaultCharacterSet :: Lens' DBEngineVersion (Maybe CharacterSet)

The default character set for new instances of this engine version, if the CharacterSetName parameter of the CreateDBInstance API is not specified.

dbevSupportedCharacterSets :: Lens' DBEngineVersion [CharacterSet]

A list of the character sets supported by this engine for the CharacterSetName parameter of the CreateDBInstance API.

DBInstance

data DBInstance

Contains the result of a successful invocation of the following actions: CreateDBInstance DeleteDBInstance ModifyDBInstance This data type is used as a response element in the DescribeDBInstances action.

mkDBInstance :: DBInstance

Smart constructor for the minimum required fields to construct a valid DBInstance data type.

DBInstance is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

DBInstanceIdentifier :: Maybe Text

DBInstanceClass :: Maybe Text

Engine :: Maybe Text

DBInstanceStatus :: Maybe Text

MasterUsername :: Maybe Text

DBName :: Maybe Text

Endpoint :: Maybe Endpoint

AllocatedStorage :: Maybe Integer

InstanceCreateTime :: Maybe ISO8601

PreferredBackupWindow :: Maybe Text

BackupRetentionPeriod :: Maybe Integer

DBSecurityGroups :: [DBSecurityGroupMembership]

VpcSecurityGroups :: [VpcSecurityGroupMembership]

DBParameterGroups :: [DBParameterGroupStatus]

AvailabilityZone :: Maybe Text

DBSubnetGroup :: Maybe DBSubnetGroup

PreferredMaintenanceWindow :: Maybe Text

PendingModifiedValues :: Maybe PendingModifiedValues

LatestRestorableTime :: Maybe ISO8601

MultiAZ :: Maybe Bool

EngineVersion :: Maybe Text

AutoMinorVersionUpgrade :: Maybe Bool

ReadReplicaSourceDBInstanceIdentifier :: Maybe Text

ReadReplicaDBInstanceIdentifiers :: [Text]

LicenseModel :: Maybe Text

Iops :: Maybe Integer

OptionGroupMemberships :: [OptionGroupMembership]

CharacterSetName :: Maybe Text

SecondaryAvailabilityZone :: Maybe Text

PubliclyAccessible :: Maybe Bool

StatusInfos :: [DBInstanceStatusInfo]

dbiDBInstanceIdentifier :: Lens' DBInstance (Maybe Text)

Contains a user-supplied database identifier. This is the unique key that identifies a DB instance.

dbiDBInstanceClass :: Lens' DBInstance (Maybe Text)

Contains the name of the compute and memory capacity class of the DB instance.

dbiEngine :: Lens' DBInstance (Maybe Text)

Provides the name of the database engine to be used for this DB instance.

dbiDBInstanceStatus :: Lens' DBInstance (Maybe Text)

Specifies the current state of this database.

dbiMasterUsername :: Lens' DBInstance (Maybe Text)

Contains the master username for the DB instance.

dbiDBName :: Lens' DBInstance (Maybe Text)

The meaning of this parameter differs according to the database engine you use. For example, this value returns only MySQL information when returning values from CreateDBInstanceReadReplica since read replicas are only supported for MySQL. MySQL Contains the name of the initial database of this instance that was provided at create time, if one was specified when the DB instance was created. This same name is returned for the life of the DB instance. Type: String Oracle Contains the Oracle System ID (SID) of the created DB instance. Not shown when the returned parameters do not apply to an Oracle DB instance.

dbiEndpoint :: Lens' DBInstance (Maybe Endpoint)

Specifies the connection endpoint.

dbiAllocatedStorage :: Lens' DBInstance (Maybe Integer)

Specifies the allocated storage size specified in gigabytes.

dbiInstanceCreateTime :: Lens' DBInstance (Maybe ISO8601)

Provides the date and time the DB instance was created.

dbiPreferredBackupWindow :: Lens' DBInstance (Maybe Text)

Specifies the daily time range during which automated backups are created if automated backups are enabled, as determined by the BackupRetentionPeriod.

dbiBackupRetentionPeriod :: Lens' DBInstance (Maybe Integer)

Specifies the number of days for which automatic DB snapshots are retained.

dbiDBSecurityGroups :: Lens' DBInstance [DBSecurityGroupMembership]

Provides List of DB security group elements containing only DBSecurityGroup.Name and DBSecurityGroup.Status subelements.

dbiVpcSecurityGroups :: Lens' DBInstance [VpcSecurityGroupMembership]

Provides List of VPC security group elements that the DB instance belongs to.

dbiDBParameterGroups :: Lens' DBInstance [DBParameterGroupStatus]

Provides the list of DB parameter groups applied to this DB instance.

dbiAvailabilityZone :: Lens' DBInstance (Maybe Text)

Specifies the name of the Availability Zone the DB instance is located in.

dbiDBSubnetGroup :: Lens' DBInstance (Maybe DBSubnetGroup)

Specifies information on the subnet group associated with the DB instance, including the name, description, and subnets in the subnet group.

dbiPreferredMaintenanceWindow :: Lens' DBInstance (Maybe Text)

Specifies the weekly time range (in UTC) during which system maintenance can occur.

dbiPendingModifiedValues :: Lens' DBInstance (Maybe PendingModifiedValues)

Specifies that changes to the DB instance are pending. This element is only included when changes are pending. Specific changes are identified by subelements.

dbiLatestRestorableTime :: Lens' DBInstance (Maybe ISO8601)

Specifies the latest time to which a database can be restored with point-in-time restore.

dbiMultiAZ :: Lens' DBInstance (Maybe Bool)

Specifies if the DB instance is a Multi-AZ deployment.

dbiEngineVersion :: Lens' DBInstance (Maybe Text)

Indicates the database engine version.

dbiAutoMinorVersionUpgrade :: Lens' DBInstance (Maybe Bool)

Indicates that minor version patches are applied automatically.

dbiReadReplicaSourceDBInstanceIdentifier :: Lens' DBInstance (Maybe Text)

Contains the identifier of the source DB instance if this DB instance is a read replica.

dbiReadReplicaDBInstanceIdentifiers :: Lens' DBInstance [Text]

Contains one or more identifiers of the read replicas associated with this DB instance.

dbiLicenseModel :: Lens' DBInstance (Maybe Text)

License model information for this DB instance.

dbiIops :: Lens' DBInstance (Maybe Integer)

Specifies the Provisioned IOPS (I/O operations per second) value.

dbiOptionGroupMemberships :: Lens' DBInstance [OptionGroupMembership]

Provides the list of option group memberships for this DB instance.

dbiCharacterSetName :: Lens' DBInstance (Maybe Text)

If present, specifies the name of the character set that this instance is associated with.

dbiSecondaryAvailabilityZone :: Lens' DBInstance (Maybe Text)

If present, specifies the name of the secondary Availability Zone for a DB instance with multi-AZ support.

dbiPubliclyAccessible :: Lens' DBInstance (Maybe Bool)

Specifies the accessibility options for the DB instance. A value of true specifies an Internet-facing instance with a publicly resolvable DNS name, which resolves to a public IP address. A value of false specifies an internal instance with a DNS name that resolves to a private IP address. Default: The default behavior varies depending on whether a VPC has been requested or not. The following list shows the default behavior in each case. Default VPC:true VPC:false If no DB subnet group has been specified as part of the request and the PubliclyAccessible value has not been set, the DB instance will be publicly accessible. If a specific DB subnet group has been specified as part of the request and the PubliclyAccessible value has not been set, the DB instance will be private.

dbiStatusInfos :: Lens' DBInstance [DBInstanceStatusInfo]

The status of a read replica. If the instance is not a read replica, this will be blank.

DBInstanceStatusInfo

data DBInstanceStatusInfo

Provides a list of status information for a DB instance.

mkDBInstanceStatusInfo :: DBInstanceStatusInfo

Smart constructor for the minimum required fields to construct a valid DBInstanceStatusInfo data type to populate a request.

The fields accessible through corresponding lenses are:

StatusType :: Maybe Text

Normal :: Maybe Bool

Status :: Maybe Text

Message :: Maybe Text

dbisiStatusType :: Lens' DBInstanceStatusInfo (Maybe Text)

This value is currently "read replication.".

dbisiNormal :: Lens' DBInstanceStatusInfo (Maybe Bool)

Boolean value that is true if the instance is operating normally, or false if the instance is in an error state.

dbisiStatus :: Lens' DBInstanceStatusInfo (Maybe Text)

Status of the DB instance. For a StatusType of read replica, the values can be replicating, error, stopped, or terminated.

dbisiMessage :: Lens' DBInstanceStatusInfo (Maybe Text)

Details of the error if there is an error for the instance. If the instance is not in an error state, this value is blank.

DBParameterGroup

data DBParameterGroup

Contains the result of a successful invocation of the CreateDBParameterGroup action. This data type is used as a request parameter in the DeleteDBParameterGroup action, and as a response element in the DescribeDBParameterGroups action.

mkDBParameterGroup :: DBParameterGroup

Smart constructor for the minimum required fields to construct a valid DBParameterGroup data type.

DBParameterGroup is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

DBParameterGroupName :: Maybe Text

DBParameterGroupFamily :: Maybe Text

Description :: Maybe Text

dbpgDBParameterGroupName :: Lens' DBParameterGroup (Maybe Text)

Provides the name of the DB parameter group.

dbpgDBParameterGroupFamily :: Lens' DBParameterGroup (Maybe Text)

Provides the name of the DB parameter group family that this DB parameter group is compatible with.

dbpgDescription :: Lens' DBParameterGroup (Maybe Text)

Provides the customer-specified description for this DB parameter group.

DBParameterGroupStatus

data DBParameterGroupStatus

The status of the DB parameter group. This data type is used as a response element in the following actions: CreateDBInstance CreateDBInstanceReadReplica DeleteDBInstance ModifyDBInstance RebootDBInstance RestoreDBInstanceFromDBSnapshot.

mkDBParameterGroupStatus :: DBParameterGroupStatus

Smart constructor for the minimum required fields to construct a valid DBParameterGroupStatus data type to populate a request.

The fields accessible through corresponding lenses are:

DBParameterGroupName :: Maybe Text

ParameterApplyStatus :: Maybe Text

dbpgsDBParameterGroupName :: Lens' DBParameterGroupStatus (Maybe Text)

The name of the DP parameter group.

dbpgsParameterApplyStatus :: Lens' DBParameterGroupStatus (Maybe Text)

The status of parameter updates.

DBSecurityGroup

data DBSecurityGroup

Contains the result of a successful invocation of the following actions: DescribeDBSecurityGroups AuthorizeDBSecurityGroupIngress CreateDBSecurityGroup RevokeDBSecurityGroupIngress This data type is used as a response element in the DescribeDBSecurityGroups action.

mkDBSecurityGroup :: DBSecurityGroup

Smart constructor for the minimum required fields to construct a valid DBSecurityGroup data type.

DBSecurityGroup is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

OwnerId :: Maybe Text

DBSecurityGroupName :: Maybe Text

DBSecurityGroupDescription :: Maybe Text

VpcId :: Maybe Text

EC2SecurityGroups :: [EC2SecurityGroup]

IPRanges :: [IPRange]

dbsgOwnerId :: Lens' DBSecurityGroup (Maybe Text)

Provides the AWS ID of the owner of a specific DB security group.

dbsgDBSecurityGroupName :: Lens' DBSecurityGroup (Maybe Text)

Specifies the name of the DB security group.

dbsgDBSecurityGroupDescription :: Lens' DBSecurityGroup (Maybe Text)

Provides the description of the DB security group.

dbsgVpcId :: Lens' DBSecurityGroup (Maybe Text)

Provides the VpcId of the DB security group.

dbsgEC2SecurityGroups :: Lens' DBSecurityGroup [EC2SecurityGroup]

Contains a list of EC2SecurityGroup elements.

dbsgIPRanges :: Lens' DBSecurityGroup [IPRange]

Contains a list of IPRange elements.

DBSecurityGroupMembership

data DBSecurityGroupMembership

This data type is used as a response element in the following actions: ModifyDBInstance RebootDBInstance RestoreDBInstanceFromDBSnapshot RestoreDBInstanceToPointInTime.

mkDBSecurityGroupMembership :: DBSecurityGroupMembership

Smart constructor for the minimum required fields to construct a valid DBSecurityGroupMembership data type to populate a request.

The fields accessible through corresponding lenses are:

DBSecurityGroupName :: Maybe Text

Status :: Maybe Text

dbsgmDBSecurityGroupName :: Lens' DBSecurityGroupMembership (Maybe Text)

The name of the DB security group.

dbsgmStatus :: Lens' DBSecurityGroupMembership (Maybe Text)

The status of the DB security group.

DBSnapshot

data DBSnapshot

Contains the result of a successful invocation of the following actions: CreateDBSnapshot DeleteDBSnapshot This data type is used as a response element in the DescribeDBSnapshots action.

mkDBSnapshot :: DBSnapshot

Smart constructor for the minimum required fields to construct a valid DBSnapshot data type.

DBSnapshot is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

DBSnapshotIdentifier :: Maybe Text

DBInstanceIdentifier :: Maybe Text

SnapshotCreateTime :: Maybe ISO8601

Engine :: Maybe Text

AllocatedStorage :: Maybe Integer

Status :: Maybe Text

Port :: Maybe Integer

AvailabilityZone :: Maybe Text

VpcId :: Maybe Text

InstanceCreateTime :: Maybe ISO8601

MasterUsername :: Maybe Text

EngineVersion :: Maybe Text

LicenseModel :: Maybe Text

SnapshotType :: Maybe Text

Iops :: Maybe Integer

OptionGroupName :: Maybe Text

PercentProgress :: Maybe Integer

SourceRegion :: Maybe Text

dbsDBSnapshotIdentifier :: Lens' DBSnapshot (Maybe Text)

Specifies the identifier for the DB snapshot.

dbsDBInstanceIdentifier :: Lens' DBSnapshot (Maybe Text)

Specifies the DB instance identifier of the DB instance this DB snapshot was created from.

dbsSnapshotCreateTime :: Lens' DBSnapshot (Maybe ISO8601)

Provides the time (UTC) when the snapshot was taken.

dbsEngine :: Lens' DBSnapshot (Maybe Text)

Specifies the name of the database engine.

dbsAllocatedStorage :: Lens' DBSnapshot (Maybe Integer)

Specifies the allocated storage size in gigabytes (GB).

dbsStatus :: Lens' DBSnapshot (Maybe Text)

Specifies the status of this DB snapshot.

dbsPort :: Lens' DBSnapshot (Maybe Integer)

Specifies the port that the database engine was listening on at the time of the snapshot.

dbsAvailabilityZone :: Lens' DBSnapshot (Maybe Text)

Specifies the name of the Availability Zone the DB instance was located in at the time of the DB snapshot.

dbsVpcId :: Lens' DBSnapshot (Maybe Text)

Provides the Vpc Id associated with the DB snapshot.

dbsInstanceCreateTime :: Lens' DBSnapshot (Maybe ISO8601)

Specifies the time (UTC) when the snapshot was taken.

dbsMasterUsername :: Lens' DBSnapshot (Maybe Text)

Provides the master username for the DB snapshot.

dbsEngineVersion :: Lens' DBSnapshot (Maybe Text)

Specifies the version of the database engine.

dbsLicenseModel :: Lens' DBSnapshot (Maybe Text)

License model information for the restored DB instance.

dbsSnapshotType :: Lens' DBSnapshot (Maybe Text)

Provides the type of the DB snapshot.

dbsIops :: Lens' DBSnapshot (Maybe Integer)

Specifies the Provisioned IOPS (I/O operations per second) value of the DB instance at the time of the snapshot.

dbsOptionGroupName :: Lens' DBSnapshot (Maybe Text)

Provides the option group name for the DB snapshot.

dbsPercentProgress :: Lens' DBSnapshot (Maybe Integer)

The percentage of the estimated data that has been transferred.

dbsSourceRegion :: Lens' DBSnapshot (Maybe Text)

The region that the DB snapshot was created in or copied from.

DBSubnetGroup

data DBSubnetGroup

Specifies information on the subnet group associated with the DB instance, including the name, description, and subnets in the subnet group.

mkDBSubnetGroup :: DBSubnetGroup

Smart constructor for the minimum required fields to construct a valid DBSubnetGroup data type to populate a request.

The fields accessible through corresponding lenses are:

DBSubnetGroupName :: Maybe Text

DBSubnetGroupDescription :: Maybe Text

VpcId :: Maybe Text

SubnetGroupStatus :: Maybe Text

Subnets :: [Subnet]

dbsgrDBSubnetGroupName :: Lens' DBSubnetGroup (Maybe Text)

Specifies the name of the DB subnet group.

dbsgrDBSubnetGroupDescription :: Lens' DBSubnetGroup (Maybe Text)

Provides the description of the DB subnet group.

dbsgrVpcId :: Lens' DBSubnetGroup (Maybe Text)

Provides the VpcId of the DB subnet group.

dbsgrSubnetGroupStatus :: Lens' DBSubnetGroup (Maybe Text)

Provides the status of the DB subnet group.

dbsgrSubnets :: Lens' DBSubnetGroup [Subnet]

Contains a list of Subnet elements.

DescribeDBLogFilesDetails

data DescribeDBLogFilesDetails

This data type is used as a response element to DescribeDBLogFiles.

mkDescribeDBLogFilesDetails :: DescribeDBLogFilesDetails

Smart constructor for the minimum required fields to construct a valid DescribeDBLogFilesDetails data type.

DescribeDBLogFilesDetails is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

LogFileName :: Maybe Text

LastWritten :: Maybe Integer

Size :: Maybe Integer

ddblfdLogFileName :: Lens' DescribeDBLogFilesDetails (Maybe Text)

The name of the log file for the specified DB instance.

ddblfdLastWritten :: Lens' DescribeDBLogFilesDetails (Maybe Integer)

A POSIX timestamp when the last log entry was written.

ddblfdSize :: Lens' DescribeDBLogFilesDetails (Maybe Integer)

The size, in bytes, of the log file for the specified DB instance.

EC2SecurityGroup

data EC2SecurityGroup

This data type is used as a response element in the following actions: AuthorizeDBSecurityGroupIngress DescribeDBSecurityGroups RevokeDBSecurityGroupIngress.

mkEC2SecurityGroup :: EC2SecurityGroup

Smart constructor for the minimum required fields to construct a valid EC2SecurityGroup data type to populate a request.

The fields accessible through corresponding lenses are:

Status :: Maybe Text

EC2SecurityGroupName :: Maybe Text

EC2SecurityGroupId :: Maybe Text

EC2SecurityGroupOwnerId :: Maybe Text

ecsgStatus :: Lens' EC2SecurityGroup (Maybe Text)

Provides the status of the EC2 security group. Status can be "authorizing", "authorized", "revoking", and "revoked".

ecsgEC2SecurityGroupName :: Lens' EC2SecurityGroup (Maybe Text)

Specifies the name of the EC2 security group.

ecsgEC2SecurityGroupId :: Lens' EC2SecurityGroup (Maybe Text)

Specifies the id of the EC2 security group.

ecsgEC2SecurityGroupOwnerId :: Lens' EC2SecurityGroup (Maybe Text)

Specifies the AWS ID of the owner of the EC2 security group specified in the EC2SecurityGroupName field.

Endpoint

data Endpoint

Specifies the connection endpoint.

mkEndpoint :: Endpoint

Smart constructor for the minimum required fields to construct a valid Endpoint data type to populate a request.

The fields accessible through corresponding lenses are:

Address :: Maybe Text

Port :: Maybe Integer

eAddress :: Lens' Endpoint (Maybe Text)

Specifies the DNS address of the DB instance.

ePort :: Lens' Endpoint (Maybe Integer)

Specifies the port that the database engine is listening on.

EngineDefaults

data EngineDefaults

Contains the result of a successful invocation of the DescribeEngineDefaultParameters action.

mkEngineDefaults :: EngineDefaults

Smart constructor for the minimum required fields to construct a valid EngineDefaults data type.

EngineDefaults is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

DBParameterGroupFamily :: Maybe Text

Marker :: Maybe Text

Parameters :: [Parameter]

edDBParameterGroupFamily :: Lens' EngineDefaults (Maybe Text)

Specifies the name of the DB parameter group family which the engine default parameters apply to.

edMarker :: Lens' EngineDefaults (Maybe Text)

An optional pagination token provided by a previous EngineDefaults request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

edParameters :: Lens' EngineDefaults [Parameter]

Contains a list of engine default parameters.

Event

data Event

This data type is used as a response element in the DescribeEvents action.

mkEvent :: Event

Smart constructor for the minimum required fields to construct a valid Event data type.

Event is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

SourceIdentifier :: Maybe Text

SourceType :: Maybe SourceType

Message :: Maybe Text

EventCategories :: [Text]

Date :: Maybe ISO8601

erSourceIdentifier :: Lens' Event (Maybe Text)

Provides the identifier for the source of the event.

erSourceType :: Lens' Event (Maybe SourceType)

Specifies the source type for this event.

erMessage :: Lens' Event (Maybe Text)

Provides the text of this event.

erEventCategories :: Lens' Event [Text]

Specifies the category for the event.

erDate :: Lens' Event (Maybe ISO8601)

Specifies the date and time of the event.

EventCategoriesMap

data EventCategoriesMap

Contains the results of a successful invocation of the DescribeEventCategories action.

mkEventCategoriesMap :: EventCategoriesMap

Smart constructor for the minimum required fields to construct a valid EventCategoriesMap data type.

EventCategoriesMap is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

SourceType :: Maybe Text

EventCategories :: [Text]

ecmSourceType :: Lens' EventCategoriesMap (Maybe Text)

The source type that the returned categories belong to.

ecmEventCategories :: Lens' EventCategoriesMap [Text]

The event categories for the specified source type.

EventSubscription

data EventSubscription

Contains the results of a successful invocation of the DescribeEventSubscriptions action.

mkEventSubscription :: EventSubscription

Smart constructor for the minimum required fields to construct a valid EventSubscription data type.

EventSubscription is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

CustomerAwsId :: Maybe Text

CustSubscriptionId :: Maybe Text

SnsTopicArn :: Maybe Text

Status :: Maybe Text

SubscriptionCreationTime :: Maybe Text

SourceType :: Maybe Text

SourceIdsList :: [Text]

EventCategoriesList :: [Text]

Enabled :: Maybe Bool

esCustomerAwsId :: Lens' EventSubscription (Maybe Text)

The AWS customer account associated with the RDS event notification subscription.

esCustSubscriptionId :: Lens' EventSubscription (Maybe Text)

The RDS event notification subscription Id.

esSnsTopicArn :: Lens' EventSubscription (Maybe Text)

The topic ARN of the RDS event notification subscription.

esStatus :: Lens' EventSubscription (Maybe Text)

The status of the RDS event notification subscription. Constraints: Can be one of the following: creating | modifying | deleting | active | no-permission | topic-not-exist The status "no-permission" indicates that RDS no longer has permission to post to the SNS topic. The status "topic-not-exist" indicates that the topic was deleted after the subscription was created.

esSubscriptionCreationTime :: Lens' EventSubscription (Maybe Text)

The time the RDS event notification subscription was created.

esSourceType :: Lens' EventSubscription (Maybe Text)

The source type for the RDS event notification subscription.

esSourceIdsList :: Lens' EventSubscription [Text]

A list of source Ids for the RDS event notification subscription.

esEventCategoriesList :: Lens' EventSubscription [Text]

A list of event categories for the RDS event notification subscription.

esEnabled :: Lens' EventSubscription (Maybe Bool)

A Boolean value indicating if the subscription is enabled. True indicates the subscription is enabled.

IPRange

data IPRange

This data type is used as a response element in the DescribeDBSecurityGroups action.

mkIPRange :: IPRange

Smart constructor for the minimum required fields to construct a valid IPRange data type to populate a request.

The fields accessible through corresponding lenses are:

Status :: Maybe Text

CIDRIP :: Maybe Text

iprStatus :: Lens' IPRange (Maybe Text)

Specifies the status of the IP range. Status can be "authorizing", "authorized", "revoking", and "revoked".

iprCIDRIP :: Lens' IPRange (Maybe Text)

Specifies the IP range.

Option

data Option

Option details.

mkOption :: Option

Smart constructor for the minimum required fields to construct a valid Option data type to populate a request.

The fields accessible through corresponding lenses are:

OptionName :: Maybe Text

OptionDescription :: Maybe Text

Persistent :: Maybe Bool

Permanent :: Maybe Bool

Port :: Maybe Integer

OptionSettings :: [OptionSetting]

DBSecurityGroupMemberships :: [DBSecurityGroupMembership]

VpcSecurityGroupMemberships :: [VpcSecurityGroupMembership]

oOptionName :: Lens' Option (Maybe Text)

The name of the option.

oOptionDescription :: Lens' Option (Maybe Text)

The description of the option.

oPersistent :: Lens' Option (Maybe Bool)

Indicate if this option is persistent.

oPermanent :: Lens' Option (Maybe Bool)

Indicate if this option is permanent.

oPort :: Lens' Option (Maybe Integer)

If required, the port configured for this option to use.

oOptionSettings :: Lens' Option [OptionSetting]

The option settings for this option.

oDBSecurityGroupMemberships :: Lens' Option [DBSecurityGroupMembership]

If the option requires access to a port, then this DB security group allows access to the port.

oVpcSecurityGroupMemberships :: Lens' Option [VpcSecurityGroupMembership]

If the option requires access to a port, then this VPC security group allows access to the port.

OptionConfiguration

data OptionConfiguration

A list of all available options.

mkOptionConfiguration

Arguments

:: Text
ocOptionName

-> OptionConfiguration

Smart constructor for the minimum required fields to construct a valid OptionConfiguration data type to populate a request.

The fields accessible through corresponding lenses are:

OptionName :: Text

Port :: Maybe Integer

DBSecurityGroupMemberships :: [Text]

VpcSecurityGroupMemberships :: [Text]

OptionSettings :: [OptionSetting]

ocOptionName :: Lens' OptionConfiguration Text

The configuration of options to include in a group.

ocPort :: Lens' OptionConfiguration (Maybe Integer)

The optional port for the option.

ocDBSecurityGroupMemberships :: Lens' OptionConfiguration [Text]

A list of DBSecurityGroupMemebrship name strings used for this option.

ocVpcSecurityGroupMemberships :: Lens' OptionConfiguration [Text]

A list of VpcSecurityGroupMemebrship name strings used for this option.

ocOptionSettings :: Lens' OptionConfiguration [OptionSetting]

The option settings to include in an option group.

OptionGroup

data OptionGroup

mkOptionGroup :: OptionGroup

Smart constructor for the minimum required fields to construct a valid OptionGroup data type.

OptionGroup is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

OptionGroupName :: Maybe Text

OptionGroupDescription :: Maybe Text

EngineName :: Maybe Text

MajorEngineVersion :: Maybe Text

Options :: [Option]

AllowsVpcAndNonVpcInstanceMemberships :: Maybe Bool

VpcId :: Maybe Text

ogOptionGroupName :: Lens' OptionGroup (Maybe Text)

Specifies the name of the option group.

ogOptionGroupDescription :: Lens' OptionGroup (Maybe Text)

Provides the description of the option group.

ogEngineName :: Lens' OptionGroup (Maybe Text)

Engine name that this option group can be applied to.

ogMajorEngineVersion :: Lens' OptionGroup (Maybe Text)

Indicates the major engine version associated with this option group.

ogOptions :: Lens' OptionGroup [Option]

Indicates what options are available in the option group.

ogAllowsVpcAndNonVpcInstanceMemberships :: Lens' OptionGroup (Maybe Bool)

Indicates whether this option group can be applied to both VPC and non-VPC instances. The value true indicates the option group can be applied to both VPC and non-VPC instances.

ogVpcId :: Lens' OptionGroup (Maybe Text)

If AllowsVpcAndNonVpcInstanceMemberships is false, this field is blank. If AllowsVpcAndNonVpcInstanceMemberships is true and this field is blank, then this option group can be applied to both VPC and non-VPC instances. If this field contains a value, then this option group can only be applied to instances that are in the VPC indicated by this field.

OptionGroupMembership

data OptionGroupMembership

Provides information on the option groups the DB instance is a member of.

mkOptionGroupMembership :: OptionGroupMembership

Smart constructor for the minimum required fields to construct a valid OptionGroupMembership data type to populate a request.

The fields accessible through corresponding lenses are:

OptionGroupName :: Maybe Text

Status :: Maybe Text

ogmOptionGroupName :: Lens' OptionGroupMembership (Maybe Text)

The name of the option group that the instance belongs to.

ogmStatus :: Lens' OptionGroupMembership (Maybe Text)

The status of the DB instance's option group membership (e.g. in-sync, pending, pending-maintenance, applying).

OptionGroupOption

data OptionGroupOption

Available option.

mkOptionGroupOption :: OptionGroupOption

Smart constructor for the minimum required fields to construct a valid OptionGroupOption data type.

OptionGroupOption is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Name :: Maybe Text

Description :: Maybe Text

EngineName :: Maybe Text

MajorEngineVersion :: Maybe Text

MinimumRequiredMinorEngineVersion :: Maybe Text

PortRequired :: Maybe Bool

DefaultPort :: Maybe Integer

OptionsDependedOn :: [Text]

Persistent :: Maybe Bool

Permanent :: Maybe Bool

OptionGroupOptionSettings :: [OptionGroupOptionSetting]

ogoName :: Lens' OptionGroupOption (Maybe Text)

The name of the option.

ogoDescription :: Lens' OptionGroupOption (Maybe Text)

The description of the option.

ogoEngineName :: Lens' OptionGroupOption (Maybe Text)

Engine name that this option can be applied to.

ogoMajorEngineVersion :: Lens' OptionGroupOption (Maybe Text)

Indicates the major engine version that the option is available for.

ogoMinimumRequiredMinorEngineVersion :: Lens' OptionGroupOption (Maybe Text)

The minimum required engine version for the option to be applied.

ogoPortRequired :: Lens' OptionGroupOption (Maybe Bool)

Specifies whether the option requires a port.

ogoDefaultPort :: Lens' OptionGroupOption (Maybe Integer)

If the option requires a port, specifies the default port for the option.

ogoOptionsDependedOn :: Lens' OptionGroupOption [Text]

List of all options that are prerequisites for this option.

ogoPersistent :: Lens' OptionGroupOption (Maybe Bool)

A persistent option cannot be removed from the option group once the option group is used, but this option can be removed from the db instance while modifying the related data and assigning another option group without this option.

ogoPermanent :: Lens' OptionGroupOption (Maybe Bool)

A permanent option cannot be removed from the option group once the option group is used, and it cannot be removed from the db instance after assigning an option group with this permanent option.

ogoOptionGroupOptionSettings :: Lens' OptionGroupOption [OptionGroupOptionSetting]

Specifies the option settings that are available (and the default value) for each option in an option group.

OptionGroupOptionSetting

data OptionGroupOptionSetting

option group option settings are used to display settings available for each option with their default values and other information. These values are used with the DescribeOptionGroupOptions action.

mkOptionGroupOptionSetting :: OptionGroupOptionSetting

Smart constructor for the minimum required fields to construct a valid OptionGroupOptionSetting data type to populate a request.

The fields accessible through corresponding lenses are:

SettingName :: Maybe Text

SettingDescription :: Maybe Text

DefaultValue :: Maybe Text

ApplyType :: Maybe Text

AllowedValues :: Maybe Text

IsModifiable :: Maybe Bool

ogosSettingName :: Lens' OptionGroupOptionSetting (Maybe Text)

The name of the option group option.

ogosSettingDescription :: Lens' OptionGroupOptionSetting (Maybe Text)

The description of the option group option.

ogosDefaultValue :: Lens' OptionGroupOptionSetting (Maybe Text)

The default value for the option group option.

ogosApplyType :: Lens' OptionGroupOptionSetting (Maybe Text)

The DB engine specific parameter type for the option group option.

ogosAllowedValues :: Lens' OptionGroupOptionSetting (Maybe Text)

Indicates the acceptable values for the option group option.

ogosIsModifiable :: Lens' OptionGroupOptionSetting (Maybe Bool)

Boolean value where true indicates that this option group option can be changed from the default value.

OptionSetting

data OptionSetting

Option settings are the actual settings being applied or configured for that option. It is used when you modify an option group or describe option groups. For example, the NATIVE_NETWORK_ENCRYPTION option has a setting called SQLNET.ENCRYPTION_SERVER that can have several different values.

mkOptionSetting :: OptionSetting

Smart constructor for the minimum required fields to construct a valid OptionSetting data type to populate a request.

The fields accessible through corresponding lenses are:

Name :: Maybe Text

Value :: Maybe Text

DefaultValue :: Maybe Text

Description :: Maybe Text

ApplyType :: Maybe Text

DataType :: Maybe Text

AllowedValues :: Maybe Text

IsModifiable :: Maybe Bool

IsCollection :: Maybe Bool

osName :: Lens' OptionSetting (Maybe Text)

The name of the option that has settings that you can set.

osValue :: Lens' OptionSetting (Maybe Text)

The current value of the option setting.

osDefaultValue :: Lens' OptionSetting (Maybe Text)

The default value of the option setting.

osDescription :: Lens' OptionSetting (Maybe Text)

The description of the option setting.

osApplyType :: Lens' OptionSetting (Maybe Text)

The DB engine specific parameter type.

osDataType :: Lens' OptionSetting (Maybe Text)

The data type of the option setting.

osAllowedValues :: Lens' OptionSetting (Maybe Text)

The allowed values of the option setting.

osIsModifiable :: Lens' OptionSetting (Maybe Bool)

A Boolean value that, when true, indicates the option setting can be modified from the default.

osIsCollection :: Lens' OptionSetting (Maybe Bool)

Indicates if the option setting is part of a collection.

OrderableDBInstanceOption

data OrderableDBInstanceOption

Contains a list of available options for a DB instance This data type is used as a response element in the DescribeOrderableDBInstanceOptions action.

mkOrderableDBInstanceOption :: OrderableDBInstanceOption

Smart constructor for the minimum required fields to construct a valid OrderableDBInstanceOption data type.

OrderableDBInstanceOption is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Engine :: Maybe Text

EngineVersion :: Maybe Text

DBInstanceClass :: Maybe Text

LicenseModel :: Maybe Text

AvailabilityZones :: [AvailabilityZone]

MultiAZCapable :: Maybe Bool

ReadReplicaCapable :: Maybe Bool

Vpc :: Maybe Bool

odbioEngine :: Lens' OrderableDBInstanceOption (Maybe Text)

The engine type of the orderable DB instance.

odbioEngineVersion :: Lens' OrderableDBInstanceOption (Maybe Text)

The engine version of the orderable DB instance.

odbioDBInstanceClass :: Lens' OrderableDBInstanceOption (Maybe Text)

The DB instance Class for the orderable DB instance.

odbioLicenseModel :: Lens' OrderableDBInstanceOption (Maybe Text)

The license model for the orderable DB instance.

odbioAvailabilityZones :: Lens' OrderableDBInstanceOption [AvailabilityZone]

A list of availability zones for the orderable DB instance.

odbioMultiAZCapable :: Lens' OrderableDBInstanceOption (Maybe Bool)

Indicates whether this orderable DB instance is multi-AZ capable.

odbioReadReplicaCapable :: Lens' OrderableDBInstanceOption (Maybe Bool)

Indicates whether this orderable DB instance can have a read replica.

odbioVpc :: Lens' OrderableDBInstanceOption (Maybe Bool)

Indicates whether this is a VPC orderable DB instance.

Parameter

data Parameter

This data type is used as a request parameter in the ModifyDBParameterGroup and ResetDBParameterGroup actions. This data type is used as a response element in the DescribeEngineDefaultParameters and DescribeDBParameters actions.

mkParameter :: Parameter

Smart constructor for the minimum required fields to construct a valid Parameter data type to populate a request.

The fields accessible through corresponding lenses are:

ParameterName :: Maybe Text

ParameterValue :: Maybe Text

Description :: Maybe Text

Source :: Maybe Text

ApplyType :: Maybe Text

DataType :: Maybe Text

AllowedValues :: Maybe Text

IsModifiable :: Maybe Bool

MinimumEngineVersion :: Maybe Text

ApplyMethod :: Maybe ApplyMethod

pParameterName :: Lens' Parameter (Maybe Text)

Specifies the name of the parameter.

pParameterValue :: Lens' Parameter (Maybe Text)

Specifies the value of the parameter.

pDescription :: Lens' Parameter (Maybe Text)

Provides a description of the parameter.

pSource :: Lens' Parameter (Maybe Text)

Indicates the source of the parameter value.

pApplyType :: Lens' Parameter (Maybe Text)

Specifies the engine specific parameters type.

pDataType :: Lens' Parameter (Maybe Text)

Specifies the valid data type for the parameter.

pAllowedValues :: Lens' Parameter (Maybe Text)

Specifies the valid range of values for the parameter.

pIsModifiable :: Lens' Parameter (Maybe Bool)

Indicates whether (true) or not (false) the parameter can be modified. Some parameters have security or operational implications that prevent them from being changed.

pMinimumEngineVersion :: Lens' Parameter (Maybe Text)

The earliest engine version to which the parameter can apply.

pApplyMethod :: Lens' Parameter (Maybe ApplyMethod)

Indicates when to apply parameter updates.

PendingModifiedValues

data PendingModifiedValues

Specifies that changes to the DB instance are pending. This element is only included when changes are pending. Specific changes are identified by subelements.

mkPendingModifiedValues :: PendingModifiedValues

Smart constructor for the minimum required fields to construct a valid PendingModifiedValues data type to populate a request.

The fields accessible through corresponding lenses are:

DBInstanceClass :: Maybe Text

AllocatedStorage :: Maybe Integer

MasterUserPassword :: Maybe Text

Port :: Maybe Integer

BackupRetentionPeriod :: Maybe Integer

MultiAZ :: Maybe Bool

EngineVersion :: Maybe Text

Iops :: Maybe Integer

DBInstanceIdentifier :: Maybe Text

pmvDBInstanceClass :: Lens' PendingModifiedValues (Maybe Text)

Contains the new DBInstanceClass for the DB instance that will be applied or is in progress.

pmvAllocatedStorage :: Lens' PendingModifiedValues (Maybe Integer)

Contains the new AllocatedStorage size for the DB instance that will be applied or is in progress.

pmvMasterUserPassword :: Lens' PendingModifiedValues (Maybe Text)

Contains the pending or in-progress change of the master credentials for the DB instance.

pmvPort :: Lens' PendingModifiedValues (Maybe Integer)

Specifies the pending port for the DB instance.

pmvBackupRetentionPeriod :: Lens' PendingModifiedValues (Maybe Integer)

Specifies the pending number of days for which automated backups are retained.

pmvMultiAZ :: Lens' PendingModifiedValues (Maybe Bool)

Indicates that the Single-AZ DB instance is to change to a Multi-AZ deployment.

pmvEngineVersion :: Lens' PendingModifiedValues (Maybe Text)

Indicates the database engine version.

pmvIops :: Lens' PendingModifiedValues (Maybe Integer)

Specifies the new Provisioned IOPS value for the DB instance that will be applied or is being applied.

pmvDBInstanceIdentifier :: Lens' PendingModifiedValues (Maybe Text)

Contains the new DBInstanceIdentifier for the DB instance that will be applied or is in progress.

RecurringCharge

data RecurringCharge

This data type is used as a response element in the DescribeReservedDBInstances and DescribeReservedDBInstancesOfferings actions.

mkRecurringCharge :: RecurringCharge

Smart constructor for the minimum required fields to construct a valid RecurringCharge data type to populate a request.

The fields accessible through corresponding lenses are:

RecurringChargeAmount :: Maybe Double

RecurringChargeFrequency :: Maybe Text

rcRecurringChargeAmount :: Lens' RecurringCharge (Maybe Double)

The amount of the recurring charge.

rcRecurringChargeFrequency :: Lens' RecurringCharge (Maybe Text)

The frequency of the recurring charge.

ReservedDBInstance

data ReservedDBInstance

This data type is used as a response element in the DescribeReservedDBInstances and PurchaseReservedDBInstancesOffering actions.

mkReservedDBInstance :: ReservedDBInstance

Smart constructor for the minimum required fields to construct a valid ReservedDBInstance data type.

ReservedDBInstance is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

ReservedDBInstanceId :: Maybe Text

ReservedDBInstancesOfferingId :: Maybe Text

DBInstanceClass :: Maybe Text

StartTime :: Maybe ISO8601

Duration :: Maybe Integer

FixedPrice :: Maybe Double

UsagePrice :: Maybe Double

CurrencyCode :: Maybe Text

DBInstanceCount :: Maybe Integer

ProductDescription :: Maybe Text

OfferingType :: Maybe Text

MultiAZ :: Maybe Bool

State :: Maybe Text

RecurringCharges :: [RecurringCharge]

rdbiReservedDBInstanceId :: Lens' ReservedDBInstance (Maybe Text)

The unique identifier for the reservation.

rdbiReservedDBInstancesOfferingId :: Lens' ReservedDBInstance (Maybe Text)

The offering identifier.

rdbiDBInstanceClass :: Lens' ReservedDBInstance (Maybe Text)

The DB instance class for the reserved DB instance.

rdbiStartTime :: Lens' ReservedDBInstance (Maybe ISO8601)

The time the reservation started.

rdbiDuration :: Lens' ReservedDBInstance (Maybe Integer)

The duration of the reservation in seconds.

rdbiFixedPrice :: Lens' ReservedDBInstance (Maybe Double)

The fixed price charged for this reserved DB instance.

rdbiUsagePrice :: Lens' ReservedDBInstance (Maybe Double)

The hourly price charged for this reserved DB instance.

rdbiCurrencyCode :: Lens' ReservedDBInstance (Maybe Text)

The currency code for the reserved DB instance.

rdbiDBInstanceCount :: Lens' ReservedDBInstance (Maybe Integer)

The number of reserved DB instances.

rdbiProductDescription :: Lens' ReservedDBInstance (Maybe Text)

The description of the reserved DB instance.

rdbiOfferingType :: Lens' ReservedDBInstance (Maybe Text)

The offering type of this reserved DB instance.

rdbiMultiAZ :: Lens' ReservedDBInstance (Maybe Bool)

Indicates if the reservation applies to Multi-AZ deployments.

rdbiState :: Lens' ReservedDBInstance (Maybe Text)

The state of the reserved DB instance.

rdbiRecurringCharges :: Lens' ReservedDBInstance [RecurringCharge]

The recurring price charged to run this reserved DB instance.

ReservedDBInstancesOffering

data ReservedDBInstancesOffering

This data type is used as a response element in the DescribeReservedDBInstancesOfferings action.

mkReservedDBInstancesOffering :: ReservedDBInstancesOffering

Smart constructor for the minimum required fields to construct a valid ReservedDBInstancesOffering data type.

ReservedDBInstancesOffering is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

ReservedDBInstancesOfferingId :: Maybe Text

DBInstanceClass :: Maybe Text

Duration :: Maybe Integer

FixedPrice :: Maybe Double

UsagePrice :: Maybe Double

CurrencyCode :: Maybe Text

ProductDescription :: Maybe Text

OfferingType :: Maybe Text

MultiAZ :: Maybe Bool

RecurringCharges :: [RecurringCharge]

rdbioReservedDBInstancesOfferingId :: Lens' ReservedDBInstancesOffering (Maybe Text)

The offering identifier.

rdbioDBInstanceClass :: Lens' ReservedDBInstancesOffering (Maybe Text)

The DB instance class for the reserved DB instance.

rdbioDuration :: Lens' ReservedDBInstancesOffering (Maybe Integer)

The duration of the offering in seconds.

rdbioFixedPrice :: Lens' ReservedDBInstancesOffering (Maybe Double)

The fixed price charged for this offering.

rdbioUsagePrice :: Lens' ReservedDBInstancesOffering (Maybe Double)

The hourly price charged for this offering.

rdbioCurrencyCode :: Lens' ReservedDBInstancesOffering (Maybe Text)

The currency code for the reserved DB instance offering.

rdbioProductDescription :: Lens' ReservedDBInstancesOffering (Maybe Text)

The database engine used by the offering.

rdbioOfferingType :: Lens' ReservedDBInstancesOffering (Maybe Text)

The offering type.

rdbioMultiAZ :: Lens' ReservedDBInstancesOffering (Maybe Bool)

Indicates if the offering applies to Multi-AZ deployments.

rdbioRecurringCharges :: Lens' ReservedDBInstancesOffering [RecurringCharge]

The recurring price charged to run this reserved DB instance.

Subnet

data Subnet

This data type is used as a response element in the DescribeDBSubnetGroups action.

mkSubnet :: Subnet

Smart constructor for the minimum required fields to construct a valid Subnet data type to populate a request.

The fields accessible through corresponding lenses are:

SubnetIdentifier :: Maybe Text

SubnetAvailabilityZone :: Maybe AvailabilityZone

SubnetStatus :: Maybe Text

sSubnetIdentifier :: Lens' Subnet (Maybe Text)

Specifies the identifier of the subnet.

sSubnetAvailabilityZone :: Lens' Subnet (Maybe AvailabilityZone)

Contains Availability Zone information. This data type is used as an element in the following data type: OrderableDBInstanceOption.

sSubnetStatus :: Lens' Subnet (Maybe Text)

Specifies the status of the subnet.

Tag

data Tag

Metadata assigned to an Amazon RDS resource consisting of a key-value pair.

mkTag :: Tag

Smart constructor for the minimum required fields to construct a valid Tag data type to populate a request.

The fields accessible through corresponding lenses are:

Key :: Maybe Text

Value :: Maybe Text

tKey :: Lens' Tag (Maybe Text)

A key is the required name of the tag. The string value can be from 1 to 128 Unicode characters in length and cannot be prefixed with "aws:" or "rds:". The string may only contain only the set of Unicode letters, digits, white-space, '_', ., /, '=', +, - (Java regex: "^([\p{L}\p{Z}\p{N}_.:/=+\-]*)$").

tValue :: Lens' Tag (Maybe Text)

A value is the optional value of the tag. The string value can be from 1 to 256 Unicode characters in length and cannot be prefixed with "aws:" or "rds:". The string may only contain only the set of Unicode letters, digits, white-space, '_', ., /, '=', +, - (Java regex: "^([\p{L}\p{Z}\p{N}_.:/=+\-]*)$").

VpcSecurityGroupMembership

data VpcSecurityGroupMembership

This data type is used as a response element for queries on VPC security group membership.

mkVpcSecurityGroupMembership :: VpcSecurityGroupMembership

Smart constructor for the minimum required fields to construct a valid VpcSecurityGroupMembership data type to populate a request.

The fields accessible through corresponding lenses are:

VpcSecurityGroupId :: Maybe Text

Status :: Maybe Text

vsgmVpcSecurityGroupId :: Lens' VpcSecurityGroupMembership (Maybe Text)

The name of the VPC security group.

vsgmStatus :: Lens' VpcSecurityGroupMembership (Maybe Text)

The status of the VPC security group.