Network.AWS.RDS.AddSourceIdentifierToSubscription

Description

Adds a source identifier to an existing RDS event notification subscription.

http://docs.aws.amazon.com/AmazonRDS/latest/APIReference/API_AddSourceIdentifierToSubscription.html

Request

data AddSourceIdentifierToSubscription

Request constructor

addSourceIdentifierToSubscription

Arguments

:: Text
asitsSubscriptionName

-> Text
asitsSourceIdentifier

-> AddSourceIdentifierToSubscription

AddSourceIdentifierToSubscription constructor.

The fields accessible through corresponding lenses are:

asitsSourceIdentifier :: Text

asitsSubscriptionName :: Text

Request lenses

asitsSourceIdentifier :: Lens' AddSourceIdentifierToSubscription Text

The identifier of the event source to be added. An identifier must begin with a letter and must contain only ASCII letters, digits, and hyphens; it cannot end with a hyphen or contain two consecutive hyphens.

Constraints:

If the source type is a DB instance, then a DBInstanceIdentifier must be supplied. If the source type is a DB security group, a DBSecurityGroupName must be supplied. If the source type is a DB parameter group, a DBParameterGroupName must be supplied. If the source type is a DB snapshot, a DBSnapshotIdentifier must be supplied.

asitsSubscriptionName :: Lens' AddSourceIdentifierToSubscription Text

The name of the RDS event notification subscription you want to add a source identifier to.

Response

data AddSourceIdentifierToSubscriptionResponse

Response constructor

addSourceIdentifierToSubscriptionResponse :: AddSourceIdentifierToSubscriptionResponse

AddSourceIdentifierToSubscriptionResponse constructor.

The fields accessible through corresponding lenses are:

asitsrEventSubscription :: Maybe EventSubscription

Response lenses

asitsrEventSubscription :: Lens' AddSourceIdentifierToSubscriptionResponse (Maybe EventSubscription)