Network.AWS.RDS.ApplyPendingMaintenanceAction

Description

Applies a pending maintenance action to a resource (for example, to a DB instance).

Creating a Request

applyPendingMaintenanceAction

Arguments

:: Text
apmaResourceIdentifier

-> Text
apmaApplyAction

-> Text
apmaOptInType

-> ApplyPendingMaintenanceAction

Creates a value of ApplyPendingMaintenanceAction with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

apmaResourceIdentifier - The RDS Amazon Resource Name (ARN) of the resource that the pending maintenance action applies to. For information about creating an ARN, see Constructing an RDS Amazon Resource Name (ARN) .

apmaApplyAction - The pending maintenance action to apply to this resource. Valid values: system-update , db-upgrade

apmaOptInType - A value that specifies the type of opt-in request, or undoes an opt-in request. An opt-in request of type immediate cannot be undone. Valid values: * immediate - Apply the maintenance action immediately. * next-maintenance - Apply the maintenance action during the next maintenance window for the resource. * undo-opt-in - Cancel any existing next-maintenance opt-in requests.

data ApplyPendingMaintenanceAction

See: applyPendingMaintenanceAction smart constructor.

Request Lenses

apmaResourceIdentifier :: Lens' ApplyPendingMaintenanceAction Text

The RDS Amazon Resource Name (ARN) of the resource that the pending maintenance action applies to. For information about creating an ARN, see Constructing an RDS Amazon Resource Name (ARN) .

apmaApplyAction :: Lens' ApplyPendingMaintenanceAction Text

The pending maintenance action to apply to this resource. Valid values: system-update , db-upgrade

apmaOptInType :: Lens' ApplyPendingMaintenanceAction Text

A value that specifies the type of opt-in request, or undoes an opt-in request. An opt-in request of type immediate cannot be undone. Valid values: * immediate - Apply the maintenance action immediately. * next-maintenance - Apply the maintenance action during the next maintenance window for the resource. * undo-opt-in - Cancel any existing next-maintenance opt-in requests.

Destructuring the Response

applyPendingMaintenanceActionResponse

Arguments

:: Int
apmarsResponseStatus

-> ApplyPendingMaintenanceActionResponse

Creates a value of ApplyPendingMaintenanceActionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

apmarsResourcePendingMaintenanceActions - Undocumented member.

apmarsResponseStatus - -- | The response status code.

data ApplyPendingMaintenanceActionResponse

See: applyPendingMaintenanceActionResponse smart constructor.

Response Lenses

apmarsResourcePendingMaintenanceActions :: Lens' ApplyPendingMaintenanceActionResponse (Maybe ResourcePendingMaintenanceActions)

Undocumented member.

apmarsResponseStatus :: Lens' ApplyPendingMaintenanceActionResponse Int

- | The response status code.