Network.AWS.RDS.CopyDBParameterGroup

Description

Copies the specified DB parameter group.

Creating a Request

copyDBParameterGroup

Arguments

:: Text
cdpgSourceDBParameterGroupIdentifier

-> Text
cdpgTargetDBParameterGroupIdentifier

-> Text
cdpgTargetDBParameterGroupDescription

-> CopyDBParameterGroup

Creates a value of CopyDBParameterGroup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdpgTags - Undocumented member.

cdpgSourceDBParameterGroupIdentifier - The identifier or ARN for the source DB parameter group. For information about creating an ARN, see Constructing an RDS Amazon Resource Name (ARN) . Constraints: * Must specify a valid DB parameter group. * Must specify a valid DB parameter group identifier, for example my-db-param-group , or a valid ARN.

cdpgTargetDBParameterGroupIdentifier - The identifier for the copied DB parameter group. Constraints: * Cannot be null, empty, or blank * Must contain from 1 to 255 alphanumeric characters or hyphens * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens Example: my-db-parameter-group

cdpgTargetDBParameterGroupDescription - A description for the copied DB parameter group.

data CopyDBParameterGroup

See: copyDBParameterGroup smart constructor.

Request Lenses

cdpgTags :: Lens' CopyDBParameterGroup [Tag]

Undocumented member.

cdpgSourceDBParameterGroupIdentifier :: Lens' CopyDBParameterGroup Text

The identifier or ARN for the source DB parameter group. For information about creating an ARN, see Constructing an RDS Amazon Resource Name (ARN) . Constraints: * Must specify a valid DB parameter group. * Must specify a valid DB parameter group identifier, for example my-db-param-group , or a valid ARN.

cdpgTargetDBParameterGroupIdentifier :: Lens' CopyDBParameterGroup Text

The identifier for the copied DB parameter group. Constraints: * Cannot be null, empty, or blank * Must contain from 1 to 255 alphanumeric characters or hyphens * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens Example: my-db-parameter-group

cdpgTargetDBParameterGroupDescription :: Lens' CopyDBParameterGroup Text

A description for the copied DB parameter group.

Destructuring the Response

copyDBParameterGroupResponse

Arguments

:: Int
cdbpgrsResponseStatus

-> CopyDBParameterGroupResponse

Creates a value of CopyDBParameterGroupResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdbpgrsDBParameterGroup - Undocumented member.

cdbpgrsResponseStatus - -- | The response status code.

data CopyDBParameterGroupResponse

See: copyDBParameterGroupResponse smart constructor.

Response Lenses

cdbpgrsDBParameterGroup :: Lens' CopyDBParameterGroupResponse (Maybe DBParameterGroup)

Undocumented member.

cdbpgrsResponseStatus :: Lens' CopyDBParameterGroupResponse Int

- | The response status code.