Network.AWS.RDS.CopyDBSnapshot

Description

Copies the specified DBSnapshot. The source DBSnapshot must be in the "available" state.

http://docs.aws.amazon.com/AmazonRDS/latest/APIReference/API_CopyDBSnapshot.html

Request

data CopyDBSnapshot

Request constructor

copyDBSnapshot

Arguments

:: Text
cdbsSourceDBSnapshotIdentifier

-> Text
cdbsTargetDBSnapshotIdentifier

-> CopyDBSnapshot

CopyDBSnapshot constructor.

The fields accessible through corresponding lenses are:

cdbsSourceDBSnapshotIdentifier :: Text

cdbsTags :: [Tag]

cdbsTargetDBSnapshotIdentifier :: Text

Request lenses

cdbsSourceDBSnapshotIdentifier :: Lens' CopyDBSnapshot Text

The identifier for the source DB snapshot.

Constraints:

Must specify a valid system snapshot in the "available" state. If the source snapshot is in the same region as the copy, specify a valid DB snapshot identifier. If the source snapshot is in a different region than the copy, specify a valid DB snapshot ARN. For more information, go to Copying aDB Snapshot. Example: 'rds:mydb-2012-04-02-00-01'

Example: 'arn:aws:rds:rr-regn-1:123456789012:snapshot:mysql-instance1-snapshot-20130805'

cdbsTags :: Lens' CopyDBSnapshot [Tag]

cdbsTargetDBSnapshotIdentifier :: Lens' CopyDBSnapshot Text

The identifier for the copied snapshot.

Constraints:

Cannot be null, empty, or blank Must contain from 1 to 255 alphanumeric characters or hyphens First character must be a letter Cannot end with a hyphen or contain two consecutive hyphens Example: 'my-db-snapshot'

Response

data CopyDBSnapshotResponse

Response constructor

copyDBSnapshotResponse :: CopyDBSnapshotResponse

CopyDBSnapshotResponse constructor.

The fields accessible through corresponding lenses are:

cdbsrDBSnapshot :: Maybe DBSnapshot

Response lenses

cdbsrDBSnapshot :: Lens' CopyDBSnapshotResponse (Maybe DBSnapshot)