Network.AWS.RDS.CopyOptionGroup

Description

Copies the specified option group.

Creating a Request

copyOptionGroup

Arguments

:: Text
cSourceOptionGroupIdentifier

-> Text
cTargetOptionGroupIdentifier

-> Text
cTargetOptionGroupDescription

-> CopyOptionGroup

Creates a value of CopyOptionGroup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cTags - Undocumented member.

cSourceOptionGroupIdentifier - The identifier or ARN for the source option group. For information about creating an ARN, see Constructing an RDS Amazon Resource Name (ARN) . Constraints: * Must specify a valid option group. * If the source option group is in the same AWS Region as the copy, specify a valid option group identifier, for example my-option-group , or a valid ARN. * If the source option group is in a different AWS Region than the copy, specify a valid option group ARN, for example arn:aws:rds:us-west-2:123456789012:og:special-options .

cTargetOptionGroupIdentifier - The identifier for the copied option group. Constraints: * Cannot be null, empty, or blank * Must contain from 1 to 255 alphanumeric characters or hyphens * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens Example: my-option-group

cTargetOptionGroupDescription - The description for the copied option group.

data CopyOptionGroup

See: copyOptionGroup smart constructor.

Request Lenses

cTags :: Lens' CopyOptionGroup [Tag]

Undocumented member.

cSourceOptionGroupIdentifier :: Lens' CopyOptionGroup Text

The identifier or ARN for the source option group. For information about creating an ARN, see Constructing an RDS Amazon Resource Name (ARN) . Constraints: * Must specify a valid option group. * If the source option group is in the same AWS Region as the copy, specify a valid option group identifier, for example my-option-group , or a valid ARN. * If the source option group is in a different AWS Region than the copy, specify a valid option group ARN, for example arn:aws:rds:us-west-2:123456789012:og:special-options .

cTargetOptionGroupIdentifier :: Lens' CopyOptionGroup Text

The identifier for the copied option group. Constraints: * Cannot be null, empty, or blank * Must contain from 1 to 255 alphanumeric characters or hyphens * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens Example: my-option-group

cTargetOptionGroupDescription :: Lens' CopyOptionGroup Text

The description for the copied option group.

Destructuring the Response

copyOptionGroupResponse

Arguments

:: Int
cogrsResponseStatus

-> CopyOptionGroupResponse

Creates a value of CopyOptionGroupResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cogrsOptionGroup - Undocumented member.

cogrsResponseStatus - -- | The response status code.

data CopyOptionGroupResponse

See: copyOptionGroupResponse smart constructor.

Response Lenses

cogrsOptionGroup :: Lens' CopyOptionGroupResponse (Maybe OptionGroup)

Undocumented member.

cogrsResponseStatus :: Lens' CopyOptionGroupResponse Int

- | The response status code.