Network.AWS.RDS.CreateDBParameterGroup

Description

Creates a new DB parameter group.

A DB parameter group is initially created with the default parameters for the database engine used by the DB instance. To provide custom values for any of the parameters, you must modify the group after creating it using ModifyDBParameterGroup . Once you've created a DB parameter group, you need to associate it with your DB instance using ModifyDBInstance . When you associate a new DB parameter group with a running DB instance, you need to reboot the DB instance without failover for the new DB parameter group and associated settings to take effect.

Important: After you create a DB parameter group, you should wait at least 5 minutes before creating your first DB instance that uses that DB parameter group as the default parameter group. This allows Amazon RDS to fully complete the create action before the parameter group is used as the default for a new DB instance. This is especially important for parameters that are critical when creating the default database for a DB instance, such as the character set for the default database defined by the character_set_database parameter. You can use the Parameter Groups option of the Amazon RDS console or the DescribeDBParameters command to verify that your DB parameter group has been created or modified.

Creating a Request

createDBParameterGroup

Arguments

:: Text
cdbpgDBParameterGroupName

-> Text
cdbpgDBParameterGroupFamily

-> Text
cdbpgDescription

-> CreateDBParameterGroup

Creates a value of CreateDBParameterGroup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdbpgTags - Undocumented member.

cdbpgDBParameterGroupName - The name of the DB parameter group. Constraints: * Must be 1 to 255 alphanumeric characters * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens

cdbpgDBParameterGroupFamily - The DB parameter group family name. A DB parameter group can be associated with one and only one DB parameter group family, and can be applied only to a DB instance running a database engine and engine version compatible with that DB parameter group family.

cdbpgDescription - The description for the DB parameter group.

data CreateDBParameterGroup

See: createDBParameterGroup smart constructor.

Request Lenses

cdbpgTags :: Lens' CreateDBParameterGroup [Tag]

Undocumented member.

cdbpgDBParameterGroupName :: Lens' CreateDBParameterGroup Text

The name of the DB parameter group. Constraints: * Must be 1 to 255 alphanumeric characters * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens

cdbpgDBParameterGroupFamily :: Lens' CreateDBParameterGroup Text

The DB parameter group family name. A DB parameter group can be associated with one and only one DB parameter group family, and can be applied only to a DB instance running a database engine and engine version compatible with that DB parameter group family.

cdbpgDescription :: Lens' CreateDBParameterGroup Text

The description for the DB parameter group.

Destructuring the Response

createDBParameterGroupResponse

Arguments

:: Int
cdpgrsResponseStatus

-> CreateDBParameterGroupResponse

Creates a value of CreateDBParameterGroupResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdpgrsDBParameterGroup - Undocumented member.

cdpgrsResponseStatus - -- | The response status code.

data CreateDBParameterGroupResponse

See: createDBParameterGroupResponse smart constructor.

Response Lenses

cdpgrsDBParameterGroup :: Lens' CreateDBParameterGroupResponse (Maybe DBParameterGroup)

Undocumented member.

cdpgrsResponseStatus :: Lens' CreateDBParameterGroupResponse Int

- | The response status code.