Network.AWS.RDS.CreateDBParameterGroup

Description

Creates a new DB parameter group.

A DB parameter group is initially created with the default parameters for the database engine used by the DB instance. To provide custom values for any of the parameters, you must modify the group after creating it using ModifyDBParameterGroup. Once you've created a DB parameter group, you need to associate it with your DB instance using ModifyDBInstance. When you associate a new DB parameter group with a running DB instance, you need to reboot the DB instance without failover for the new DB parameter group and associated settings to take effect.

After you create a DB parameter group, you should wait at least 5 minutes before creating your first DB instance that uses that DB parameter group as the default parameter group. This allows Amazon RDS to fully complete the create action before the parameter group is used as the default for a new DB instance. This is especially important for parameters that are critical when creating the default database for a DB instance, such as the character set for the default database defined by the character_set_database parameter. You can use the Parameter Groups option of the Amazon RDS console or the DescribeDBParameters command to verify that your DB parameter group has been created or modified.

http://docs.aws.amazon.com/AmazonRDS/latest/APIReference/API_CreateDBParameterGroup.html

Request

data CreateDBParameterGroup

Request constructor

createDBParameterGroup

Arguments

:: Text
cdbpg1DBParameterGroupName

-> Text
cdbpg1DBParameterGroupFamily

-> Text
cdbpg1Description

-> CreateDBParameterGroup

CreateDBParameterGroup constructor.

The fields accessible through corresponding lenses are:

cdbpg1DBParameterGroupFamily :: Text

cdbpg1DBParameterGroupName :: Text

cdbpg1Description :: Text

cdbpg1Tags :: [Tag]

Request lenses

cdbpg1DBParameterGroupFamily :: Lens' CreateDBParameterGroup Text

The DB parameter group family name. A DB parameter group can be associated with one and only one DB parameter group family, and can be applied only to a DB instance running a database engine and engine version compatible with that DB parameter group family.

cdbpg1DBParameterGroupName :: Lens' CreateDBParameterGroup Text

The name of the DB parameter group.

Constraints:

Must be 1 to 255 alphanumeric characters First character must be a letter Cannot end with a hyphen or contain two consecutive hyphens This value is stored as a lower-case string.

cdbpg1Description :: Lens' CreateDBParameterGroup Text

The description for the DB parameter group.

cdbpg1Tags :: Lens' CreateDBParameterGroup [Tag]

Response

data CreateDBParameterGroupResponse

Response constructor

createDBParameterGroupResponse :: CreateDBParameterGroupResponse

CreateDBParameterGroupResponse constructor.

The fields accessible through corresponding lenses are:

cdbpgrDBParameterGroup :: Maybe DBParameterGroup

Response lenses

cdbpgrDBParameterGroup :: Lens' CreateDBParameterGroupResponse (Maybe DBParameterGroup)