Network.AWS.RDS.CreateDBSecurityGroup

Description

Creates a new DB security group. DB security groups control access to a DB instance.

Creating a Request

createDBSecurityGroup

Arguments

:: Text
cdsgDBSecurityGroupName

-> Text
cdsgDBSecurityGroupDescription

-> CreateDBSecurityGroup

Creates a value of CreateDBSecurityGroup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdsgTags - Undocumented member.

cdsgDBSecurityGroupName - The name for the DB security group. This value is stored as a lowercase string. Constraints: * Must be 1 to 255 alphanumeric characters * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens * Must not be Default Example: mysecuritygroup

cdsgDBSecurityGroupDescription - The description for the DB security group.

data CreateDBSecurityGroup

See: createDBSecurityGroup smart constructor.

Request Lenses

cdsgTags :: Lens' CreateDBSecurityGroup [Tag]

Undocumented member.

cdsgDBSecurityGroupName :: Lens' CreateDBSecurityGroup Text

The name for the DB security group. This value is stored as a lowercase string. Constraints: * Must be 1 to 255 alphanumeric characters * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens * Must not be Default Example: mysecuritygroup

cdsgDBSecurityGroupDescription :: Lens' CreateDBSecurityGroup Text

The description for the DB security group.

Destructuring the Response

createDBSecurityGroupResponse

Arguments

:: Int
cdbsgrsResponseStatus

-> CreateDBSecurityGroupResponse

Creates a value of CreateDBSecurityGroupResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdbsgrsDBSecurityGroup - Undocumented member.

cdbsgrsResponseStatus - -- | The response status code.

data CreateDBSecurityGroupResponse

See: createDBSecurityGroupResponse smart constructor.

Response Lenses

cdbsgrsDBSecurityGroup :: Lens' CreateDBSecurityGroupResponse (Maybe DBSecurityGroup)

Undocumented member.

cdbsgrsResponseStatus :: Lens' CreateDBSecurityGroupResponse Int

- | The response status code.