Network.AWS.RDS.CreateEventSubscription

Description

Creates an RDS event notification subscription. This action requires a topic ARN (Amazon Resource Name) created by either the RDS console, the SNS console, or the SNS API. To obtain an ARN with SNS, you must create a topic in Amazon SNS and subscribe to the topic. The ARN is displayed in the SNS console.

You can specify the type of source (SourceType) you want to be notified of, provide a list of RDS sources (SourceIds) that triggers the events, and provide a list of event categories (EventCategories) for events you want to be notified of. For example, you can specify SourceType = db-instance, SourceIds = mydbinstance1, mydbinstance2 and EventCategories = Availability, Backup.

If you specify both the SourceType and SourceIds, such as SourceType = db-instance and SourceIdentifier = myDBInstance1, you will be notified of all the db-instance events for the specified source. If you specify a SourceType but do not specify a SourceIdentifier, you will receive notice of the events for that source type for all your RDS sources. If you do not specify either the SourceType nor the SourceIdentifier, you will be notified of events generated from all RDS sources belonging to your customer account.

Creating a Request

createEventSubscription

Arguments

:: Text
cesSubscriptionName

-> Text
cesSNSTopicARN

-> CreateEventSubscription

Creates a value of CreateEventSubscription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cesEnabled - A Boolean value; set to true to activate the subscription, set to false to create the subscription but not active it.

cesSourceType - The type of source that will be generating the events. For example, if you want to be notified of events generated by a DB instance, you would set this parameter to db-instance. if this value is not specified, all events are returned. Valid values: db-instance | db-cluster | db-parameter-group | db-security-group | db-snapshot | db-cluster-snapshot

cesEventCategories - A list of event categories for a SourceType that you want to subscribe to. You can see a list of the categories for a given SourceType in the Events topic in the Amazon RDS User Guide or by using the DescribeEventCategories action.

cesSourceIds - The list of identifiers of the event sources for which events will be returned. If not specified, then all sources are included in the response. An identifier must begin with a letter and must contain only ASCII letters, digits, and hyphens; it cannot end with a hyphen or contain two consecutive hyphens. Constraints: * If SourceIds are supplied, SourceType must also be provided. * If the source type is a DB instance, then a DBInstanceIdentifier must be supplied. * If the source type is a DB security group, a DBSecurityGroupName must be supplied. * If the source type is a DB parameter group, a DBParameterGroupName must be supplied. * If the source type is a DB snapshot, a DBSnapshotIdentifier must be supplied.

cesTags - Undocumented member.

cesSubscriptionName - The name of the subscription. Constraints: The name must be less than 255 characters.

cesSNSTopicARN - The Amazon Resource Name (ARN) of the SNS topic created for event notification. The ARN is created by Amazon SNS when you create a topic and subscribe to it.

data CreateEventSubscription

See: createEventSubscription smart constructor.

Request Lenses

cesEnabled :: Lens' CreateEventSubscription (Maybe Bool)

A Boolean value; set to true to activate the subscription, set to false to create the subscription but not active it.

cesSourceType :: Lens' CreateEventSubscription (Maybe Text)

The type of source that will be generating the events. For example, if you want to be notified of events generated by a DB instance, you would set this parameter to db-instance. if this value is not specified, all events are returned. Valid values: db-instance | db-cluster | db-parameter-group | db-security-group | db-snapshot | db-cluster-snapshot

cesEventCategories :: Lens' CreateEventSubscription [Text]

A list of event categories for a SourceType that you want to subscribe to. You can see a list of the categories for a given SourceType in the Events topic in the Amazon RDS User Guide or by using the DescribeEventCategories action.

cesSourceIds :: Lens' CreateEventSubscription [Text]

The list of identifiers of the event sources for which events will be returned. If not specified, then all sources are included in the response. An identifier must begin with a letter and must contain only ASCII letters, digits, and hyphens; it cannot end with a hyphen or contain two consecutive hyphens. Constraints: * If SourceIds are supplied, SourceType must also be provided. * If the source type is a DB instance, then a DBInstanceIdentifier must be supplied. * If the source type is a DB security group, a DBSecurityGroupName must be supplied. * If the source type is a DB parameter group, a DBParameterGroupName must be supplied. * If the source type is a DB snapshot, a DBSnapshotIdentifier must be supplied.

cesTags :: Lens' CreateEventSubscription [Tag]

Undocumented member.

cesSubscriptionName :: Lens' CreateEventSubscription Text

The name of the subscription. Constraints: The name must be less than 255 characters.

cesSNSTopicARN :: Lens' CreateEventSubscription Text

The Amazon Resource Name (ARN) of the SNS topic created for event notification. The ARN is created by Amazon SNS when you create a topic and subscribe to it.

Destructuring the Response

createEventSubscriptionResponse

Arguments

:: Int
cesrsResponseStatus

-> CreateEventSubscriptionResponse

Creates a value of CreateEventSubscriptionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cesrsEventSubscription - Undocumented member.

cesrsResponseStatus - -- | The response status code.

data CreateEventSubscriptionResponse

See: createEventSubscriptionResponse smart constructor.

Response Lenses

cesrsEventSubscription :: Lens' CreateEventSubscriptionResponse (Maybe EventSubscription)

Undocumented member.

cesrsResponseStatus :: Lens' CreateEventSubscriptionResponse Int

- | The response status code.