Network.AWS.RDS.CreateOptionGroup

Description

Creates a new option group. You can create up to 20 option groups.

Creating a Request

createOptionGroup

Arguments

:: Text
cogOptionGroupName

-> Text
cogEngineName

-> Text
cogMajorEngineVersion

-> Text
cogOptionGroupDescription

-> CreateOptionGroup

Creates a value of CreateOptionGroup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cogTags - Undocumented member.

cogOptionGroupName - Specifies the name of the option group to be created. Constraints: * Must be 1 to 255 alphanumeric characters or hyphens * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens Example: myoptiongroup

cogEngineName - Specifies the name of the engine that this option group should be associated with.

cogMajorEngineVersion - Specifies the major version of the engine that this option group should be associated with.

cogOptionGroupDescription - The description of the option group.

data CreateOptionGroup

See: createOptionGroup smart constructor.

Request Lenses

cogTags :: Lens' CreateOptionGroup [Tag]

Undocumented member.

cogOptionGroupName :: Lens' CreateOptionGroup Text

Specifies the name of the option group to be created. Constraints: * Must be 1 to 255 alphanumeric characters or hyphens * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens Example: myoptiongroup

cogEngineName :: Lens' CreateOptionGroup Text

Specifies the name of the engine that this option group should be associated with.

cogMajorEngineVersion :: Lens' CreateOptionGroup Text

Specifies the major version of the engine that this option group should be associated with.

cogOptionGroupDescription :: Lens' CreateOptionGroup Text

The description of the option group.

Destructuring the Response

createOptionGroupResponse

Arguments

:: Int
crsResponseStatus

-> CreateOptionGroupResponse

Creates a value of CreateOptionGroupResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

crsOptionGroup - Undocumented member.

crsResponseStatus - -- | The response status code.

data CreateOptionGroupResponse

See: createOptionGroupResponse smart constructor.

Response Lenses

crsOptionGroup :: Lens' CreateOptionGroupResponse (Maybe OptionGroup)

Undocumented member.

crsResponseStatus :: Lens' CreateOptionGroupResponse Int

- | The response status code.