Network.AWS.RDS.DeleteDBInstance

Description

The DeleteDBInstance action deletes a previously provisioned DB instance. When you delete a DB instance, all automated backups for that instance are deleted and cannot be recovered. Manual DB snapshots of the DB instance to be deleted by DeleteDBInstance are not deleted.

If you request a final DB snapshot the status of the Amazon RDS DB instance is deleting until the DB snapshot is created. The API action DescribeDBInstance is used to monitor the status of this operation. The action cannot be canceled or reverted once submitted.

Note that when a DB instance is in a failure state and has a status of failed , incompatible-restore , or incompatible-network , you can only delete it when the SkipFinalSnapshot parameter is set to true .

If the specified DB instance is part of an Amazon Aurora DB cluster, you cannot delete the DB instance if the following are true:

The DB cluster is a Read Replica of another Amazon Aurora DB cluster.

The DB instance is the only instance in the DB cluster.

To delete a DB instance in this case, first call the PromoteReadReplicaDBCluster API action to promote the DB cluster so it's no longer a Read Replica. After the promotion completes, then call the DeleteDBInstance API action to delete the final instance in the DB cluster.

Creating a Request

deleteDBInstance

Arguments

:: Text
ddiDBInstanceIdentifier

-> DeleteDBInstance

Creates a value of DeleteDBInstance with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddiFinalDBSnapshotIdentifier - The DBSnapshotIdentifier of the new DBSnapshot created when SkipFinalSnapshot is set to false . Constraints: * Must be 1 to 255 alphanumeric characters * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens * Cannot be specified when deleting a Read Replica.

ddiSkipFinalSnapshot - Determines whether a final DB snapshot is created before the DB instance is deleted. If true is specified, no DBSnapshot is created. If false is specified, a DB snapshot is created before the DB instance is deleted. Note that when a DB instance is in a failure state and has a status of failed, 'incompatible-restore', or 'incompatible-network', it can only be deleted when the SkipFinalSnapshot parameter is set to "true". Specify true when deleting a Read Replica. Default: false

ddiDBInstanceIdentifier - The DB instance identifier for the DB instance to be deleted. This parameter isn't case-sensitive. Constraints: * Must contain from 1 to 63 alphanumeric characters or hyphens * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens

data DeleteDBInstance

See: deleteDBInstance smart constructor.

Request Lenses

ddiFinalDBSnapshotIdentifier :: Lens' DeleteDBInstance (Maybe Text)

The DBSnapshotIdentifier of the new DBSnapshot created when SkipFinalSnapshot is set to false . Constraints: * Must be 1 to 255 alphanumeric characters * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens * Cannot be specified when deleting a Read Replica.

ddiSkipFinalSnapshot :: Lens' DeleteDBInstance (Maybe Bool)

Determines whether a final DB snapshot is created before the DB instance is deleted. If true is specified, no DBSnapshot is created. If false is specified, a DB snapshot is created before the DB instance is deleted. Note that when a DB instance is in a failure state and has a status of failed, 'incompatible-restore', or 'incompatible-network', it can only be deleted when the SkipFinalSnapshot parameter is set to "true". Specify true when deleting a Read Replica. Default: false

ddiDBInstanceIdentifier :: Lens' DeleteDBInstance Text

The DB instance identifier for the DB instance to be deleted. This parameter isn't case-sensitive. Constraints: * Must contain from 1 to 63 alphanumeric characters or hyphens * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens

Destructuring the Response

deleteDBInstanceResponse

Arguments

:: Int
ddirsResponseStatus

-> DeleteDBInstanceResponse

Creates a value of DeleteDBInstanceResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddirsDBInstance - Undocumented member.

ddirsResponseStatus - -- | The response status code.

data DeleteDBInstanceResponse

See: deleteDBInstanceResponse smart constructor.

Response Lenses

ddirsDBInstance :: Lens' DeleteDBInstanceResponse (Maybe DBInstance)

Undocumented member.

ddirsResponseStatus :: Lens' DeleteDBInstanceResponse Int

- | The response status code.