Network.AWS.RDS.DescribeCertificates

Description

Lists the set of CA certificates provided by Amazon RDS for this AWS account.

Creating a Request

describeCertificates :: DescribeCertificates

Creates a value of DescribeCertificates with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcFilters - This parameter is not currently supported.

dcCertificateIdentifier - The user-supplied certificate identifier. If this parameter is specified, information for only the identified certificate is returned. This parameter isn't case-sensitive. Constraints: * Must contain from 1 to 63 alphanumeric characters or hyphens * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens

dcMarker - An optional pagination token provided by a previous DescribeCertificates request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

dcMaxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: Minimum 20, maximum 100.

data DescribeCertificates

See: describeCertificates smart constructor.

Request Lenses

dcFilters :: Lens' DescribeCertificates [Filter]

This parameter is not currently supported.

dcCertificateIdentifier :: Lens' DescribeCertificates (Maybe Text)

The user-supplied certificate identifier. If this parameter is specified, information for only the identified certificate is returned. This parameter isn't case-sensitive. Constraints: * Must contain from 1 to 63 alphanumeric characters or hyphens * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens

dcMarker :: Lens' DescribeCertificates (Maybe Text)

An optional pagination token provided by a previous DescribeCertificates request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

dcMaxRecords :: Lens' DescribeCertificates (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: Minimum 20, maximum 100.

Destructuring the Response

describeCertificatesResponse

Arguments

:: Int
dcrsResponseStatus

-> DescribeCertificatesResponse

Creates a value of DescribeCertificatesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcrsCertificates - The list of Certificate objects for the AWS account.

dcrsMarker - An optional pagination token provided by a previous DescribeCertificates request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

dcrsResponseStatus - -- | The response status code.

data DescribeCertificatesResponse

Data returned by the DescribeCertificates action.

See: describeCertificatesResponse smart constructor.

Response Lenses

dcrsCertificates :: Lens' DescribeCertificatesResponse [Certificate]

The list of Certificate objects for the AWS account.

dcrsMarker :: Lens' DescribeCertificatesResponse (Maybe Text)

An optional pagination token provided by a previous DescribeCertificates request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

dcrsResponseStatus :: Lens' DescribeCertificatesResponse Int

- | The response status code.