Network.AWS.RDS.DescribeDBEngineVersions

Description

Returns a list of the available DB engines.

This operation returns paginated results.

Creating a Request

describeDBEngineVersions :: DescribeDBEngineVersions

Creates a value of DescribeDBEngineVersions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddevEngineVersion - The database engine version to return. Example: 5.1.49

ddevListSupportedTimezones - If this parameter is specified and the requested engine supports the TimeZone parameter for CreateDBInstance , the response includes a list of supported time zones for each engine version.

ddevDefaultOnly - Indicates that only the default version of the specified engine or engine and major version combination is returned.

ddevFilters - Not currently supported.

ddevEngine - The database engine to return.

ddevDBParameterGroupFamily - The name of a specific DB parameter group family to return details for. Constraints: * Must be 1 to 255 alphanumeric characters * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens

ddevListSupportedCharacterSets - If this parameter is specified and the requested engine supports the CharacterSetName parameter for CreateDBInstance , the response includes a list of supported character sets for each engine version.

ddevMarker - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

ddevMaxRecords - The maximum number of records to include in the response. If more than the MaxRecords value is available, a pagination token called a marker is included in the response so that the following results can be retrieved. Default: 100 Constraints: Minimum 20, maximum 100.

data DescribeDBEngineVersions

See: describeDBEngineVersions smart constructor.

Request Lenses

ddevEngineVersion :: Lens' DescribeDBEngineVersions (Maybe Text)

The database engine version to return. Example: 5.1.49

ddevListSupportedTimezones :: Lens' DescribeDBEngineVersions (Maybe Bool)

If this parameter is specified and the requested engine supports the TimeZone parameter for CreateDBInstance , the response includes a list of supported time zones for each engine version.

ddevDefaultOnly :: Lens' DescribeDBEngineVersions (Maybe Bool)

Indicates that only the default version of the specified engine or engine and major version combination is returned.

ddevFilters :: Lens' DescribeDBEngineVersions [Filter]

Not currently supported.

ddevEngine :: Lens' DescribeDBEngineVersions (Maybe Text)

The database engine to return.

ddevDBParameterGroupFamily :: Lens' DescribeDBEngineVersions (Maybe Text)

The name of a specific DB parameter group family to return details for. Constraints: * Must be 1 to 255 alphanumeric characters * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens

ddevListSupportedCharacterSets :: Lens' DescribeDBEngineVersions (Maybe Bool)

If this parameter is specified and the requested engine supports the CharacterSetName parameter for CreateDBInstance , the response includes a list of supported character sets for each engine version.

ddevMarker :: Lens' DescribeDBEngineVersions (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

ddevMaxRecords :: Lens' DescribeDBEngineVersions (Maybe Int)

The maximum number of records to include in the response. If more than the MaxRecords value is available, a pagination token called a marker is included in the response so that the following results can be retrieved. Default: 100 Constraints: Minimum 20, maximum 100.

Destructuring the Response

describeDBEngineVersionsResponse

Arguments

:: Int
ddevrsResponseStatus

-> DescribeDBEngineVersionsResponse

Creates a value of DescribeDBEngineVersionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddevrsMarker - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

ddevrsDBEngineVersions - A list of DBEngineVersion elements.

ddevrsResponseStatus - -- | The response status code.

data DescribeDBEngineVersionsResponse

Contains the result of a successful invocation of the DescribeDBEngineVersions action.

See: describeDBEngineVersionsResponse smart constructor.

Response Lenses

ddevrsMarker :: Lens' DescribeDBEngineVersionsResponse (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

ddevrsDBEngineVersions :: Lens' DescribeDBEngineVersionsResponse [DBEngineVersion]

A list of DBEngineVersion elements.

ddevrsResponseStatus :: Lens' DescribeDBEngineVersionsResponse Int

- | The response status code.