Network.AWS.RDS.DescribeDBLogFiles

Description

Returns a list of DB log files for the DB instance.

This operation returns paginated results.

Creating a Request

describeDBLogFiles

Arguments

:: Text
ddlfDBInstanceIdentifier

-> DescribeDBLogFiles

Creates a value of DescribeDBLogFiles with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddlfFilenameContains - Filters the available log files for log file names that contain the specified string.

ddlfFilters - This parameter is not currently supported.

ddlfFileSize - Filters the available log files for files larger than the specified size.

ddlfFileLastWritten - Filters the available log files for files written since the specified date, in POSIX timestamp format with milliseconds.

ddlfMarker - The pagination token provided in the previous request. If this parameter is specified the response includes only records beyond the marker, up to MaxRecords.

ddlfMaxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

ddlfDBInstanceIdentifier - The customer-assigned name of the DB instance that contains the log files you want to list. Constraints: * Must contain from 1 to 63 alphanumeric characters or hyphens * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens

data DescribeDBLogFiles

See: describeDBLogFiles smart constructor.

Request Lenses

ddlfFilenameContains :: Lens' DescribeDBLogFiles (Maybe Text)

Filters the available log files for log file names that contain the specified string.

ddlfFilters :: Lens' DescribeDBLogFiles [Filter]

This parameter is not currently supported.

ddlfFileSize :: Lens' DescribeDBLogFiles (Maybe Integer)

Filters the available log files for files larger than the specified size.

ddlfFileLastWritten :: Lens' DescribeDBLogFiles (Maybe Integer)

Filters the available log files for files written since the specified date, in POSIX timestamp format with milliseconds.

ddlfMarker :: Lens' DescribeDBLogFiles (Maybe Text)

The pagination token provided in the previous request. If this parameter is specified the response includes only records beyond the marker, up to MaxRecords.

ddlfMaxRecords :: Lens' DescribeDBLogFiles (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

ddlfDBInstanceIdentifier :: Lens' DescribeDBLogFiles Text

The customer-assigned name of the DB instance that contains the log files you want to list. Constraints: * Must contain from 1 to 63 alphanumeric characters or hyphens * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens

Destructuring the Response

describeDBLogFilesResponse

Arguments

:: Int
ddlfrsResponseStatus

-> DescribeDBLogFilesResponse

Creates a value of DescribeDBLogFilesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddlfrsDescribeDBLogFiles - The DB log files returned.

ddlfrsMarker - A pagination token that can be used in a subsequent DescribeDBLogFiles request.

ddlfrsResponseStatus - -- | The response status code.

data DescribeDBLogFilesResponse

The response from a call to DescribeDBLogFiles .

See: describeDBLogFilesResponse smart constructor.

Response Lenses

ddlfrsDescribeDBLogFiles :: Lens' DescribeDBLogFilesResponse [DescribeDBLogFilesDetails]

The DB log files returned.

ddlfrsMarker :: Lens' DescribeDBLogFilesResponse (Maybe Text)

A pagination token that can be used in a subsequent DescribeDBLogFiles request.

ddlfrsResponseStatus :: Lens' DescribeDBLogFilesResponse Int

- | The response status code.