Network.AWS.RDS.DescribeDBParameterGroups

Description

Returns a list of DBParameterGroup descriptions. If a DBParameterGroupName is specified, the list will contain only the description of the specified DB parameter group.

This operation returns paginated results.

Creating a Request

describeDBParameterGroups :: DescribeDBParameterGroups

Creates a value of DescribeDBParameterGroups with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddpgFilters - This parameter is not currently supported.

ddpgDBParameterGroupName - The name of a specific DB parameter group to return details for. Constraints: * Must be 1 to 255 alphanumeric characters * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens

ddpgMarker - An optional pagination token provided by a previous DescribeDBParameterGroups request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

ddpgMaxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: Minimum 20, maximum 100.

data DescribeDBParameterGroups

See: describeDBParameterGroups smart constructor.

Request Lenses

ddpgFilters :: Lens' DescribeDBParameterGroups [Filter]

This parameter is not currently supported.

ddpgDBParameterGroupName :: Lens' DescribeDBParameterGroups (Maybe Text)

The name of a specific DB parameter group to return details for. Constraints: * Must be 1 to 255 alphanumeric characters * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens

ddpgMarker :: Lens' DescribeDBParameterGroups (Maybe Text)

An optional pagination token provided by a previous DescribeDBParameterGroups request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

ddpgMaxRecords :: Lens' DescribeDBParameterGroups (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: Minimum 20, maximum 100.

Destructuring the Response

describeDBParameterGroupsResponse

Arguments

:: Int
ddpgrsResponseStatus

-> DescribeDBParameterGroupsResponse

Creates a value of DescribeDBParameterGroupsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddpgrsMarker - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

ddpgrsDBParameterGroups - A list of DBParameterGroup instances.

ddpgrsResponseStatus - -- | The response status code.

data DescribeDBParameterGroupsResponse

Contains the result of a successful invocation of the DescribeDBParameterGroups action.

See: describeDBParameterGroupsResponse smart constructor.

Response Lenses

ddpgrsMarker :: Lens' DescribeDBParameterGroupsResponse (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

ddpgrsDBParameterGroups :: Lens' DescribeDBParameterGroupsResponse [DBParameterGroup]

A list of DBParameterGroup instances.

ddpgrsResponseStatus :: Lens' DescribeDBParameterGroupsResponse Int

- | The response status code.