Network.AWS.RDS.DescribeDBParameters

Description

Returns the detailed parameter list for a particular DB parameter group.

This operation returns paginated results.

Creating a Request

describeDBParameters

Arguments

:: Text
ddpDBParameterGroupName

-> DescribeDBParameters

Creates a value of DescribeDBParameters with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddpFilters - This parameter is not currently supported.

ddpMarker - An optional pagination token provided by a previous DescribeDBParameters request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

ddpMaxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: Minimum 20, maximum 100.

ddpSource - The parameter types to return. Default: All parameter types returned Valid Values: user | system | engine-default

ddpDBParameterGroupName - The name of a specific DB parameter group to return details for. Constraints: * Must be 1 to 255 alphanumeric characters * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens

data DescribeDBParameters

See: describeDBParameters smart constructor.

Request Lenses

ddpFilters :: Lens' DescribeDBParameters [Filter]

This parameter is not currently supported.

ddpMarker :: Lens' DescribeDBParameters (Maybe Text)

An optional pagination token provided by a previous DescribeDBParameters request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

ddpMaxRecords :: Lens' DescribeDBParameters (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: Minimum 20, maximum 100.

ddpSource :: Lens' DescribeDBParameters (Maybe Text)

The parameter types to return. Default: All parameter types returned Valid Values: user | system | engine-default

ddpDBParameterGroupName :: Lens' DescribeDBParameters Text

The name of a specific DB parameter group to return details for. Constraints: * Must be 1 to 255 alphanumeric characters * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens

Destructuring the Response

describeDBParametersResponse

Arguments

:: Int
ddprsResponseStatus

-> DescribeDBParametersResponse

Creates a value of DescribeDBParametersResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddprsMarker - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

ddprsParameters - A list of Parameter values.

ddprsResponseStatus - -- | The response status code.

data DescribeDBParametersResponse

Contains the result of a successful invocation of the DescribeDBParameters action.

See: describeDBParametersResponse smart constructor.

Response Lenses

ddprsMarker :: Lens' DescribeDBParametersResponse (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

ddprsParameters :: Lens' DescribeDBParametersResponse [Parameter]

A list of Parameter values.

ddprsResponseStatus :: Lens' DescribeDBParametersResponse Int

- | The response status code.