Network.AWS.RDS.DescribeDBSecurityGroups

Description

Returns a list of DBSecurityGroup descriptions. If a DBSecurityGroupName is specified, the list will contain only the descriptions of the specified DB security group.

This operation returns paginated results.

Creating a Request

describeDBSecurityGroups :: DescribeDBSecurityGroups

Creates a value of DescribeDBSecurityGroups with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddbsgFilters - This parameter is not currently supported.

ddbsgMarker - An optional pagination token provided by a previous DescribeDBSecurityGroups request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

ddbsgMaxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: Minimum 20, maximum 100.

ddbsgDBSecurityGroupName - The name of the DB security group to return details for.

data DescribeDBSecurityGroups

See: describeDBSecurityGroups smart constructor.

Request Lenses

ddbsgFilters :: Lens' DescribeDBSecurityGroups [Filter]

This parameter is not currently supported.

ddbsgMarker :: Lens' DescribeDBSecurityGroups (Maybe Text)

An optional pagination token provided by a previous DescribeDBSecurityGroups request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

ddbsgMaxRecords :: Lens' DescribeDBSecurityGroups (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: Minimum 20, maximum 100.

ddbsgDBSecurityGroupName :: Lens' DescribeDBSecurityGroups (Maybe Text)

The name of the DB security group to return details for.

Destructuring the Response

describeDBSecurityGroupsResponse

Arguments

:: Int
ddbsgrsResponseStatus

-> DescribeDBSecurityGroupsResponse

Creates a value of DescribeDBSecurityGroupsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddbsgrsDBSecurityGroups - A list of DBSecurityGroup instances.

ddbsgrsMarker - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

ddbsgrsResponseStatus - -- | The response status code.

data DescribeDBSecurityGroupsResponse

Contains the result of a successful invocation of the DescribeDBSecurityGroups action.

See: describeDBSecurityGroupsResponse smart constructor.

Response Lenses

ddbsgrsDBSecurityGroups :: Lens' DescribeDBSecurityGroupsResponse [DBSecurityGroup]

A list of DBSecurityGroup instances.

ddbsgrsMarker :: Lens' DescribeDBSecurityGroupsResponse (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

ddbsgrsResponseStatus :: Lens' DescribeDBSecurityGroupsResponse Int

- | The response status code.