Network.AWS.RDS.DescribeDBSecurityGroups

Description

Returns a list of DBSecurityGroup descriptions. If a DBSecurityGroupName is specified, the list will contain only the descriptions of the specified DB security group.

http://docs.aws.amazon.com/AmazonRDS/latest/APIReference/API_DescribeDBSecurityGroups.html

Request

data DescribeDBSecurityGroups

Request constructor

describeDBSecurityGroups :: DescribeDBSecurityGroups

DescribeDBSecurityGroups constructor.

The fields accessible through corresponding lenses are:

ddbsg1DBSecurityGroupName :: Maybe Text

ddbsg1Filters :: [Filter]

ddbsg1Marker :: Maybe Text

ddbsg1MaxRecords :: Maybe Int

Request lenses

ddbsg1DBSecurityGroupName :: Lens' DescribeDBSecurityGroups (Maybe Text)

The name of the DB security group to return details for.

ddbsg1Filters :: Lens' DescribeDBSecurityGroups [Filter]

This parameter is not currently supported.

ddbsg1Marker :: Lens' DescribeDBSecurityGroups (Maybe Text)

An optional pagination token provided by a previous DescribeDBSecurityGroups request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

ddbsg1MaxRecords :: Lens' DescribeDBSecurityGroups (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results may be retrieved.

Default: 100

Constraints: minimum 20, maximum 100

Response

data DescribeDBSecurityGroupsResponse

Response constructor

describeDBSecurityGroupsResponse :: DescribeDBSecurityGroupsResponse

DescribeDBSecurityGroupsResponse constructor.

The fields accessible through corresponding lenses are:

ddbsgr1DBSecurityGroups :: [DBSecurityGroup]

ddbsgr1Marker :: Maybe Text

Response lenses

ddbsgr1DBSecurityGroups :: Lens' DescribeDBSecurityGroupsResponse [DBSecurityGroup]

A list of DBSecurityGroup instances.

ddbsgr1Marker :: Lens' DescribeDBSecurityGroupsResponse (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.