Network.AWS.RDS.DescribeDBSnapshots

Description

Returns information about DB snapshots. This API action supports pagination.

This operation returns paginated results.

Creating a Request

describeDBSnapshots :: DescribeDBSnapshots

Creates a value of DescribeDBSnapshots with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddsIncludeShared - Set this value to true to include shared manual DB snapshots from other AWS accounts that this AWS account has been given permission to copy or restore, otherwise set this value to false . The default is false . You can give an AWS account permission to restore a manual DB snapshot from another AWS account by using the ModifyDBSnapshotAttribute API action.

ddsFilters - This parameter is not currently supported.

ddsDBSnapshotIdentifier - A specific DB snapshot identifier to describe. This parameter cannot be used in conjunction with DBInstanceIdentifier . This value is stored as a lowercase string. Constraints: * Must be 1 to 255 alphanumeric characters. * First character must be a letter. * Cannot end with a hyphen or contain two consecutive hyphens. * If this identifier is for an automated snapshot, the SnapshotType parameter must also be specified.

ddsSnapshotType - The type of snapshots to be returned. You can specify one of the following values: * automated - Return all DB snapshots that have been automatically taken by Amazon RDS for my AWS account. * manual - Return all DB snapshots that have been taken by my AWS account. * shared - Return all manual DB snapshots that have been shared to my AWS account. * public - Return all DB snapshots that have been marked as public. If you don't specify a SnapshotType value, then both automated and manual snapshots are returned. Shared and public DB snapshots are not included in the returned results by default. You can include shared snapshots with these results by setting the IncludeShared parameter to true . You can include public snapshots with these results by setting the IncludePublic parameter to true . The IncludeShared and IncludePublic parameters don't apply for SnapshotType values of manual or automated . The IncludePublic parameter doesn't apply when SnapshotType is set to shared . The IncludeShared parameter doesn't apply when SnapshotType is set to public .

ddsDBInstanceIdentifier - The ID of the DB instance to retrieve the list of DB snapshots for. This parameter cannot be used in conjunction with DBSnapshotIdentifier . This parameter is not case-sensitive. Constraints: * Must contain from 1 to 63 alphanumeric characters or hyphens * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens

ddsMarker - An optional pagination token provided by a previous DescribeDBSnapshots request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

ddsMaxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: Minimum 20, maximum 100.

ddsIncludePublic - Set this value to true to include manual DB snapshots that are public and can be copied or restored by any AWS account, otherwise set this value to false . The default is false . You can share a manual DB snapshot as public by using the ModifyDBSnapshotAttribute API.

data DescribeDBSnapshots

See: describeDBSnapshots smart constructor.

Request Lenses

ddsIncludeShared :: Lens' DescribeDBSnapshots (Maybe Bool)

Set this value to true to include shared manual DB snapshots from other AWS accounts that this AWS account has been given permission to copy or restore, otherwise set this value to false . The default is false . You can give an AWS account permission to restore a manual DB snapshot from another AWS account by using the ModifyDBSnapshotAttribute API action.

ddsFilters :: Lens' DescribeDBSnapshots [Filter]

This parameter is not currently supported.

ddsDBSnapshotIdentifier :: Lens' DescribeDBSnapshots (Maybe Text)

A specific DB snapshot identifier to describe. This parameter cannot be used in conjunction with DBInstanceIdentifier . This value is stored as a lowercase string. Constraints: * Must be 1 to 255 alphanumeric characters. * First character must be a letter. * Cannot end with a hyphen or contain two consecutive hyphens. * If this identifier is for an automated snapshot, the SnapshotType parameter must also be specified.

ddsSnapshotType :: Lens' DescribeDBSnapshots (Maybe Text)

The type of snapshots to be returned. You can specify one of the following values: * automated - Return all DB snapshots that have been automatically taken by Amazon RDS for my AWS account. * manual - Return all DB snapshots that have been taken by my AWS account. * shared - Return all manual DB snapshots that have been shared to my AWS account. * public - Return all DB snapshots that have been marked as public. If you don't specify a SnapshotType value, then both automated and manual snapshots are returned. Shared and public DB snapshots are not included in the returned results by default. You can include shared snapshots with these results by setting the IncludeShared parameter to true . You can include public snapshots with these results by setting the IncludePublic parameter to true . The IncludeShared and IncludePublic parameters don't apply for SnapshotType values of manual or automated . The IncludePublic parameter doesn't apply when SnapshotType is set to shared . The IncludeShared parameter doesn't apply when SnapshotType is set to public .

ddsDBInstanceIdentifier :: Lens' DescribeDBSnapshots (Maybe Text)

The ID of the DB instance to retrieve the list of DB snapshots for. This parameter cannot be used in conjunction with DBSnapshotIdentifier . This parameter is not case-sensitive. Constraints: * Must contain from 1 to 63 alphanumeric characters or hyphens * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens

ddsMarker :: Lens' DescribeDBSnapshots (Maybe Text)

An optional pagination token provided by a previous DescribeDBSnapshots request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

ddsMaxRecords :: Lens' DescribeDBSnapshots (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: Minimum 20, maximum 100.

ddsIncludePublic :: Lens' DescribeDBSnapshots (Maybe Bool)

Set this value to true to include manual DB snapshots that are public and can be copied or restored by any AWS account, otherwise set this value to false . The default is false . You can share a manual DB snapshot as public by using the ModifyDBSnapshotAttribute API.

Destructuring the Response

describeDBSnapshotsResponse

Arguments

:: Int
ddsrsResponseStatus

-> DescribeDBSnapshotsResponse

Creates a value of DescribeDBSnapshotsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddsrsMarker - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

ddsrsDBSnapshots - A list of DBSnapshot instances.

ddsrsResponseStatus - -- | The response status code.

data DescribeDBSnapshotsResponse

Contains the result of a successful invocation of the DescribeDBSnapshots action.

See: describeDBSnapshotsResponse smart constructor.

Response Lenses

ddsrsMarker :: Lens' DescribeDBSnapshotsResponse (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

ddsrsDBSnapshots :: Lens' DescribeDBSnapshotsResponse [DBSnapshot]

A list of DBSnapshot instances.

ddsrsResponseStatus :: Lens' DescribeDBSnapshotsResponse Int

- | The response status code.