Network.AWS.RDS.DescribeDBSubnetGroups

Description

Returns a list of DBSubnetGroup descriptions. If a DBSubnetGroupName is specified, the list will contain only the descriptions of the specified DBSubnetGroup.

For an overview of CIDR ranges, go to the Wikipedia Tutorial.

http://docs.aws.amazon.com/AmazonRDS/latest/APIReference/API_DescribeDBSubnetGroups.html

Request

data DescribeDBSubnetGroups

Request constructor

describeDBSubnetGroups :: DescribeDBSubnetGroups

DescribeDBSubnetGroups constructor.

The fields accessible through corresponding lenses are:

ddbsgDBSubnetGroupName :: Maybe Text

ddbsgFilters :: [Filter]

ddbsgMarker :: Maybe Text

ddbsgMaxRecords :: Maybe Int

Request lenses

ddbsgDBSubnetGroupName :: Lens' DescribeDBSubnetGroups (Maybe Text)

The name of the DB subnet group to return details for.

ddbsgFilters :: Lens' DescribeDBSubnetGroups [Filter]

This parameter is not currently supported.

ddbsgMarker :: Lens' DescribeDBSubnetGroups (Maybe Text)

An optional pagination token provided by a previous DescribeDBSubnetGroups request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

ddbsgMaxRecords :: Lens' DescribeDBSubnetGroups (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results may be retrieved.

Default: 100

Constraints: minimum 20, maximum 100

Response

data DescribeDBSubnetGroupsResponse

Response constructor

describeDBSubnetGroupsResponse :: DescribeDBSubnetGroupsResponse

DescribeDBSubnetGroupsResponse constructor.

The fields accessible through corresponding lenses are:

ddbsgrDBSubnetGroups :: [DBSubnetGroup]

ddbsgrMarker :: Maybe Text

Response lenses

ddbsgrDBSubnetGroups :: Lens' DescribeDBSubnetGroupsResponse [DBSubnetGroup]

A list of DBSubnetGroup instances.

ddbsgrMarker :: Lens' DescribeDBSubnetGroupsResponse (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.