Network.AWS.RDS.DescribeEventCategories

Description

Displays a list of categories for all event source types, or, if specified, for a specified source type. You can see a list of the event categories and source types in the Events topic in the Amazon RDS User Guide.

Creating a Request

describeEventCategories :: DescribeEventCategories

Creates a value of DescribeEventCategories with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

decSourceType - The type of source that will be generating the events. Valid values: db-instance | db-parameter-group | db-security-group | db-snapshot

decFilters - This parameter is not currently supported.

data DescribeEventCategories

See: describeEventCategories smart constructor.

Request Lenses

decSourceType :: Lens' DescribeEventCategories (Maybe Text)

The type of source that will be generating the events. Valid values: db-instance | db-parameter-group | db-security-group | db-snapshot

decFilters :: Lens' DescribeEventCategories [Filter]

This parameter is not currently supported.

Destructuring the Response

describeEventCategoriesResponse

Arguments

:: Int
decrsResponseStatus

-> DescribeEventCategoriesResponse

Creates a value of DescribeEventCategoriesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

decrsEventCategoriesMapList - A list of EventCategoriesMap data types.

decrsResponseStatus - -- | The response status code.

data DescribeEventCategoriesResponse

Data returned from the DescribeEventCategories action.

See: describeEventCategoriesResponse smart constructor.

Response Lenses

decrsEventCategoriesMapList :: Lens' DescribeEventCategoriesResponse [EventCategoriesMap]

A list of EventCategoriesMap data types.

decrsResponseStatus :: Lens' DescribeEventCategoriesResponse Int

- | The response status code.