Network.AWS.RDS.DescribeEventCategories

Description

Displays a list of categories for all event source types, or, if specified, for a specified source type. You can see a list of the event categories and source types in the Events topic in the Amazon RDS User Guide.

http://docs.aws.amazon.com/AmazonRDS/latest/APIReference/API_DescribeEventCategories.html

Request

data DescribeEventCategories

Request constructor

describeEventCategories :: DescribeEventCategories

DescribeEventCategories constructor.

The fields accessible through corresponding lenses are:

decFilters :: [Filter]

decSourceType :: Maybe Text

Request lenses

decFilters :: Lens' DescribeEventCategories [Filter]

This parameter is not currently supported.

decSourceType :: Lens' DescribeEventCategories (Maybe Text)

The type of source that will be generating the events.

Valid values: db-instance | db-parameter-group | db-security-group | db-snapshot

Response

data DescribeEventCategoriesResponse

Response constructor

describeEventCategoriesResponse :: DescribeEventCategoriesResponse

DescribeEventCategoriesResponse constructor.

The fields accessible through corresponding lenses are:

decrEventCategoriesMapList :: [EventCategoriesMap]

Response lenses

decrEventCategoriesMapList :: Lens' DescribeEventCategoriesResponse [EventCategoriesMap]

A list of EventCategoriesMap data types.