Network.AWS.RDS.DescribeEvents

Description

Returns events related to DB instances, DB security groups, DB snapshots, and DB parameter groups for the past 14 days. Events specific to a particular DB instance, DB security group, database snapshot, or DB parameter group can be obtained by providing the name as a parameter. By default, the past hour of events are returned.

This operation returns paginated results.

Creating a Request

describeEvents :: DescribeEvents

Creates a value of DescribeEvents with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

deStartTime - The beginning of the time interval to retrieve events for, specified in ISO 8601 format. For more information about ISO 8601, go to the ISO8601 Wikipedia page. Example: 2009-07-08T18:00Z

deSourceType - The event source to retrieve events for. If no value is specified, all events are returned.

deFilters - This parameter is not currently supported.

deSourceIdentifier - The identifier of the event source for which events will be returned. If not specified, then all sources are included in the response. Constraints: * If SourceIdentifier is supplied, SourceType must also be provided. * If the source type is DBInstance , then a DBInstanceIdentifier must be supplied. * If the source type is DBSecurityGroup , a DBSecurityGroupName must be supplied. * If the source type is DBParameterGroup , a DBParameterGroupName must be supplied. * If the source type is DBSnapshot , a DBSnapshotIdentifier must be supplied. * Cannot end with a hyphen or contain two consecutive hyphens.

deEventCategories - A list of event categories that trigger notifications for a event notification subscription.

deMarker - An optional pagination token provided by a previous DescribeEvents request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

deMaxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: Minimum 20, maximum 100.

deEndTime - The end of the time interval for which to retrieve events, specified in ISO 8601 format. For more information about ISO 8601, go to the ISO8601 Wikipedia page. Example: 2009-07-08T18:00Z

deDuration - The number of minutes to retrieve events for. Default: 60

data DescribeEvents

See: describeEvents smart constructor.

Request Lenses

deStartTime :: Lens' DescribeEvents (Maybe UTCTime)

The beginning of the time interval to retrieve events for, specified in ISO 8601 format. For more information about ISO 8601, go to the ISO8601 Wikipedia page. Example: 2009-07-08T18:00Z

deSourceType :: Lens' DescribeEvents (Maybe SourceType)

The event source to retrieve events for. If no value is specified, all events are returned.

deFilters :: Lens' DescribeEvents [Filter]

This parameter is not currently supported.

deSourceIdentifier :: Lens' DescribeEvents (Maybe Text)

The identifier of the event source for which events will be returned. If not specified, then all sources are included in the response. Constraints: * If SourceIdentifier is supplied, SourceType must also be provided. * If the source type is DBInstance , then a DBInstanceIdentifier must be supplied. * If the source type is DBSecurityGroup , a DBSecurityGroupName must be supplied. * If the source type is DBParameterGroup , a DBParameterGroupName must be supplied. * If the source type is DBSnapshot , a DBSnapshotIdentifier must be supplied. * Cannot end with a hyphen or contain two consecutive hyphens.

deEventCategories :: Lens' DescribeEvents [Text]

A list of event categories that trigger notifications for a event notification subscription.

deMarker :: Lens' DescribeEvents (Maybe Text)

An optional pagination token provided by a previous DescribeEvents request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

deMaxRecords :: Lens' DescribeEvents (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: Minimum 20, maximum 100.

deEndTime :: Lens' DescribeEvents (Maybe UTCTime)

The end of the time interval for which to retrieve events, specified in ISO 8601 format. For more information about ISO 8601, go to the ISO8601 Wikipedia page. Example: 2009-07-08T18:00Z

deDuration :: Lens' DescribeEvents (Maybe Int)

The number of minutes to retrieve events for. Default: 60

Destructuring the Response

describeEventsResponse

Arguments

:: Int
dersResponseStatus

-> DescribeEventsResponse

Creates a value of DescribeEventsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dersEvents - A list of Event instances.

dersMarker - An optional pagination token provided by a previous Events request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

dersResponseStatus - -- | The response status code.

data DescribeEventsResponse

Contains the result of a successful invocation of the DescribeEvents action.

See: describeEventsResponse smart constructor.

Response Lenses

dersEvents :: Lens' DescribeEventsResponse [Event]

A list of Event instances.

dersMarker :: Lens' DescribeEventsResponse (Maybe Text)

An optional pagination token provided by a previous Events request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

dersResponseStatus :: Lens' DescribeEventsResponse Int

- | The response status code.