Network.AWS.RDS.DescribeEvents

Description

Returns events related to DB instances, DB security groups, DB snapshots, and DB parameter groups for the past 14 days. Events specific to a particular DB instance, DB security group, database snapshot, or DB parameter group can be obtained by providing the name as a parameter. By default, the past hour of events are returned.

http://docs.aws.amazon.com/AmazonRDS/latest/APIReference/API_DescribeEvents.html

Request

data DescribeEvents

Request constructor

describeEvents :: DescribeEvents

DescribeEvents constructor.

The fields accessible through corresponding lenses are:

deDuration :: Maybe Int

deEndTime :: Maybe UTCTime

deEventCategories :: [Text]

deFilters :: [Filter]

deMarker :: Maybe Text

deMaxRecords :: Maybe Int

deSourceIdentifier :: Maybe Text

deSourceType :: Maybe SourceType

deStartTime :: Maybe UTCTime

Request lenses

deDuration :: Lens' DescribeEvents (Maybe Int)

The number of minutes to retrieve events for.

Default: 60

deEndTime :: Lens' DescribeEvents (Maybe UTCTime)

The end of the time interval for which to retrieve events, specified in ISO 8601 format. For more information about ISO 8601, go to the ISO8601 Wikipediapage.

Example: 2009-07-08T18:00Z

deEventCategories :: Lens' DescribeEvents [Text]

A list of event categories that trigger notifications for a event notification subscription.

deFilters :: Lens' DescribeEvents [Filter]

This parameter is not currently supported.

deMarker :: Lens' DescribeEvents (Maybe Text)

An optional pagination token provided by a previous DescribeEvents request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

deMaxRecords :: Lens' DescribeEvents (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results may be retrieved.

Default: 100

Constraints: minimum 20, maximum 100

deSourceIdentifier :: Lens' DescribeEvents (Maybe Text)

The identifier of the event source for which events will be returned. If not specified, then all sources are included in the response.

Constraints:

If SourceIdentifier is supplied, SourceType must also be provided. If the source type is DBInstance, then a DBInstanceIdentifier must be supplied. If the source type is DBSecurityGroup, a DBSecurityGroupName must be supplied. If the source type is DBParameterGroup, a DBParameterGroupName must be supplied. If the source type is DBSnapshot, a DBSnapshotIdentifier must be supplied. Cannot end with a hyphen or contain two consecutive hyphens.

deSourceType :: Lens' DescribeEvents (Maybe SourceType)

The event source to retrieve events for. If no value is specified, all events are returned.

deStartTime :: Lens' DescribeEvents (Maybe UTCTime)

The beginning of the time interval to retrieve events for, specified in ISO 8601 format. For more information about ISO 8601, go to the ISO8601 Wikipediapage.

Example: 2009-07-08T18:00Z

Response

data DescribeEventsResponse

Response constructor

describeEventsResponse :: DescribeEventsResponse

DescribeEventsResponse constructor.

The fields accessible through corresponding lenses are:

derEvents :: [Event]

derMarker :: Maybe Text

Response lenses

derEvents :: Lens' DescribeEventsResponse [Event]

A list of Event instances.

derMarker :: Lens' DescribeEventsResponse (Maybe Text)

An optional pagination token provided by a previous Events request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .