Network.AWS.RDS.DescribeOrderableDBInstanceOptions

Description

Returns a list of orderable DB instance options for the specified engine.

This operation returns paginated results.

Creating a Request

describeOrderableDBInstanceOptions

Arguments

:: Text
dodioEngine

-> DescribeOrderableDBInstanceOptions

Creates a value of DescribeOrderableDBInstanceOptions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dodioEngineVersion - The engine version filter value. Specify this parameter to show only the available offerings matching the specified engine version.

dodioFilters - This parameter is not currently supported.

dodioDBInstanceClass - The DB instance class filter value. Specify this parameter to show only the available offerings matching the specified DB instance class.

dodioLicenseModel - The license model filter value. Specify this parameter to show only the available offerings matching the specified license model.

dodioMarker - An optional pagination token provided by a previous DescribeOrderableDBInstanceOptions request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

dodioMaxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: Minimum 20, maximum 100.

dodioVPC - The VPC filter value. Specify this parameter to show only the available VPC or non-VPC offerings.

dodioEngine - The name of the engine to retrieve DB instance options for.

data DescribeOrderableDBInstanceOptions

See: describeOrderableDBInstanceOptions smart constructor.

Request Lenses

dodioEngineVersion :: Lens' DescribeOrderableDBInstanceOptions (Maybe Text)

The engine version filter value. Specify this parameter to show only the available offerings matching the specified engine version.

dodioFilters :: Lens' DescribeOrderableDBInstanceOptions [Filter]

This parameter is not currently supported.

dodioDBInstanceClass :: Lens' DescribeOrderableDBInstanceOptions (Maybe Text)

The DB instance class filter value. Specify this parameter to show only the available offerings matching the specified DB instance class.

dodioLicenseModel :: Lens' DescribeOrderableDBInstanceOptions (Maybe Text)

The license model filter value. Specify this parameter to show only the available offerings matching the specified license model.

dodioMarker :: Lens' DescribeOrderableDBInstanceOptions (Maybe Text)

An optional pagination token provided by a previous DescribeOrderableDBInstanceOptions request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

dodioMaxRecords :: Lens' DescribeOrderableDBInstanceOptions (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: Minimum 20, maximum 100.

dodioVPC :: Lens' DescribeOrderableDBInstanceOptions (Maybe Bool)

The VPC filter value. Specify this parameter to show only the available VPC or non-VPC offerings.

dodioEngine :: Lens' DescribeOrderableDBInstanceOptions Text

The name of the engine to retrieve DB instance options for.

Destructuring the Response

describeOrderableDBInstanceOptionsResponse

Arguments

:: Int
dodiorsResponseStatus

-> DescribeOrderableDBInstanceOptionsResponse

Creates a value of DescribeOrderableDBInstanceOptionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dodiorsOrderableDBInstanceOptions - An OrderableDBInstanceOption structure containing information about orderable options for the DB instance.

dodiorsMarker - An optional pagination token provided by a previous OrderableDBInstanceOptions request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

dodiorsResponseStatus - -- | The response status code.

data DescribeOrderableDBInstanceOptionsResponse

Contains the result of a successful invocation of the DescribeOrderableDBInstanceOptions action.

See: describeOrderableDBInstanceOptionsResponse smart constructor.

Response Lenses

dodiorsOrderableDBInstanceOptions :: Lens' DescribeOrderableDBInstanceOptionsResponse [OrderableDBInstanceOption]

An OrderableDBInstanceOption structure containing information about orderable options for the DB instance.

dodiorsMarker :: Lens' DescribeOrderableDBInstanceOptionsResponse (Maybe Text)

An optional pagination token provided by a previous OrderableDBInstanceOptions request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

dodiorsResponseStatus :: Lens' DescribeOrderableDBInstanceOptionsResponse Int

- | The response status code.