Network.AWS.RDS.DescribeOrderableDBInstanceOptions

Description

Returns a list of orderable DB instance options for the specified engine.

http://docs.aws.amazon.com/AmazonRDS/latest/APIReference/API_DescribeOrderableDBInstanceOptions.html

Request

data DescribeOrderableDBInstanceOptions

Request constructor

describeOrderableDBInstanceOptions

Arguments

:: Text
dodbioEngine

-> DescribeOrderableDBInstanceOptions

DescribeOrderableDBInstanceOptions constructor.

The fields accessible through corresponding lenses are:

dodbioDBInstanceClass :: Maybe Text

dodbioEngine :: Text

dodbioEngineVersion :: Maybe Text

dodbioFilters :: [Filter]

dodbioLicenseModel :: Maybe Text

dodbioMarker :: Maybe Text

dodbioMaxRecords :: Maybe Int

dodbioVpc :: Maybe Bool

Request lenses

dodbioDBInstanceClass :: Lens' DescribeOrderableDBInstanceOptions (Maybe Text)

The DB instance class filter value. Specify this parameter to show only the available offerings matching the specified DB instance class.

dodbioEngine :: Lens' DescribeOrderableDBInstanceOptions Text

The name of the engine to retrieve DB instance options for.

dodbioEngineVersion :: Lens' DescribeOrderableDBInstanceOptions (Maybe Text)

The engine version filter value. Specify this parameter to show only the available offerings matching the specified engine version.

dodbioFilters :: Lens' DescribeOrderableDBInstanceOptions [Filter]

This parameter is not currently supported.

dodbioLicenseModel :: Lens' DescribeOrderableDBInstanceOptions (Maybe Text)

The license model filter value. Specify this parameter to show only the available offerings matching the specified license model.

dodbioMarker :: Lens' DescribeOrderableDBInstanceOptions (Maybe Text)

An optional pagination token provided by a previous DescribeOrderableDBInstanceOptions request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

dodbioMaxRecords :: Lens' DescribeOrderableDBInstanceOptions (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: minimum 20, maximum 100

dodbioVpc :: Lens' DescribeOrderableDBInstanceOptions (Maybe Bool)

The VPC filter value. Specify this parameter to show only the available VPC or non-VPC offerings.

Response

data DescribeOrderableDBInstanceOptionsResponse

Response constructor

describeOrderableDBInstanceOptionsResponse :: DescribeOrderableDBInstanceOptionsResponse

DescribeOrderableDBInstanceOptionsResponse constructor.

The fields accessible through corresponding lenses are:

dodbiorMarker :: Maybe Text

dodbiorOrderableDBInstanceOptions :: [OrderableDBInstanceOption]

Response lenses

dodbiorMarker :: Lens' DescribeOrderableDBInstanceOptionsResponse (Maybe Text)

An optional pagination token provided by a previous OrderableDBInstanceOptions request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

dodbiorOrderableDBInstanceOptions :: Lens' DescribeOrderableDBInstanceOptionsResponse [OrderableDBInstanceOption]

An OrderableDBInstanceOption structure containing information about orderable options for the DB instance.