Network.AWS.RDS.DescribeReservedDBInstances

Description

Returns information about reserved DB instances for this account, or about a specified reserved DB instance.

This operation returns paginated results.

Creating a Request

describeReservedDBInstances :: DescribeReservedDBInstances

Creates a value of DescribeReservedDBInstances with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drdiProductDescription - The product description filter value. Specify this parameter to show only those reservations matching the specified product description.

drdiFilters - This parameter is not currently supported.

drdiReservedDBInstanceId - The reserved DB instance identifier filter value. Specify this parameter to show only the reservation that matches the specified reservation ID.

drdiDBInstanceClass - The DB instance class filter value. Specify this parameter to show only those reservations matching the specified DB instances class.

drdiMarker - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

drdiMaxRecords - The maximum number of records to include in the response. If more than the MaxRecords value is available, a pagination token called a marker is included in the response so that the following results can be retrieved. Default: 100 Constraints: Minimum 20, maximum 100.

drdiMultiAZ - The Multi-AZ filter value. Specify this parameter to show only those reservations matching the specified Multi-AZ parameter.

drdiReservedDBInstancesOfferingId - The offering identifier filter value. Specify this parameter to show only purchased reservations matching the specified offering identifier.

drdiOfferingType - The offering type filter value. Specify this parameter to show only the available offerings matching the specified offering type. Valid Values: "Partial Upfront" | "All Upfront" | "No Upfront"

drdiDuration - The duration filter value, specified in years or seconds. Specify this parameter to show only reservations for this duration. Valid Values: 1 | 3 | 31536000 | 94608000

data DescribeReservedDBInstances

See: describeReservedDBInstances smart constructor.

Request Lenses

drdiProductDescription :: Lens' DescribeReservedDBInstances (Maybe Text)

The product description filter value. Specify this parameter to show only those reservations matching the specified product description.

drdiFilters :: Lens' DescribeReservedDBInstances [Filter]

This parameter is not currently supported.

drdiReservedDBInstanceId :: Lens' DescribeReservedDBInstances (Maybe Text)

The reserved DB instance identifier filter value. Specify this parameter to show only the reservation that matches the specified reservation ID.

drdiDBInstanceClass :: Lens' DescribeReservedDBInstances (Maybe Text)

The DB instance class filter value. Specify this parameter to show only those reservations matching the specified DB instances class.

drdiMarker :: Lens' DescribeReservedDBInstances (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

drdiMaxRecords :: Lens' DescribeReservedDBInstances (Maybe Int)

The maximum number of records to include in the response. If more than the MaxRecords value is available, a pagination token called a marker is included in the response so that the following results can be retrieved. Default: 100 Constraints: Minimum 20, maximum 100.

drdiMultiAZ :: Lens' DescribeReservedDBInstances (Maybe Bool)

The Multi-AZ filter value. Specify this parameter to show only those reservations matching the specified Multi-AZ parameter.

drdiReservedDBInstancesOfferingId :: Lens' DescribeReservedDBInstances (Maybe Text)

The offering identifier filter value. Specify this parameter to show only purchased reservations matching the specified offering identifier.

drdiOfferingType :: Lens' DescribeReservedDBInstances (Maybe Text)

The offering type filter value. Specify this parameter to show only the available offerings matching the specified offering type. Valid Values: "Partial Upfront" | "All Upfront" | "No Upfront"

drdiDuration :: Lens' DescribeReservedDBInstances (Maybe Text)

The duration filter value, specified in years or seconds. Specify this parameter to show only reservations for this duration. Valid Values: 1 | 3 | 31536000 | 94608000

Destructuring the Response

describeReservedDBInstancesResponse

Arguments

:: Int
drdirsResponseStatus

-> DescribeReservedDBInstancesResponse

Creates a value of DescribeReservedDBInstancesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drdirsReservedDBInstances - A list of reserved DB instances.

drdirsMarker - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

drdirsResponseStatus - -- | The response status code.

data DescribeReservedDBInstancesResponse

Contains the result of a successful invocation of the DescribeReservedDBInstances action.

See: describeReservedDBInstancesResponse smart constructor.

Response Lenses

drdirsReservedDBInstances :: Lens' DescribeReservedDBInstancesResponse [ReservedDBInstance]

A list of reserved DB instances.

drdirsMarker :: Lens' DescribeReservedDBInstancesResponse (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

drdirsResponseStatus :: Lens' DescribeReservedDBInstancesResponse Int

- | The response status code.