Network.AWS.RDS.DescribeReservedDBInstances

Description

Returns information about reserved DB instances for this account, or about a specified reserved DB instance.

http://docs.aws.amazon.com/AmazonRDS/latest/APIReference/API_DescribeReservedDBInstances.html

Request

data DescribeReservedDBInstances

Request constructor

describeReservedDBInstances :: DescribeReservedDBInstances

DescribeReservedDBInstances constructor.

The fields accessible through corresponding lenses are:

drdbiDBInstanceClass :: Maybe Text

drdbiDuration :: Maybe Text

drdbiFilters :: [Filter]

drdbiMarker :: Maybe Text

drdbiMaxRecords :: Maybe Int

drdbiMultiAZ :: Maybe Bool

drdbiOfferingType :: Maybe Text

drdbiProductDescription :: Maybe Text

drdbiReservedDBInstanceId :: Maybe Text

drdbiReservedDBInstancesOfferingId :: Maybe Text

Request lenses

drdbiDBInstanceClass :: Lens' DescribeReservedDBInstances (Maybe Text)

The DB instance class filter value. Specify this parameter to show only those reservations matching the specified DB instances class.

drdbiDuration :: Lens' DescribeReservedDBInstances (Maybe Text)

The duration filter value, specified in years or seconds. Specify this parameter to show only reservations for this duration.

Valid Values: '1 | 3 | 31536000 | 94608000'

drdbiFilters :: Lens' DescribeReservedDBInstances [Filter]

This parameter is not currently supported.

drdbiMarker :: Lens' DescribeReservedDBInstances (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

drdbiMaxRecords :: Lens' DescribeReservedDBInstances (Maybe Int)

The maximum number of records to include in the response. If more than the MaxRecords value is available, a pagination token called a marker is included in the response so that the following results can be retrieved.

Default: 100

Constraints: minimum 20, maximum 100

drdbiMultiAZ :: Lens' DescribeReservedDBInstances (Maybe Bool)

The Multi-AZ filter value. Specify this parameter to show only those reservations matching the specified Multi-AZ parameter.

drdbiOfferingType :: Lens' DescribeReservedDBInstances (Maybe Text)

The offering type filter value. Specify this parameter to show only the available offerings matching the specified offering type.

Valid Values: '"Light Utilization" | "Medium Utilization" | HeavyUtilization '

drdbiProductDescription :: Lens' DescribeReservedDBInstances (Maybe Text)

The product description filter value. Specify this parameter to show only those reservations matching the specified product description.

drdbiReservedDBInstanceId :: Lens' DescribeReservedDBInstances (Maybe Text)

The reserved DB instance identifier filter value. Specify this parameter to show only the reservation that matches the specified reservation ID.

drdbiReservedDBInstancesOfferingId :: Lens' DescribeReservedDBInstances (Maybe Text)

The offering identifier filter value. Specify this parameter to show only purchased reservations matching the specified offering identifier.

Response

data DescribeReservedDBInstancesResponse

Response constructor

describeReservedDBInstancesResponse :: DescribeReservedDBInstancesResponse

DescribeReservedDBInstancesResponse constructor.

The fields accessible through corresponding lenses are:

drdbirMarker :: Maybe Text

drdbirReservedDBInstances :: [ReservedDBInstance]

Response lenses

drdbirMarker :: Lens' DescribeReservedDBInstancesResponse (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

drdbirReservedDBInstances :: Lens' DescribeReservedDBInstancesResponse [ReservedDBInstance]

A list of reserved DB instances.