Network.AWS.RDS.DownloadDBLogFilePortion

Description

Downloads all or a portion of the specified log file, up to 1 MB in size.

This operation returns paginated results.

Creating a Request

downloadDBLogFilePortion

Arguments

:: Text
ddlfpDBInstanceIdentifier

-> Text
ddlfpLogFileName

-> DownloadDBLogFilePortion

Creates a value of DownloadDBLogFilePortion with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddlfpNumberOfLines - The number of lines to download. If the number of lines specified results in a file over 1 MB in size, the file will be truncated at 1 MB in size. If the NumberOfLines parameter is specified, then the block of lines returned can be from the beginning or the end of the log file, depending on the value of the Marker parameter. * If neither Marker or NumberOfLines are specified, the entire log file is returned up to a maximum of 10000 lines, starting with the most recent log entries first. * If NumberOfLines is specified and Marker is not specified, then the most recent lines from the end of the log file are returned. * If Marker is specified as "0", then the specified number of lines from the beginning of the log file are returned. * You can download the log file in blocks of lines by specifying the size of the block using the NumberOfLines parameter, and by specifying a value of "0" for the Marker parameter in your first request. Include the Marker value returned in the response as the Marker value for the next request, continuing until the AdditionalDataPending response element returns false.

ddlfpMarker - The pagination token provided in the previous request or "0". If the Marker parameter is specified the response includes only records beyond the marker until the end of the file or up to NumberOfLines.

ddlfpDBInstanceIdentifier - The customer-assigned name of the DB instance that contains the log files you want to list. Constraints: * Must contain from 1 to 63 alphanumeric characters or hyphens * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens

ddlfpLogFileName - The name of the log file to be downloaded.

data DownloadDBLogFilePortion

See: downloadDBLogFilePortion smart constructor.

Request Lenses

ddlfpNumberOfLines :: Lens' DownloadDBLogFilePortion (Maybe Int)

The number of lines to download. If the number of lines specified results in a file over 1 MB in size, the file will be truncated at 1 MB in size. If the NumberOfLines parameter is specified, then the block of lines returned can be from the beginning or the end of the log file, depending on the value of the Marker parameter. * If neither Marker or NumberOfLines are specified, the entire log file is returned up to a maximum of 10000 lines, starting with the most recent log entries first. * If NumberOfLines is specified and Marker is not specified, then the most recent lines from the end of the log file are returned. * If Marker is specified as "0", then the specified number of lines from the beginning of the log file are returned. * You can download the log file in blocks of lines by specifying the size of the block using the NumberOfLines parameter, and by specifying a value of "0" for the Marker parameter in your first request. Include the Marker value returned in the response as the Marker value for the next request, continuing until the AdditionalDataPending response element returns false.

ddlfpMarker :: Lens' DownloadDBLogFilePortion (Maybe Text)

The pagination token provided in the previous request or "0". If the Marker parameter is specified the response includes only records beyond the marker until the end of the file or up to NumberOfLines.

ddlfpDBInstanceIdentifier :: Lens' DownloadDBLogFilePortion Text

The customer-assigned name of the DB instance that contains the log files you want to list. Constraints: * Must contain from 1 to 63 alphanumeric characters or hyphens * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens

ddlfpLogFileName :: Lens' DownloadDBLogFilePortion Text

The name of the log file to be downloaded.

Destructuring the Response

downloadDBLogFilePortionResponse

Arguments

:: Int
ddlfprsResponseStatus

-> DownloadDBLogFilePortionResponse

Creates a value of DownloadDBLogFilePortionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddlfprsLogFileData - Entries from the specified log file.

ddlfprsAdditionalDataPending - Boolean value that if true, indicates there is more data to be downloaded.

ddlfprsMarker - A pagination token that can be used in a subsequent DownloadDBLogFilePortion request.

ddlfprsResponseStatus - -- | The response status code.

data DownloadDBLogFilePortionResponse

This data type is used as a response element to DownloadDBLogFilePortion .

See: downloadDBLogFilePortionResponse smart constructor.

Response Lenses

ddlfprsLogFileData :: Lens' DownloadDBLogFilePortionResponse (Maybe Text)

Entries from the specified log file.

ddlfprsAdditionalDataPending :: Lens' DownloadDBLogFilePortionResponse (Maybe Bool)

Boolean value that if true, indicates there is more data to be downloaded.

ddlfprsMarker :: Lens' DownloadDBLogFilePortionResponse (Maybe Text)

A pagination token that can be used in a subsequent DownloadDBLogFilePortion request.

ddlfprsResponseStatus :: Lens' DownloadDBLogFilePortionResponse Int

- | The response status code.