Network.AWS.RDS.ModifyDBParameterGroup

Description

Modifies the parameters of a DB parameter group. To modify more than one parameter, submit a list of the following: ParameterName , ParameterValue , and ApplyMethod . A maximum of 20 parameters can be modified in a single request.

Important: After you modify a DB parameter group, you should wait at least 5 minutes before creating your first DB instance that uses that DB parameter group as the default parameter group. This allows Amazon RDS to fully complete the modify action before the parameter group is used as the default for a new DB instance. This is especially important for parameters that are critical when creating the default database for a DB instance, such as the character set for the default database defined by the character_set_database parameter. You can use the Parameter Groups option of the Amazon RDS console or the DescribeDBParameters command to verify that your DB parameter group has been created or modified.

Creating a Request

modifyDBParameterGroup

Arguments

:: Text
mdpgDBParameterGroupName

-> ModifyDBParameterGroup

Creates a value of ModifyDBParameterGroup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mdpgDBParameterGroupName - The name of the DB parameter group. Constraints: * Must be the name of an existing DB parameter group * Must be 1 to 255 alphanumeric characters * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens

mdpgParameters - An array of parameter names, values, and the apply method for the parameter update. At least one parameter name, value, and apply method must be supplied; subsequent arguments are optional. A maximum of 20 parameters can be modified in a single request. Valid Values (for the application method): immediate | pending-reboot

data ModifyDBParameterGroup

See: modifyDBParameterGroup smart constructor.

Request Lenses

mdpgDBParameterGroupName :: Lens' ModifyDBParameterGroup Text

The name of the DB parameter group. Constraints: * Must be the name of an existing DB parameter group * Must be 1 to 255 alphanumeric characters * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens

mdpgParameters :: Lens' ModifyDBParameterGroup [Parameter]

An array of parameter names, values, and the apply method for the parameter update. At least one parameter name, value, and apply method must be supplied; subsequent arguments are optional. A maximum of 20 parameters can be modified in a single request. Valid Values (for the application method): immediate | pending-reboot

Destructuring the Response

dbParameterGroupNameMessage :: DBParameterGroupNameMessage

Creates a value of DBParameterGroupNameMessage with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dpgnmDBParameterGroupName - Provides the name of the DB parameter group.

data DBParameterGroupNameMessage

Contains the result of a successful invocation of the ModifyDBParameterGroup or ResetDBParameterGroup action.

See: dbParameterGroupNameMessage smart constructor.

Response Lenses

dpgnmDBParameterGroupName :: Lens' DBParameterGroupNameMessage (Maybe Text)

Provides the name of the DB parameter group.