Network.AWS.RDS.RebootDBInstance

Description

Rebooting a DB instance restarts the database engine service. A reboot also applies to the DB instance any modifications to the associated DB parameter group that were pending. Rebooting a DB instance results in a momentary outage of the instance, during which the DB instance status is set to rebooting. If the RDS instance is configured for MultiAZ, it is possible that the reboot will be conducted through a failover. An Amazon RDS event is created when the reboot is completed.

If your DB instance is deployed in multiple Availability Zones, you can force a failover from one AZ to the other during the reboot. You might force a failover to test the availability of your DB instance deployment or to restore operations to the original AZ after a failover occurs.

The time required to reboot is a function of the specific database engine's crash recovery process. To improve the reboot time, we recommend that you reduce database activities as much as possible during the reboot process to reduce rollback activity for in-transit transactions.

Creating a Request

rebootDBInstance

Arguments

:: Text
rdiDBInstanceIdentifier

-> RebootDBInstance

Creates a value of RebootDBInstance with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rdiForceFailover - When true , the reboot will be conducted through a MultiAZ failover. Constraint: You cannot specify true if the instance is not configured for MultiAZ.

rdiDBInstanceIdentifier - The DB instance identifier. This parameter is stored as a lowercase string. Constraints: * Must contain from 1 to 63 alphanumeric characters or hyphens * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens

data RebootDBInstance

See: rebootDBInstance smart constructor.

Request Lenses

rdiForceFailover :: Lens' RebootDBInstance (Maybe Bool)

When true , the reboot will be conducted through a MultiAZ failover. Constraint: You cannot specify true if the instance is not configured for MultiAZ.

rdiDBInstanceIdentifier :: Lens' RebootDBInstance Text

The DB instance identifier. This parameter is stored as a lowercase string. Constraints: * Must contain from 1 to 63 alphanumeric characters or hyphens * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens

Destructuring the Response

rebootDBInstanceResponse

Arguments

:: Int
rdirsResponseStatus

-> RebootDBInstanceResponse

Creates a value of RebootDBInstanceResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rdirsDBInstance - Undocumented member.

rdirsResponseStatus - -- | The response status code.

data RebootDBInstanceResponse

See: rebootDBInstanceResponse smart constructor.

Response Lenses

rdirsDBInstance :: Lens' RebootDBInstanceResponse (Maybe DBInstance)

Undocumented member.

rdirsResponseStatus :: Lens' RebootDBInstanceResponse Int

- | The response status code.