Network.AWS.RDS.RemoveSourceIdentifierFromSubscription

Description

Removes a source identifier from an existing RDS event notification subscription.

Creating a Request

removeSourceIdentifierFromSubscription

Arguments

:: Text
rsifsSubscriptionName

-> Text
rsifsSourceIdentifier

-> RemoveSourceIdentifierFromSubscription

Creates a value of RemoveSourceIdentifierFromSubscription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rsifsSubscriptionName - The name of the RDS event notification subscription you want to remove a source identifier from.

rsifsSourceIdentifier - The source identifier to be removed from the subscription, such as the DB instance identifier for a DB instance or the name of a security group.

data RemoveSourceIdentifierFromSubscription

See: removeSourceIdentifierFromSubscription smart constructor.

Request Lenses

rsifsSubscriptionName :: Lens' RemoveSourceIdentifierFromSubscription Text

The name of the RDS event notification subscription you want to remove a source identifier from.

rsifsSourceIdentifier :: Lens' RemoveSourceIdentifierFromSubscription Text

The source identifier to be removed from the subscription, such as the DB instance identifier for a DB instance or the name of a security group.

Destructuring the Response

removeSourceIdentifierFromSubscriptionResponse

Arguments

:: Int
rsifsrsResponseStatus

-> RemoveSourceIdentifierFromSubscriptionResponse

Creates a value of RemoveSourceIdentifierFromSubscriptionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rsifsrsEventSubscription - Undocumented member.

rsifsrsResponseStatus - -- | The response status code.

data RemoveSourceIdentifierFromSubscriptionResponse

See: removeSourceIdentifierFromSubscriptionResponse smart constructor.

Response Lenses

rsifsrsEventSubscription :: Lens' RemoveSourceIdentifierFromSubscriptionResponse (Maybe EventSubscription)

Undocumented member.

rsifsrsResponseStatus :: Lens' RemoveSourceIdentifierFromSubscriptionResponse Int

- | The response status code.