Network.AWS.RDS.RestoreDBInstanceFromDBSnapshot

Description

Creates a new DB instance from a DB snapshot. The target database is created from the source database restore point with the most of original configuration with the default security group and the default DB parameter group. By default, the new DB instance is created as a single-AZ deployment except when the instance is a SQL Server instance that has an option group that is associated with mirroring; in this case, the instance becomes a mirrored AZ deployment and not a single-AZ deployment.

If your intent is to replace your original DB instance with the new, restored DB instance, then rename your original DB instance before you call the RestoreDBInstanceFromDBSnapshot action. RDS does not allow two DB instances with the same name. Once you have renamed your original DB instance with a different identifier, then you can pass the original name of the DB instance as the DBInstanceIdentifier in the call to the RestoreDBInstanceFromDBSnapshot action. The result is that you will replace the original DB instance with the DB instance created from the snapshot.

If you are restoring from a shared manual DB snapshot, the DBSnapshotIdentifier must be the ARN of the shared DB snapshot.

Creating a Request

restoreDBInstanceFromDBSnapshot

Arguments

:: Text
rdifdsDBInstanceIdentifier

-> Text
rdifdsDBSnapshotIdentifier

-> RestoreDBInstanceFromDBSnapshot

Creates a value of RestoreDBInstanceFromDBSnapshot with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rdifdsPubliclyAccessible - Specifies the accessibility options for the DB instance. A value of true specifies an Internet-facing instance with a publicly resolvable DNS name, which resolves to a public IP address. A value of false specifies an internal instance with a DNS name that resolves to a private IP address. Default: The default behavior varies depending on whether a VPC has been requested or not. The following list shows the default behavior in each case. * Default VPC: true * VPC: false If no DB subnet group has been specified as part of the request and the PubliclyAccessible value has not been set, the DB instance will be publicly accessible. If a specific DB subnet group has been specified as part of the request and the PubliclyAccessible value has not been set, the DB instance will be private.

rdifdsAutoMinorVersionUpgrade - Indicates that minor version upgrades will be applied automatically to the DB instance during the maintenance window.

rdifdsDBSubnetGroupName - The DB subnet group name to use for the new instance. Constraints: Must contain no more than 255 alphanumeric characters, periods, underscores, spaces, or hyphens. Must not be default. Example: mySubnetgroup

rdifdsIOPS - Specifies the amount of provisioned IOPS for the DB instance, expressed in I/O operations per second. If this parameter is not specified, the IOPS value will be taken from the backup. If this parameter is set to 0, the new instance will be converted to a non-PIOPS instance, which will take additional time, though your DB instance will be available for connections before the conversion starts. Constraints: Must be an integer greater than 1000. SQL Server Setting the IOPS value for the SQL Server database engine is not supported.

rdifdsDomain - Specify the Active Directory Domain to restore the instance in.

rdifdsEngine - The database engine to use for the new instance. Default: The same as source Constraint: Must be compatible with the engine of the source. You can restore a MariaDB 10.1 DB instance from a MySQL 5.6 snapshot. Valid Values: MySQL | mariadb | oracle-se1 | oracle-se | oracle-ee | sqlserver-ee | sqlserver-se | sqlserver-ex | sqlserver-web | postgres | aurora

rdifdsTDECredentialPassword - The password for the given ARN from the Key Store in order to access the device.

rdifdsDBInstanceClass - The compute and memory capacity of the Amazon RDS DB instance. Valid Values: db.t1.micro | db.m1.small | db.m1.medium | db.m1.large | db.m1.xlarge | db.m2.2xlarge | db.m2.4xlarge | db.m3.medium | db.m3.large | db.m3.xlarge | db.m3.2xlarge | db.m4.large | db.m4.xlarge | db.m4.2xlarge | db.m4.4xlarge | db.m4.10xlarge | db.r3.large | db.r3.xlarge | db.r3.2xlarge | db.r3.4xlarge | db.r3.8xlarge | db.t2.micro | db.t2.small | db.t2.medium | db.t2.large

rdifdsLicenseModel - License model information for the restored DB instance. Default: Same as source. Valid values: license-included | bring-your-own-license | general-public-license

rdifdsAvailabilityZone - The EC2 Availability Zone that the database instance will be created in. Default: A random, system-chosen Availability Zone. Constraint: You cannot specify the AvailabilityZone parameter if the MultiAZ parameter is set to true . Example: us-east-1a

rdifdsMultiAZ - Specifies if the DB instance is a Multi-AZ deployment. Constraint: You cannot specify the AvailabilityZone parameter if the MultiAZ parameter is set to true .

rdifdsOptionGroupName - The name of the option group to be used for the restored DB instance. Permanent options, such as the TDE option for Oracle Advanced Security TDE, cannot be removed from an option group, and that option group cannot be removed from a DB instance once it is associated with a DB instance

rdifdsCopyTagsToSnapshot - True to copy all tags from the restored DB instance to snapshots of the DB instance; otherwise false. The default is false.

rdifdsTDECredentialARN - The ARN from the Key Store with which to associate the instance for TDE encryption.

rdifdsDomainIAMRoleName - Specify the name of the IAM role to be used when making API calls to the Directory Service.

rdifdsTags - Undocumented member.

rdifdsPort - The port number on which the database accepts connections. Default: The same port as the original DB instance Constraints: Value must be 1150-65535

rdifdsEnableIAMDatabaseAuthentication - True to enable mapping of AWS Identity and Access Management (IAM) accounts to database accounts; otherwise false. You can enable IAM database authentication for the following database engines * For MySQL 5.6, minor version 5.6.34 or higher * For MySQL 5.7, minor version 5.7.16 or higher * Aurora 5.6 or higher. Default: false

rdifdsStorageType - Specifies the storage type to be associated with the DB instance. Valid values: standard | gp2 | io1 If you specify io1 , you must also include a value for the Iops parameter. Default: io1 if the Iops parameter is specified; otherwise standard

rdifdsDBName - The database name for the restored DB instance.

rdifdsDBInstanceIdentifier - Name of the DB instance to create from the DB snapshot. This parameter isn't case-sensitive. Constraints: * Must contain from 1 to 63 alphanumeric characters or hyphens * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens Example: my-snapshot-id

rdifdsDBSnapshotIdentifier - The identifier for the DB snapshot to restore from. Constraints: * Must contain from 1 to 255 alphanumeric characters or hyphens * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens If you are restoring from a shared manual DB snapshot, the DBSnapshotIdentifier must be the ARN of the shared DB snapshot.

data RestoreDBInstanceFromDBSnapshot

See: restoreDBInstanceFromDBSnapshot smart constructor.

Request Lenses

rdifdsPubliclyAccessible :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Bool)

Specifies the accessibility options for the DB instance. A value of true specifies an Internet-facing instance with a publicly resolvable DNS name, which resolves to a public IP address. A value of false specifies an internal instance with a DNS name that resolves to a private IP address. Default: The default behavior varies depending on whether a VPC has been requested or not. The following list shows the default behavior in each case. * Default VPC: true * VPC: false If no DB subnet group has been specified as part of the request and the PubliclyAccessible value has not been set, the DB instance will be publicly accessible. If a specific DB subnet group has been specified as part of the request and the PubliclyAccessible value has not been set, the DB instance will be private.

rdifdsAutoMinorVersionUpgrade :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Bool)

Indicates that minor version upgrades will be applied automatically to the DB instance during the maintenance window.

rdifdsDBSubnetGroupName :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Text)

The DB subnet group name to use for the new instance. Constraints: Must contain no more than 255 alphanumeric characters, periods, underscores, spaces, or hyphens. Must not be default. Example: mySubnetgroup

rdifdsIOPS :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Int)

Specifies the amount of provisioned IOPS for the DB instance, expressed in I/O operations per second. If this parameter is not specified, the IOPS value will be taken from the backup. If this parameter is set to 0, the new instance will be converted to a non-PIOPS instance, which will take additional time, though your DB instance will be available for connections before the conversion starts. Constraints: Must be an integer greater than 1000. SQL Server Setting the IOPS value for the SQL Server database engine is not supported.

rdifdsDomain :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Text)

Specify the Active Directory Domain to restore the instance in.

rdifdsEngine :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Text)

The database engine to use for the new instance. Default: The same as source Constraint: Must be compatible with the engine of the source. You can restore a MariaDB 10.1 DB instance from a MySQL 5.6 snapshot. Valid Values: MySQL | mariadb | oracle-se1 | oracle-se | oracle-ee | sqlserver-ee | sqlserver-se | sqlserver-ex | sqlserver-web | postgres | aurora

rdifdsTDECredentialPassword :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Text)

The password for the given ARN from the Key Store in order to access the device.

rdifdsDBInstanceClass :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Text)

The compute and memory capacity of the Amazon RDS DB instance. Valid Values: db.t1.micro | db.m1.small | db.m1.medium | db.m1.large | db.m1.xlarge | db.m2.2xlarge | db.m2.4xlarge | db.m3.medium | db.m3.large | db.m3.xlarge | db.m3.2xlarge | db.m4.large | db.m4.xlarge | db.m4.2xlarge | db.m4.4xlarge | db.m4.10xlarge | db.r3.large | db.r3.xlarge | db.r3.2xlarge | db.r3.4xlarge | db.r3.8xlarge | db.t2.micro | db.t2.small | db.t2.medium | db.t2.large

rdifdsLicenseModel :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Text)

License model information for the restored DB instance. Default: Same as source. Valid values: license-included | bring-your-own-license | general-public-license

rdifdsAvailabilityZone :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Text)

The EC2 Availability Zone that the database instance will be created in. Default: A random, system-chosen Availability Zone. Constraint: You cannot specify the AvailabilityZone parameter if the MultiAZ parameter is set to true . Example: us-east-1a

rdifdsMultiAZ :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Bool)

Specifies if the DB instance is a Multi-AZ deployment. Constraint: You cannot specify the AvailabilityZone parameter if the MultiAZ parameter is set to true .

rdifdsOptionGroupName :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Text)

The name of the option group to be used for the restored DB instance. Permanent options, such as the TDE option for Oracle Advanced Security TDE, cannot be removed from an option group, and that option group cannot be removed from a DB instance once it is associated with a DB instance

rdifdsCopyTagsToSnapshot :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Bool)

True to copy all tags from the restored DB instance to snapshots of the DB instance; otherwise false. The default is false.

rdifdsTDECredentialARN :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Text)

The ARN from the Key Store with which to associate the instance for TDE encryption.

rdifdsDomainIAMRoleName :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Text)

Specify the name of the IAM role to be used when making API calls to the Directory Service.

rdifdsTags :: Lens' RestoreDBInstanceFromDBSnapshot [Tag]

Undocumented member.

rdifdsPort :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Int)

The port number on which the database accepts connections. Default: The same port as the original DB instance Constraints: Value must be 1150-65535

rdifdsEnableIAMDatabaseAuthentication :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Bool)

True to enable mapping of AWS Identity and Access Management (IAM) accounts to database accounts; otherwise false. You can enable IAM database authentication for the following database engines * For MySQL 5.6, minor version 5.6.34 or higher * For MySQL 5.7, minor version 5.7.16 or higher * Aurora 5.6 or higher. Default: false

rdifdsStorageType :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Text)

Specifies the storage type to be associated with the DB instance. Valid values: standard | gp2 | io1 If you specify io1 , you must also include a value for the Iops parameter. Default: io1 if the Iops parameter is specified; otherwise standard

rdifdsDBName :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Text)

The database name for the restored DB instance.

rdifdsDBInstanceIdentifier :: Lens' RestoreDBInstanceFromDBSnapshot Text

Name of the DB instance to create from the DB snapshot. This parameter isn't case-sensitive. Constraints: * Must contain from 1 to 63 alphanumeric characters or hyphens * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens Example: my-snapshot-id

rdifdsDBSnapshotIdentifier :: Lens' RestoreDBInstanceFromDBSnapshot Text

The identifier for the DB snapshot to restore from. Constraints: * Must contain from 1 to 255 alphanumeric characters or hyphens * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens If you are restoring from a shared manual DB snapshot, the DBSnapshotIdentifier must be the ARN of the shared DB snapshot.

Destructuring the Response

restoreDBInstanceFromDBSnapshotResponse

Arguments

:: Int
rdifdsrsResponseStatus

-> RestoreDBInstanceFromDBSnapshotResponse

Creates a value of RestoreDBInstanceFromDBSnapshotResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rdifdsrsDBInstance - Undocumented member.

rdifdsrsResponseStatus - -- | The response status code.

data RestoreDBInstanceFromDBSnapshotResponse

See: restoreDBInstanceFromDBSnapshotResponse smart constructor.

Response Lenses

rdifdsrsDBInstance :: Lens' RestoreDBInstanceFromDBSnapshotResponse (Maybe DBInstance)

Undocumented member.

rdifdsrsResponseStatus :: Lens' RestoreDBInstanceFromDBSnapshotResponse Int

- | The response status code.