Network.AWS.RDS.RestoreDBInstanceFromDBSnapshot

Description

Creates a new DB instance from a DB snapshot. The target database is created from the source database restore point with the same configuration as the original source database, except that the new RDS instance is created with the default security group.

If your intent is to replace your original DB instance with the new, restored DB instance, then rename your original DB instance before you call the RestoreDBInstanceFromDBSnapshot action. RDS does not allow two DB instances with the same name. Once you have renamed your original DB instance with a different identifier, then you can pass the original name of the DB instance as the DBInstanceIdentifier in the call to the RestoreDBInstanceFromDBSnapshot action. The result is that you will replace the original DB instance with the DB instance created from the snapshot.

http://docs.aws.amazon.com/AmazonRDS/latest/APIReference/API_RestoreDBInstanceFromDBSnapshot.html

Request

data RestoreDBInstanceFromDBSnapshot

Request constructor

restoreDBInstanceFromDBSnapshot

Arguments

:: Text
rdbifdbsDBInstanceIdentifier

-> Text
rdbifdbsDBSnapshotIdentifier

-> RestoreDBInstanceFromDBSnapshot

RestoreDBInstanceFromDBSnapshot constructor.

The fields accessible through corresponding lenses are:

rdbifdbsAutoMinorVersionUpgrade :: Maybe Bool

rdbifdbsAvailabilityZone :: Maybe Text

rdbifdbsDBInstanceClass :: Maybe Text

rdbifdbsDBInstanceIdentifier :: Text

rdbifdbsDBName :: Maybe Text

rdbifdbsDBSnapshotIdentifier :: Text

rdbifdbsDBSubnetGroupName :: Maybe Text

rdbifdbsEngine :: Maybe Text

rdbifdbsIops :: Maybe Int

rdbifdbsLicenseModel :: Maybe Text

rdbifdbsMultiAZ :: Maybe Bool

rdbifdbsOptionGroupName :: Maybe Text

rdbifdbsPort :: Maybe Int

rdbifdbsPubliclyAccessible :: Maybe Bool

rdbifdbsStorageType :: Maybe Text

rdbifdbsTags :: [Tag]

rdbifdbsTdeCredentialArn :: Maybe Text

rdbifdbsTdeCredentialPassword :: Maybe Text

Request lenses

rdbifdbsAutoMinorVersionUpgrade :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Bool)

Indicates that minor version upgrades will be applied automatically to the DB instance during the maintenance window.

rdbifdbsAvailabilityZone :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Text)

The EC2 Availability Zone that the database instance will be created in.

Default: A random, system-chosen Availability Zone.

Constraint: You cannot specify the AvailabilityZone parameter if the MultiAZ parameter is set to true.

Example: 'us-east-1a'

rdbifdbsDBInstanceClass :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Text)

The compute and memory capacity of the Amazon RDS DB instance.

Valid Values: 'db.t1.micro | db.m1.small | db.m1.medium | db.m1.large |db.m1.xlarge | db.m2.2xlarge | db.m2.4xlarge | db.m3.medium | db.m3.large |db.m3.xlarge | db.m3.2xlarge | db.r3.large | db.r3.xlarge | db.r3.2xlarge |db.r3.4xlarge | db.r3.8xlarge | db.t2.micro | db.t2.small | db.t2.medium'

rdbifdbsDBInstanceIdentifier :: Lens' RestoreDBInstanceFromDBSnapshot Text

Name of the DB instance to create from the DB snapshot. This parameter isn't case sensitive.

Constraints:

Must contain from 1 to 255 alphanumeric characters or hyphens First character must be a letter Cannot end with a hyphen or contain two consecutive hyphens Example: 'my-snapshot-id'

rdbifdbsDBName :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Text)

The database name for the restored DB instance.

This parameter doesn't apply to the MySQL engine.

rdbifdbsDBSnapshotIdentifier :: Lens' RestoreDBInstanceFromDBSnapshot Text

The identifier for the DB snapshot to restore from.

Constraints:

Must contain from 1 to 63 alphanumeric characters or hyphens First character must be a letter Cannot end with a hyphen or contain two consecutive hyphens

rdbifdbsDBSubnetGroupName :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Text)

The DB subnet group name to use for the new instance.

rdbifdbsEngine :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Text)

The database engine to use for the new instance.

Default: The same as source

Constraint: Must be compatible with the engine of the source

Valid Values: MySQL | 'oracle-se1' | 'oracle-se' | 'oracle-ee' | 'sqlserver-ee' | 'sqlserver-se' | 'sqlserver-ex' | 'sqlserver-web' | postgres

rdbifdbsIops :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Int)

Specifies the amount of provisioned IOPS for the DB instance, expressed in I/O operations per second. If this parameter is not specified, the IOPS value will be taken from the backup. If this parameter is set to 0, the new instance will be converted to a non-PIOPS instance, which will take additional time, though your DB instance will be available for connections before the conversion starts.

Constraints: Must be an integer greater than 1000.

SQL Server

Setting the IOPS value for the SQL Server database engine is not supported.

rdbifdbsLicenseModel :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Text)

License model information for the restored DB instance.

Default: Same as source.

Valid values: 'license-included' | 'bring-your-own-license' | 'general-public-license'

rdbifdbsMultiAZ :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Bool)

Specifies if the DB instance is a Multi-AZ deployment.

Constraint: You cannot specify the AvailabilityZone parameter if the MultiAZ parameter is set to true.

rdbifdbsOptionGroupName :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Text)

The name of the option group to be used for the restored DB instance.

Permanent options, such as the TDE option for Oracle Advanced Security TDE, cannot be removed from an option group, and that option group cannot be removed from a DB instance once it is associated with a DB instance

rdbifdbsPort :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Int)

The port number on which the database accepts connections.

Default: The same port as the original DB instance

Constraints: Value must be '1150-65535'

rdbifdbsPubliclyAccessible :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Bool)

Specifies the accessibility options for the DB instance. A value of true specifies an Internet-facing instance with a publicly resolvable DNS name, which resolves to a public IP address. A value of false specifies an internal instance with a DNS name that resolves to a private IP address.

Default: The default behavior varies depending on whether a VPC has been requested or not. The following list shows the default behavior in each case.

Default VPC: true VPC: false If no DB subnet group has been specified as part of the request and the PubliclyAccessible value has not been set, the DB instance will be publicly accessible. If a specific DB subnet group has been specified as part of the request and the PubliclyAccessible value has not been set, the DB instance will be private.

rdbifdbsStorageType :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Text)

Specifies the storage type to be associated with the DB instance.

Valid values: 'standard | gp2 | io1'

If you specify io1, you must also include a value for the Iops parameter.

Default: io1 if the Iops parameter is specified; otherwise standard

rdbifdbsTags :: Lens' RestoreDBInstanceFromDBSnapshot [Tag]

rdbifdbsTdeCredentialArn :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Text)

The ARN from the Key Store with which to associate the instance for TDE encryption.

rdbifdbsTdeCredentialPassword :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Text)

The password for the given ARN from the Key Store in order to access the device.

Response

data RestoreDBInstanceFromDBSnapshotResponse

Response constructor

restoreDBInstanceFromDBSnapshotResponse :: RestoreDBInstanceFromDBSnapshotResponse

RestoreDBInstanceFromDBSnapshotResponse constructor.

The fields accessible through corresponding lenses are:

rdbifdbsrDBInstance :: Maybe DBInstance

Response lenses

rdbifdbsrDBInstance :: Lens' RestoreDBInstanceFromDBSnapshotResponse (Maybe DBInstance)