Network.AWS.RDS.RestoreDBInstanceToPointInTime

Description

Restores a DB instance to an arbitrary point in time. You can restore to any point in time before the time identified by the LatestRestorableTime property. You can restore to a point up to the number of days specified by the BackupRetentionPeriod property.

The target database is created with most of the original configuration, but in a system-selected availability zone, with the default security group, the default subnet group, and the default DB parameter group. By default, the new DB instance is created as a single-AZ deployment except when the instance is a SQL Server instance that has an option group that is associated with mirroring; in this case, the instance becomes a mirrored deployment and not a single-AZ deployment.

Creating a Request

restoreDBInstanceToPointInTime

Arguments

:: Text
rditpitSourceDBInstanceIdentifier

-> Text
rditpitTargetDBInstanceIdentifier

-> RestoreDBInstanceToPointInTime

Creates a value of RestoreDBInstanceToPointInTime with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rditpitUseLatestRestorableTime - Specifies whether (true ) or not (false ) the DB instance is restored from the latest backup time. Default: false Constraints: Cannot be specified if RestoreTime parameter is provided.

rditpitPubliclyAccessible - Specifies the accessibility options for the DB instance. A value of true specifies an Internet-facing instance with a publicly resolvable DNS name, which resolves to a public IP address. A value of false specifies an internal instance with a DNS name that resolves to a private IP address. Default: The default behavior varies depending on whether a VPC has been requested or not. The following list shows the default behavior in each case. * Default VPC: true * VPC: false If no DB subnet group has been specified as part of the request and the PubliclyAccessible value has not been set, the DB instance will be publicly accessible. If a specific DB subnet group has been specified as part of the request and the PubliclyAccessible value has not been set, the DB instance will be private.

rditpitAutoMinorVersionUpgrade - Indicates that minor version upgrades will be applied automatically to the DB instance during the maintenance window.

rditpitDBSubnetGroupName - The DB subnet group name to use for the new instance. Constraints: Must contain no more than 255 alphanumeric characters, periods, underscores, spaces, or hyphens. Must not be default. Example: mySubnetgroup

rditpitRestoreTime - The date and time to restore from. Valid Values: Value must be a time in Universal Coordinated Time (UTC) format Constraints: * Must be before the latest restorable time for the DB instance * Cannot be specified if UseLatestRestorableTime parameter is true Example: 2009-09-07T23:45:00Z

rditpitIOPS - The amount of Provisioned IOPS (input/output operations per second) to be initially allocated for the DB instance. Constraints: Must be an integer greater than 1000. SQL Server Setting the IOPS value for the SQL Server database engine is not supported.

rditpitDomain - Specify the Active Directory Domain to restore the instance in.

rditpitEngine - The database engine to use for the new instance. Default: The same as source Constraint: Must be compatible with the engine of the source Valid Values: MySQL | mariadb | oracle-se1 | oracle-se | oracle-ee | sqlserver-ee | sqlserver-se | sqlserver-ex | sqlserver-web | postgres | aurora

rditpitTDECredentialPassword - The password for the given ARN from the Key Store in order to access the device.

rditpitDBInstanceClass - The compute and memory capacity of the Amazon RDS DB instance. Valid Values: db.t1.micro | db.m1.small | db.m1.medium | db.m1.large | db.m1.xlarge | db.m2.2xlarge | db.m2.4xlarge | db.m3.medium | db.m3.large | db.m3.xlarge | db.m3.2xlarge | db.m4.large | db.m4.xlarge | db.m4.2xlarge | db.m4.4xlarge | db.m4.10xlarge | db.r3.large | db.r3.xlarge | db.r3.2xlarge | db.r3.4xlarge | db.r3.8xlarge | db.t2.micro | db.t2.small | db.t2.medium | db.t2.large Default: The same DBInstanceClass as the original DB instance.

rditpitLicenseModel - License model information for the restored DB instance. Default: Same as source. Valid values: license-included | bring-your-own-license | general-public-license

rditpitAvailabilityZone - The EC2 Availability Zone that the database instance will be created in. Default: A random, system-chosen Availability Zone. Constraint: You cannot specify the AvailabilityZone parameter if the MultiAZ parameter is set to true. Example: us-east-1a

rditpitMultiAZ - Specifies if the DB instance is a Multi-AZ deployment. Constraint: You cannot specify the AvailabilityZone parameter if the MultiAZ parameter is set to true .

rditpitOptionGroupName - The name of the option group to be used for the restored DB instance. Permanent options, such as the TDE option for Oracle Advanced Security TDE, cannot be removed from an option group, and that option group cannot be removed from a DB instance once it is associated with a DB instance

rditpitCopyTagsToSnapshot - True to copy all tags from the restored DB instance to snapshots of the DB instance; otherwise false. The default is false.

rditpitTDECredentialARN - The ARN from the Key Store with which to associate the instance for TDE encryption.

rditpitDomainIAMRoleName - Specify the name of the IAM role to be used when making API calls to the Directory Service.

rditpitTags - Undocumented member.

rditpitPort - The port number on which the database accepts connections. Constraints: Value must be 1150-65535 Default: The same port as the original DB instance.

rditpitEnableIAMDatabaseAuthentication - True to enable mapping of AWS Identity and Access Management (IAM) accounts to database accounts; otherwise false. You can enable IAM database authentication for the following database engines * For MySQL 5.6, minor version 5.6.34 or higher * For MySQL 5.7, minor version 5.7.16 or higher * Aurora 5.6 or higher. Default: false

rditpitStorageType - Specifies the storage type to be associated with the DB instance. Valid values: standard | gp2 | io1 If you specify io1 , you must also include a value for the Iops parameter. Default: io1 if the Iops parameter is specified; otherwise standard

rditpitDBName - The database name for the restored DB instance.

rditpitSourceDBInstanceIdentifier - The identifier of the source DB instance from which to restore. Constraints: * Must be the identifier of an existing database instance * Must contain from 1 to 63 alphanumeric characters or hyphens * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens

rditpitTargetDBInstanceIdentifier - The name of the new database instance to be created. Constraints: * Must contain from 1 to 63 alphanumeric characters or hyphens * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens

data RestoreDBInstanceToPointInTime

See: restoreDBInstanceToPointInTime smart constructor.

Request Lenses

rditpitUseLatestRestorableTime :: Lens' RestoreDBInstanceToPointInTime (Maybe Bool)

Specifies whether (true ) or not (false ) the DB instance is restored from the latest backup time. Default: false Constraints: Cannot be specified if RestoreTime parameter is provided.

rditpitPubliclyAccessible :: Lens' RestoreDBInstanceToPointInTime (Maybe Bool)

Specifies the accessibility options for the DB instance. A value of true specifies an Internet-facing instance with a publicly resolvable DNS name, which resolves to a public IP address. A value of false specifies an internal instance with a DNS name that resolves to a private IP address. Default: The default behavior varies depending on whether a VPC has been requested or not. The following list shows the default behavior in each case. * Default VPC: true * VPC: false If no DB subnet group has been specified as part of the request and the PubliclyAccessible value has not been set, the DB instance will be publicly accessible. If a specific DB subnet group has been specified as part of the request and the PubliclyAccessible value has not been set, the DB instance will be private.

rditpitAutoMinorVersionUpgrade :: Lens' RestoreDBInstanceToPointInTime (Maybe Bool)

Indicates that minor version upgrades will be applied automatically to the DB instance during the maintenance window.

rditpitDBSubnetGroupName :: Lens' RestoreDBInstanceToPointInTime (Maybe Text)

The DB subnet group name to use for the new instance. Constraints: Must contain no more than 255 alphanumeric characters, periods, underscores, spaces, or hyphens. Must not be default. Example: mySubnetgroup

rditpitRestoreTime :: Lens' RestoreDBInstanceToPointInTime (Maybe UTCTime)

The date and time to restore from. Valid Values: Value must be a time in Universal Coordinated Time (UTC) format Constraints: * Must be before the latest restorable time for the DB instance * Cannot be specified if UseLatestRestorableTime parameter is true Example: 2009-09-07T23:45:00Z

rditpitIOPS :: Lens' RestoreDBInstanceToPointInTime (Maybe Int)

The amount of Provisioned IOPS (input/output operations per second) to be initially allocated for the DB instance. Constraints: Must be an integer greater than 1000. SQL Server Setting the IOPS value for the SQL Server database engine is not supported.

rditpitDomain :: Lens' RestoreDBInstanceToPointInTime (Maybe Text)

Specify the Active Directory Domain to restore the instance in.

rditpitEngine :: Lens' RestoreDBInstanceToPointInTime (Maybe Text)

The database engine to use for the new instance. Default: The same as source Constraint: Must be compatible with the engine of the source Valid Values: MySQL | mariadb | oracle-se1 | oracle-se | oracle-ee | sqlserver-ee | sqlserver-se | sqlserver-ex | sqlserver-web | postgres | aurora

rditpitTDECredentialPassword :: Lens' RestoreDBInstanceToPointInTime (Maybe Text)

The password for the given ARN from the Key Store in order to access the device.

rditpitDBInstanceClass :: Lens' RestoreDBInstanceToPointInTime (Maybe Text)

The compute and memory capacity of the Amazon RDS DB instance. Valid Values: db.t1.micro | db.m1.small | db.m1.medium | db.m1.large | db.m1.xlarge | db.m2.2xlarge | db.m2.4xlarge | db.m3.medium | db.m3.large | db.m3.xlarge | db.m3.2xlarge | db.m4.large | db.m4.xlarge | db.m4.2xlarge | db.m4.4xlarge | db.m4.10xlarge | db.r3.large | db.r3.xlarge | db.r3.2xlarge | db.r3.4xlarge | db.r3.8xlarge | db.t2.micro | db.t2.small | db.t2.medium | db.t2.large Default: The same DBInstanceClass as the original DB instance.

rditpitLicenseModel :: Lens' RestoreDBInstanceToPointInTime (Maybe Text)

License model information for the restored DB instance. Default: Same as source. Valid values: license-included | bring-your-own-license | general-public-license

rditpitAvailabilityZone :: Lens' RestoreDBInstanceToPointInTime (Maybe Text)

The EC2 Availability Zone that the database instance will be created in. Default: A random, system-chosen Availability Zone. Constraint: You cannot specify the AvailabilityZone parameter if the MultiAZ parameter is set to true. Example: us-east-1a

rditpitMultiAZ :: Lens' RestoreDBInstanceToPointInTime (Maybe Bool)

Specifies if the DB instance is a Multi-AZ deployment. Constraint: You cannot specify the AvailabilityZone parameter if the MultiAZ parameter is set to true .

rditpitOptionGroupName :: Lens' RestoreDBInstanceToPointInTime (Maybe Text)

The name of the option group to be used for the restored DB instance. Permanent options, such as the TDE option for Oracle Advanced Security TDE, cannot be removed from an option group, and that option group cannot be removed from a DB instance once it is associated with a DB instance

rditpitCopyTagsToSnapshot :: Lens' RestoreDBInstanceToPointInTime (Maybe Bool)

True to copy all tags from the restored DB instance to snapshots of the DB instance; otherwise false. The default is false.

rditpitTDECredentialARN :: Lens' RestoreDBInstanceToPointInTime (Maybe Text)

The ARN from the Key Store with which to associate the instance for TDE encryption.

rditpitDomainIAMRoleName :: Lens' RestoreDBInstanceToPointInTime (Maybe Text)

Specify the name of the IAM role to be used when making API calls to the Directory Service.

rditpitTags :: Lens' RestoreDBInstanceToPointInTime [Tag]

Undocumented member.

rditpitPort :: Lens' RestoreDBInstanceToPointInTime (Maybe Int)

The port number on which the database accepts connections. Constraints: Value must be 1150-65535 Default: The same port as the original DB instance.

rditpitEnableIAMDatabaseAuthentication :: Lens' RestoreDBInstanceToPointInTime (Maybe Bool)

True to enable mapping of AWS Identity and Access Management (IAM) accounts to database accounts; otherwise false. You can enable IAM database authentication for the following database engines * For MySQL 5.6, minor version 5.6.34 or higher * For MySQL 5.7, minor version 5.7.16 or higher * Aurora 5.6 or higher. Default: false

rditpitStorageType :: Lens' RestoreDBInstanceToPointInTime (Maybe Text)

Specifies the storage type to be associated with the DB instance. Valid values: standard | gp2 | io1 If you specify io1 , you must also include a value for the Iops parameter. Default: io1 if the Iops parameter is specified; otherwise standard

rditpitDBName :: Lens' RestoreDBInstanceToPointInTime (Maybe Text)

The database name for the restored DB instance.

rditpitSourceDBInstanceIdentifier :: Lens' RestoreDBInstanceToPointInTime Text

The identifier of the source DB instance from which to restore. Constraints: * Must be the identifier of an existing database instance * Must contain from 1 to 63 alphanumeric characters or hyphens * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens

rditpitTargetDBInstanceIdentifier :: Lens' RestoreDBInstanceToPointInTime Text

The name of the new database instance to be created. Constraints: * Must contain from 1 to 63 alphanumeric characters or hyphens * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens

Destructuring the Response

restoreDBInstanceToPointInTimeResponse

Arguments

:: Int
rditpitrsResponseStatus

-> RestoreDBInstanceToPointInTimeResponse

Creates a value of RestoreDBInstanceToPointInTimeResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rditpitrsDBInstance - Undocumented member.

rditpitrsResponseStatus - -- | The response status code.

data RestoreDBInstanceToPointInTimeResponse

See: restoreDBInstanceToPointInTimeResponse smart constructor.

Response Lenses

rditpitrsDBInstance :: Lens' RestoreDBInstanceToPointInTimeResponse (Maybe DBInstance)

Undocumented member.

rditpitrsResponseStatus :: Lens' RestoreDBInstanceToPointInTimeResponse Int

- | The response status code.