Network.AWS.RDS.RestoreDBInstanceToPointInTime

Description

Restores a DB instance to an arbitrary point-in-time. Users can restore to any point in time before the LatestRestorableTime for up to BackupRetentionPeriod days. The target database is created from the source database with the same configuration as the original database except that the DB instance is created with the default DB security group.

http://docs.aws.amazon.com/AmazonRDS/latest/APIReference/API_RestoreDBInstanceToPointInTime.html

Request

data RestoreDBInstanceToPointInTime

Request constructor

restoreDBInstanceToPointInTime

Arguments

:: Text
rdbitpitSourceDBInstanceIdentifier

-> Text
rdbitpitTargetDBInstanceIdentifier

-> RestoreDBInstanceToPointInTime

RestoreDBInstanceToPointInTime constructor.

The fields accessible through corresponding lenses are:

rdbitpitAutoMinorVersionUpgrade :: Maybe Bool

rdbitpitAvailabilityZone :: Maybe Text

rdbitpitDBInstanceClass :: Maybe Text

rdbitpitDBName :: Maybe Text

rdbitpitDBSubnetGroupName :: Maybe Text

rdbitpitEngine :: Maybe Text

rdbitpitIops :: Maybe Int

rdbitpitLicenseModel :: Maybe Text

rdbitpitMultiAZ :: Maybe Bool

rdbitpitOptionGroupName :: Maybe Text

rdbitpitPort :: Maybe Int

rdbitpitPubliclyAccessible :: Maybe Bool

rdbitpitRestoreTime :: Maybe UTCTime

rdbitpitSourceDBInstanceIdentifier :: Text

rdbitpitStorageType :: Maybe Text

rdbitpitTags :: [Tag]

rdbitpitTargetDBInstanceIdentifier :: Text

rdbitpitTdeCredentialArn :: Maybe Text

rdbitpitTdeCredentialPassword :: Maybe Text

rdbitpitUseLatestRestorableTime :: Maybe Bool

Request lenses

rdbitpitAutoMinorVersionUpgrade :: Lens' RestoreDBInstanceToPointInTime (Maybe Bool)

Indicates that minor version upgrades will be applied automatically to the DB instance during the maintenance window.

rdbitpitAvailabilityZone :: Lens' RestoreDBInstanceToPointInTime (Maybe Text)

The EC2 Availability Zone that the database instance will be created in.

Default: A random, system-chosen Availability Zone.

Constraint: You cannot specify the AvailabilityZone parameter if the MultiAZ parameter is set to true.

Example: 'us-east-1a'

rdbitpitDBInstanceClass :: Lens' RestoreDBInstanceToPointInTime (Maybe Text)

The compute and memory capacity of the Amazon RDS DB instance.

Valid Values: 'db.t1.micro | db.m1.small | db.m1.medium | db.m1.large |db.m1.xlarge | db.m2.2xlarge | db.m2.4xlarge | db.m3.medium | db.m3.large |db.m3.xlarge | db.m3.2xlarge | db.r3.large | db.r3.xlarge | db.r3.2xlarge |db.r3.4xlarge | db.r3.8xlarge | db.t2.micro | db.t2.small | db.t2.medium'

Default: The same DBInstanceClass as the original DB instance.

rdbitpitDBName :: Lens' RestoreDBInstanceToPointInTime (Maybe Text)

The database name for the restored DB instance.

This parameter is not used for the MySQL engine.

rdbitpitDBSubnetGroupName :: Lens' RestoreDBInstanceToPointInTime (Maybe Text)

The DB subnet group name to use for the new instance.

rdbitpitEngine :: Lens' RestoreDBInstanceToPointInTime (Maybe Text)

The database engine to use for the new instance.

Default: The same as source

Constraint: Must be compatible with the engine of the source

Valid Values: MySQL | 'oracle-se1' | 'oracle-se' | 'oracle-ee' | 'sqlserver-ee' | 'sqlserver-se' | 'sqlserver-ex' | 'sqlserver-web' | postgres

rdbitpitIops :: Lens' RestoreDBInstanceToPointInTime (Maybe Int)

The amount of Provisioned IOPS (input/output operations per second) to be initially allocated for the DB instance.

Constraints: Must be an integer greater than 1000.

SQL Server

Setting the IOPS value for the SQL Server database engine is not supported.

rdbitpitLicenseModel :: Lens' RestoreDBInstanceToPointInTime (Maybe Text)

License model information for the restored DB instance.

Default: Same as source.

Valid values: 'license-included' | 'bring-your-own-license' | 'general-public-license'

rdbitpitMultiAZ :: Lens' RestoreDBInstanceToPointInTime (Maybe Bool)

Specifies if the DB instance is a Multi-AZ deployment.

Constraint: You cannot specify the AvailabilityZone parameter if the MultiAZ parameter is set to true.

rdbitpitOptionGroupName :: Lens' RestoreDBInstanceToPointInTime (Maybe Text)

The name of the option group to be used for the restored DB instance.

Permanent options, such as the TDE option for Oracle Advanced Security TDE, cannot be removed from an option group, and that option group cannot be removed from a DB instance once it is associated with a DB instance

rdbitpitPort :: Lens' RestoreDBInstanceToPointInTime (Maybe Int)

The port number on which the database accepts connections.

Constraints: Value must be '1150-65535'

Default: The same port as the original DB instance.

rdbitpitPubliclyAccessible :: Lens' RestoreDBInstanceToPointInTime (Maybe Bool)

Specifies the accessibility options for the DB instance. A value of true specifies an Internet-facing instance with a publicly resolvable DNS name, which resolves to a public IP address. A value of false specifies an internal instance with a DNS name that resolves to a private IP address.

Default: The default behavior varies depending on whether a VPC has been requested or not. The following list shows the default behavior in each case.

Default VPC:true VPC:false If no DB subnet group has been specified as part of the request and the PubliclyAccessible value has not been set, the DB instance will be publicly accessible. If a specific DB subnet group has been specified as part of the request and the PubliclyAccessible value has not been set, the DB instance will be private.

rdbitpitRestoreTime :: Lens' RestoreDBInstanceToPointInTime (Maybe UTCTime)

The date and time to restore from.

Valid Values: Value must be a UTC time

Constraints:

Must be before the latest restorable time for the DB instance Cannot be specified if UseLatestRestorableTime parameter is true Example: '2009-09-07T23:45:00Z'

rdbitpitSourceDBInstanceIdentifier :: Lens' RestoreDBInstanceToPointInTime Text

The identifier of the source DB instance from which to restore.

Constraints:

Must be the identifier of an existing database instance Must contain from 1 to 63 alphanumeric characters or hyphens First character must be a letter Cannot end with a hyphen or contain two consecutive hyphens

rdbitpitStorageType :: Lens' RestoreDBInstanceToPointInTime (Maybe Text)

Specifies the storage type to be associated with the DB instance.

Valid values: 'standard | gp2 | io1'

If you specify io1, you must also include a value for the Iops parameter.

Default: io1 if the Iops parameter is specified; otherwise standard

rdbitpitTags :: Lens' RestoreDBInstanceToPointInTime [Tag]

rdbitpitTargetDBInstanceIdentifier :: Lens' RestoreDBInstanceToPointInTime Text

The name of the new database instance to be created.

Constraints:

Must contain from 1 to 63 alphanumeric characters or hyphens First character must be a letter Cannot end with a hyphen or contain two consecutive hyphens

rdbitpitTdeCredentialArn :: Lens' RestoreDBInstanceToPointInTime (Maybe Text)

The ARN from the Key Store with which to associate the instance for TDE encryption.

rdbitpitTdeCredentialPassword :: Lens' RestoreDBInstanceToPointInTime (Maybe Text)

The password for the given ARN from the Key Store in order to access the device.

rdbitpitUseLatestRestorableTime :: Lens' RestoreDBInstanceToPointInTime (Maybe Bool)

Specifies whether (true) or not (false) the DB instance is restored from the latest backup time.

Default: false

Constraints: Cannot be specified if RestoreTime parameter is provided.

Response

data RestoreDBInstanceToPointInTimeResponse

Response constructor

restoreDBInstanceToPointInTimeResponse :: RestoreDBInstanceToPointInTimeResponse

RestoreDBInstanceToPointInTimeResponse constructor.

The fields accessible through corresponding lenses are:

rdbitpitrDBInstance :: Maybe DBInstance

Response lenses

rdbitpitrDBInstance :: Lens' RestoreDBInstanceToPointInTimeResponse (Maybe DBInstance)