Network.AWS.RDS.V2013_09_09.CreateDBSnapshot

Description

Creates a DBSnapshot. The source DBInstance must be in "available" state. https:/rds.amazonaws.com ?Action=CreateDBSnapshot &DBInstanceIdentifier=simcoprod01 &DBSnapshotIdentifier=mydbsnapshot &Version=2013-05-15 &SignatureVersion=2&SignatureMethod=HmacSHA256 &Timestamp=2011-05-23T06%3A27%3A42.551Z &AWSAccessKeyId= &Signature= 3306 mysql creating us-east-1a general-public-license 2011-05-23T06:06:43.110Z 10 simcoprod01 5.1.50 mydbsnapshot manual master c4181d1d-8505-11e0-90aa-eb648410240d.

Documentation

createDBSnapshot

Arguments

:: Text
_cdbsnDBSnapshotIdentifier

-> Text
_cdbsnDBInstanceIdentifier

-> CreateDBSnapshot

Minimum specification for a CreateDBSnapshot request.

data CreateDBSnapshot

Constructors

CreateDBSnapshot

Fields

_cdbsnDBSnapshotIdentifier :: Text

The identifier for the DB snapshot. Constraints: Cannot be null, empty, or blank Must contain from 1 to 255 alphanumeric characters or hyphens First character must be a letter Cannot end with a hyphen or contain two consecutive hyphens Example: my-snapshot-id.

_cdbsnDBInstanceIdentifier :: Text

The DB instance identifier. This is the unique key that identifies a DB instance. This parameter isn't case sensitive. Constraints: Must contain from 1 to 63 alphanumeric characters or hyphens First character must be a letter Cannot end with a hyphen or contain two consecutive hyphens.

_cdbsnTags :: [Tag]

A list of tags.

cdbsnTags :: Lens' CreateDBSnapshot [Tag]

cdbsnDBSnapshotIdentifier :: Lens' CreateDBSnapshot Text

cdbsnDBInstanceIdentifier :: Lens' CreateDBSnapshot Text

data CreateDBSnapshotResponse

Constructors

CreateDBSnapshotResponse

Fields

_dbsxDBSnapshot :: Maybe DBSnapshot

Contains the result of a successful invocation of the following actions: CreateDBSnapshot DeleteDBSnapshot This data type is used as a response element in the DescribeDBSnapshots action.

dbsxDBSnapshot :: Iso' CreateDBSnapshotResponse (Maybe DBSnapshot)