Network.AWS.RDS.V2013_09_09.DescribeDBParameters

Description

Returns the detailed parameter list for a particular DB parameter group. https:/rds.amazonaws.com ?Action=DescribeDBParameters &DBParameterGroupName=mydbparametergroup &Source=system &MaxRecords=100 &Version=2013-05-15 &SignatureVersion=2 &SignatureMethod=HmacSHA256 &Timestamp=2011-05-11T19%3A31%3A42.262Z &AWSAccessKeyId= &Signature= rdsdbbinmysql string system false The MySQL installation base directory. static basedir 32768 integer system true The size of the cache to hold the SQL statements for the binary log during a transaction. dynamic 4096-9223372036854775807 binlog_cache_size 8743f2cf-bf41-11de-8c8e-49155882c409.

Request

data DescribeDBParameters

Request constructor

mkDescribeDBParameters

Arguments

:: Text
ddbpDBParameterGroupName

-> DescribeDBParameters

Smart constructor for the minimum required parameters to construct a valid DescribeDBParameters request.

Request lenses

ddbpDBParameterGroupName :: Lens' DescribeDBParameters Text

The name of a specific DB parameter group to return details for. Constraints: Must be 1 to 255 alphanumeric characters First character must be a letter Cannot end with a hyphen or contain two consecutive hyphens.

ddbpSource :: Lens' DescribeDBParameters (Maybe Text)

The parameter types to return. Default: All parameter types returned Valid Values: user | system | engine-default.

ddbpMaxRecords :: Lens' DescribeDBParameters (Maybe Integer)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results may be retrieved. Default: 100 Constraints: minimum 20, maximum 100.

ddbpMarker :: Lens' DescribeDBParameters (Maybe Text)

An optional pagination token provided by a previous DescribeDBParameters request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

Response

data DescribeDBParametersResponse

Contains the result of a successful invocation of the DescribeDBParameters action.

Response lenses

ddbprsParameters :: Lens' DescribeDBParametersResponse [Parameter]

A list of Parameter values.

ddbprsMarker :: Lens' DescribeDBParametersResponse (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.