Network.AWS.RDS.V2013_09_09.DescribeEngineDefaultParameters

Description

Returns the default engine and system parameter information for the specified database engine. https:/rds.amazonaws.com ?Action=DescribeEngineDefaultParameters &DBParameterGroupFamily=mysql5.1 &Version=2013-05-15 &MaxRecords=100 &SignatureVersion=2 &SignatureMethod=HmacSHA256 &Timestamp=2011-02-15T19%3A10%3A03.510Z &AWSAccessKeyId= &Signature= bG93ZXJfY2FzZV90YWJsZV9uYW1lcw== mysql5.1 boolean engine-default false Controls whether user-defined functions that have only an xxx symbol for the main function can be loaded static 0,1 allow-suspicious-udfs integer engine-default true Intended for use with master-to-master replication, and can be used to control the operation of AUTO_INCREMENT columns dynamic 1-65535 auto_increment_increment integer engine-default true Determines the starting point for the AUTO_INCREMENT column value dynamic 1-65535 auto_increment_offset 6c1341eb-a124-11df-bf5c-973b09643c5d.

Request

data DescribeEngineDefaultParameters

Request constructor

mkDescribeEngineDefaultParameters

Arguments

:: Text
dedpDBParameterGroupFamily

-> DescribeEngineDefaultParameters

Smart constructor for the minimum required parameters to construct a valid DescribeEngineDefaultParameters request.

Request lenses

dedpDBParameterGroupFamily :: Lens' DescribeEngineDefaultParameters Text

The name of the DB parameter group family.

dedpMaxRecords :: Lens' DescribeEngineDefaultParameters (Maybe Integer)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results may be retrieved. Default: 100 Constraints: minimum 20, maximum 100.

dedpMarker :: Lens' DescribeEngineDefaultParameters (Maybe Text)

An optional pagination token provided by a previous DescribeEngineDefaultParameters request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

Response

data DescribeEngineDefaultParametersResponse

Response constructor

mkDescribeEngineDefaultParametersResponse

Arguments

:: EngineDefaults
dedprEngineDefaults

-> DescribeEngineDefaultParametersResponse

Smart constructor for the minimum required parameters to construct a valid DescribeEngineDefaultParametersResponse response.

This constructor is provided for convenience and testing purposes.

Response lenses

dedprEngineDefaults :: Lens' DescribeEngineDefaultParametersResponse EngineDefaults

Contains the result of a successful invocation of the DescribeEngineDefaultParameters action.