Network.AWS.RDS.V2013_09_09.DescribeReservedDBInstancesOfferings

Description

Lists available reserved DB instance offerings. https:/rds.amazonaws.com ?Action=DescribeReservedDBInstancesOfferings &ReservedDBInstancesOfferingId=438012d3-4052-4cc7-b2e3-8d3372e0e706 &SignatureVersion=2 &SignatureMethod=HmacSHA256 &Timestamp=2011-12-18T18%3A31%3A36.118Z &AWSAccessKeyId= &Signature= c2012-04-02"> 31536000 Heavy Utilization USD Hourly 0.123 162.0 mysql 0.0 false SampleOfferingId db.m1.small 521b420a-2961-11e1-bd06-6fe008f046c3.

Request

data DescribeReservedDBInstancesOfferings

Request constructor

mkDescribeReservedDBInstancesOfferings :: DescribeReservedDBInstancesOfferings

Smart constructor for the minimum required parameters to construct a valid DescribeReservedDBInstancesOfferings request.

The fields accessible through corresponding lenses are:

ReservedDBInstancesOfferingId :: Maybe Text

DBInstanceClass :: Maybe Text

Duration :: Maybe Text

ProductDescription :: Maybe Text

OfferingType :: Maybe Text

MultiAZ :: Maybe Bool

MaxRecords :: Maybe Integer

Marker :: Maybe Text

Request lenses

drdbioReservedDBInstancesOfferingId :: Lens' DescribeReservedDBInstancesOfferings (Maybe Text)

The offering identifier filter value. Specify this parameter to show only the available offering that matches the specified reservation identifier. Example: 438012d3-4052-4cc7-b2e3-8d3372e0e706.

drdbioDBInstanceClass :: Lens' DescribeReservedDBInstancesOfferings (Maybe Text)

The DB instance class filter value. Specify this parameter to show only the available offerings matching the specified DB instance class.

drdbioDuration :: Lens' DescribeReservedDBInstancesOfferings (Maybe Text)

Duration filter value, specified in years or seconds. Specify this parameter to show only reservations for this duration. Valid Values: 1 | 3 | 31536000 | 94608000.

drdbioProductDescription :: Lens' DescribeReservedDBInstancesOfferings (Maybe Text)

Product description filter value. Specify this parameter to show only the available offerings matching the specified product description.

drdbioOfferingType :: Lens' DescribeReservedDBInstancesOfferings (Maybe Text)

The offering type filter value. Specify this parameter to show only the available offerings matching the specified offering type. Valid Values: "Light Utilization" | "Medium Utilization" | "Heavy Utilization".

drdbioMultiAZ :: Lens' DescribeReservedDBInstancesOfferings (Maybe Bool)

The Multi-AZ filter value. Specify this parameter to show only the available offerings matching the specified Multi-AZ parameter.

drdbioMaxRecords :: Lens' DescribeReservedDBInstancesOfferings (Maybe Integer)

The maximum number of records to include in the response. If more than the MaxRecords value is available, a pagination token called a marker is included in the response so that the following results can be retrieved. Default: 100 Constraints: minimum 20, maximum 100.

drdbioMarker :: Lens' DescribeReservedDBInstancesOfferings (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

Response

data DescribeReservedDBInstancesOfferingsResponse

Contains the result of a successful invocation of the DescribeReservedDBInstancesOfferings action.

Response constructor

mkDescribeReservedDBInstancesOfferingsResponse :: DescribeReservedDBInstancesOfferingsResponse

Smart constructor for the minimum required parameters to construct a valid DescribeReservedDBInstancesOfferingsResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Marker :: Maybe Text

ReservedDBInstancesOfferings :: [ReservedDBInstancesOffering]

Response lenses

drdbiorMarker :: Lens' DescribeReservedDBInstancesOfferingsResponse (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

drdbiorReservedDBInstancesOfferings :: Lens' DescribeReservedDBInstancesOfferingsResponse [ReservedDBInstancesOffering]

A list of reserved DB instance offerings.