Network.AWS.RDS.V2013_09_09.RemoveSourceIdentifierFromSubscription

Description

Removes a source identifier from an existing RDS event notification subscription. https:/rds.us-east-1.amazonaws.com ?Action=RemoveSourceIdentifierFromSubscription &SubscriptionName=EventSubscription01 &SourceIdentifier=dbinstance01 &Version=2013-01-10 &SignatureVersion=4 &SignatureMethod=HmacSHA256 &Timestamp=20130128T012415Z &AWSAccessKeyId= &Signature= true 012345678901 db-instance modifying 2013-01-28 00:29:23.736 creation deletion EventSubscription01 arn:aws:sns:us-east-1:012345678901:EventSubscription01 6f0b82bf-68e9-11e2-b97b-43c6362ec60d.

Request

data RemoveSourceIdentifierFromSubscription

Request constructor

mkRemoveSourceIdentifierFromSubscription

Arguments

:: Text
rsifsSubscriptionName

-> Text
rsifsSourceIdentifier

-> RemoveSourceIdentifierFromSubscription

Smart constructor for the minimum required parameters to construct a valid RemoveSourceIdentifierFromSubscription request.

Request lenses

rsifsSubscriptionName :: Lens' RemoveSourceIdentifierFromSubscription Text

The name of the RDS event notification subscription you want to remove a source identifier from.

rsifsSourceIdentifier :: Lens' RemoveSourceIdentifierFromSubscription Text

The source identifier to be removed from the subscription, such as the DB instance identifier for a DB instance or the name of a security group.

Response

data RemoveSourceIdentifierFromSubscriptionResponse

Response lenses

rsifsrsEventSubscription :: Lens' RemoveSourceIdentifierFromSubscriptionResponse (Maybe EventSubscription)

Contains the results of a successful invocation of the DescribeEventSubscriptions action.